### 2025-11-16 #9424 @editorialbot
[PRE REVIEW]: Molass Library: A Python Package for SEC-SAXS Data Analysis
**Submitting author:** @nshimizu0721 (<a href="http://orcid.org/0000-0002-3636-1663">Nobutaka Shimizu</a>)
**Repository:** https://github.com/biosaxs-dev/molass-library
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.7.1
