### 2025-11-19 #9424 @nshimizu0721
To better reflect the methodological and application context of our work, we conducted a targeted search and identified five JOSS papers that are directly relevant. These papers present open-source tools for chromatography, SAXS, or related biophysical data analysis, and overlap with workflows similar to those addressed by our submission:

1. PeakPerformance
https://joss.theoj.org/papers/10.21105/joss.07313
(Automated peak detection and quantification for chromatography and related data)

 2. idpflex
https://joss.theoj.org/papers/10.21105/joss.01007
(Flexible analysis of intrinsically disordered proteins using SAXS and molecular simulations)

3. A GPU-Accelerated Open-Source Python Package
https://joss.theoj.org/papers/10.21105/joss.06024
(Open-source toolkit for powder diffraction, small-angle, and total scattering data analysis)

4. CADET-Core 5.0
https://joss.theoj.org/papers/10.21105/joss.07881
(Open-source tool for chromatography simulation and analysis)

5. hplc-py
https://joss.theoj.org/papers/10.21105/joss.06270
(Python package for HPLC data analysis)

We believe these five papers are more representative of the current state of open-source development in our field and provide a more appropriate context for our submission.
