### 2025-11-21 2d46a60
Update dependencies to confine to the latest version of molass_legacy
