+++ pyproject.toml
@@ -12,13 +12,12 @@ dependencies = [
   "matplotlib",
   "tqdm",
   "pybaselines>=1.2.0",
-  "molass_legacy>=0.3.2",
+  "molass_legacy>=0.4.0",
   "statsmodels",
   "scikit-learn",
   "toml",
   "psutil",
   "seaborn",
-  "openpyxl",
   "numba",
   "numba",
   "ruptures",
