### 2025-11-26 6e6f8cf
Add timescale argument to SDM model
