+++ molass/LowRank/Decomposition.py
@@ -375,7 +375,7 @@ class Decomposition:
         scd_colors = ['green' if rank == 1 else 'red' for rank in ranks]
         return peak_top_xes, scd_colors
     
-    def optimize_with_model(self, model_name, debug=False):
+    def optimize_with_model(self, model_name, model_params=None, debug=False):
         """
         Optimize the decomposition with a model.
 
@@ -388,6 +388,9 @@ class Decomposition:
 
             - ``SDM``: `Stochastic Dispersive Model <https://biosaxs-dev.github.io/molass-essence/chapters/60/stochastic-theory.html#stochastic-dispersive-model>`_
             - ``EDM``: `Equilibrium Dispersive Model <https://biosaxs-dev.github.io/molass-essence/chapters/60/kinetic-theory.html#equilibrium-dispersive-model>`_
+
+        model_params : dict, optional
+            The parameters for the model.
 
         debug : bool, optional
             If True, enable debug mode.
@@ -402,7 +405,7 @@ class Decomposition:
             reload(molass.SEC.ModelFactory)
         from molass.SEC.ModelFactory import create_model
         model = create_model(model_name, debug=debug)
-        return model.optimize_decomposition(self, debug=debug)
+        return model.optimize_decomposition(self, model_params=model_params, debug=debug)
     
     def rigorous_decomposition(self, rgcurve, debug=False):
         """
