### 2025-11-27 863175b
Improve SDM init_params estimation
