### 2025-11-28 e552344
Refactor to implement rigorous optimization
