### 2025-12-01 50f5334
Improve sub titles of the decomposition plot
