### 2025-12-01 86448be
Working on rigorous optimization
