+++ molass/LowRank/Decomposition.py
@@ -434,7 +434,7 @@ class Decomposition:
         model = create_model(model_name, debug=debug)
         return model.optimize_decomposition(self, model_params=model_params, debug=debug)
     
-    def make_rigorous_initparams(self, base_curves, debug=False):
+    def make_rigorous_initparams(self, baseparams, debug=False):
         """
         Make initial parameters for rigorous optimization.
 
@@ -449,43 +449,10 @@ class Decomposition:
             The initial parameters for rigorous optimization.
         """
         if debug:
-            from importlib import reload
-            import molass.Rigorous.EghParams
-            reload(molass.Rigorous.EghParams)
-        from molass.Rigorous.EghParams import make_egh_initparams
-
-        # XR initial parameters
-        xr_params = []
-        for ccurve in self.xr_ccurves:
-            xr_params.append(ccurve.get_params())
-        xr_params = np.array(xr_params)
-        # XR baseline parameters
-        xr_baseparams = base_curves[0].get_params()
-
-        # Rg parameters
-        rg_params = self.get_rgs()
-
-        # Mapping parameters
-        a, b = self.ssd.get_mapping()
-
-        # UV initial parameters
-        uv_params = []
-        for uv_ccurve, xr_ccurve in zip(self.uv_ccurves, self.xr_ccurves):
-            uv_params.append(uv_ccurve.get_params()[0]/xr_ccurve.get_params()[0])
-
-        # UV baseline parameters
-        uv_baseparams = base_curves[1].get_params()
-
-        # SecCol parameters
-        x = self.ssd.xr_icurve.x
-        init_mappable_range = (x[0], x[-1])
-
-        # SecCol parameters
-        from molass_legacy.SecTheory.SecEstimator import guess_initial_secparams
-        Npc, rp, tI, t0, P, m = guess_initial_secparams(xr_params, rg_params)
-        seccol_params = np.array([Npc, rp, tI, t0, P, m])
-
-        return np.concatenate([xr_params.flatten(), xr_baseparams, rg_params, uv_params, uv_baseparams, init_mappable_range, seccol_params])
+            import molass.Rigorous.RigorousImplement
+            reload(molass.Rigorous.RigorousImplement)
+        from molass.Rigorous.RigorousImplement import make_rigorous_initparams_impl
+        return make_rigorous_initparams_impl(self, baseparams, debug=debug)
 
     def optimize_rigorously(self, rgcurve=None, debug=False):
         """
