### 2025-12-03 e3e48b6
First version of rigorous optimization using EGH
