+++ molass/Rigorous/RigorousImplement.py
@@ -1,26 +1,24 @@
 """
 LowRank.RigorousImplement
 """
+import os
 import numpy as np
 from importlib import reload
 
-class DummyBasecurve:
-    def __init__(self, x):
-        self.x = x
-        self.y = np.zeros(len(x))
-
-    def __call__(self, x, params, y_, cy_list):
-        return self.y
-
 def make_dsets_from_decomposition(decomposition, rg_curve, debug=False):
     from molass_legacy.Optimizer.OptDataSets import OptDataSets
+    from molass_legacy.SecSaxs.ElCurve import ElCurve
+    if debug:
+        import molass.Bridge.LegacyRgCurve
+        reload(molass.Bridge.LegacyRgCurve)
+    from molass.Bridge.LegacyRgCurve import LegacyRgCurve
     ssd = decomposition.ssd
-    xr_curve = decomposition.xr_icurve
+    xr_curve = ElCurve(*decomposition.xr_icurve.get_xy())
     D = ssd.xr.M
     E = ssd.xr.E
-    uv_curve = decomposition.uv_icurve
+    uv_curve = ElCurve(*decomposition.uv_icurve.get_xy())
     U = ssd.uv.M
-    dsets = ((xr_curve, D), rg_curve, (uv_curve, U))
+    dsets = ((xr_curve, D), LegacyRgCurve(xr_curve, rg_curve), (uv_curve, U))
     return OptDataSets(None, None, dsets=dsets, E=E)
 
 def make_basecurves_from_decomposition(decomposition, debug=False):
@@ -36,7 +34,7 @@ def make_basecurves_from_decomposition(decomposition, debug=False):
     baseline_type = 1
     return make_basecurves_from_sd(sd, baseline_type, debug=debug)
 
-def construct_legacy_optimizer(dsets, base_curves, spectral_vectors, num_components=3, model="EGH", method="BH", debug=False):
+def construct_legacy_optimizer(dsets, baseline_objects, spectral_vectors, num_components=3, model="EGH", method="BH", debug=False):
     from molass_legacy.Optimizer.OptimizerUtils import get_function_code
     from molass_legacy.Optimizer.FuncImporter import import_objective_function
     function_code = get_function_code(model)
@@ -44,8 +42,8 @@ def construct_legacy_optimizer(dsets, base_curves, spectral_vectors, num_compone
     optimizer = function_class(
         dsets,
         num_components + 1,
-        xr_base_curve=base_curves[0],
-        uv_base_curve=base_curves[1],
+        xr_base_curve=baseline_objects[1],
+        uv_base_curve=baseline_objects[0],
         qvector=spectral_vectors[0],
         wvector=spectral_vectors[1],
         )
@@ -68,8 +66,8 @@ def make_rigorous_initparams_impl(decomposition, baseparams, debug=False):
 
     # UV initial parameters
     uv_params = []
-    for uv_ccurve, xr_ccurve in zip(decomposition.uv_ccurves, decomposition.xr_ccurves):
-        uv_params.append(uv_ccurve.get_params()[0]/xr_ccurve.get_params()[0])
+    for uv_ccurve in decomposition.uv_ccurves:
+        uv_params.append(uv_ccurve.get_params()[0])
 
     # UV baseline parameters
     uv_baseparams = baseparams[0]
@@ -79,13 +77,16 @@ def make_rigorous_initparams_impl(decomposition, baseparams, debug=False):
     init_mappable_range = (x[0], x[-1])
 
     # SecCol parameters
+    if debug:
+        import molass_legacy.SecTheory.SecEstimator
+        reload(molass_legacy.SecTheory.SecEstimator)
[…]
