### 2025-12-04 7d7c97b
First version of rigorous optimization using SDM
