+++ molass/Rigorous/RigorousImplement.py
@@ -5,88 +5,7 @@ import os
 import numpy as np
 from importlib import reload
 
-def make_dsets_from_decomposition(decomposition, rg_curve, debug=False):
-    from molass_legacy.Optimizer.OptDataSets import OptDataSets
-    from molass_legacy.SecSaxs.ElCurve import ElCurve
-    if debug:
-        import molass.Bridge.LegacyRgCurve
-        reload(molass.Bridge.LegacyRgCurve)
-    from molass.Bridge.LegacyRgCurve import LegacyRgCurve
-    ssd = decomposition.ssd
-    xr_curve = ElCurve(*decomposition.xr_icurve.get_xy())
-    D = ssd.xr.M
-    E = ssd.xr.E
-    uv_curve = ElCurve(*decomposition.uv_icurve.get_xy())
-    U = ssd.uv.M
-    dsets = ((xr_curve, D), LegacyRgCurve(xr_curve, rg_curve), (uv_curve, U))
-    return OptDataSets(None, None, dsets=dsets, E=E)
-
-def make_basecurves_from_decomposition(decomposition, debug=False):
-    if debug:
-        import molass.Bridge.SdProxy
-        reload(molass.Bridge.SdProxy)
-        import molass.Bridge.LegacyBaselines
-        reload(molass.Bridge.LegacyBaselines)
-    from molass.Bridge.SdProxy import SdProxy
-    from molass.Bridge.LegacyBaselines import make_basecurves_from_sd
-    ssd = decomposition.ssd
-    sd = SdProxy(ssd)
-    baseline_type = 1
-    return make_basecurves_from_sd(sd, baseline_type, debug=debug)
-
-def construct_legacy_optimizer(dsets, baseline_objects, spectral_vectors, num_components=3, model="EGH", method="BH", debug=False):
-    from molass_legacy.Optimizer.OptimizerUtils import get_function_code
-    from molass_legacy.Optimizer.FuncImporter import import_objective_function
-    function_code = get_function_code(model)
-    function_class = import_objective_function(function_code)
-    optimizer = function_class(
-        dsets,
-        num_components + 1,
-        xr_base_curve=baseline_objects[1],
-        uv_base_curve=baseline_objects[0],
-        qvector=spectral_vectors[0],
-        wvector=spectral_vectors[1],
-        )
-    return optimizer
-
-def make_rigorous_initparams_impl(decomposition, baseparams, debug=False):
-    # XR initial parameters
-    xr_params = []
-    for ccurve in decomposition.xr_ccurves:
-        xr_params.append(ccurve.get_params())
-    xr_params = np.array(xr_params)
-    # XR baseline parameters
-    xr_baseparams = baseparams[1]
-
-    # Rg parameters
-    rg_params = decomposition.get_rgs()
-
-    # Mapping parameters
-    a, b = decomposition.ssd.get_mapping()
-
-    # UV initial parameters
-    uv_params = []
-    for uv_ccurve in decomposition.uv_ccurves:
-        uv_params.append(uv_ccurve.get_params()[0])
-
-    # UV baseline parameters
-    uv_baseparams = baseparams[0]
-
-    # SecCol parameters
-    x = decomposition.ssd.xr.get_icurve().x
-    init_mappable_range = (x[0], x[-1])
-
-    # SecCol parameters
-    if debug:
-        import molass_legacy.SecTheory.SecEstimator
-        reload(molass_legacy.SecTheory.SecEstimator)
-    from molass_legacy.SecTheory.SecEstimator import guess_initial_secparams
-    Npc, rp, tI, t0, P, m = guess_initial_secparams(xr_params, rg_params, poresize=70)
-    seccol_params = np.array([Npc, rp, tI, t0, P, m])
-
[…]
