### 2025-12-04 c69d341
First version of rigorous optimization using EDM
