### 2025-12-05 22de193
Update to ignore multiple temp folders
