### 2025-12-10 607ca1c
Uniformly keep uv components with UvComponent object
