+++ molass/Rigorous/RunInfo.py
@@ -3,7 +3,17 @@ Rigorous.RunInfo.py
 """
 
 class RunInfo:
-    def __init__(self, optimizer, dsets, init_params):
+    def __init__(self, ssd, optimizer, dsets, init_params):
+        self.ssd = ssd
         self.optimizer = optimizer
         self.dsets = dsets
         self.init_params = init_params
+
+    def get_current_decomposition(self, **kwargs):
+        debug = kwargs.get('debug', False)
+        if debug:
+            from importlib import reload
+            import molass.Rigorous.CurrentStateUtils
+            reload(molass.Rigorous.CurrentStateUtils)
+        from molass.Rigorous.CurrentStateUtils import construct_decomposition_from_results
+        return construct_decomposition_from_results(self, **kwargs)
