### 2025-12-11 23077a4
Fix degrade from the refactoring using UvComponent
