### 2025-12-16 3b2e81e
Add contents around levy aware implementation
