+++ study/src/felinger1999_stochastic_dispersive.py
@@ -450,6 +450,19 @@ class StochasticDispersiveChromatography:
     """
     Unified stochastic-dispersive chromatography model (Felinger 1999).
     
+    LÉVY-THEOREM-FIRST IMPLEMENTATION:
+    ───────────────────────────────────
+    This implementation DIRECTLY INVOKES Lévy-Khintchine and Lévy-Itô theorems
+    as mathematical shortcuts, rather than deriving from chromatographic first
+    principles.
+    
+    Key design principle: Trust the theorems!
+    ─────────────────────────────────────────
+    1. Lévy-Khintchine → CF structure is guaranteed correct
+    2. Lévy-Itô decomposition → Independent components multiply/add
+    3. Infinite divisibility → Any (γ, σ², ν) triplet is valid
+    4. Cumulant generating function → Moments without PDF
+    
     Combines:
     - Random adsorption-desorption (stochastic)
     - Axial dispersion (Gaussian spreading)
@@ -486,6 +499,104 @@ class StochasticDispersiveChromatography:
         self.sorption = sorption
         self.n_ads = n_ads
     
+    # ════════════════════════════════════════════════════════════════════════
+    # LÉVY PROCESS STRUCTURE: Direct Access to Theoretical Components
+    # ════════════════════════════════════════════════════════════════════════
+    
+    @property
+    def levy_triplet(self) -> dict:
+        """Extract Lévy-Khintchine triplet (γ, σ², ν).
+        
+        By Lévy-Khintchine theorem, this triplet UNIQUELY determines the
+        entire stochastic process. All properties can be derived from this.
+        
+        Returns
+        -------
+        dict with:
+            'gamma': Drift parameter (hold-up time)
+            'sigma_squared': Brownian variance parameter
+            'lambda_rate': Jump process intensity
+            'levy_measure': Sorption model (defines ν)
+        """
+        return {
+            'gamma': self.column.t0,
+            'sigma_squared': 2 * self.column.D * self.column.t0 / self.column.u,
+            'lambda_rate': self.n_ads,
+            'levy_measure': self.sorption
+        }
+    
+    def validate_infinite_divisibility(self, n_test: int = 10, omega_max: float = 10.0) -> dict:
+        """Verify φ(ω) = [φ(ω/n)]ⁿ (Lévy-Khintchine requirement).
+        
+        This tests whether the characteristic function satisfies infinite
+        divisibility, which is REQUIRED for any Lévy process.
+        
+        Parameters
+        ----------
+        n_test : int
+            Divisibility parameter (e.g., n=10 checks if φ = (φ/10)^10)
+        omega_max : float
+            Maximum frequency to test
+            
+        Returns
+        -------
+        dict with validation results
+        """
+        # Use smaller omega range to avoid numerical issues
+        omega = np.linspace(-omega_max, omega_max, 100)
+        omega = omega[omega != 0]  # Exclude zero to avoid trivial case
+        
+        phi_omega = self.characteristic_function(omega)
+        phi_omega_n = self.characteristic_function(omega / n_test)
+        
+        # Check if φ(ω) ≈ [φ(ω/n)]ⁿ
+        # Use log to avoid numerical overflow: log φ(ω) ≈ n·log φ(ω/n)
+        log_phi = np.log(phi_omega + 1e-300)  # Avoid log(0)
+        log_phi_n = np.log(phi_omega_n + 1e-300)
+        
+        ratio = np.exp(log_phi - n_test * log_phi_n)
+        
[…]
