+++ study/README-STUDY.md
@@ -3,6 +3,11 @@ This folder holds contents to study Stochastic Theory of Chromatograpy.
 
 Include the following papers as needed to cosult Copilot.
 
+* 1951, J. Gil-Pelaez.pdf
+* 1955, J. C. Giddings.pdf
 * 1999, A. Felinger.pdf
 * 2002, Francesco Dondi.pdf
 * 2005, Luisa Pasti.pdf
+* 2014a, Annamária Sepsey.pdf
+* 2018, Viktor Witkovský.pdf
+* 2022, Jamil UR REHMAN.pdf
