### 2025-12-16 7734877
Squashed 'molass-library/' changes from a59c443..12a713e
12a713e Merge master into joss-paper to synchronize branches
b2f1d39 Update .gitignore
3b2e81e Add contents around levy aware implementation
50216a7 Add Pasti 2005 study contents
0db382f Update denss to 1.8.7
23077a4 Fix degrade from the refactoring using UvComponent
607ca1c Uniformly keep uv components with UvComponent object
2a829d7 Update to separate temporary analysis folders
22de193 Update to ignore multiple temp folders
c69d341 First version of rigorous optimization using EDM
7d7c97b First version of rigorous optimization using SDM
[…]
