+++ review-discussions/anticipated-review-questions.md
@@ -1,0 +1,336 @@
+# Anticipated JOSS Review Questions
+
+**Purpose**: Proactive preparation for upcoming JOSS review  
+**Created**: December 16, 2025  
+**Updated**: December 16, 2025 (after analyzing 5 similar JOSS reviews)  
+**Status**: Pre-review preparation  
+
+**Based on actual reviews of similar papers**:
+- PeakPerformance (#7313) - LC-MS/MS chromatography peak fitting
+- hplc-py (#6270) - HPLC chromatography analysis
+- DebyeCalculator (#6024) - Small-angle scattering (SAXS-related)
+- idpflex (#1007) - SAXS for disordered proteins (2018 - older review format)
+- CADET-Core (#7881) - Chromatography simulation (C++ project)
+
+---
+
+## 1. Software Functionality & Scope
+
+### Q1.1: Coverage of SEC-SAXS Workflow
+**Question**: The paper states that Molass Library implements steps 3-7 of the 8-step SEC-SAXS workflow. Why were steps 1-2 and 8 excluded? Could these be integrated in the future?
+
+### Q1.2: Comparison with Existing Tools
+**Question**: The paper mentions ATSAS and BioXTAS RAW as alternatives. Can you provide a more detailed comparison of features, particularly:
+- What specific functionality does Molass Library provide that these tools lack?
+- In what scenarios would a user choose Molass Library over these established tools?
+- Are there any benchmark comparisons available?
+
+### Q1.3: Low-Rank Factorization Implementation
+**Question**: The paper describes low-rank factorization as a "central feature." How does your implementation compare to standard matrix factorization methods in scipy or scikit-learn? What makes your approach specialized for SEC-SAXS?
+
+### Q1.4: Elution Curve Models
+**Question**: Three elution curve models are mentioned (EGH, SDM, EDM). How does a user choose between these models? Are there guidelines or automated model selection?
+
+---
+
+## 2. Installation & Dependencies
+
+### Q2.1: Python Version Support
+**Question**: The pyproject.toml specifies `requires-python = ">=3.9,<3.13"`. 
+- Why is Python 3.13 excluded?
+- Is Python 3.8 support planned (it's still widely used)?
+- Are there specific dependency constraints preventing broader support?
+
+### Q2.2: Dependency on molass_legacy
+**Question**: There's a dependency on `molass_legacy>=0.3.2`. What functionality is provided by this legacy package? Is there a plan to eliminate this dependency?
+
+### Q2.3: Platform Support
+**Question**: README mentions testing on Windows 11 and Ubuntu 22.04. What about:
+- macOS support?
+- Other Linux distributions?
+- Python 3.9, 3.10, 3.11 testing (only 3.12 mentioned)?
+
+### Q2.4: Optional Dependencies
+**Question**: The `excel` extra requires `pywin32` (Windows-only). How critical is this feature? Are there alternatives for non-Windows users?
+
+---
+
+## 3. Documentation
+
+### Q3.1: Documentation Structure
+**Question**: Four separate documentation sites are mentioned (Tutorial, Essence, Technical, Reference). Why this separation? Could this be confusing for new users?
+
+### Q3.2: API Documentation
+**Question**: Is the API fully documented with docstrings? Are there auto-generated API docs (e.g., using Sphinx)?
+
+### Q3.3: Installation Instructions
+**Question**: Are there detailed installation instructions for different platforms? What about troubleshooting common installation issues?
+
+### Q3.4: Example Data
+**Question**: Are example datasets provided for users to test the software? Where can users find real SEC-SAXS data to practice with?
+
+### Q3.5: Contribution Guidelines
[…]
