### 2025-12-16 667adb6
Update the Python verion requirement to restrict to the tested versions
