### 2025-12-17 85b085b
Add study stuff which explores the GEC monopore CF, etc.
