+++ study/verify_gec_assumptions.py
@@ -1,0 +1,466 @@
+"""
+Verify GEC Model Assumptions from Animation Data
+
+This script runs the SEC animation with statistics tracking enabled and tests:
+1. Exponential distribution of adsorption durations (Dondi Eq. 41)
+2. Poisson distribution of adsorption counts (Dondi Eq. 42)
+3. Comparison of empirical vs theoretical characteristic functions
+
+Usage:
+    python verify_gec_assumptions.py
+"""
+
+import sys
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy import stats as sp_stats
+from pathlib import Path
+
+# Add molass to path
+molass_path = Path(__file__).parent.parent
+sys.path.insert(0, str(molass_path))
+
+from molass.SEC.ColumnSimulation import get_animation
+
+def run_animation_with_tracking(num_frames=400, seed=42):
+    """
+    Run animation and collect statistics.
+    
+    Returns
+    -------
+    stats : dict
+        Dictionary containing adsorption statistics
+    """
+    print(f"Running animation with {num_frames} frames, seed={seed}...")
+    anim, stats_dict = get_animation(
+        num_frames=num_frames,
+        seed=seed,
+        close_plot=True,
+        track_statistics=True,
+        use_tqdm=True,
+        blit=False
+    )
+    
+    # Must save the animation to actually execute all frames
+    print("Executing animation frames to collect statistics...")
+    # import tempfile
+    # with tempfile.NamedTemporaryFile(suffix='.gif', delete=True) as tmp:
+    #     anim.save(tmp.name, writer='pillow', fps=10)
+    anim.save("temp.gif")
+    
+    print("Animation complete.")
+    return stats_dict
+
+
+def test_exponential_egress(stats, particle_type=None, alpha=0.05):
+    """
+    Test if adsorption durations follow exponential distribution.
+    
+    Parameters
+    ----------
+    stats : dict
+        Statistics from animation
+    particle_type : int, optional
+        0=large, 1=medium, 2=small. If None, test all particles.
+    alpha : float
+        Significance level for KS test
+    
+    Returns
+    -------
+    result : dict
+        Test results with p-values and fitted parameters
+    """
+    print("\n" + "="*70)
+    print("TEST 1: Exponential Egress Time Distribution (Dondi Eq. 41)")
+    print("="*70)
+    
+    # Select particles
+    if particle_type is not None:
+        type_names = ['Large (green)', 'Medium (blue)', 'Small (red)']
+        type_keys = ['large_indeces', 'middle_indeces', 'small_indeces']
+        particle_indeces = stats[type_keys[particle_type]]
+        print(f"Testing {type_names[particle_type]} particles only")
+    else:
+        particle_indeces = np.arange(len(stats['ptype_indeces']))
+        print("Testing all particles combined")
+    
+    # Collect all durations for selected particles
+    all_durations = []
+    for k in particle_indeces:
+        all_durations.extend(stats['adsorption_durations_list'][k])
+    
+    all_durations = np.array(all_durations)
+    print(f"Total number of adsorption events: {len(all_durations)}")
+    
+    if len(all_durations) == 0:
+        print("WARNING: No adsorption events recorded!")
+        return None
+    
+    # Fit exponential distribution
+    tau_bar = np.mean(all_durations)
+    print(f"Mean adsorption time: τ̄ = {tau_bar:.6f}")
[…]
