### 2025-12-18 b557332
Add the test of independence
