+++ study/verify_gec_assumptions.py
@@ -382,6 +382,135 @@ def compare_characteristic_functions(stats, particle_type=None):
     print(f"Mean squared error: {mse:.6f}")
 
 
+def test_independence(stats, particle_type=None, alpha=0.05):
+    """
+    Test independence between adsorption count (r_M) and total adsorbed time (t_S).
+    
+    For a true Compound Poisson Process, r_M and t_S should be independent.
+    This tests whether the simulation violates this critical assumption.
+    
+    Parameters
+    ----------
+    stats : dict
+        Statistics from animation
+    particle_type : int, optional
+        0=large, 1=medium, 2=small. If None, test all particles.
+    alpha : float
+        Significance level for correlation test
+    
+    Returns
+    -------
+    result : dict
+        Test results with correlation coefficient and p-value
+    """
+    print("\n" + "="*70)
+    print("TEST 4: Independence of r_M and t_S (CPP Assumption)")
+    print("="*70)
+    
+    # Select particles
+    if particle_type is not None:
+        type_names = ['Large (green)', 'Medium (blue)', 'Small (red)']
+        type_keys = ['large_indeces', 'middle_indeces', 'small_indeces']
+        particle_indeces = stats[type_keys[particle_type]]
+        print(f"Testing {type_names[particle_type]} particles only")
+    else:
+        particle_indeces = np.arange(len(stats['ptype_indeces']))
+        print("Testing all particles combined")
+    
+    # Get data
+    r_M = stats['adsorption_counts'][particle_indeces]
+    t_S = stats['total_adsorbed_time'][particle_indeces]
+    
+    # Filter out particles with no adsorptions
+    valid = r_M > 0
+    r_M_valid = r_M[valid]
+    t_S_valid = t_S[valid]
+    
+    print(f"Number of particles: {len(particle_indeces)}")
+    print(f"Particles with adsorptions: {len(r_M_valid)}")
+    
+    if len(r_M_valid) < 3:
+        print("WARNING: Too few particles with adsorptions for correlation test!")
+        return None
+    
+    # Pearson correlation
+    corr_coef, p_value = sp_stats.pearsonr(r_M_valid, t_S_valid)
+    
+    print(f"\nPearson Correlation:")
+    print(f"  r_M vs t_S: ρ = {corr_coef:.4f}")
+    print(f"  p-value: {p_value:.4f}")
+    
+    if p_value > alpha:
+        print(f"  ✓ PASS: Cannot reject independence (p > {alpha})")
+        print(f"  → r_M and t_S appear independent (CPP assumption holds)")
+    else:
+        print(f"  ✗ FAIL: Reject independence (p ≤ {alpha})")
+        if corr_coef > 0:
+            print(f"  → Positive correlation: More adsorptions → longer total time")
+        else:
+            print(f"  → Negative correlation: More adsorptions → shorter total time")
+    
+    # Spearman correlation (robust to nonlinearity)
+    spearman_corr, spearman_p = sp_stats.spearmanr(r_M_valid, t_S_valid)
+    print(f"\nSpearman Rank Correlation (robust):")
+    print(f"  ρ_s = {spearman_corr:.4f}, p = {spearman_p:.4f}")
+    
+    # Plot scatter
+    fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(12, 5))
+    
+    # Scatter plot
+    ax1.scatter(r_M_valid, t_S_valid, alpha=0.5, s=20, color='steelblue')
+    
+    # Add regression line if significant
+    if p_value < alpha:
+        z = np.polyfit(r_M_valid, t_S_valid, 1)
+        p = np.poly1d(z)
[…]
