+++ CURRENT-STATUS.md
@@ -1,38 +1,59 @@
 # CURRENT STATUS
 
-
-**📅 Last Updated: November 17, 2025**  
+**📅 Last Updated: December 19, 2025**  
 **👤 Updated By: Copilot (Takahashi session)**
 
 ---
 
+## 🎯 Current Phase: PRE-REVIEW (Waiting for Editor Assignment)
 
-## 🎯 What's Happening Now?
-**JOSS review has started. Tracking reviewer/editor comments and internal issues.**
+**Status**: Paper submitted and in queue - **No active review yet**
 
-**📅 Last Updated: November 17, 2025**  
-**👤 Updated By: Copilot (Takahashi session)**
+### Timeline:
+- ✅ **Nov 16, 2025**: Paper submitted to JOSS
+- ✅ **Nov 16, 2025**: Editorial bot completed automated checks
+- ✅ **Nov 16-17, 2025**: Managing EiC (Kevin M. Moerman) asked about BPPB publication differences
+- ✅ **Nov 17, 2025**: Shimizu provided detailed response
+- ✅ **Nov 18, 2025**: Shimizu fixed DOI formatting issues
+- ✅ **Nov 19, 2025**: Shimizu suggested 5 relevant JOSS papers as reviewer candidates
+- ⏳ **Currently**: Waiting for editor assignment
+
+### What This Means:
+- **Editor**: Not yet assigned
+- **Reviewers**: Not yet assigned
+- **Review Status**: Has NOT started yet
+- **Next Step**: Wait for editor to be assigned, then reviewers
+
+---
+
 ## 🔗 JOSS Review Issue (External)
-Once the JOSS review starts, the official review will take place at a GitHub issue in the JOSS repository. Track the link here:
-
 
 **JOSS Review Issue URL:** [https://github.com/openjournals/joss-reviews/issues/9424](https://github.com/openjournals/joss-reviews/issues/9424)
-**JOSS review in progress. Tracking reviewer/editor comments and internal issues.**
-_When the review issue is created, paste the URL above for easy access during the review process._
+
+**Current JOSS Status Badge**: REVIEW PENDING
 
 ---
 
-**JOSS Review Issue URL:** [https://github.com/openjournals/joss-reviews/issues/9424](https://github.com/openjournals/joss-reviews/issues/9424)
+## 👥 Team Status
+
+### **Shimizu** (Lead Responder):
+- ✅ Submitted paper
+- ✅ Responded to EiC's pre-review questions
+- ✅ Fixed DOI formatting issues
+- ✅ Suggested relevant reviewer candidates
+- ⏳ Waiting for editor/reviewer assignment
+
+### **Takahashi** (Technical Support):
+- ✅ Repository setup complete
+- ⏳ On standby for technical questions when review starts
+
 ---
 
-## 👥 Who's Doing What?
-- **Takahashi** (Programmer): Repository setup ✅ Done
-1. **Track and respond to reviewer/editor comments in the JOSS review issue**
-2. **Summarize and address all open internal GitHub Issues** (see below)
+## 📋 Open Internal GitHub Issues (Tracking Pre-Review Activity)
 
-## 📋 Open Internal GitHub Issues (as of Nov 17, 2025)
+**Active Issues**: 4 open issues tracking pre-review discussion and preparation
 
-1. [Required Reply: Differences from the time of BPPB publication](https://github.com/biosaxs-dev/molass-review/issues/7)
-
----
-5. [Email from JOSS editorial robot.](https://github.com/biosaxs-dev/molass-review/issues/3)
+1. [Similar JOSS papers for the reviewers](https://github.com/biosaxs-dev/molass-review/issues/6) - Discussion about reviewer candidates
[…]
