### 2025-12-23 cc91200
Add a functionalty which tracks the trajectory of a single particle
