+++ study/visualize_levy_trajectory.py
@@ -1,0 +1,437 @@
+"""
+Visualize Single Particle Trajectory as a Lévy Process
+
+This script creates a 3D animation showing how a single particle's trajectory
+through the SEC column forms a Compound Poisson Process (a type of Lévy process).
+
+The 3D visualization shows:
+- X, Y axes: Spatial position in the column
+- Z axis: Cumulative adsorbed time t_S(t) ← This is the Lévy process!
+
+Key features:
+- Horizontal segments: Particle is mobile (free diffusion, no time penalty)
+- Vertical jumps: Particle is adsorbed (random "delays" accumulate)
+- The result is a staircase-like 3D trajectory that visualizes the CPP structure
+
+Usage:
+    python visualize_levy_trajectory.py --particle 500 --frames 400
+    python visualize_levy_trajectory.py --particle 1000 --type small
+"""
+
+import sys
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.animation import FuncAnimation, PillowWriter
+from mpl_toolkits.mplot3d import Axes3D
+from pathlib import Path
+import argparse
+
+# Add molass to path
+molass_path = Path(__file__).parent.parent
+sys.path.insert(0, str(molass_path))
+
+from molass.SEC.ColumnSimulation import get_animation
+
+
+def run_with_trajectory_tracking(particle_id, num_frames=400, seed=42):
+    """
+    Run animation and collect single particle trajectory.
+    
+    Parameters
+    ----------
+    particle_id : int
+        Index of particle to track (0-indexed)
+    num_frames : int
+        Number of frames to simulate
+    seed : int
+        Random seed for reproducibility
+    
+    Returns
+    -------
+    trajectory_data : dict
+        Dictionary containing trajectory information
+    """
+    print(f"Running animation with trajectory tracking...")
+    print(f"  Particle ID: {particle_id}")
+    print(f"  Frames: {num_frames}")
+    print(f"  Seed: {seed}")
+    
+    anim, stats = get_animation(
+        num_frames=num_frames,
+        seed=seed,
+        close_plot=True,
+        track_particle_id=particle_id,
+        use_tqdm=True,
+        blit=False
+    )
+    
+    # Execute all frames to collect trajectory data
+    # Simply iterate through all frames without saving
+    print("Executing animation frames to collect trajectory...")
+    try:
+        # Force execution by rendering each frame
+        for i in range(num_frames):
+            anim._func(i)
+    except Exception as e:
+        print(f"Note: Some frames may have had issues: {e}")
+        print("Continuing with collected data...")
+    
+    print("Trajectory collection complete.")
+    return stats['trajectory']
+
+
+def create_3d_trajectory_plot(trajectory, output_path=None, interactive=True):
+    """
+    Create static 3D plot of the trajectory.
+    
+    Parameters
+    ----------
+    trajectory : dict
+        Trajectory data from animation
+    output_path : str or Path, optional
+        If provided, save the plot to this path
+    interactive : bool
+        If True, show interactive plot
+    """
+    positions = trajectory['positions']
+    states = trajectory['states']
+    cum_time = trajectory['cumulative_adsorbed_time']
+    particle_type = trajectory['particle_type']
+    
+    type_names = ['Large (green)', 'Medium (blue)', 'Small (red)']
+    type_colors = ['green', 'blue', 'red']
+    
+    print(f"\nTrajectory Statistics:")
[…]
