+++ molass/SEC/ColumnSimulation.py
@@ -12,7 +12,7 @@ plt.rcParams['animation.embed_limit'] = 512     # https://stackoverflow.com/ques
 from .ColumnElements import Particle
 from .ColumnStructure import plot_column_structure
 
-def get_animation(num_frames=None, interval=100, seed=None, close_plot=True, return_init=False, fig_check=False, blit=False, use_tqdm=True, large_only=False, debug=False, track_statistics=False):
+def get_animation(num_frames=None, interval=100, seed=None, close_plot=True, return_init=False, fig_check=False, blit=False, use_tqdm=True, large_only=False, debug=False, track_statistics=False, track_particle_id=None):
     """
     Create an animation of particles moving through a column structure.
     
@@ -40,13 +40,16 @@ def get_animation(num_frames=None, interval=100, seed=None, close_plot=True, ret
         If True, enable debug mode with additional output. Default is False.
     track_statistics : bool, optional
         If True, track adsorption statistics for theoretical verification. Default is False.
+    track_particle_id : int, optional
+        If provided, track the full trajectory of this specific particle (0-indexed).
+        Stores position, state, and cumulative adsorbed time at each frame.
 
     Returns
     -------
     anim : matplotlib.animation.FuncAnimation
         The animation object.
     stats : dict, optional
-        If track_statistics=True, also return statistics dictionary with adsorption data.
+        If track_statistics=True or track_particle_id is set, also return statistics dictionary.
     """
     ymin, ymax = 0, 1
     xmin, xmax = 0.35, 0.65
@@ -175,6 +178,14 @@ def get_animation(num_frames=None, interval=100, seed=None, close_plot=True, ret
         adsorption_durations_list = [[] for _ in range(num_particles)]  # List of durations for each particle
         adsorption_counts = np.zeros(num_particles, dtype=int)  # Total number of adsorptions
         total_adsorbed_time = np.zeros(num_particles, dtype=float)  # Sum of all adsorption times
+    
+    # Single particle trajectory tracking for Lévy process visualization
+    if track_particle_id is not None:
+        trajectory_positions = []  # List of (x, y) tuples
+        trajectory_states = []  # List of state (True=mobile, False=adsorbed)
+        trajectory_cumulative_adsorbed = []  # Cumulative adsorbed time at each frame
+        trajectory_cumulative_adsorbed_time = 0.0
+        trajectory_particle_exited = False  # Flag to stop tracking after particle exits
 
     def touchable_indeces(inmobile_states, last_pxv, last_pyv, debug=False):
         indeces = []
@@ -227,6 +238,8 @@ def get_animation(num_frames=None, interval=100, seed=None, close_plot=True, ret
 
     def compute_next_positions(debug=False, frame_number=None):
         nonlocal pxv, pyv, cancel_debug
+        if track_particle_id is not None:
+            nonlocal trajectory_cumulative_adsorbed_time
         if debug:
             print("(2) inmobile_states=", ''.join(map(lambda b: '%d' % b, inmobile_states)))
         
@@ -269,6 +282,24 @@ def get_animation(num_frames=None, interval=100, seed=None, close_plot=True, ret
 
         exceed_top = pxv > ymax
         pyv[exceed_top] += -2*dyv[exceed_top]
+        
+        # Track single particle trajectory
+        if track_particle_id is not None and frame_number is not None:
+            nonlocal trajectory_particle_exited
+            pid = track_particle_id
+            # Only track while particle is still in column (y > 0)
+            if not trajectory_particle_exited and pyv[pid] > 0:
+                # Store position
+                trajectory_positions.append((pxv[pid], pyv[pid]))
+                # Store state
+                trajectory_states.append(inmobile_states[pid])
+                # Update cumulative adsorbed time
+                if not inmobile_states[pid]:  # False = adsorbed
[…]
