+++ study/visualize_levy_trajectory.py
@@ -271,6 +271,472 @@ def create_animated_trajectory(trajectory, output_path=None, fps=20):
     plt.close()
 
 
+def create_levy_khintchine_visualization(trajectory, output_path=None):
+    """
+    Create visualization showing the Lévy-Khintchine representation.
+    
+    This shows how the trajectory relates to the mathematical description:
+    φ(θ) = exp(t[∫(e^(iθτ) - 1)Π(dτ)])
+    
+    Components visualized:
+    1. Empirical Lévy measure Π(dτ) = λF(dτ) from observed jumps
+    2. Characteristic function φ(θ) in complex plane
+    3. Jump size distribution F(τ)
+    4. Cumulative Lévy measure and its relation to k'
+    """
+    positions = trajectory['positions']
+    states = trajectory['states']
+    cum_time = trajectory['cumulative_adsorbed_time']
+    particle_type = trajectory['particle_type']
+    
+    type_names = ['Large (green)', 'Medium (blue)', 'Small (red)']
+    
+    # Extract ACTUAL adsorption durations (not frame-by-frame increments!)
+    # Need to find state transitions: mobile→adsorbed (start) and adsorbed→mobile (end)
+    jump_sizes = []
+    current_adsorption_start = None
+    
+    for i in range(len(states)):
+        is_mobile = states[i]
+        
+        # Transition: mobile → adsorbed (start of adsorption event)
+        if i > 0 and states[i-1] and not is_mobile:
+            current_adsorption_start = i
+        
+        # Transition: adsorbed → mobile (end of adsorption event)
+        if i > 0 and not states[i-1] and is_mobile:
+            if current_adsorption_start is not None:
+                # Compute duration of this adsorption event
+                duration = cum_time[i-1] - cum_time[current_adsorption_start-1]
+                if duration > 0:
+                    jump_sizes.append(duration)
+                current_adsorption_start = None
+    
+    # Handle case where particle is still adsorbed at the end
+    if current_adsorption_start is not None and current_adsorption_start < len(cum_time):
+        duration = cum_time[-1] - cum_time[current_adsorption_start-1]
+        if duration > 0:
+            jump_sizes.append(duration)
+    
+    jump_sizes = np.array(jump_sizes)
+    
+    if len(jump_sizes) == 0:
+        print("No adsorption events detected. Cannot create Lévy-Khintchine visualization.")
+        return
+    
+    # Calculate total observation time (in real time, not frames)
+    frame_indices = np.arange(len(cum_time))
+    total_frames = len(cum_time)
+    
+    # Estimate jump rate λ (jumps per frame)
+    lambda_rate = len(jump_sizes) / total_frames
+    
+    print(f"\nJump size statistics:")
+    print(f"  Number of jumps: {len(jump_sizes)}")
+    print(f"  Min jump size: {np.min(jump_sizes):.6f}")
+    print(f"  Max jump size: {np.max(jump_sizes):.6f}")
+    print(f"  Mean jump size: {np.mean(jump_sizes):.6f}")
+    print(f"  Std jump size: {np.std(jump_sizes):.6f}")
+    print(f"  Unique values: {len(np.unique(jump_sizes))}")
+    
+    # Create histogram for jump size distribution F(τ)
[…]
