### 2025-12-25 aecaa10
Add synchronization to initialization workflow
- Updated COPILOT-INIT-ACTION.md to include molass-library subtree sync as first step
- Created sync-molass-library.ps1 helper script for easy synchronization
- Updated TAKAHASHI-QUICK-START.md to document automatic sync during init
