### 2025-12-25 0ee4b69
Clarify synchronization is Takahashi-only
- Updated COPILOT-INIT-ACTION.md to specify sync happens only for Takahashi
- Updated SYNC-GUIDE.md to explain this is a Takahashi-only feature
- Shimizu doesn't need to manage the molass-library subtree
