+++ SYNC-GUIDE.md
@@ -1,0 +1,241 @@
+# 🔄 Synchronization Guide
+
+**Automatic synchronization is now integrated into Takahashi's initialization process!**
+
+---
+
+## 🚀 Quick Start: Automatic Sync (Takahashi Only)
+
+**Note**: This synchronization process is **only for Takahashi**. Shimizu doesn't need to manage the molass-library subtree.
+
+Every time Takahashi initializes Copilot, synchronization happens automatically:
+
+```
+Please read COPILOT-INIT.md to initialize
+```
+
+Copilot will:
+1. ✅ Sync molass-library subtree with joss-paper branch
+2. ✅ Check JOSS review issue for new comments
+3. ✅ Report status changes
+4. ✅ Alert you to any actionable items
+
+**You don't need to do anything manually!**
+
+---
+
+## 📦 What Gets Synchronized
+
+### 1. molass-library Subtree
+- **Source**: biosaxs-dev/molass-library (joss-paper branch)
+- **Destination**: molass-library/ folder in this repo
+- **Method**: git subtree pull (squashed)
+- **Frequency**: Every initialization
+
+### 2. JOSS Review Issue
+- **Source**: openjournals/joss-reviews#9424
+- **What's checked**: New reviewer/editor comments
+- **Action**: Alert if actionable comments found
+- **Frequency**: Every initialization
+
+---
+
+## 🛠️ Manual Synchronization (If Needed)
+
+### Option 1: PowerShell Script (Recommended)
+
+```powershell
+# Dry run (preview only)
+.\scripts\sync-molass-library.ps1 -DryRun
+
+# Actual sync
+.\scripts\sync-molass-library.ps1
+```
+
+**Note**: If you get an execution policy error, use Option 2 instead.
+
+### Option 2: Direct Git Commands
+
+```bash
+# Fetch latest changes
+git fetch molass-upstream
+
+# Check what's new (optional)
+git log HEAD..molass-upstream/joss-paper --oneline
+
+# Pull updates
+git subtree pull --prefix=molass-library molass-upstream joss-paper --squash
+```
+
+---
+
+## 🔍 Understanding the Sync Process
+
+### Git Subtree Basics
+
+The `molass-library/` folder is a **subtree**, not a submodule:
+
+- ✅ **Pros**: 
+  - Files are part of this repo (no separate clone needed)
+  - Simpler for most workflows
+  - Can make changes here and push back to source
+
+- ⚠️ **Important**:
+  - Changes in this repo need manual sync with source
+  - Must use git subtree commands (not regular git pull)
+
+### Remote Configuration
+
+The remote `molass-upstream` points to:
+```
+https://github.com/biosaxs-dev/molass-library.git
+```
+
+This is automatically configured. You can verify with:
+```bash
+git remote -v | Select-String "molass"
+```
+
+---
+
+## 📊 Sync Status Indicators
+
+After initialization, Copilot reports:
+
+### ✅ Up-to-date
+```
+✅ molass-library is already up-to-date!
+   No synchronization needed.
+```
+
+### 🔄 Updates Applied
+```
+✅ Synchronization completed successfully!
+   molass-library has been updated with latest changes from joss-paper branch.
+   
+   Changes:
+   - 16 files changed
+   - 4945 insertions
+```
+
+### ⚠️ Conflicts Detected
+```
+⚠️  Synchronization encountered conflicts.
+   Please resolve conflicts manually and commit.
+```
+
+---
+
+## 🔄 Workflow Integration
+
+### Typical Session Flow
+
+1. **Open VS Code**
+2. **Say**: "Please read COPILOT-INIT.md to initialize"
+3. **Copilot does**:
+   - Syncs molass-library subtree ← **Automatic!**
+   - Checks JOSS review for updates ← **Automatic!**
[…]
