### 2025-12-25 b5d121b
Clarify PowerShell script is optional due to execution policy
- Reordered manual sync options: git commands first (always works)
- Added detailed explanation of execution policy issue
- Provided 3 solutions with clear recommendations
- Emphasized that Copilot uses git commands directly (not affected)
- Script is kept as optional convenience with nice formatting
