### 2025-12-27 30fd784
Add a trouble shooting markdown
