+++ study/jupyter-kernel-troubleshooting.md
@@ -1,0 +1,261 @@
+# Jupyter Kernel Troubleshooting Guide
+
+## Problem: "ipykernel package required" Error Loop
+
+### Symptoms
+- VS Code shows: "Running cells with 'Python X.XX' requires the ipykernel package to be installed or requires an update"
+- Clicking "Install" causes an installation loop
+- Error persists even after installing ipykernel
+- Kernel crashes with import errors like: `ImportError: zmq Cython backend has not been compiled`
+
+### Root Cause
+Multiple Python installations on the same system with ambiguous Jupyter kernel specifications, causing:
+- Kernel starts with one Python version
+- But tries to import modules from a different Python version
+- Results in version mismatch and import failures
+
+---
+
+## Diagnosis Steps
+
+### 1. Check Python Installations
+```powershell
+where.exe python
+```
+This shows all Python executables in your PATH. Look for multiple versions.
+
+### 2. Verify Which Python You're Using
+```powershell
+python -V
+```
+
+### 3. Check ipykernel Installation
+```powershell
+# For system Python
+python -m pip show ipykernel
+
+# For specific Python installation
+& "C:\Program Files\Python313\python.exe" -m pip show ipykernel
+```
+
+### 4. List Jupyter Kernel Specs
+```powershell
+# Using default Python
+python -m jupyter kernelspec list
+
+# Using specific Python
+& "C:\Program Files\Python313\python.exe" -m jupyter kernelspec list
+```
+
+### 5. Inspect Kernel Configuration
+```powershell
+# Check the kernel.json file
+Get-Content "C:\Users\<USERNAME>\AppData\Roaming\Python\share\jupyter\kernels\python3\kernel.json"
+```
+
+Look for the `"argv"` field - if it just says `"python"` instead of a full path, that's the problem!
+
+---
+
+## Solution
+
+### Step 1: Update Kernel Spec to Use Explicit Path
+
+**Replace the generic kernel spec:**
+
+```powershell
+# Create updated kernel.json with explicit Python path
+$json = @'
+{
+ "argv": [
+  "C:\\Program Files\\Python313\\python.exe",
+  "-m",
+  "ipykernel_launcher",
+  "-f",
+  "{connection_file}"
+ ],
+ "display_name": "Python 3.13.11",
+ "language": "python",
+ "metadata": {
+  "debugger": true
+ }
+}
+'@
+
+# Write to kernel spec file
+$json | Out-File -Encoding utf8 "C:\Users\$env:USERNAME\AppData\Roaming\Python\share\jupyter\kernels\python3\kernel.json"
+```
+
+**Important:** Replace `C:\\Program Files\\Python313\\python.exe` with your actual Python path. Use double backslashes (`\\`) in the JSON.
+
+### Step 2: Clear Jupyter and VS Code Caches
+
+```powershell
+# Clear Jupyter runtime cache
+Remove-Item -Path "$env:APPDATA\jupyter\runtime\*" -Recurse -Force -ErrorAction SilentlyContinue
+Write-Host "Jupyter runtime cache cleared"
+
+# Clear VS Code Jupyter extension cache
+Remove-Item -Path "$env:APPDATA\Code - Insiders\User\globalStorage\ms-toolsai.jupyter\*" -Recurse -Force -ErrorAction SilentlyContinue
+Write-Host "VS Code Jupyter cache cleared"
+
+# For regular VS Code (not Insiders)
+Remove-Item -Path "$env:APPDATA\Code\User\globalStorage\ms-toolsai.jupyter\*" -Recurse -Force -ErrorAction SilentlyContinue
+```
+
+### Step 3: Reinstall ipykernel (if needed)
+
+```powershell
[…]
