### 2025-12-27 bb11156
Update the trouble shooting markdown
