### 2025-12-27 88dd07b
Migration of legacy lognormal pore implementation
