### 2025-12-27 4919736
Implement SDM Lognormal pore versions
