+++ molass/SEC/Models/LognormalPore.py
@@ -4,6 +4,7 @@ SEC.Models.LognormalPore.py
 import numpy as np
 from scipy.stats import lognorm
 from molass.MathUtils.IntegrateUtils import complex_quadrature_vec
+from molass.MathUtils.FftUtils import FftInvPdf
 
 def compute_mode(mu, sigma):
     return np.exp(mu - sigma**2)
@@ -40,8 +41,273 @@ def gec_lognormal_pore_pdf(x, scale, N, T, me, mp, mu, sigma, Rg, x0):
     return scale*gec_lognormal_pore_pdf_impl(x - x0, N, T, me, mp, mu, sigma, Rg, 0)  # not always the same as below
     # return scale*gec_lognormal_pore_pdf_impl(x, N, T, me, mp, mu, sigma, Rg, x0)
 
-"""
-task:
-- sdm_lognormal_pore_gamma_cf
-- sdm_lognormal_pore_gamma_pdf
-"""
+# ============================================================================
+# SDM Extensions: Adding Mobile Phase Dispersion (Brownian component)
+# 
+# The following implementations were proposed by GitHub Copilot (Claude Sonnet 4.5)
+# on December 27, 2025, extending the GEC lognormal pore model with:
+# 1. Mobile phase dispersion (SDM framework)
+# 2. Gamma-distributed residence times (kinetic heterogeneity)
+# ============================================================================
+
+def sdm_lognormal_pore_cf(w, N, T, me, mp, mu, sigma, Rg, N0, t0, const_rg_limit=False):
+    """
+    SDM with lognormal pore distribution and exponential residence time.
+    
+    Adds mobile phase dispersion (Brownian term) to GEC lognormal model.
+    
+    Parameters
+    ----------
+    w : array
+        Frequency array
+    N : float
+        Pore interaction scale parameter
+    T : float
+        Residence time scale parameter
+    me : float
+        Pore entry exponent
+    mp : float
+        Pore residence exponent
+    mu : float
+        Log-mean of pore size distribution
+    sigma : float
+        Log-std of pore size distribution
+    Rg : float
+        Molecule radius of gyration (lower integration limit)
+    N0 : float
+        Plate number (mobile phase dispersion parameter)
+    t0 : float
+        Mobile phase hold-up time (drift term)
+    const_rg_limit : bool, optional
+        Use constant integration limit (default: False)
+    
+    Returns
+    -------
+    complex array
+        Characteristic function values
+        
+    Notes
+    -----
+    CF structure: φ(ω) = exp(Z + Z²/(2*N0))
+    where Z = [lognormal pore integral] + iω*t0
+    
+    The Z²/(2*N0) term represents axial dispersion in mobile phase.
+    """
+    if const_rg_limit:
+        max_rg = PORESIZE_INTEG_LIMIT
+    else:
+        mode = compute_mode(mu, sigma)
+        stdev = compute_stdev(mu, sigma)
+        max_rg = min(PORESIZE_INTEG_LIMIT, mode + 5*stdev)
+
+    # Integrate over lognormal pore distribution (same as GEC)
+    integrated = complex_quadrature_vec(
+        lambda r: gec_lognormal_pore_integrand_impl(r, w, N, T, me, mp, mu, sigma, Rg), 
+        Rg, max_rg
+    )[0]
+    
+    # Add drift term to get Z
+    Z = integrated + 1j*w*t0
+    
+    # Add Brownian dispersion term
+    return np.exp(Z + Z**2/(2*N0))
+
+sdm_lognormal_pore_pdf_impl = FftInvPdf(sdm_lognormal_pore_cf)
+
+def sdm_lognormal_pore_pdf(x, scale, N, T, me, mp, mu, sigma, Rg, N0, t0):
+    """
+    PDF for SDM with lognormal pore distribution.
[…]
