### 2026-01-05 e57ccc0
Add Gamma distribution superiority analysis to the simulation
