### 2026-01-06 d423a07
Update to set today dynamically in review-start estimate
