### 2026-01-06 c4ce172
Working for XR only optimization
