+++ molass/Rigorous/LegacyBridgeUtils.py
@@ -16,8 +16,13 @@ def make_dsets_from_decomposition(decomposition, rg_curve, debug=False):
     xr_curve = ElCurve(*decomposition.xr_icurve.get_xy())
     D = ssd.xr.M
     E = ssd.xr.E
-    uv_curve = ElCurve(*decomposition.uv_icurve.get_xy())
-    U = ssd.uv.M
+    if decomposition.uv is None:
+        # temporary work-around for the case without UV data
+        uv_curve = xr_curve
+        U = D.copy()
+    else:
+        uv_curve = ElCurve(*decomposition.uv_icurve.get_xy())
+        U = ssd.uv.M
     dsets = ((xr_curve, D), LegacyRgCurve(xr_curve, rg_curve), (uv_curve, U))
     return OptDataSets(None, None, dsets=dsets, E=E)
 
@@ -88,6 +93,18 @@ def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, debug
         os.makedirs(optimizer_folder)
     if not os.path.exists(rg_folder):
         os.makedirs(rg_folder)
+
+    temp_in_folder = os.path.abspath(os.path.join(analysis_folder, "temp_in_folder"))
+    in_folder = get_setting('in_folder')
+    if in_folder is None:
+        in_folder = temp_in_folder
+        print(f"Exporting SecSaxsData to temporary folder: {in_folder}")
+        set_setting('in_folder', in_folder)
+        if not os.path.exists(in_folder):
+            os.makedirs(in_folder)
+    if os.path.exists(temp_in_folder):
+        assert in_folder == temp_in_folder
+        decomposition.ssd.export(temp_in_folder)
 
     # make datasets and basecurves
     from molass_legacy.RgProcess.RgCurve import check_rg_folder
