### 2026-01-07 e48cdf0
Working for XR only optimization
