+++ _sources/source/molass.SEC.Models.GEC.rst
@@ -1,0 +1,7 @@
+SEC.Models.GEC
+============================
+
+.. automodule:: molass.SEC.Models.GEC
+   :members:
+   :undoc-members:
+   :show-inheritance:
