+++ _sources/source/molass.Bridge.LegacyBaselines.rst
@@ -1,0 +1,7 @@
+Bridge.LegacyBaselines
+====================================
+
+.. automodule:: molass.Bridge.LegacyBaselines
+   :members:
+   :undoc-members:
+   :show-inheritance:
