+++ _sources/source/molass.Rigorous.RigorousEdmParams.rst
@@ -1,0 +1,7 @@
+Rigorous.RigorousEdmParams
+========================================
+
+.. automodule:: molass.Rigorous.RigorousEdmParams
+   :members:
+   :undoc-members:
+   :show-inheritance:
