+++ _sources/source/molass.Rigorous.RigorousSdmParams.rst
@@ -1,0 +1,7 @@
+Rigorous.RigorousSdmParams
+========================================
+
+.. automodule:: molass.Rigorous.RigorousSdmParams
+   :members:
+   :undoc-members:
+   :show-inheritance:
