+++ _sources/source/molass.Bridge.LegacyRgCurve.rst
@@ -1,0 +1,7 @@
+Bridge.LegacyRgCurve
+==================================
+
+.. automodule:: molass.Bridge.LegacyRgCurve
+   :members:
+   :undoc-members:
+   :show-inheritance:
