+++ _sources/source/molass.Rigorous.rst
@@ -1,0 +1,21 @@
+Rigorous
+=======================
+
+.. automodule:: molass.Rigorous
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   molass.Rigorous.CurrentStateUtils
+   molass.Rigorous.LegacyBridgeUtils
+   molass.Rigorous.RigorousEdmParams
+   molass.Rigorous.RigorousEghParams
+   molass.Rigorous.RigorousImplement
+   molass.Rigorous.RigorousSdmParams
+   molass.Rigorous.RunInfo
