### 2026-01-08 ee1adea
Update for document generation
