+++ molass/PlotUtils/GuinierPlot.py
@@ -4,35 +4,53 @@ PlotUtils.GuinierPlot.py
 import numpy as np
 import matplotlib.pyplot as plt
 
-def guinier_plot_impl(sg, axes=None, debug=False):
+def guinier_plot_general(q, I, e, guinier_region, plot_region, axes=None, debug=False):
     if axes is None:
-        fig, axes = plt.subplots(ncols=2, figsize=(12,5))
+        fig, axes = plt.subplots(nrows=2, ncols=2, figsize=(12,10))
     else:
-        fig = axes[0].figure
-    q = sg.x_
-    I = sg.y_
+        fig = axes[0,0].figure
+
     lnI = np.log(I)
     q2 = q**2
+    plot_start, plot_stop = plot_region
+    print(f"Guinier plot range: start={plot_start}, stop={plot_stop}")
+    ax1, ax2 = axes[0,:]
+    ax1.plot(q, I, 'o', markersize=3, label='Data')
+    ax1.set_title('Linear Plot')
+    ax1.set_xlabel(r"$q [\AA^{-1}]$")
+    ax1.set_ylabel(r'$Intensity$')
+    ax2.set_yscale('log')
+    ax2.set_xscale('log')
+    ax2.set_title('Logarithmic Plot')
+    ax2.set_xlabel(r"$q [\AA^{-1}]$")
+    ax2.plot(q, I, 'o', markersize=3, label='Data')
+    plot_slice = slice(plot_start, plot_stop)
+    ax3, ax4 = axes[1,:]
+    ax3.plot(q[plot_slice], I[plot_slice], 'o', markersize=3, label='Data')
+    gindeces = np.array(guinier_region)
+    ax3.axvspan(*q[gindeces], color='red', alpha=0.2, label='Guinier region')
+    ax3.set_title('Linear Plot')
+    ax3.set_xlabel(r"$q [\AA^{-1}]$")
+    ax3.set_ylabel(r'$Intensity$')
+    ax4.plot(q2[plot_slice], lnI[plot_slice], 'o', markersize=3, label='Data')
+    ax4.axvspan(*q2[gindeces], color='red', alpha=0.2, label='Guinier region')
+    ax4.set_xlabel(r"$q^2 [\AA^{-2}]$")
+    ax4.set_ylabel(r'$\ln(Intensity)$')
+    ax4.set_title('Guinier Plot')
+    ax4.grid()
+    ax4.legend()
+    fig.tight_layout()
+    return fig, axes
+
+def guinier_plot_impl(sg, axes=None, debug=False):
+    q = sg.x_
     if sg.Rg is None:
         start= 0
         stop = len(q)//8
     else:
         start = sg.guinier_start
         stop = sg.guinier_stop
-    print(f"Guinier plot range: start={start}, stop={stop}, Rg={sg.Rg}")
-    gslice = slice(start, stop)
-    ax1, ax2 = axes
-    ax1.plot(q[gslice], I[gslice], 'o', markersize=3, label='Data')
-    ax1.set_title('Linear Plot')
-    ax1.set_xlabel(r"$q [\AA^{-1}]$")
-    ax1.set_ylabel(r'$Intensity$')
-    ax2.plot(q2[gslice], lnI[gslice], 'o', markersize=3, label='Data')
-    ax2.set_xlabel(r"$q^2 [\AA^{-2}]$")
-    ax2.set_ylabel(r'$\ln(Intensity)$')
-    ax2.set_title('Guinier Plot')
-    ax2.grid()
-    ax2.legend()
-    return fig, axes
+    guinier_plot_general(q, sg.y_, sg.e_, (start, stop), (start, stop), axes=axes, debug=debug)
 
 def inspect_guinier_plot(sg, debug=False):
     if debug:
[…]
