+++ molass/Guinier/SimpleFallback.py
@@ -1,0 +1,312 @@
+"""
+Guinier.SimpleFallback.py
+"""
+import numpy as np
+import matplotlib.pyplot as plt
+MIN_RG = 10.0   # Angstrom
+MAX_RG = 100.0   # Angstrom
+
+def compute_rg(qw2, lnI, weights):
+    """
+    Compute radius of gyration (Rg) using weighted linear regression.
+    """
+    W = np.diag(weights)
+    A = np.vstack([qw2, np.ones(len(qw2))]).T
+    Aw = W @ A
+    lnIw = W @ lnI
+    coeffs, residuals, rank, s = np.linalg.lstsq(Aw, lnIw, rcond=None)
+    slope = coeffs[0]
+    Rg = np.sqrt(-3 * slope)
+    return Rg
+
+def compute_r_squared(qw2, lnI, weights):
+    """
+    Compute R² (coefficient of determination) for weighted linear regression.
+    
+    Parameters
+    ----------
+    qw2 : ndarray
+        q² values
+    lnI : ndarray
+        Natural log of intensity
+    weights : ndarray
+        Weights for each point
+    
+    Returns
+    -------
+    float
+        R² value (closer to 1.0 means better linearity)
+    """
+    W = np.diag(weights)
+    A = np.vstack([qw2, np.ones(len(qw2))]).T
+    Aw = W @ A
+    lnIw = W @ lnI
+    
+    coeffs, _, _, _ = np.linalg.lstsq(Aw, lnIw, rcond=None)
+    
+    # Compute fitted values
+    lnI_fit = coeffs[0] * qw2 + coeffs[1]
+    
+    # Compute R² with weights
+    residuals = lnI - lnI_fit
+    ss_res = np.sum(weights * residuals**2)
+    
+    lnI_mean = np.sum(weights * lnI) / np.sum(weights)
+    ss_tot = np.sum(weights * (lnI - lnI_mean)**2)
+    
+    if ss_tot == 0:
+        return 0.0
+    
+    r_squared = 1 - (ss_res / ss_tot)
+    return r_squared
+
+def estimate_rg_simply(data, rg_range=None, min_num_points=5, q_rg_limit=1.3, initial_q_max=0.025):
+    """
+    A simple fallback function for Guinier analysis.
+    
+    This function uses a multi-step approach:
+    1. Sliding window to find the most linear region and exclude problematic initial points
+    2. Initial conservative fit with the best window to estimate Rg
+    3. Range refinement based on q*Rg < q_rg_limit criterion
+    4. Final fit with the refined range
+    
+    Parameters
+    ----------
+    data : ndarray
+        SAXS data with columns [q, I, error]
+    rg_range : tuple, optional
+        Expected range of Rg values (min_rg, max_rg) in Angstroms.
+        Default is (MIN_RG, MAX_RG) = (10.0, 80.0)
+    min_num_points : int, optional
+        Minimum number of points required in the Guinier region (default: 5)
+    q_rg_limit : float, optional
+        Guinier criterion limit for q*Rg (default: 1.3)
+    initial_q_max : float, optional
+        Initial conservative q maximum for searching the best starting window (default: 0.025 Å⁻¹)
+    
+    Returns
+    -------
+    dict
+        Dictionary containing:
+        - 'Rg': Radius of gyration in Angstroms
+        - 'I0': Forward scattering intensity
+        - 'q_start': Starting index of the fitted range
+        - 'q_end': Ending index of the fitted range
+        - 'q_min': Minimum q value used in fit (Å⁻¹)
+        - 'q_max': Maximum q value used in fit (Å⁻¹)
+        - 'n_points': Number of points used in fit
+        - 'q_rg_max': Maximum q*Rg value in the fitted range
+        - 'r_squared': R² value indicating quality of linear fit
[…]

[…]
