### 2026-01-08 e7868bd
Improve Rg estimation
