### 2026-01-08 1dd2960
Change xr_only set timing
