### 2026-01-09 39479fa
Fix uv baseline issue in xr_only mode
