### 2026-01-13 ce56536
Temporary fix for DENSS 1.8.7
