+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "molass"
-version = "0.8.0"
+version = "0.8.1"
 
 dependencies = [
   "numpy",
@@ -12,7 +12,7 @@ dependencies = [
   "matplotlib",
   "tqdm",
   "pybaselines>=1.2.0",
-  "molass_legacy>=0.4.3",
+  "molass_legacy>=0.5.0",
   "statsmodels",
   "scikit-learn",
   "toml",
