### 2026-01-14 6911c85
Update for rigorous SDM optimization
