### 2026-01-14 a22e098
Add monitor to RunInfo
