### 2026-01-16 fd9440b
Update for reorganization of plot results
