### 2026-01-19 e1d84ea
Add a note for PDF reading
