### 2026-01-19 f79073a
Add three required JOSS sections to paper.md
- Add Software Design section with sustainability rationale and AI-maintainability design philosophy
- Add Research Impact Statement positioning against CHROMIXS, EFAMIX, and REGALS limitations
- Add AI usage disclosure with transparent reporting of AI tool usage
- Add BibTeX entries for Panjkovich 2018 (CHROMIXS) and Konarev 2021 (EFAMIX)
- Reorganize sections to follow JOSS example structure: Statement of Need → Software Design → Research Impact Statement
- All citations verified against paper.bib

Addresses GitHub issue #9424 requirements for new
[…]
