+++ molass-library/paper.bib
@@ -348,7 +348,30 @@ pages={2069-2080},
 doi={10.2298/TSCI201229179U},
 url= {https://doi.org/10.2298/TSCI201229179U},
 }
+@article{Panjkovich:2018,
+author = "Panjkovich, Alejandro and Svergun, Dmitri I.",
+title = "{CHROMIXS: automatic and interactive analysis of chromatography-coupled small-angle X-ray scattering data}",
+journal = "Bioinformatics",
+year = "2018",
+volume = "34",
+number = "11",
+pages = "1944--1946",
+month = "June",
+doi = {10.1093/bioinformatics/btx846},
+url = {https://doi.org/10.1093/bioinformatics/btx846}
+}
 
+@article{Konarev:2021,
+author = "Konarev, Petr V. and Graewert, Melissa A. and Jeffries, Cy M. and Fukuda, Masakazu and Cheremnykh, Taisiia A. and Volkov, Vladimir V. and Svergun, Dmitri I.",
+title = "{EFAMIX, a tool to decompose inline chromatography SAXS data from partially overlapping components}",
+journal = "Protein Science",
+year = "2021",
+volume = "31",
+pages = "269--282",
+month = "Dec",
+doi = {10.1002/pro.4237},
+url = {https://doi.org/10.1002/pro.4237}
+}
 @article{Meisburger:mf5050,
 author = "Meisburger, Steve P. and Xu, Da and Ando, Nozomi",
 title = "{{\it REGALS}: a general method to deconvolve X-ray scattering data from evolving mixtures}",
