### 2026-01-19 6e049fd
Delete the sentence on planned paper
