### 2026-01-20 6242cbf
Update paper.md: Revise Research Impact Statement and Software Design section
- Remove 'forthcoming publication' paragraph from Research Impact Statement
- Add concrete details about rigorous optimization with SEC/SAXS theory constraints
- Change 'retiring within one year' to 'approaching retirement' for accuracy
- Strengthen pedagogical aspect in Software Design section
