+++ paper.md
@@ -56,11 +56,11 @@ Molass Library currently implements steps 3–7. For steps 1 and 2, users may em
 
 # Software Design
 
-Molass Library was designed to address a specific sustainability challenge: the primary developer is retiring within one year, with no confirmed successor. Rather than contributing to existing GUI-based tools like ATSAS or BioXTAS RAW, we chose to build a Python-first, Jupyter-centric solution optimized for AI-assisted maintenance and researcher extensibility. This design philosophy prioritizes explicit, readable code over graphical interfaces, enabling future researchers to leverage AI coding assistants for maintenance and extension. The architecture emphasizes modularity and clear separation of concerns: elution curve models (EGH, SDM, EDM) are isolated as parametric functions; low-rank factorization uses standard linear algebra (Moore-Penrose pseudoinverse); and visualization is decoupled from computation. By integrating established packages (NumPy, SciPy, pybaselines, ruptures) rather than reimplementing core functionality, we reduce custom code volume and enhance long-term viability. This approach explicitly optimizes for scenarios where AI tools—rather than human experts—must interpret, debug, and extend the codebase.
+Molass Library was designed to address a specific sustainability challenge: the primary developer is approaching retirement, with no confirmed successor. Rather than contributing to existing GUI-based tools like ATSAS or BioXTAS RAW, we chose to build a Python-first, Jupyter-centric solution optimized for AI-assisted maintenance and researcher extensibility. This design philosophy prioritizes explicit, readable code over graphical interfaces, making the analysis methodology transparent for both learning and AI-assisted maintenance. The architecture emphasizes modularity and clear separation of concerns: elution curve models (EGH, SDM, EDM) are isolated as parametric functions; low-rank factorization uses standard linear algebra (Moore-Penrose pseudoinverse); and visualization is decoupled from computation. By integrating established packages (NumPy, SciPy, pybaselines, ruptures) rather than reimplementing core functionality, we reduce custom code volume and enhance long-term viability. This approach explicitly optimizes for scenarios where AI tools—rather than human experts—must interpret, debug, and extend the codebase.
 
 # Research Impact Statement
 
-Molass Library has demonstrated credible research significance through its predecessor's track record [@Yonezawa:2023] and its positioning against documented limitations of alternative approaches. While CHROMIXS explicitly defers analysis of overlapping chromatographic peaks to "other methods" [@Panjkovich:2018], and EFAMIX quantifies failure thresholds (signal-to-noise ratio ≥10³ required for three-component separation; peak asymmetry parameter τ≤2; baseline width separation ≥2× for reliable decomposition) [@Konarev:2021], Molass Library addresses these scenarios through explicit parametric modeling of elution curves. REGALS employs a two-stage approach combining EFA with L1 regularization [@Meisburger:mf5050], inheriting EFA's fundamental limitations in overlapping peak scenarios.
[…]
