### 2026-01-20 c2930ba
Refine Research Impact Statement wording
- Replace 'assumption-dependent methods' with more nuanced phrasing
- Emphasize widening scope of tractable cases rather than claiming complete solution
- Maintain accurate representation of capabilities and limitations
