### 2026-01-20 1f2e869
Clarify Software Design section wording
- Replace ambiguous 'AI tools rather than human experts' phrase
- Use clearer language: 'allows domain researchers to maintain and extend code using AI-assisted tools'
- Avoid misinterpretation that AI works independently of humans
