+++ paper.md
@@ -56,7 +56,7 @@ Molass Library currently implements steps 3–7. For steps 1 and 2, users may em
 
 # Software Design
 
-Molass Library was designed to address a specific sustainability challenge: the primary developer is approaching retirement, with no confirmed successor. Rather than contributing to existing GUI-based tools like ATSAS or BioXTAS RAW, we chose to build a Python-first, Jupyter-centric solution optimized for AI-assisted maintenance and researcher extensibility. This design philosophy prioritizes explicit, readable code over graphical interfaces, making the analysis methodology transparent for both learning and AI-assisted maintenance. The architecture emphasizes modularity and clear separation of concerns: elution curve models (EGH, SDM, EDM) are isolated as parametric functions; low-rank factorization uses standard linear algebra (Moore-Penrose pseudoinverse); and visualization is decoupled from computation. By integrating established packages (NumPy, SciPy, pybaselines, ruptures) rather than reimplementing core functionality, we reduce custom code volume and enhance long-term viability. This approach explicitly optimizes for scenarios where AI tools—rather than human experts—must interpret, debug, and extend the codebase.
+Molass Library was designed to address a specific sustainability challenge: the primary developer is approaching retirement, with no confirmed successor. Rather than contributing to existing GUI-based tools like ATSAS or BioXTAS RAW, we chose to build a Python-first, Jupyter-centric solution optimized for AI-assisted maintenance and researcher extensibility. This design philosophy prioritizes explicit, readable code over graphical interfaces, making the analysis methodology transparent for both learning and AI-assisted maintenance. The architecture emphasizes modularity and clear separation of concerns: elution curve models (EGH, SDM, EDM) are isolated as parametric functions; low-rank factorization uses standard linear algebra (Moore-Penrose pseudoinverse); and visualization is decoupled from computation. By integrating established packages (NumPy, SciPy, pybaselines, ruptures) rather than reimplementing core functionality, we reduce custom code volume and enhance long-term viability. This approach allows domain researchers to maintain and extend the code using AI-assisted development tools.
 
 # Research Impact Statement
 
