### 2026-01-20 39fffad
Restructure Research Impact Statement for better flow
- Group all alternative tool limitations together (CHROMIXS, EFAMIX, REGALS)
- Present Molass response as unified solution after all limitations
- Eliminate awkward back-and-forth between alternatives and Molass
- Improve logical progression and readability
