+++ molass-library/paper.md
@@ -60,7 +60,7 @@ Molass Library was designed to address a specific sustainability challenge: the 
 
 # Research Impact Statement
 
-Molass Library has demonstrated credible research significance through its predecessor's track record [@Yonezawa:2023] and its positioning against documented limitations of alternative approaches. While CHROMIXS explicitly defers analysis of overlapping chromatographic peaks to "other methods" [@Panjkovich:2018], and EFAMIX quantifies failure thresholds (signal-to-noise ratio ≥10³ required for three-component separation; peak asymmetry parameter τ≤2; baseline width separation ≥2× for reliable decomposition) [@Konarev:2021], Molass Library addresses these scenarios through explicit parametric modeling of elution curves. REGALS employs a two-stage approach combining EFA with L1 regularization [@Meisburger:mf5050], inheriting EFA's fundamental limitations in overlapping peak scenarios. The library implements rigorous global optimization methods that constrain solutions using SEC theory (size-dependent elution) and SAXS theory (radius of gyration consistency), widening the scope of tractable cases for overlapping peak decomposition beyond what existing methods can reliably address.
+Molass Library has demonstrated credible research significance through its predecessor's track record [@Yonezawa:2023] and its positioning against documented limitations of alternative approaches. While CHROMIXS explicitly defers analysis of overlapping chromatographic peaks to "other methods" [@Panjkovich:2018], EFAMIX quantifies failure thresholds (signal-to-noise ratio ≥10³ required for three-component separation; peak asymmetry parameter τ≤2; baseline width separation ≥2× for reliable decomposition) [@Konarev:2021], and REGALS employs a two-stage approach combining EFA with L1 regularization [@Meisburger:mf5050], inheriting EFA's fundamental limitations in overlapping peak scenarios. Molass Library addresses these documented limitations through explicit parametric modeling of elution curves and rigorous global optimization methods that constrain solutions using SEC theory (size-dependent elution) and SAXS theory (radius of gyration consistency), widening the scope of tractable cases for overlapping peak decomposition beyond what existing methods can reliably address.
 
 # Notable package dependencies
 
