### 2026-01-20 15b4bfc
Revise Software Design section: AI-assisted maintenance claim accuracy
Changes to paper.md:
- Changed "optimized for" to "designed to support" AI-assisted maintenance
- Added sentence about active enhancement through systematic usability testing
- Maintained "allows" (not "enables") for appropriate modesty

Context documented in review-discussions/response-drafts/paper_revision_context.md

This revision reflects empirical testing results showing the library is not yet "optimized" but has systematic verification framework in place for achieving that goal.
