+++ review-discussions/response-drafts/paper_revision_context.md
@@ -1,0 +1,308 @@
+# Paper Revision Context: Software Design Section
+
+**Date**: January 20, 2026  
+**Repository**: biosaxs-dev/modeling-vs-model_free  
+**Related Documents**: 
+- `AI_assisted_maintenance_framework.md` (verification framework)
+- `molass/paper.md` (JOSS submission draft)
+- `evidence/efa_original/EFA_limitations_verification.ipynb` (verification notebook)
+
+---
+
+## Summary of Changes
+
+### Section Modified
+**Software Design** section of JOSS paper submission
+
+### Change Type
+Accuracy improvement: Changed claim from "achieved state" to "ongoing effort with active verification"
+
+---
+
+## What Was Changed
+
+### Before (Original Claim)
+> "Rather than contributing to existing GUI-based tools like ATSAS or BioXTAS RAW, we chose to build a Python-first, Jupyter-centric solution **optimized for AI-assisted maintenance** and researcher extensibility."
+
+### After (Revised Claim)
+> "Rather than contributing to existing GUI-based tools like ATSAS or BioXTAS RAW, we chose to build a Python-first, Jupyter-centric solution **designed to support AI-assisted maintenance** and researcher extensibility. [...] **We are actively enhancing AI-readiness through systematic usability testing with AI agents and iterative improvements to API discoverability and inline documentation.**"
+
+### Key Differences
+1. **"optimized for"** → **"designed to support"** (acknowledges current state)
+2. **Added evidence** of active verification and improvement process
+3. **Maintained forward-looking tone** while being honest about status
+
+---
+
+## Why This Change Was Necessary
+
+### Empirical Testing Revealed Gap
+During verification of JOSS claims, an AI assistant (Claude 3.5 Sonnet) attempted to use Molass Library for a basic task:
+
+**Task**: "Load SAMPLE1 from molass_data and measure SNR from real SEC-SAXS data"
+
+**Result**:
+- ❌ **6 attempts** required to complete task
+- ❌ Required **external documentation lookup**
+- ❌ Needed **workaround code** for dimension mismatch
+- ❌ No helpful error messages to guide toward solution
+- ❌ API not discoverable without prior knowledge
+
+**Conclusion**: Library is NOT currently "optimized" for AI-assisted maintenance.
+
+### Specific Friction Points
+1. **Q-vector access**: Tried `xr_data.sv`, `.qv`, `.q` - all failed
+2. **Method discovery**: `get_spectral_vectors()` not intuitive
+3. **Dimension mismatch**: Returned vector length ≠ matrix shape (1028 vs 242)
+4. **No inline guidance**: Error messages didn't suggest solutions
+5. **Tutorial gap**: Common operations not explicitly documented
+
+### Full Documentation
+See `AI_assisted_maintenance_framework.md` Section 1 "Pilot Case Study" for complete testing transcript.
+
+---
+
+## Evidence Supporting Revision
+
+### Quantitative Metrics from Testing
+- **Success on first attempt**: 0% (baseline task failed)
+- **Iterations to completion**: 6 (target: ≤2 for "optimized")
+- **Documentation lookups**: 1 (target: 0 for "optimized")  
+- **Workarounds required**: 1 (target: 0 for "optimized")
+
+### Qualitative Assessment
+**What worked well** ✅:
+- Standard Python conventions (`.M` for matrix)
+- Logical object hierarchy (`SecSaxsData → xr → M`)
+- Documentation exists and is accessible
+
+**What needs improvement** ⚠️:
+- API discoverability (method names not intuitive)
+- Error guidance (no hints when operations fail)
+- Edge case handling (dimension mismatches unexplained)
+
+### Conclusion
[…]
