### 2026-01-20 6b93f69
Squashed 'molass-library/' changes from cad635f..fe37f85
fe37f85 Revise Software Design section: AI-assisted maintenance claim accuracy
9a89e04 Restructure Research Impact Statement for better flow
1f2e869 Clarify Software Design section wording
eb9fb19 Refine Research Impact Statement wording
6242cbf Update paper.md: Revise Research Impact Statement and Software Design section
7907b21 Delete the sentence on planned paper
5d04ebb Add three required JOSS sections to paper.md
ff868e0 Merge commit '761fa2bd7168f849aaeb57aca400fbb3a853926d'
a20b361 Merge commit '77348770b44f3f1cf36c8434610e2b9a2b7
[…]
