### 2026-01-21 259613a
Improve documentation to loosen the role division and memorize github account mapping
