### 2026-01-21 0cd673e
Address Shimizu's feedback: clarify facilities and soften sustainability language
