### 2026-01-21 a5ec1e6
Improve docs: clarify subtree workflow to prevent branch confusion
