### 2026-01-23 c16d06e
Update Copilot guidelines to add ai-friendlyness
