### 2026-01-29 1c6fd46
Add supporting evidence citation and future AI disclosure to JOSS paper
- Added companion repository citation in Research Impact Statement
- Added time-bound AI usage disclosure for future development
- Added bibliography entry for modeling-vs-model_free repository
- Added Pandoc citation formatting markers
