### 2026-01-29 9615b0e
Add notational caveat to theoretical foundation section
Clarify that M = P·C is conceptual; in practice methods minimize ||M - PC||² plus regularization rather than achieving exact equality.
