### 2026-01-29 3cdb53c
Squashed 'molass-library/' changes from 6b7c64d..8d819c3
8d819c3 Add notational caveat to theoretical foundation section
1c6fd46 Add supporting evidence citation and future AI disclosure to JOSS paper

git-subtree-dir: molass-library
git-subtree-split: 8d819c3a68adec4cf35bb113e112f1e88b11de9a
