+++ molass-library/paper.bib
@@ -430,12 +430,13 @@ url = {https://www.sciencedirect.com/science/article/pii/S0079610763800152},
 author = {O. Kratky}
 }
 
-@misc{takahashi2026modeling,
-  author = {Takahashi, Masatsuyo and Shimizu, Nobutaka},
-  title = {Modeling vs Model-Free SEC-SAXS Analysis: Mathematical Foundations and Practical Limitations of Matrix Factorization Methods},
-  year = {2026},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/biosaxs-dev/modeling-vs-model_free},
-  note = {Research-in-progress with AI assistance (GitHub Copilot)}
-}
+% REMOVED (Feb 2, 2026): Citation to modeling-vs-model_free repository (now private, may be restored later)
+%@misc{takahashi2026modeling,
+%  author = {Takahashi, Masatsuyo and Shimizu, Nobutaka},
+%  title = {Modeling vs Model-Free SEC-SAXS Analysis: Mathematical Foundations and Practical Limitations of Matrix Factorization Methods},
+%  year = {2026},
+%  publisher = {GitHub},
+%  journal = {GitHub repository},
+%  url = {https://github.com/biosaxs-dev/modeling-vs-model_free},
+%  note = {Research-in-progress with AI assistance (GitHub Copilot)}
+%}
