### 2026-02-02 461f75b
Remove references to private modeling-vs-model_free repository
- Commented out citation in paper.bib
- Removed sentence citing companion repository in Research Impact section
- Added CHANGELOG documenting changes for potential future restoration
- Reason: JOSS requires all cited materials to be publicly accessible
