### 2026-02-02 a25bdc7
Remove problematic comments to fix JOSS PDF generation
- Remove HTML comment containing unresolved citation reference from paper.md
- Remove BibTeX comment block from paper.bib
- CHANGELOG_2026-02-02.md and git history preserve what was removed
- This should resolve pandoc PDF generation failure
