+++ paper.md
@@ -62,10 +62,6 @@ The MOLASS library is designed to support sustainable software development by pr
 
 Molass Library has demonstrated credible research significance through its predecessor's track record [@Yonezawa:2023] and its positioning against documented limitations of alternative approaches. While CHROMIXS explicitly defers analysis of overlapping chromatographic peaks to "other methods" [@Panjkovich:2018], EFAMIX quantifies failure thresholds (signal-to-noise ratio ≥10³ required for three-component separation; peak asymmetry parameter τ≤2; baseline width separation ≥2× for reliable decomposition) [@Konarev:2021], and REGALS employs a two-stage approach combining EFA with L1 regularization [@Meisburger:mf5050], inheriting EFA's fundamental limitations in overlapping peak scenarios. Molass Library addresses these documented limitations through explicit parametric modeling of elution curves and rigorous global optimization methods that constrain solutions using SEC theory (size-dependent elution) and SAXS theory (radius of gyration consistency), widening the scope of tractable cases for overlapping peak decomposition beyond what existing methods can reliably address.
 
-<!-- REMOVED (Feb 2, 2026): Reference to modeling-vs-model_free repository (now private, may be restored later)
-Detailed evidence extraction, mathematical analysis of regularization behavior, and computational validation of these documented limitations is provided in a companion research repository [@takahashi2026modeling].
--->
-
 # Notable package dependencies
 
 Molass Library is built on robust scientific Python libraries, including NumPy, SciPy, and Matplotlib. It further integrates:
