### 2026-02-02 cb771b1
Update status: Paper sections complete, private repo references removed
- Updated date to Feb 2, 2026
- Marked all paper sections as complete
- Documented removal of private repository references
- Status: Awaiting Shimizu's review of latest changes before JOSS notification
