### 2026-02-03 a65af60
Squashed 'molass-library/' changes from 8d819c3..a25bdc7
a25bdc7 Remove problematic comments to fix JOSS PDF generation
461f75b Remove references to private modeling-vs-model_free repository

git-subtree-dir: molass-library
git-subtree-split: a25bdc7b13362a39f367b6b49d8698effaaf1dba
