+++ CHANGELOG_2026-02-02.md
@@ -1,0 +1,23 @@
+# Changes to paper.md and paper.bib (February 2, 2026)
+
+## Summary
+
+Removed reference to the `modeling-vs-model_free` repository, which has been made private.
+
+## Files Modified
+
+### paper.md
+- **Research Impact Statement section**: Removed sentence citing the companion research repository
+- Change preserved in HTML comment for potential future restoration
+
+### paper.bib
+- **Citation entry**: Commented out `@misc{takahashi2026modeling}` 
+- Change preserved in BibTeX comment for potential future restoration
+
+## Reason
+
+The referenced repository has been changed from public to private visibility. JOSS requires all cited materials to be publicly accessible.
+
+## Status
+
+Changes may be restored if repository visibility changes in the future. All modifications are marked with restoration notes.
