### 2026-02-09 47afae9
Add a response draft
