+++ review-discussions/response-drafts/joss-paper-update-notification.md
@@ -1,0 +1,134 @@
+# Response to JOSS: Paper Updated with Required Sections
+
+**Date**: February 9, 2026  
+**Author**: Shimizu (to be posted as nshimizu0721)  
+**Target**: JOSS Review Issue #9424  
+**Status**: DRAFT - Ready for technical review by Takahashi
+
+---
+
+## Draft Message to Post
+
+---
+
+@arfon Thank you for your patience and for the clear guidance on the updated requirements.
+
+We have updated our paper to include all four required sections:
+
+1. **State of the Field**: Integrated within our existing "Statement of Need" section, which compares Molass Library to existing tools (ATSAS, BioXTAS RAW) and provides our "build vs. contribute" justification.
+
+2. **Software Design** (~200 words): Added. We explain our design philosophy—choosing a Python-first, Jupyter-centric architecture to support AI-assisted maintenance and researcher extensibility. The section describes our modular architecture and integration of established scientific Python packages.
+
+3. **Research Impact Statement** (~200 words): Added. We document how Molass Library addresses documented limitations in existing methods (CHROMIXS, EFAMIX, REGALS) for overlapping chromatographic peak decomposition, grounded in both our predecessor's track record and specific technical advantages.
+
+4. **AI Usage Disclosure**: Added. We provide transparent disclosure of generative AI use in paper text (grammar improvement), code generation (standard boilerplate only), and documentation (initial drafting). All AI-generated content was reviewed and validated by the human authors.
+
+The updated paper is available in the `joss-paper` branch of our repository: [https://github.com/biosaxs-dev/molass-library/blob/joss-paper/paper.md](https://github.com/biosaxs-dev/molass-library/blob/joss-paper/paper.md)
+
+We believe the updated paper demonstrates the scholarly contribution and research significance required by JOSS's updated scope criteria. We are ready for your assessment and, if appropriate, assignment of an editor to begin formal review.
+
+Please let us know if any clarifications or additional information would be helpful.
+
+---
+
+## Internal Notes (Not for JOSS)
+
+### Pre-Posting Checklist
+
+**Before posting, I (Shimizu) need to verify:**
+
+- [ ] **All four sections present**: ✅ Confirmed (see paper.md)
+- [ ] **Word counts appropriate**: Need to verify (~200 words each)
+- [ ] **Citations accurate**: Need to check CHROMIXS, EFAMIX, REGALS claims
+- [ ] **Link correct**: Verify joss-paper branch has latest changes
+- [ ] **Tone appropriate**: Professional, respectful, not defensive
+
+### Questions for Takahashi (Technical Review)
+
+**Before I post this, I need Takahashi to confirm:**
+
+1. **Link verification**: Does the joss-paper branch link work correctly? Can reviewers access it?
+
+2. **Technical claims**: Are the competitor limitation claims in Research Impact Statement accurate?
+   - CHROMIXS "explicitly defers analysis of overlapping chromatographic peaks"
+   - EFAMIX thresholds: SNR ≥10³, τ≤2, baseline width separation ≥2×
+   - REGALS "inheriting EFA's fundamental limitations"
+
+3. **AI disclosure accuracy**: Is the AI usage disclosure complete and accurate? Did we use AI for anything else?
+
+4. **Software Design section**: Is the technical architecture description accurate? Any misrepresentations of design decisions?
+
+5. **Timing**: Is now the right time to post? Any other updates needed first?
+
+### Risk Assessment
+
+**Potential JOSS concerns:**
+
+1. **"Approaching retirement" statement**: Might raise sustainability questions
[…]
