+++ paper.bib
@@ -429,3 +429,15 @@ issn = {0079-6107},
 url = {https://www.sciencedirect.com/science/article/pii/S0079610763800152},
 author = {O. Kratky}
 }
+
+@article{Matsui:2024,
+author = {Matsui, Tsutomu and Rajkovic, Ivan and Mooers, Blaine H. M. and Liu, Ping and Weiss, Thomas M.},
+title = {Adaptable {SEC}-{SAXS} data collection for higher quality structure analysis in solution},
+journal = {Protein Science},
+year = {2024},
+volume = {33},
+number = {4},
+pages = {e4946},
+doi = {10.1002/pro.4946},
+url = {https://doi.org/10.1002/pro.4946}
+}
