### 2026-02-13 59c2427
Squashed 'molass-library/' changes from a25bdc7..11325ff
11325ff Modify for Research Impact Statement
68cb23e Add Matsui 2024 citation to Research Impact Statement

git-subtree-dir: molass-library
git-subtree-split: 11325ff83b9abf891917be9a6d4eca3e4af1dd12
