+++ paper.bib
@@ -429,3 +429,41 @@ issn = {0079-6107},
 url = {https://www.sciencedirect.com/science/article/pii/S0079610763800152},
 author = {O. Kratky}
 }
+
+@article{Jiang:2023,
+author = {Jiang, Xuguang and Ogawa, Tadayuki and Yonezawa, Kento and Danev, Radostin and Kikkawa, Masahiro},
+title = {The two-step cargo recognition mechanism of heterotrimeric kinesin},
+journal = {EMBO Reports},
+year = {2023},
+volume = {24},
+number = {11},
+pages = {e56864},
+month = {Nov},
+doi = {10.15252/embr.202356864},
+url = {https://doi.org/10.15252/embr.202356864}
+}
+
+@article{Furukawa:2025,
+author = {Furukawa, Ayako and Yonezawa, Kento and Negami, Tatsuki and Yoshimura, Yuriko and Hayashi, Aki and Nakayama, Jun-ichi and Adachi, Naruhiko and Senda, Toshiya and Shimizu, Kentaro and Terada, Tohru and Shimizu, Nobutaka},
+title = {A dynamic structural unit of phase-separated heterochromatin protein 1α as revealed by integrative structural analyses},
+journal = {Nucleic Acids Research},
+year = {2025},
+volume = {53},
+number = {6},
+pages = {gkaf154},
+month = {Mar},
+doi = {10.1093/nar/gkaf154},
+url = {https://doi.org/10.1093/nar/gkaf154}
+}
+
+@article{Matsui:2024,
+author = {Matsui, Tsutomu and Rajkovic, Ivan and Mooers, Blaine H. M. and Liu, Ping and Weiss, Thomas M.},
+title = {Adaptable {SEC}-{SAXS} data collection for higher quality structure analysis in solution},
+journal = {Protein Science},
+year = {2024},
+volume = {33},
+number = {4},
+pages = {e4946},
+doi = {10.1002/pro.4946},
+url = {https://doi.org/10.1002/pro.4946}
+}
