+++ paper.md
@@ -60,7 +60,7 @@ The MOLASS library is designed to support sustainable software development by pr
 
 # Research Impact Statement
 
-Molass Library has demonstrated credible research significance through its predecessor's track record [@Yonezawa:2023] and its positioning against documented limitations of alternative approaches. While CHROMIXS explicitly defers analysis of overlapping chromatographic peaks to "other methods" [@Panjkovich:2018], EFAMIX quantifies failure thresholds (signal-to-noise ratio ≥10³ required for three-component separation; peak asymmetry parameter τ≤2; baseline width separation ≥2× for reliable decomposition) [@Konarev:2021], and REGALS employs a two-stage approach combining EFA with L1 regularization [@Meisburger:mf5050], inheriting EFA's fundamental limitations in overlapping peak scenarios. Molass Library addresses these documented limitations through explicit parametric modeling of elution curves and rigorous global optimization methods that constrain solutions using SEC theory (size-dependent elution) and SAXS theory (radius of gyration consistency), widening the scope of tractable cases for overlapping peak decomposition beyond what existing methods can reliably address.
+The MOLASS library has demonstrated its scientific value through research conducted using its predecessor GUI tool [@Jiang:2023; @Furukawa:2025], as well as through its contribution to the development of SEC-SAXS data analysis systems at synchrotron radiation facilities [@Yonezawa:2023]. In SEC-SAXS experiments, component decomposition techniques are required when significant interparticle interference occurs or when chromatographic peaks overlap. MOLASS is one of several software packages developed to address this challenge and is currently regarded as providing one of the most rigorous peak decomposition frameworks available [@Matsui:2024]. While CHROMIXS explicitly defers the analysis of overlapping chromatographic peaks to “other methods” [@Panjkovich:2018], EFAMIX defines quantitative conditions under which reliable separation can be achieved (signal-to-noise ratio ≥10³ for three-component separation; peak asymmetry parameter τ≤2; baseline width separation ≥2× for reliable decomposition) [@Konarev:2021]. REGALS employs a two-stage framework combining EFA with L1 regularization [@Meisburger:mf5050] and is therefore guided by the assumptions underlying EFA in cases of strongly overlapping peaks. The MOLASS library approaches overlapping peak decomposition through explicit parametric modeling of elution curves combined with global optimization. The solution space is constrained using physicochemical consistency derived from SEC theory (size-dependent elution behavior) and SAXS analysis (consistency of radius of gyration across elution fractions). This approach offers an alternative and complementary algorithmic strategy for modeling overlapping chromatographic peaks under physically motivated constraints.
 
 # Notable package dependencies
 
