### 2026-02-15 af90e24
Squashed 'molass-library/' changes from 68cb23e..51e7ad5
51e7ad5 Modify for Research Impact Statement #2
11325ff Modify for Research Impact Statement

git-subtree-dir: molass-library
git-subtree-split: 51e7ad5576bda3464386482dd96003282044db3e
