+++ CURRENT-STATUS.md
@@ -1,13 +1,13 @@
 # CURRENT STATUS
 
-**📅 Last Updated: February 2, 2026**  
+**📅 Last Updated: February 16, 2026**  
 **👤 Updated By: Takahashi**
 
 ---
 
-## 🎯 Current Phase: PRE-REVIEW (Updating Paper for New JOSS Requirements)
+## 🎯 Current Phase: PRE-REVIEW (Awaiting Final Paper Approval)
 
-**Status**: Responding to JOSS policy change - **Updating paper format before review can proceed**
+**Status**: Paper sections complete with Shimizu's improvements - **Ready for final review and JOSS notification**
 
 ### Timeline:
 - ✅ **Nov 16, 2025**: Paper submitted to JOSS
@@ -19,13 +19,16 @@
 - 🔔 **Jan 20, 2026**: JOSS announced policy change requiring new paper sections
 - 🔄 **Jan 19-29, 2026**: Team working on adding required sections (Issue #8)
 - ✅ **Feb 2, 2026**: Removed references to private companion repository from paper
-- ⏳ **Currently**: Paper updates complete, preparing to notify JOSS
+- ✅ **Feb 16, 2026**: Shimizu updated Research Impact Statement with new evidence (Jiang 2023, Furukawa 2025)
+- ✅ **Feb 16, 2026**: Verified new citations - both papers legitimately used MOLASS
+- ⏳ **Currently**: Paper updates complete (192 words Research Impact Statement), awaiting Shimizu's final approval to notify JOSS
 
 ### What This Means:
-- **Editor**: Not yet assigned (waiting for paper update confirmation)
+- **Editor**: Not yet assigned (waiting for paper update notification)
 - **Reviewers**: Not yet assigned
-- **Review Status**: Has NOT started yet (blocked on paper update notification)
-- **Next Step**: Shimizu to review latest changes and notify JOSS that paper has been updated
+- **Review Status**: Has NOT started yet (JOSS waiting for authors to confirm updates complete)
+- **Last JOSS Activity**: January 20, 2026 (policy change announcement)
+- **Next Step**: Shimizu to give final approval and notify JOSS that paper has been updated per new requirements
 
 ---
 
@@ -44,16 +47,20 @@
 - ✅ Responded to EiC's pre-review questions
 - ✅ Fixed DOI formatting issues
 - ✅ Suggested relevant reviewer candidates
-- ⏳ Reviewing paper updates from Takahashi (Issue #8 - updated Feb 2, 2026)
-- ⏳ Will notify JOSS when updates are approved
+- ✅ Updated Research Impact Statement with stronger evidence (Feb 16, 2026)
+- ✅ Added citations: Jiang 2023, Furukawa 2025 (both verified as legitimate MOLASS usage)
+- ⏳ Final approval pending, then will notify JOSS
 
 ### **Takahashi** (Technical Support):
 - ✅ Repository setup complete
 - ✅ Drafted Software Design section (Issue #8)
-- ✅ Drafted Research Impact Statement (Issue #8)
+- ✅ Drafted initial Research Impact Statement (Issue #8)
 - ✅ Added AI Usage Disclosure (Issue #8)
 - ✅ Removed private repository references from paper (Feb 2, 2026)
-- ⏳ Awaiting Shimizu's final approval to respond to JOSS
+- ✅ Synced molass-library subtree with latest joss-paper updates (Feb 16, 2026)
+- ✅ Reviewed Shimizu's improved Research Impact Statement (192 words, well within limit)
+- ✅ Verified new paper citations (Jiang 2023, Furukawa 2025) - both used MOLASS for SEC-SAXS analysis
+- ⏳ Identified PR #7 (joss-paper→master) - awaiting clarification from Shimizu
 
 ---
 
@@ -61,10 +68,22 @@
 
[…]
