+++ COPILOT-INIT.md
@@ -1,0 +1,318 @@
+<!-- AI Context Standard: v0.1 -->
+# AI Assistant Initialization Guide — molass-library
+
+**Purpose**: Initialize AI context for coding work in this repository  
+**Created**: February 19, 2026  
+**Magic phrase**: **"Please read COPILOT-INIT.md to initialize"**
+
+> **Note**: This file follows the AI Context Standard (v0.1). It provides **technical context** (architecture, call chains, conventions).  
+> For **behavioral rules** (response style, user types, source priorities), see `Copilot/copilot-guidelines.md` (loaded automatically as Priority ⭐ 2 below).  
+> For current research status and task history, see `modeling-vs-model_free/PROJECT_STATUS.md`.
+
+---
+
+## 📚 Documents to Read
+
+| Priority | File | Purpose |
+|----------|------|---------|
+| ⭐ 1 | `README.md` | Project overview & entry points |
+| ⭐ 2 | `Copilot/copilot-guidelines.md` | Project policies and user-type rules |
+| 3 | `pyproject.toml` | Package version, dependencies, test config |
+| 4 | `tests/tutorial/01-quick_start.py` | Canonical usage example (start here for API) |
+| 5 | `tests/tutorial/05-lrf.py` | Low-rank factorization (LRF) tutorial |
+| 6 | `tests/tutorial/11-rigorous_optimization.py` | Rigorous optimization tutorial |
+
+---
+
+## 🏗️ Architecture Overview
+
+### What Molass Does
+
+Molass decomposes SEC-SAXS data: the measured 2D matrix $M$ (q-values × elution frames) is factored into scattering profiles $P$ and elution curves $C$ via $M \approx PC$.
+
+The data flows through three main stages:
+
+```
+Raw data folder / molass_data sample
+        ↓
+   SecSaxsData (DataObjects/)          ← Load, trim, correct
+        ↓
+   quick_decomposition()               ← Low-rank factorization (LRF)
+        ↓
+   Decomposition object (LowRank/)     ← Holds xr_ccurves, uv_ccurves, components
+        ↓
+   make_rigorous_decomposition()       ← (Optional) Physics-constrained refinement
+        ↓
+   Reports / DENSS                     ← Output
+```
+
+### Key Packages
+
+| Package | Role |
+|---------|------|
+| `DataObjects/` | Core data containers: `SecSaxsData`, `XrData`, `UvData`, `SsMatrixData`, `Curve` |
+| `LowRank/` | Matrix factorization engine: `Decomposition`, `CurveDecomposer`, `CoupledAdjuster`, `QuickImplement` |
+| `Rigorous/` | Physics-constrained optimization (EGH/SDM/EDM elution models + Rg-consistency); bridges to `molass-legacy` |
+| `SEC/Models/` | Column elution models: `EDM.py`, `SDM.py`, `Simple.py` (+ Gaussian, lognormal pore distributions) |
+| `Guinier/` | Rg estimation: `RgEstimator`, `RgCurve`, `RgCurveUtils` |
+| `Peaks/` | Peak recognition: `Recognizer`, `PeakSimilarity` |
+| `Decompose/` | Helper utilities for decomposition variants (proportional, XR-only) |
+| `Global/` | Global optimization options (`Options.py`) |
+| `Baseline/` | Baseline correction |
+| `Trimming/` | Data trimming |
+| `SAXS/` | DENSS integration (`DenssTools`), MRC viewer (`MrcViewer`) |
+| `Shapes/` | Geometric shapes for scattering models (`Ellipsoid`, `Sphere`) |
+| `Testing/` | Test infrastructure (`control_matplotlib_plot` decorator, etc.) |
+| `Legacy/` | Bridge to `molass-legacy` package |
+
+### Important Relationships
+
[…]
