### 2026-02-20 718735c
Add bounded LRF as default for rank 2 component
