### 2026-03-03 8efb320
Update CURRENT-STATUS.md: awaiting arfon's assessment of updated paper
