### 2026-03-06 ca7240e
Apply proposed 5 improvements from molass-researcher work
