+++ molass/LowRank/Decomposition.py
@@ -141,13 +141,101 @@ class Decomposition:
 
         Returns
         -------
-        list of float
-            The list of Rg values for each XR component.
+        rgs : list of float, length n_components
+            Radius of gyration in **Ångströms (Å)** for each XR component,
+            in the same order as ``get_xr_components()``.
+            If Guinier fitting fails for a component, ``float('nan')`` is
+            returned for that position (never ``None``), so the result is
+            always safe to use in numeric / numpy operations.
+
+            Guard pattern::
+
+                import math
+                for i, rg in enumerate(decomp.get_rgs()):
+                    if math.isnan(rg):
+                        print(f"Component {i+1}: Guinier fit failed")
+                    else:
+                        print(f"Component {i+1}: Rg = {rg:.2f} Å")
         """
-        return [sv.Rg for sv in self.get_guinier_objects()]
+        return [sv.Rg if sv.Rg is not None else float('nan')
+                for sv in self.get_guinier_objects()]
+
+    def get_P_at(self, q_target, normalize=False):
+        """Return the XR scattering matrix P interpolated onto *q_target*.
+
+        Parameters
+        ----------
+        q_target : array-like, shape (m,)
+            Target q-values in Å⁻¹.
+        normalize : bool, optional
+            If ``True``, each component column is divided by its maximum so
+            that all columns peak at 1.  Default ``False``.
+
+        Returns
+        -------
+        P_interp : np.ndarray, shape (m, n_components)
+            Scattering matrix P evaluated at *q_target*.
+        """
+        from molass.LowRank.AlignDecompositions import get_P_at
+        return get_P_at(self, q_target, normalize=normalize)
+
+    def component_quality_scores(self):
+        """
+        Compute a per-component reliability score in [0, 1].
+
+        Scores blend Rg distinctiveness (70 %) and area proportion (30 %).
+        A score of 0.0 means Guinier fitting failed or Rg is indistinguishable
+        from another component's.  A score near 1.0 means the component is
+        well-separated in Rg and carries a non-trivial fraction of the signal.
+
+        Returns
+        -------
+        scores : list of float
+            Reliability score for each component, in the same order as
+            ``get_rgs()`` and ``get_proportions()``.
+
+        See Also
+        --------
+        is_component_reliable : threshold-based boolean version.
+
+        Examples
+        --------
+        ::
+
+            scores = decomp.component_quality_scores()
+            for i, s in enumerate(scores):
+                print(f"Component {i+1}: reliability = {s:.2f}")
+        """
+        from molass.LowRank.ComponentReliability import component_quality_scores
+        return component_quality_scores(self)
+
+    def is_component_reliable(self, index, threshold=0.5):
+        """
+        Return ``True`` if component *index* has a quality score ≥ *threshold*.
+
+        Parameters
+        ----------
+        index : int
+            Zero-based component index.
+        threshold : float, optional
+            Minimum score considered reliable.  Default 0.5.
+
+        Returns
+        -------
+        bool
+
+        Examples
+        --------
+        ::
+
+            if not decomp.is_component_reliable(1):
+                print("Component 2 may be a noise artifact.")
+        """
+        from molass.LowRank.ComponentReliability import is_component_reliable
+        return is_component_reliable(self, index, threshold=threshold)
 
     def plot_components(self, **kwargs):
-        """decomposition.plot_components(title=None, **kwargs)
+        """decomposition.plot_components(title=None, fig=None, axes=None, **kwargs)
 
         Plot the components.
 
@@ -155,6 +243,26 @@ class Decomposition:
         ----------
         title : str, optional
[…]
