+++ Copilot/API_IMPROVEMENTS.md
@@ -1,0 +1,135 @@
+# API Improvement Proposals
+
+**Source**: Friction points identified during AI-assisted research in `molass-researcher` (Experiment 01, March 6, 2026)  
+**Context**: Running MOLASS on 6 SEC-SAXS datasets in a Jupyter notebook workflow, comparing original vs pre-averaged data. The issues below caused code failures or workarounds during that work.
+
+---
+
+## 1. `get_rgs()` should not return `None` silently
+
+**Current behaviour**: When Guinier fitting fails for a component (e.g. because the component is a noise artifact from forced `num_components`), `get_rgs()` returns `None` in the list for that position.
+
+**Problem**: This silently breaks all downstream numeric code. Example crash:
+```python
+for rg in decomp.get_rgs():
+    print(f"Rg = {rg:.2f} Å")   # TypeError if rg is None
+```
+The failure mode is a `TypeError` at the format string, with no contextual message about which component failed or why.
+
+**Recommendations** (in order of preference):
+- Return `float('nan')` instead of `None` — composable with numpy math, and `math.isnan(rg)` is an explicit guard
+- OR raise a named exception with context: `GuinierFailedError(component=2, reason="not enough valid data points")`
+- OR add a companion `get_rg_flags()` → `["ok", "failed", "failed"]` so callers can inspect quality without try/except
+
+**Workaround used**:
+```python
+rg_str = f"{rg:.2f}" if rg is not None else "N/A (Guinier failed)"
+```
+
+---
+
+## 2. `plot_components()` is not axis-injectable
+
+**Current behaviour**: `plot_components()` always creates its own `plt.figure()` internally. Calling `plt.sca(ax)` before it has no effect.
+
+**Problem**: Cannot be embedded in a subplot grid. Any attempt to build a multi-panel figure with `plt.subplots()` and then call `plot_components()` into a specific axis silently produces a separate floating figure (and a blank subplot).
+
+**Recommendation**: Add optional `fig=None` / `axes=None` parameters following the standard matplotlib convention:
+```python
+def plot_components(self, title=None, fig=None, axes=None):
+    if fig is None:
+        fig, axes = plt.subplots(...)
+    # draw into provided axes
+```
+
+**Workaround used**: Called `plot_components()` standalone for each dataset and saved each figure individually, rather than composing a comparison grid.
+
+---
+
+## 3. No q-grid alignment utility
+
+**Current behaviour**: `get_xr_matrices()` returns `P` with shape `(n_q, n_components)`, where `n_q` differs between datasets because trimming leaves different numbers of q-points. There is no built-in way to compare P(q) profiles across datasets.
+
+**Problem**: `scipy.stats.pearsonr(P_a, P_b)` raises a `ValueError` when the arrays have different lengths (e.g. 966 vs 963 points). This is a routine need whenever comparing two decompositions.
+
+**Recommendation**: Add a utility to project a scattering profile onto a target q-grid:
+```python
+decomp.P_at(q_target)   # interpolated P(q) array at given q values
+# or
+molass.align_decompositions(decomp_a, decomp_b)  # returns both P arrays on common grid
+```
+
+**Workaround used**:
+```python
+q_common = np.linspace(max(q_a[0], q_b[0]), min(q_a[-1], q_b[-1]), 500)
+P_a_interp = np.interp(q_common, q_a, P_a[:, 0])
+P_b_interp = np.interp(q_common, q_b, P_b[:, 0])
+r, _ = pearsonr(P_a_interp, P_b_interp)
+```
+
+---
+
+## 4. No component reliability indicator for forced decompositions
+
+**Current behaviour**: `quick_decomposition(num_components=N)` forces exactly N components regardless of data quality. There is no returned quality metric to distinguish real from spurious components.
+
[…]
