+++ README.md
@@ -35,10 +35,11 @@ Especially for testing, see the first two sections in
 
 ## Copilot Usage
 
-Before starting a Copilot chat session with this repository, please use the following magic phrase to ensure Copilot follows project rules:
-For details on Copilot rules and usage, see [`Copilot/copilot-guidelines.md`](https://github.com/biosaxs-dev/molass-library/blob/master/Copilot/copilot-guidelines.md).
+Before starting a Copilot chat session with this repository, use the following magic phrase to load conventions and architecture:
 
-> “Please follow the Copilot guidelines in this project for all advice and responses.”
+> "Please read COPILOT-INIT.md to initialize"
+
+For behavioral rules and user-type guidance, see [`Copilot/copilot-guidelines.md`](https://github.com/biosaxs-dev/molass-library/blob/master/Copilot/copilot-guidelines.md).
 
 ## Optional Features
 
