+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "molass"
-version = "0.8.2"
+version = "0.8.3"
 
 dependencies = [
   "numpy",

[…]
