### 2026-03-06 834021e
Squashed 'molass-library/' changes from 51e7ad5..df8d773
df8d773 Merge branch 'master' into joss-paper
ca7240e Apply proposed 5 improvements from molass-researcher work
718735c Add bounded LRF as default for rank 2 component
8d9a528 Improve documentation
bf3fba6 Add a documented version of moment derivation
fd9440b Update for reorganization of plot results
a22e098 Add monitor to RunInfo
6911c85 Update for rigorous SDM optimization
ce56536 Temporary fix for DENSS 1.8.7
39479fa Fix uv baseline issue in xr_only mode
1dd2960 Change xr_only set timing
e7868bd Improve Rg estimation
ee1adea Update fo
[…]
