+++ CURRENT-STATUS.md
@@ -1,7 +1,8 @@
 # CURRENT STATUS
 
-**📅 Last Updated: February 20, 2026**  
-**👤 Updated By: Copilot (auto-updated at session init)**
+**📅 Last Updated: March 9, 2026**  
+**👤 Updated By: Copilot (auto-updated at session init)**  
+**✅ Status verified: No change since Feb 20, 2026 — still awaiting arfon's assessment and editor assignment**
 
 ---
 
