### 2026-03-09 a14d923
Add ComponentCurve.y, trimmed_copy(nsigmas=), auto-mapping in TrimmingUtils; bump to v0.8.4
Closes #13

- ComponentCurve: add @property y delegating to get_y(self.x)
- SecSaxsData.trimmed_copy: accept nsigmas= and forward to make_trimming()
- TrimmingUtils.make_trimming_impl: auto-compute mapping when jranges is given
