### 2026-03-09 211f611
Add uv_pickat support: SSD(uv_pickat=290) for non-280nm samples; bump to v0.8.5
Closes #14

- UvData: add self.pickat instance attr (default 280); override copy() to
  preserve it; get_icurve() uses self.pickat when called without explicit pickat
- SecSaxsData.__init__: add uv_pickat=None parameter; sets self.uv.pickat
  after construction, propagating automatically through all pipeline calls
  (trimming, mapping, decomposition)
