+++ molass/DataObjects/UvData.py
@@ -36,6 +36,12 @@ class UvData(SsMatrixData):
         """
         super().__init__(iv, jv, M, E, **kwargs)
         self.wv = iv
+        self.pickat = PICKAT
+
+    def copy(self, slices=None):
+        result = super().copy(slices=slices)
+        result.pickat = self.pickat
+        return result
 
     def get_ipickvalues(self):
         """Get the default pickvalues for i-curves.
@@ -46,7 +52,7 @@ class UvData(SsMatrixData):
         """
         return PICKVALUES
 
-    def get_icurve(self, pickat=PICKAT):
+    def get_icurve(self, pickat=None):
         """uv_data.get_icurve(pickat=280)
         
         Returns an i-curve from the UV matrix data.
@@ -54,16 +60,20 @@ class UvData(SsMatrixData):
         Parameters
         ----------
         pickat : float, optional
-            Specifies the value in ssd.qv where to pick an i-curve.
+            Specifies the wavelength (nm) at which to pick an i-curve.
             The i-curve will be made from self.M[i,:] where
             the picking index i will be determined to satisfy
-                self.wv[i-1] <= pickat < self.vec.wv[i]
+                self.wv[i-1] <= pickat < self.wv[i]
             according to bisect_right.
+            If None, uses self.pickat (default 280 nm, or the value set
+            via SSD(uv_pickat=...)).
 
         Examples
         --------
         >>> curve = uv_data.get_icurve()
         """
+        if pickat is None:
+            pickat = self.pickat
         return super().get_icurve(pickat)
 
     def get_flowchange_points(self, pickvalues=PICKVALUES, return_also_curves=False):
