### 2026-03-10 3b43397
Issue #20: add get_rg_curve() and fix scipy/Python 3.13 Guinier performance
- Decomposition.get_rg_curve() delegates to xr.compute_rgcurve()
- RgCurveUtils.py: cache inspect.getfullargspec to fix 1000x slowdown on Py3.13
- test_030_get_rgs.py: test_040 updated (RgCurve.y stores None for failed fits)
- API_IMPROVEMENTS.md: mark #20 Done
