### 2026-03-10 5e8c2ff
Issue #21: use xrdata.jv[j] for original frame numbers in RgCurve.x
Both compute_rgcurve_info() and compute_rgcurve_info_atsas() now store
int(jv[j]) instead of j, so rgcurve.x reflects the true acquisition
frame numbers after trimmed_copy()/corrected_copy().
