### 2026-03-10 9737f7e
DataObjects and PlotUtils improvements; add study scripts
- SsMatrixData: add .data alias for .M
- UvData: add .wavelengths and .frames aliases (issue #15)
- MatrixPlot: various plot utility updates
- test_010_SSD: expanded tests for new aliases
- copilot-guidelines: updated preferences
- study/scipy_inspect_issue.ipynb: scipy inspect caching investigation
- read_nb_outputs.py, run_rg_test.py: helper scripts
