### 2026-03-11 0f4f5e6
fix: adaptive p_final in compute_lpm_baseline (issue #4)
Replicate legacy correct_all_q_planes behaviour:
- SsMatrixData.get_baseline2d() pre-computes size_sigma from the
  elution curve peak width via ElutionBaseCurve.compute_size_sigma()
  and passes it in method_kwargs
- LpmBaseline.compute_lpm_baseline() calls new helper
  _compute_adaptive_p_final(x, y, size_sigma) when size_sigma is
  provided; computes per-row noisiness via LSQUnivariateSpline then
  calls base_percentile_offset(noisiness, size_sigma) to get
  adaptive p_final (~28.6%) instead of hardcoded PERCENTILE_FINAL=10

For SAMPL
[…]
