### 2026-03-11 71f1228
fix: suppress stdout from ElutionBaseCurve in get_baseline2d() (issue #23)
Wrap ElutionBaseCurve instantiation and compute_size_sigma() call with
contextlib.redirect_stdout(io.StringIO()) to swallow legacy diagnostic
prints ('recognize_num_peaks', 'peak_width= ...').

Also adds test_08_get_baseline2d_no_stdout using pytest capsys.

Fixes #23
