+++ molass/DataObjects/SsMatrixData.py
@@ -148,9 +148,11 @@ class SsMatrixData:
         debug = kwargs.get('debug', False)
         counter = [0, 0, 0] if debug else None
         if self.baseline_method in ['linear', 'uvdiff', 'integral']:
+            import io, contextlib
             from molass_legacy.SerialAnalyzer.ElutionBaseCurve import ElutionBaseCurve as _EBC
-            _ecurve = _EBC(self.M.sum(axis=0))
-            _size_sigma = _ecurve.compute_size_sigma()
+            with contextlib.redirect_stdout(io.StringIO()):
+                _ecurve = _EBC(self.M.sum(axis=0))
+                _size_sigma = _ecurve.compute_size_sigma()
             default_kwargs = dict(jv=self.jv, ssmatrix=self, counter=counter, size_sigma=_size_sigma)
             if self.baseline_method == 'uvdiff':
                 from molass.Baseline.UvdiffBaseline import get_uvdiff_baseline_info
