### 2026-03-11 047403d
feat: add 'bufmask' baseline method — buffer-mask linear polyfit (issue #24)
New method uses the summed elution (high-SNR) to classify buffer frames,
then anchors a linear fit on those frames via np.polyfit. Avoids the
iterative percentile descent of LPM entirely.

Files:
- molass/Baseline/BufmaskBaseline.py: new compute_bufmask_baseline()
- molass/Baseline/Baseline2D.py: _bufmask_individual_axes_impl + assertion
- molass/DataObjects/SsMatrixData.py: compute buffer_mask from elution_norm
  when baseline_method == 'bufmask', pass via default_kwargs
- tests: test_09_bufmask_baseline verifies sha
[…]
