### 2026-03-11 c80f70e
docs: add concept explanation and Copilot attribution to BufmaskBaseline
