+++ molass/Baseline/BufmaskBaseline.py
@@ -3,10 +3,40 @@
 
     Buffer-mask linear polyfit baseline.
 
-    Uses the summed elution curve (high SNR) to classify buffer frames,
-    then fits a straight line through those frames only via np.polyfit.
-    This is more direct than the iterative LPM percentile descent and
-    yields a lower (more correct) positive_ratio in buffer regions.
+    Concept and origin
+    ------------------
+    This method was proposed by GitHub Copilot (Claude Sonnet 4.5) during an
+    interactive session on 2026-03-11, as a direct improvement over the
+    Legacy Percentile Method (LPM).
+
+    The key insight (quoted from the session):
+
+        "LPM's core premise is correct: buffer frames have lower intensity
+        than protein frames.  But LPM classifies buffer frames implicitly,
+        per q-row, under low per-row SNR.  The fix is to lift the
+        classification one level up: M.sum(axis=0) collapses all q-rows
+        into one curve, gaining √N in SNR.  In that summed elution the
+        buffer region is unambiguously flat and low.  Classify buffer frames
+        once from this high-SNR oracle, then apply that single mask to
+        every q-row for the polyfit."
+
+    In short: the classification question ("is frame j buffer or protein?")
+    is a column-level (global) question; LPM answered it row-by-row (local).
+    Bufmask answers it once, using the full-matrix aggregate, and reuses
+    that answer across all q-rows.
+
+    Performance (SAMPLE1, 2026-03-11)
+    ----------------------------------
+    Method                          mean positive_ratio
+    ----------------------------------
+    Old library (p_final fixed 10%) 0.899
+    Adaptive p_final / legacy sim   0.716
+    Bufmask (threshold = 0.10)      0.578   ← this module
+    Ideal (perfect baseline)        ~0.5
+
+    Reference
+    ---------
+    Implemented in molass-library issue #24.
 """
 import numpy as np
 
