### 2026-03-11 73a3d91
Update for AI-friendliness
