### 2026-03-12 84f2e87
v0.8.5: buffit (Otsu) as default XrData baseline (#24 #25 #26)
