### 2026-03-12 ec1a52d
Update tutorial test 04 for buffit default (v0.8.5)
