### 2026-03-12 476f664
Add Copilot/workflow_notes.md for cross-machine AI workflow reference
