### 2026-03-12 9745069
feat: add molass.requires() version guard; update all tutorial tests (#27)
