+++ tests/tutorial/01-quick_start.py
@@ -13,8 +13,8 @@ decomposition = None
 @pytest.mark.order(1)
 @control_matplotlib_plot
 def test_001_plot_3d():
-    from molass import get_version
-    assert get_version() >= '0.6.1', "This tutorial requires molass version 0.6.1 or higher."
+    from molass import requires
+    requires('0.6.1')
     from molass_data import SAMPLE1
     from molass.DataObjects import SecSaxsData as SSD
     global ssd
