+++ tests/tutorial/04-data_correction.py
@@ -13,8 +13,8 @@ corrected_ssd = None
 @pytest.mark.order(1)
 @control_matplotlib_plot
 def test_001_plot_compact():
-    from molass import get_version
-    assert get_version() >= '0.8.5', "This tutorial requires molass version 0.8.5 or higher."
+    from molass import requires
+    requires('0.8.5')
     from molass_data import SAMPLE2
     from molass.DataObjects import SecSaxsData as SSD
     global ssd, trimmed_ssd
