### 2026-03-12 d34e04c
Fix _compute_adaptive_p_final: correct knot range and relative noisiness calibration
