### 2026-03-12 511795a
Update CURRENT-STATUS.md date (session init auto-update)
