+++ CURRENT-STATUS.md
@@ -1,6 +1,6 @@
 # CURRENT STATUS
 
-**📅 Last Updated: March 12, 2026**  
+**📅 Last Updated: March 13, 2026**  
 **👤 Updated By: Copilot (auto-updated at session init)**  
 **✅ Status verified: No change since Feb 17, 2026 — still awaiting arfon's assessment and editor assignment**
 
