### 2026-03-16 cce87da
Wire xr_peakpositions API + bug fixes + AI-friendliness (#31-#35)
#31: Wire xr_peakpositions into quick_decomposition() with 3-way routing
     (default/proportional/positioned). Fix 6 bugs: 0-based ElutionCurve
     in LegacyBridgeUtils, LegacyRgCurve, SdProxy; np.searchsorted in
     Proportional; positive sigma + min_sigma in PositionedDecomposer;
     temp_params before assignment + variable shadowing in Partner.
#32: ComponentCurve docstring Rg breadcrumb
#33: Decomposition.xr_components/uv_components property aliases
#34: quick_decomposition() docstring for xr_peakpositions et al.
#35: X
[…]
