+++ molass/LowRank/ComponentCurve.py
@@ -8,12 +8,22 @@ class ComponentCurve:
     """
     A class to represent a component curve.
 
+    This object stores the *elution-curve* EGH parameters (H, tR, sigma, tau)
+    and can evaluate the model curve, but it does **not** hold scattering
+    profiles and therefore cannot compute Rg directly.
+
+    To obtain Rg values, use one of:
+
+    - ``decomp.get_rgs()`` — list of Rg for all components.
+    - ``decomp.get_xr_components()`` — returns ``XrComponent`` objects,
+      each of which has ``get_guinier_object()`` for full Guinier-fit results.
+
     Attributes
     ----------
     x : array-like
         The x-values of the component curve.
     params : array-like
-        The parameters of the component curve.
+        The EGH parameters ``[H, tR, sigma, tau]`` of the component curve.
     moment : Moment or None
         The moment of the component curve. Computed on demand. If None, it has not been computed yet.
     """
