### 2026-03-17 32966ae
Issues #36, #37: detect_peaks(return_properties) + plot_peaks()
- detect_peaks(return_properties=True) returns (peaks, props) tuple
  with prominences and peak_heights arrays (backward-compatible default)
- plot_peaks(ax=None): annotated elution curve with detected peaks,
  prominence labels, 15% ylim headroom, explanation note
- Tests: test_020_detect_peaks_return_properties, test_030_plot_peaks
