+++ molass/DataObjects/XrData.py
@@ -203,7 +203,8 @@ class XrData(SsMatrixData):
             from molass.Guinier.RgCurve import construct_rgcurve_from_list
             return construct_rgcurve_from_list(rginfo, result_type='atsas')
 
-    def detect_peaks(self, prominence=0.005, distance=20, window_length=31, polyorder=3):
+    def detect_peaks(self, prominence=0.005, distance=20, window_length=31,
+                     polyorder=3, return_properties=False):
         """xr.detect_peaks()
 
         Detect peaks in the total XR elution curve using Savitzky-Golay
@@ -223,16 +224,26 @@ class XrData(SsMatrixData):
             Savitzky-Golay filter window length (must be odd). Default 31.
         polyorder : int, optional
             Savitzky-Golay filter polynomial order. Default 3.
+        return_properties : bool, optional
+            If True, return a tuple ``(peaks, properties)`` where
+            *properties* is a dict containing ``'prominences'`` and
+            ``'peak_heights'`` arrays.  Default False (backward-compatible).
 
         Returns
         -------
         list of int
             Frame numbers of detected peaks, sorted by frame number.
+            Returned alone when ``return_properties=False``.
+        tuple of (list of int, dict)
+            ``(peaks, properties)`` when ``return_properties=True``.
 
         Examples
         --------
         >>> peaks = ssd.xr.detect_peaks()
         >>> decomp = ssd.quick_decomposition(xr_peakpositions=peaks)
+
+        >>> peaks, props = ssd.xr.detect_peaks(return_properties=True)
+        >>> props['prominences']   # array of peak prominences
         """
         from scipy.signal import savgol_filter, find_peaks
         frames = self.jv
@@ -241,12 +252,88 @@ class XrData(SsMatrixData):
         if wl % 2 == 0:
             wl -= 1
         if wl < polyorder + 2:
-            # not enough data to smooth — return raw peak search
             smooth = total_elution
         else:
             smooth = savgol_filter(total_elution, window_length=wl, polyorder=polyorder)
-        peaks_idx, _ = find_peaks(smooth, prominence=smooth.max() * prominence, distance=distance)
-        return [int(frames[i]) for i in peaks_idx]
+        peaks_idx, props = find_peaks(smooth, prominence=smooth.max() * prominence, distance=distance)
+        peak_frames = [int(frames[i]) for i in peaks_idx]
+        if return_properties:
+            out_props = {
+                'prominences': props['prominences'],
+                'peak_heights': smooth[peaks_idx],
+            }
+            return peak_frames, out_props
+        return peak_frames
+
+    def plot_peaks(self, ax=None, prominence=0.005, distance=20,
+                   window_length=31, polyorder=3):
+        """xr.plot_peaks()
+
+        Visualize detected peaks on the total XR elution curve.
+
+        Parameters
+        ----------
+        ax : matplotlib Axes, optional
+            If provided, plot on this axes.  Otherwise create a new figure.
+        prominence, distance, window_length, polyorder :
+            Forwarded to :meth:`detect_peaks`.
+
+        Returns
+        -------
+        tuple of (Figure, Axes)
+        """
+        import matplotlib.pyplot as plt
+
+        peaks, props = self.detect_peaks(
+            prominence=prominence, distance=distance,
+            window_length=window_length, polyorder=polyorder,
+            return_properties=True,
+        )
+
+        frames = self.jv
+        total_elution = self.M.sum(axis=0)
+        wl = min(window_length, len(total_elution))
+        if wl % 2 == 0:
+            wl -= 1
+        if wl < polyorder + 2:
[…]
