### 2026-03-17 d80cd65
Squashed 'molass-library/' changes from 7002c84..b12c8b5
b12c8b5 Merge branch 'master' into joss-paper
32966ae Issues #36, #37: detect_peaks(return_properties) + plot_peaks()
cce87da Wire xr_peakpositions API + bug fixes + AI-friendliness (#31-#35)
6036170 The default baseline method reverted to 'linear'
d34e04c Fix _compute_adaptive_p_final: correct knot range and relative noisiness calibration
a16ae19 Fix version: 0.8.5 -> 0.8.4
d8bb0e6 Update workflow_notes: #27 closed
9745069 feat: add molass.requires() version guard; update all tutorial tests (#27)
bc12071 Add workflow_notes.md as priorit
[…]
