+++ tests/tutorial/05-lrf.py
@@ -13,8 +13,8 @@ trimmed_ssd = None
 @pytest.mark.order(1)
 @control_matplotlib_plot
 def test_001_plot_compact():
-    from molass import get_version
-    assert get_version() >= '0.6.0', "This tutorial requires molass version 0.6.0 or higher."
+    from molass import requires
+    requires('0.6.0')
     from molass_data import SAMPLE1
     from molass.DataObjects import SecSaxsData as SSD
     global ssd
