### 2026-03-18 fa07520
SNR-weighted positive_ratio and bpo_ideal (default) — issue #38
- get_snr_weights(): per-q-row SNR = mean(I)/sigma_noise
- get_positive_ratio(baseline, weighting='snr'): SNR-weighted by default
- get_bpo_ideal(weighting='snr'): per-row noisiness + SNR aggregation by default
- weighting='uniform' available for backward-compatible flat averaging
- 3 new tests (test_14–16), all 16 tests pass
