### 2026-03-18 24543eb
evaluate_baseline() + get_ideal_positive_ratio() alias -- issues #39, #40
- evaluate_baseline(baseline) returns namedtuple(positive_ratio, ideal, delta)
- get_ideal_positive_ratio() aliases opaque get_bpo_ideal()
- Tests: test_17, test_18 added and passing
