### 2026-03-18 f8e4cc5
feat: add method= kwarg to get_baseline2d() (#36)
Previously get_baseline2d() always used self.baseline_method, silently
ignoring any method= kwarg. Now the method parameter overrides the
instance's baseline_method for that call only, e.g.:
    xr.get_baseline2d(method='linear')
    xr.get_baseline2d(method='buffit')
All three internal branches (linear/uvdiff/integral, buffit) updated
to use local 'method' variable instead of self.baseline_method.
