### 2026-03-19 bf90f40
Issues #41-45: sum-based elution recognition + pickat consistency + RgCurve naming
#41: Sum-based elution recognition
- Added elution_recognition='icurve'|'sum' global option
- SsMatrixData.get_recognition_curve() base: M.sum(axis=0)
- XrData.get_recognition_curve(): checks option, returns icurve or sum
- UvData.get_recognition_curve(): always returns icurve (sum inappropriate)
- plot_compact wired to use recognition curve when sum mode
- get_baseline2d() uses recognition curve for buffer classification
- get_bpo_ideal() uses recognition curve for ideal positive ratio

#42: pickat consistency 
[…]
