+++ _sources/source/molass.Baseline.BuffitBaseline.rst
@@ -1,0 +1,7 @@
+Baseline.BuffitBaseline
+=====================================
+
+.. automodule:: molass.Baseline.BuffitBaseline
+   :members:
+   :undoc-members:
+   :show-inheritance:
