+++ _sources/source/molass.Decompose.rst
@@ -15,3 +15,4 @@ Submodules
    molass.Decompose.Partner
    molass.Decompose.Proportional
    molass.Decompose.VaryUtils
+   molass.Decompose.XrOnlyUtils
