+++ _sources/source/molass.LowRank.AlignDecompositions.rst
@@ -1,0 +1,7 @@
+LowRank.AlignDecompositions
+=========================================
+
+.. automodule:: molass.LowRank.AlignDecompositions
+   :members:
+   :undoc-members:
+   :show-inheritance:
