+++ _sources/BOUNDED_LRF_PLAN.md
@@ -1,0 +1,314 @@
+# Bounded LRF Implementation Plan
+
+## Goal
+
+Replace the current naïve rank-2 factorization (`P = M_ @ pinv(C)`) with
+Bounded LRF — a physics-constrained version that clips B(q) within the
+envelope `|B(q)| ≤ A(q) / (c₁ · (qLR)²)` and redistributes excess to A(q).
+
+Only Bounded LRF will be implemented. Synthesized LRF (logistic splice of two
+decompositions) is not needed — see `explorations/bounded_structure_factorization.ipynb`.
+
+---
+
+## Current Architecture
+
+### Data flow (rank-2 case)
+
+```
+Decomposition.get_xr_matrices()
+  → compute_lowrank_matrices(xr.M, xr_ccurves, xr.E, xr_ranks)
+      → M_ = get_denoised_data(M, svd_rank)
+      → C  = [c, c²]                    # concentration matrix
+      → P  = M_ @ pinv(C)               # naïve factorization → [A(q), B(q)]
+      → Pe = compute_propagated_error()
+      → return (M_, C_, P_, Pe)
+```
+
+### What Decomposition currently stores
+
+```python
+# Set during __init__
+self.ssd              # SecSaxsData → has self.xr.qv, self.xr.M, self.xr.E
+self.xr               # XrData (shortcut)
+self.xr_icurve        # elution i-curve
+self.xr_ccurves       # list of ComponentCurve (one per peak)
+self.xr_ranks         # None → [1,1,...]; or [1,2,...] if SCD detected
+
+# Lazily computed
+self.guinier_objects   # list of RgEstimator (from XrComponent)
+```
+
+### How rank=2 is triggered
+
+```
+decomposition.compute_scds()       → [scd_1, scd_2, ...]
+RankEstimator.scd_to_rank(scd)     → 2 if scd ≥ 5.0
+decomposition.update_xr_ranks([1, 2, ...])
+decomposition.get_xr_matrices()    → rank-2 factorization for that component
+```
+
+---
+
+## New Variables Required by Bounded LRF
+
+The `coerce_bounds` algorithm needs the following inputs:
+
+| Variable | Type | Source | Description |
+|----------|------|--------|-------------|
+| `qv` | ndarray | `self.xr.qv` | q-vector (already available) |
+| `Rg` | float | Guinier of initial P[:,0] | Radius of gyration (per component) |
+| `R` | float | `√(5/3) · Rg` | Sphere-equivalent radius |
+| `K` | float | fitted | Structure factor amplitude |
+| `L` | float | fitted | Bound-tightness parameter |
+| `c1` | float | `C[0, j_peak]` | Monomer concentration at peak |
+| `c2` | float | `C[1, j_peak]` or `c1²` | Squared concentration at peak |
+
+And it produces:
+
+| Output | Type | Description |
+|--------|------|-------------|
+| `P_corrected` | ndarray | Corrected [A(q), B(q)] matrix |
+| `bq_bounds` | tuple of ndarray | (−bound, +bound) envelope |
+| `coerced_bq` | ndarray | Clipped B(q) |
+| `K, L` | float | Fitted parameters (diagnostic) |
+
+---
+
+## Design Question: Where Do These Fit?
+
+Three options were considered. **Option B** was chosen, with the refinement that
+the bounding orchestration lives in `Decomposition.get_xr_matrices()` (not inside
+`compute_lowrank_matrices`), and `guinier_objects` are pre-populated from naïve P.
+
+See Section 2 below for the definitive `get_xr_matrices()` pseudo-code.
+
+### Decision: Option B with shared Rg
+
+Start with a single `self.bounded_lrf_info` dict keyed by component index.
[…]
