+++ _sources/source/molass.Decompose.XrOnlyUtils.rst
@@ -1,0 +1,7 @@
+Decompose.XrOnlyUtils
+===================================
+
+.. automodule:: molass.Decompose.XrOnlyUtils
+   :members:
+   :undoc-members:
+   :show-inheritance:
