### 2026-03-20 #9424 @zhubonan
@nshimizu0721 Thanks for submitting your work to JOSS. I will handle this as the editor. Could you please start by suggesting some potential reviewers? You can leave their contact information in the comment. If you know their GitHub handle, please do not directly “@” them.
