+++ review-discussions/response-drafts/reviewer-suggestions-to-zhubonan.md
@@ -1,0 +1,52 @@
+<!-- ============================================================
+  COPY EVERYTHING BETWEEN THE LINES BELOW INTO THE GITHUB COMMENT BOX
+  (Replace the [link] placeholder first — see notes at bottom)
+  ============================================================ -->
+
+Thank you for taking on the role of handling editor, @zhubonan. I am happy to suggest some potential reviewers.
+
+I previously posted a list of candidates on November 19, 2025 — please see that comment here: [link to Nov 19 comment]
+
+In addition, I have identified the following four candidates who have experience directly relevant to molass (SEC-SAXS chromatography and scattering data analysis):
+
+- **[KedoKudo](https://github.com/KedoKudo)** (Chen Zhang) — Computational Scientist at Oak Ridge National Laboratory; focuses on scientific software for neutron/X-ray scattering data reduction and analysis, and AI-assisted scientific workflows. Reviewed a GPU-accelerated small-angle scattering Python package for JOSS ([review #6024](https://github.com/openjournals/joss-reviews/issues/6024)).
+
+- **[florian-huber](https://github.com/florian-huber)** (Florian Huber) — Professor at HSD Düsseldorf; develops Python tools for mass spectrometry data processing and similarity scoring (matchms, spec2vec), with expertise in scientific software and machine learning applied to analytical chemistry. Reviewed an HPLC chromatography analysis package for JOSS ([review #6270](https://github.com/openjournals/joss-reviews/issues/6270)).
+
+- **[Adafede](https://github.com/Adafede)** (Adriano Rutz) — Pharmaceutical Scientist at ETH Zürich; specializes in computational metabolomics and LC-MS/MS-based characterization of complex biological matrices. Reviewed a chromatographic peak detection and quantification package for JOSS ([review #7313](https://github.com/openjournals/joss-reviews/issues/7313)).
+
+- **[lazear](https://github.com/lazear)** (Michael Lazear) — Developer of Sage, a high-performance proteomics search and quantification tool; strong expertise in chromatographic data processing and scientific software engineering. Reviewed the same chromatographic peak analysis package for JOSS ([review #7313](https://github.com/openjournals/joss-reviews/issues/7313)).
+
+I am continuing to search for additional candidates and will post further suggestions if I identify strong matches.
+
+<!-- ============================================================
+  END OF COPY REGION
+  ============================================================ -->
+
+---
+
+## File Metadata
+
+**Posted by**: nshimizu0721 (Shimizu)  
+**Target**: JOSS pre-review issue #9424  
+**Date drafted**: 2026-03-23  
+**Status**: DRAFT — awaiting Shimizu's review before posting
+
+---
+
+## Notes Before Posting
+
+- Replace `[link to Nov 19 comment]` with the actual anchor URL: open issue #9424 in a browser, find the Nov 19 comment, right-click its timestamp, and select "Copy link address"
+- Do **not** add `@` before the GitHub handles in the body — zhubonan's instructions say not to directly @-mention reviewer candidates
+- The `@zhubonan` in the opening line is intentional (addressing the editor, not suggesting them as reviewer)
+
+---
+
+## Candidate Background Summary
+
+| GitHub Handle | Real Name | Affiliation | Domain | JOSS Paper Reviewed |
+|---|---|---|---|---|
+| KedoKudo | Chen Zhang | Oak Ridge National Laboratory | Neutron/X-ray scattering, AI workflows | GPU-accelerated small-angle scattering (#6024) |
+| florian-huber | Florian Huber | HSD Düsseldorf | Chromatography, ML, mass spectrometry | hplc-py HPLC analysis (#6270) |
[…]
