### 2026-03-24 109bc60
Add consolidated reviewer suggestions (6 candidates) and 7-paper research notes
