### 2026-03-24 #9424 @nshimizu0721
@zhubonan Following further research into JOSS review history, I would like to consolidate my reviewer suggestions into a single list for convenience. The candidates below were selected primarily for their expertise in chromatographic separation and scattering data analysis, but a few were also included with software API design quality in mind, as molass places particular emphasis on a clean, accessible Python interface for scientific users. These are the candidates I consider most suitable, including those I mentioned previously:

[KedoKudo](https://github.com/KedoKudo) (Chen Zhang) — Computational Scientist at Oak Ridge National Laboratory; focuses on scientific software for neutron/X-ray scattering data reduction and analysis, and AI-assisted scientific workflows. Reviewed a GPU-accelerated small-angle scattering Python package for JOSS (https://github.com/openjournals/joss-reviews/issues/6024).

[florian-huber](https://github.com/florian-huber) (Florian Huber) — Professor at HSD Düsseldorf; develops Python tools for mass spectrometry data processing and similarity scoring (matchms, spec2vec), with expertise in scientific software and machine learning applied to analytical chemistry. Reviewed an HPLC chromatography analysis package for JOSS (https://github.com/openjournals/joss-reviews/issues/6270).

[Adafede](https://github.com/Adafede) (Adriano Rutz) — Pharmaceutical Scientist at ETH Zürich; specializes in computational metabolomics and LC-MS/MS-based characterization of complex biological matrices. Reviewed a chromatographic peak detection and quantification package for JOSS (https://github.com/openjournals/joss-reviews/issues/7313).

[lazear](https://github.com/lazear) (Michael Lazear) — Developer of Sage, a high-performance proteomics search and quantification tool; strong expertise in chromatographic data processing and scientific software engineering. Reviewed the same chromatographic peak analysis package for JOSS (https://github.com/openjournals/joss-reviews/issues/7313).

[elena-pascal](https://github.com/elena-pascal) — Physicist working in scientific Python computing (UK); has hands-on experience with small-angle scattering data and diffraction analysis, including having reviewed a GPU-accelerated small-angle scattering package for JOSS (https://github.com/openjournals/joss-reviews/issues/6024). Active contributor to scikit-image. Combines domain knowledge in X-ray scattering with practical Python library development experience.

[bdice](https://github.com/bdice) (Bradley Dice) — Scientific software developer at NVIDIA RAPIDS; PhD in Physics and Scientific Computing (Glotzerlab, University of Michigan). Reviewed a high-performance chromatography simulation package for JOSS (https://github.com/openjournals/joss-reviews/issues/7881, BCM track). Strong expertise in scientific Python API design and reproducible software practices.
