### 2026-03-24 7e0ed7f
v0.8.7: endpoint_fraction, E optional, corrected_copy kwargs, (M,iv,jv) arg reorder; update PROJECT_STATUS
