+++ molass/DataObjects/UvData.py
@@ -27,23 +27,23 @@ class UvData(SsMatrixData):
     rows index wavelength, columns index frame number.
 
     """
-    def __init__(self, iv, jv, M, E, **kwargs):
+    def __init__(self, M, iv, jv, E=None, **kwargs):
         """Initialize the UvData object.
         
         Parameters
         ----------
-        iv : array-like
-            The wavelength values corresponding to the spectral axis.
-        jv : array-like
-            The values corresponding to the temporal axis.
-        M : 2D array-like
+        M : 2D array-like, shape (len(iv), len(jv))
             The 2D matrix of intensity values.
-        E : 2D array-like or None
-            The 2D matrix of error values. It can be None if errors are not available.
+        iv : array-like
+            The wavelength values corresponding to the spectral (row) axis.
+        jv : array-like or None
+            The values corresponding to the temporal (column) axis.
+        E : 2D array-like or None, optional
+            The 2D matrix of error values. Default None.
         kwargs : dict, optional
             Additional keyword arguments to pass to the SsMatrixData constructor.
         """
-        super().__init__(iv, jv, M, E, **kwargs)
+        super().__init__(M, iv, jv, E, **kwargs)
         self.wv = iv
         self.pickat = PICKAT
 
