+++ molass/DataObjects/SecSaxsData.py
@@ -139,14 +139,14 @@ class SecSaxsData:
                 xr_data = None
             else:
                 from molass.DataObjects.XrData import XrData
-                xr_data = XrData(qv, None, xrM, xrE)
+                xr_data = XrData(xrM, qv, None, xrE)
             self.xr_data = xr_data
 
             if uvM is None:
                 uv_data = None
             else:
                 from molass.DataObjects.UvData import UvData
-                uv_data = UvData(wv, None, uvM, uvE)
+                uv_data = UvData(uvM, wv, None, uvE)
     
         self.xr = xr_data
         self.uv = uv_data
@@ -492,7 +492,7 @@ class SecSaxsData:
             ret_method = (xr_method, uv_method)
         return ret_method
 
-    def corrected_copy(self, debug=False):
+    def corrected_copy(self, debug=False, **baseline_kwargs):
         """ssd.corrected_copy()
         
         Returns a deep copy of this object which has been corrected
@@ -502,20 +502,30 @@ class SecSaxsData:
         ----------
         debug : bool, optional
             If True, enables debug mode for more verbose output.
+        **baseline_kwargs :
+            Additional keyword arguments forwarded to :meth:`get_baseline2d`
+            for both XR and UV.  For example, pass ``endpoint_fraction=0.15``
+            to use the endpoint-anchored baseline for datasets with real
+            negative peaks.
 
         Returns
         -------
         SecSaxsData
             A deep copy of the SSD object with the baseline correction applied.
+
+        Examples
+        --------
+        >>> corrected = ssd.corrected_copy()                          # standard LPM
+        >>> corrected = ssd.corrected_copy(endpoint_fraction=0.15)   # for negative-peak datasets
         """
         start_time = time()
         ssd_copy = self.copy(trimmed=self.trimmed, trimming=self.trimming, datafiles=self.datafiles)
 
-        baseline = ssd_copy.xr.get_baseline2d(debug=debug)
+        baseline = ssd_copy.xr.get_baseline2d(debug=debug, **baseline_kwargs)
         ssd_copy.xr.M -= baseline
 
         if ssd_copy.uv is not None:
-            baseline = ssd_copy.uv.get_baseline2d(debug=debug)
+            baseline = ssd_copy.uv.get_baseline2d(debug=debug, **baseline_kwargs)
             ssd_copy.uv.M -= baseline
 
         ssd_copy.time_required = time() - start_time
