### 2026-03-24 3f3ae94
ci: opt into Node.js 24 for GitHub Actions to fix deprecation warning
