### 2026-03-24 51f9700
docs: update workflow_notes — add issue titling policy and complete issues #28–#49
