### 2026-03-24 bab64db
v0.8.8: allow_negative_peaks stored state (#50) — set_allow_negative_peaks(), propagated by copy()
