### 2026-03-24 2658d0d
ci: remove FORCE_JAVASCRIPT_ACTIONS_TO_NODE24 (no longer needed with v6 actions)
