+++ .github/workflows/upload_to_pypi.yml
@@ -3,9 +3,6 @@ name: Manual Upload Python Package to PyPI
 on:
   workflow_dispatch:
 
-env:
-  FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true
-
 jobs:
   pypi-publish:
     name: upload release to PyPI
