### 2026-03-24 0c080e8
docs: add molass-beginner reference to README
