### 2026-03-24 d4f7f54
standard: v0.7 — remove COPILOT-INIT.md stub
