### 2026-03-24 92f652b
Squashed 'molass-library/' changes from b12c8b5..0c080e8
0c080e8 docs: add molass-beginner reference to README
5d1291b Merge branch 'master' into joss-paper
2658d0d ci: remove FORCE_JAVASCRIPT_ACTIONS_TO_NODE24 (no longer needed with v6 actions)
e7c9b09 ci: upgrade to actions/checkout@v6 and setup-python@v6 (Node.js 24 native)
bab64db v0.8.8: allow_negative_peaks stored state (#50) — set_allow_negative_peaks(), propagated by copy()
51f9700 docs: update workflow_notes — add issue titling policy and complete issues #28–#49
3f3ae94 ci: opt into Node.js 24 for GitHub Actions to fix deprecation war
[…]
