+++ Copilot/workflow_notes.md
@@ -17,6 +17,16 @@ Notes for AI assistants (Copilot, etc.) working in this repository across multip
 - Always use `gh` CLI to create issues — do NOT attempt browser-based approaches
 - Repo: biosaxs-dev/molass-library
 
+## Issue titling policy
+
+Use the correct prefix to categorize issues:
+
+- **`AI-friendliness:`** — API is confusing specifically for an AI (or any caller without deep context): wrong arg order, silent failures, missing defaults, opaque names. Examples: #47, #49, #28.
+- **`Design:`** — workflow/philosophy issues: separation of concerns, consistency with existing patterns, architectural decisions. Example: #50.
+- **`Enhancement:`** / **`Bug:`** — standard feature additions and bug fixes.
+
+Do NOT label design or philosophy issues as AI-friendliness.
+
 ## AI-friendliness improvement series
 When fixing AI-friendliness issues, follow this pattern per issue:
 1. **Open a GitHub issue FIRST**: `gh issue create --title "AI-friendliness: ..." --label "enhancement"`
@@ -25,7 +35,7 @@ When fixing AI-friendliness issues, follow this pattern per issue:
 4. Run `python -m pytest <test_file> -v --tb=short --no-header -q` to verify
 5. Close the issue: `gh issue close <N> --comment "Fix implemented and tested."`
 
-### Issues completed (all closed as of March 12, 2026)
+### Issues completed (all closed)
 - #8: `get_rgs()` returns `None` silently → `nan`
 - #9: `plot_components()` not axis-injectable
 - #10: No q-grid alignment utility
@@ -35,10 +45,26 @@ When fixing AI-friendliness issues, follow this pattern per issue:
 - #14: `SecSaxsData` non-standard data support
 - #15: Added `uv.wavelengths` and `uv.frames` aliases to `UvData`
 - #23: `get_baseline2d()` printed `recognize_num_peaks`/`peak_width=` to stdout → suppressed with `redirect_stdout(io.StringIO())`
-- #24: `'buffit'` baseline method — buffer-frame polyfit with Otsu adaptive threshold; `positive_ratio` 0.496–0.522 (nearest to ideal 0.5) on all 7 tested datasets
+- #24: `'buffit'` baseline method — buffer-frame polyfit with Otsu adaptive threshold
 - #25: Replaced fixed `BUFFIT_THRESHOLD=0.10` with Otsu adaptive threshold in `BuffitBaseline.py`
-- #26: Made `'buffit'` the default `baseline_method` for `XrData` (v0.8.4)
-- #27: Add `molass.requires(version)` helper for version guard in notebooks/tests; updated all 10 tutorial tests and data_correction.ipynb (v0.8.4)
+- #26: Made `'buffit'` default (v0.8.4) — **later reverted to `'linear'`** (actual default in code is `'linear'`)
+- #27: Add `molass.requires(version)` helper for version guard in notebooks/tests (v0.8.4)
+- #28: `SsMatrixData.q_values` / `.frame_indices` property aliases for opaque `iv` / `jv` (with setter)
+- #29: `SsMatrixData.get_bpo_ideal()` — single-call dataset-relative ideal positive_ratio
+- #30: `compute_lpm_baseline()` `mask` parameter for frame exclusion
+- #31: Wire `xr_peakpositions` into `quick_decomposition()` API + bug fixes
+- #32: `ComponentCurve` docstring Rg breadcrumb
+- #33: `Decomposition.xr_components` / `uv_components` property aliases
[…]
