### 2026-03-25 3dad48e
standard: v0.8 — add init.prompt.md, vscode-version.txt; gitignore version file
