### 2026-03-25 683da1f
Issues #51-53: mask-based allow_negative_peaks, slice frame-number fix, SSD delegation
- #51: Redesign allow_negative_peaks to mask negative-peak frames from LPM anchor pool (SsMatrixData)
- #52: Fix slice mask to use frame numbers via np.searchsorted (not array indices)
- #53: Add set_allow_negative_peaks() delegation method to SecSaxsData
- Tests: test_050, test_052, test_053 added
