+++ molass/DataObjects/SsMatrixData.py
@@ -27,14 +27,19 @@ class SsMatrixData:
     baseline_method : str
         The method used for baseline correction. Default is 'linear'.
     allow_negative_peaks : bool
-        If True, the endpoint-anchored LPM baseline is used automatically
-        (equivalent to ``endpoint_fraction=0.15``) to avoid contamination
-        from physically real negative-peak frames.  Default False.
+        If True, frames in the negative-peak region are excluded from LPM's
+        anchor pool before baseline fitting, preventing contamination from
+        physically real negative-peak frames.  Default False.
+    negative_peak_mask : array-like of bool, slice, or None
+        Explicit mask of frames to exclude from LPM fitting when
+        ``allow_negative_peaks=True``.  If None (default), the mask is
+        derived automatically from the recognition curve (frames where y < 0).
     """
     def __init__(self, M, iv, jv, E=None,
                  moment=None,
                  baseline_method='linear',
-                 allow_negative_peaks=False):
+                 allow_negative_peaks=False,
+                 negative_peak_mask=None):
         """Initialize the SsMatrixData object.
 
         Parameters
@@ -58,6 +63,7 @@ class SsMatrixData:
         self.moment = moment
         self.baseline_method = baseline_method
         self.allow_negative_peaks = allow_negative_peaks
+        self.negative_peak_mask = negative_peak_mask  # None = auto-detect from recognition curve
 
     @property
     def data(self):
@@ -109,6 +115,7 @@ class SsMatrixData:
                             moment=None,  # note that moment is not copied
                             baseline_method=self.baseline_method,
                             allow_negative_peaks=self.allow_negative_peaks,
+                            negative_peak_mask=self.negative_peak_mask,
                             )
 
     def get_icurve(self, pickat):
@@ -183,27 +190,39 @@ class SsMatrixData:
         """Get the baseline method for this data object."""
         return self.baseline_method
 
-    def set_allow_negative_peaks(self, value=True):
+    def set_allow_negative_peaks(self, value=True, mask=None):
         """Declare that this dataset contains physically real negative peaks.
 
-        When True, ``get_baseline2d()`` automatically uses the
-        endpoint-anchored LPM (``endpoint_fraction=0.15``) instead of the
-        standard bottom-percentile anchor, which would be contaminated by
-        negative-peak frames.
-
-        This flag is propagated by ``copy()`` and ``corrected_copy()``, so
-        the typical workflow is::
-
-            ssd.set_allow_negative_peaks()
-            ssd.plot_compact(baseline=True)   # inspect endpoint-anchored baseline
-            corrected = ssd.corrected_copy()  # applies it — no extra args needed
+        When True, ``get_baseline2d()`` excludes negative-peak frames from
+        LPM's anchor pool before baseline fitting.  LPM itself is unchanged;
+        only the frame set it operates on is filtered.
 
         Parameters
         ----------
         value : bool, optional
             Default True.
+        mask : array-like of bool, slice, or None, optional
+            Explicit mask of frames to exclude (True = exclude).  If None
+            (default), the mask is derived automatically at fitting time from
+            the recognition curve (frames where y < 0).  Use a manual mask
+            when the negative-peak region is known from domain knowledge
+            (e.g. ``mask=slice(1200, 1350)``).  When a ``slice`` is given,
+            start and stop are interpreted as **frame numbers** (values in
+            ``jv``), not array indices.
+
+        Notes
+        -----
+        Both ``allow_negative_peaks`` and ``negative_peak_mask`` are
+        propagated by ``copy()`` and ``corrected_copy()``, so the typical
+        workflow is::
+
[…]
