### 2026-03-26 9300f75
AI-friendliness: axes validation (#55), detect_peaks guidance (#56)
