### 2026-03-26 14046d2
feat: scaffolded rigorous optimization API
- Decomposition: frozen_components + free_components in optimize_rigorously()
- RigorousImplement: wire frozen_components to optimizer
- RunInfo: resume(niter=) method for continuing optimization
- Update PROJECT_STATUS.md
