### 2026-03-27 c9e39f1
Issues #58, #59: folder layout docs + list_rigorous_jobs utility
- optimize_rigorously() docstring: document analysis_folder layout
- load_rigorous_result() docstring: cross-ref list_rigorous_jobs
- list_rigorous_jobs(): returns JobInfo(id, iterations, best_fv, timestamp)
- Decomposition.list_rigorous_jobs() static method
- (Earlier this session) clear_jobs param, load_rigorous_result, for_split_only support
