+++ molass/LowRank/Decomposition.py
@@ -687,7 +687,8 @@ class Decomposition:
             raise ValueError(f"Decomposition.make_rigorous_initparams: Unsupported model '{self.model}'")
 
     def optimize_rigorously(self, rgcurve=None, analysis_folder=None, method='BH', niter=20,
-                            frozen_components=None, free_components=None, debug=False):
+                            frozen_components=None, free_components=None,
+                            clear_jobs=True, debug=False):
         """
         Perform a rigorous decomposition.
 
@@ -696,7 +697,20 @@ class Decomposition:
         rgcurve : Curve
             The Rg curve to use for the decomposition.
         analysis_folder : str, optional
-            The folder to save analysis results.
+            The folder to save analysis results.  Optimization creates
+            the following layout on disk::
+
+                <analysis_folder>/
+                    optimized/
+                        jobs/
+                            000/callback.txt   # job 0
+                            001/callback.txt   # job 1 (e.g. after Resume)
+                            ...
+
+            Each ``callback.txt`` records per-iteration objective values
+            and parameter vectors.  Use ``list_rigorous_jobs()`` to
+            inspect existing jobs, or ``load_rigorous_result()`` to
+            reconstruct a ``Decomposition`` from a completed job.
         method : str, optional
             The method to use for rigorous optimization. Default is 'BH'.
         niter : int, optional
@@ -712,6 +726,10 @@ class Decomposition:
             ``frozen_components`` — use whichever is shorter.
             E.g., ``free_components=[4]`` to optimize only the main peak.
             Mutually exclusive with ``frozen_components``.
+        clear_jobs : bool, optional
+            If True (default), existing job folders are cleared before starting.
+            Set to False after a kernel restart to preserve previous job results
+            and reconstruct RunInfo without losing optimization history.
         debug : bool, optional
             If True, enable debug mode.
 
@@ -740,4 +758,71 @@ class Decomposition:
         if rgcurve is None:
             rgcurve = self.ssd.xr.compute_rgcurve()
 
-        return make_rigorous_decomposition_impl(self, rgcurve, analysis_folder=analysis_folder, method=method, niter=niter, frozen_components=frozen_components, debug=debug)
+        return make_rigorous_decomposition_impl(self, rgcurve, analysis_folder=analysis_folder, method=method, niter=niter, frozen_components=frozen_components, clear_jobs=clear_jobs, debug=debug)
+
+    def load_rigorous_result(self, analysis_folder, jobid=None, debug=False):
+        """Load a completed rigorous optimization result from disk.
+
+        This reads saved parameters without launching a new subprocess.
+        Use it to view results from a previous session after a kernel or
+        VS Code restart.
+
+        Parameters
+        ----------
+        analysis_folder : str
+            The same ``analysis_folder`` passed to ``optimize_rigorously()``.
+            See ``optimize_rigorously()`` for the folder layout.
+        jobid : str, optional
+            Specific job id (subfolder name, e.g. ``'001'``).  If None,
+            loads the latest job.  Use ``list_rigorous_jobs()`` to see
+            available jobs.
+        debug : bool, optional
+            If True, reload modules from disk.
+
+        Returns
+        -------
+        Decomposition
+            A new Decomposition with the optimized components.
+
+        Examples
+        --------
+        After kernel restart, re-run data loading and quick decomposition,
+        then::
+
+            result = decomp.load_rigorous_result("temp_analysis_scaffolded")
+            result.plot_components(rgcurve=rgcurve)
+        """
+        if debug:
+            import molass.Rigorous.CurrentStateUtils
[…]
