+++ molass/Rigorous/LegacyBridgeUtils.py
@@ -52,7 +52,7 @@ def make_basecurves_from_decomposition(decomposition, debug=False):
     baseline_type = 1
     return make_basecurves_from_sd(sd, baseline_type, xr_only=xr_only, debug=debug)
 
-def construct_legacy_optimizer(dsets, baseline_objects, spectral_vectors, num_components=3, model="EGH", method="BH", debug=False):
+def construct_legacy_optimizer(dsets, baseline_objects, spectral_vectors, num_components=3, model="EGH", method="BH", for_split_only=False, debug=False):
     from molass_legacy.Optimizer.OptimizerUtils import get_function_code
     from molass_legacy.Optimizer.FuncImporter import import_objective_function
     function_code = get_function_code(model)
@@ -64,6 +64,7 @@ def construct_legacy_optimizer(dsets, baseline_objects, spectral_vectors, num_co
         uv_base_curve=baseline_objects[0],
         qvector=spectral_vectors[0],
         wvector=spectral_vectors[1],
+        for_split_only=for_split_only,
         )
 
     # backward compatibility for estimator setting
