### 2026-03-27 645d03c
Update PROJECT_STATUS.md for March 27 session
