### 2026-03-27 35c7e3b
Negative-peak-aware rigorous optimization
- SdProxy: use get_recognition_curve() with max-normalization; mask excluded frames for ScatteringBaseline
- SecSaxsData.corrected_copy(): interpolate (not zero) excluded frames for both XR and UV via mapping
- LegacyBridgeUtils: use recognition curve in MplMonitor display; force-export corrected data when negative peaks active so subprocess sees interpolated matrices
