### 2026-03-27 729d63f
Squashed 'molass-library/' changes from 0c080e8..0ce18c6
0ce18c6 Merge branch 'main' into joss-paper
35c7e3b Negative-peak-aware rigorous optimization
645d03c Update PROJECT_STATUS.md for March 27 session
c9e39f1 Issues #58, #59: folder layout docs + list_rigorous_jobs utility
14046d2 feat: scaffolded rigorous optimization API
9300f75 AI-friendliness: axes validation (#55), detect_peaks guidance (#56)
683da1f Issues #51-53: mask-based allow_negative_peaks, slice frame-number fix, SSD delegation
3dad48e standard: v0.8 — add init.prompt.md, vscode-version.txt; gitignore version file
246b29a stan
[…]
