### 2026-03-30 #9424 @marshallmcdonnell
> Hi [@AndySAnker](https://github.com/AndySAnker), [@KedoKudo](https://github.com/KedoKudo) and [@marshallmcdonnell](https://github.com/marshallmcdonnell), I would like to invite you to review this Journal of Open Source Software (JOSS) submission. I think it should be related to your field of research.
>
> The review process of JOSS is unique as it happens on GitHub. The process is open and transparent. Interaction between the reviewers and the author are done via GitHub issues (and encouraged). You can learn more about the reviewing guideline here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.

@zhubonan Also happy to help, thanks!
