### 2026-04-02 af77b0a
AI-friendliness #60: get_positive_ratio baseline=None; Design #61: corrected_copy baseline param
