### 2026-04-03 d5f2941
Issues #63, #64: set_anomaly_mask rename + kwargs validation; plot_components recognition curve overlay
