+++ molass/DataObjects/SsMatrixData.py
@@ -27,12 +27,17 @@ class SsMatrixData:
     baseline_method : str
         The method used for baseline correction. Default is 'linear'.
     allow_negative_peaks : bool
-        If True, frames in the negative-peak region are excluded from LPM's
+        Deprecated alias for ``has_anomaly_mask``.
+    has_anomaly_mask : bool
+        If True, frames in the anomaly region are excluded from LPM's
         anchor pool before baseline fitting, preventing contamination from
-        physically real negative-peak frames.  Default False.
+        physically anomalous frames (e.g. negative-peak regions).
+        Default False.
     negative_peak_mask : array-like of bool, slice, or None
+        Deprecated alias for ``anomaly_mask``.
+    anomaly_mask : array-like of bool, slice, or None
         Explicit mask of frames to exclude from LPM fitting when
-        ``allow_negative_peaks=True``.  If None (default), the mask is
+        ``has_anomaly_mask=True``.  If None (default), the mask is
         derived automatically from the recognition curve (frames where y < 0).
     """
     def __init__(self, M, iv, jv, E=None,
@@ -62,8 +67,8 @@ class SsMatrixData:
         self.E = E      # may be None
         self.moment = moment
         self.baseline_method = baseline_method
-        self.allow_negative_peaks = allow_negative_peaks
-        self.negative_peak_mask = negative_peak_mask  # None = auto-detect from recognition curve
+        self.has_anomaly_mask = allow_negative_peaks
+        self.anomaly_mask = negative_peak_mask  # None = auto-detect from recognition curve
 
     @property
     def data(self):
@@ -114,8 +119,8 @@ class SsMatrixData:
                             Ecopy,
                             moment=None,  # note that moment is not copied
                             baseline_method=self.baseline_method,
-                            allow_negative_peaks=self.allow_negative_peaks,
-                            negative_peak_mask=self.negative_peak_mask,
+                            allow_negative_peaks=self.has_anomaly_mask,
+                            negative_peak_mask=self.anomaly_mask,
                             )
 
     def get_icurve(self, pickat):
@@ -190,39 +195,72 @@ class SsMatrixData:
         """Get the baseline method for this data object."""
         return self.baseline_method
 
-    def set_allow_negative_peaks(self, value=True, mask=None):
-        """Declare that this dataset contains physically real negative peaks.
+    def set_anomaly_mask(self, mask=None):
+        """Declare that this dataset contains anomalous frames to exclude from baseline fitting.
 
-        When True, ``get_baseline2d()`` excludes negative-peak frames from
+        When set, ``get_baseline2d()`` excludes the specified frames from
         LPM's anchor pool before baseline fitting.  LPM itself is unchanged;
         only the frame set it operates on is filtered.
 
         Parameters
         ----------
-        value : bool, optional
-            Default True.
         mask : array-like of bool, slice, or None, optional
             Explicit mask of frames to exclude (True = exclude).  If None
             (default), the mask is derived automatically at fitting time from
             the recognition curve (frames where y < 0).  Use a manual mask
-            when the negative-peak region is known from domain knowledge
+            when the anomalous region is known from domain knowledge
             (e.g. ``mask=slice(1200, 1350)``).  When a ``slice`` is given,
             start and stop are interpreted as **frame numbers** (values in
             ``jv``), not array indices.
 
         Notes
         -----
-        Both ``allow_negative_peaks`` and ``negative_peak_mask`` are
+        Both ``has_anomaly_mask`` and ``anomaly_mask`` are
[…]
