### 2026-04-06 ff3be0f
feat: EGH peeler (#72), issue fixes #68-71, version 0.9.1
