+++ tests/generic/010_DataObjects/test_010_SSD.py
@@ -44,6 +44,12 @@ def test_05_repr():
     assert "wavelengths=" in uv_repr and "frames=" in uv_repr, f"uv repr missing shape info: {uv_repr}"
     assert "nm" in uv_repr, f"uv repr should include wavelength range in nm: {uv_repr}"
 
+def test_05b_ssd_repr():
+    r = repr(ssd_instance)
+    assert "SecSaxsData" in r
+    assert "xr=242 frames" in r
+    assert "trimmed=False" in r
+
 def test_06_wavelength_range():
     uv = ssd_instance.uv
     wl_min, wl_max = uv.wavelength_range
@@ -285,3 +291,39 @@ def test_20_corrected_copy_precomputed_baseline():
 
     # XR matrices should be equal
     np.testing.assert_allclose(corrected_pre.xr.M, corrected_std.xr.M, atol=1e-12)
+
+def test_21_get_data_info():
+    """get_data_info() returns DataInfo namedtuple with UV peak wavelength and pickat mismatch."""
+    info = ssd_instance.get_data_info()
+    assert info.n_xr_frames == ssd_instance.xr.M.shape[1]
+    assert info.n_uv_frames == ssd_instance.uv.M.shape[1]
+    assert info.uv_pickat == 280
+    assert info.is_trimmed == False
+    # SAMPLE1 peak is at ~234 nm (peptide bond), below pickat 280 → no mismatch
+    assert info.uv_peak_wavelength < 250
+    assert info.pickat_mismatch == False
+    # SAMPLE1 has no significant negative XR regions (neg dip <3% of peak)
+    assert info.has_negative_xr_regions == False
+    assert info.negative_xr_fraction is not None
+    # Baseline self-test: SAMPLE1 is clean, p should be high (not significant)
+    assert info.baseline_selftest_p is not None
+    assert info.baseline_selftest_p > 0.01
+    # uv_monitor is alias for uv_pickat
+    assert info.uv_monitor == info.uv_pickat
+
+
+def test_22_uv_monitor_alias():
+    """uv_monitor alias works in SSD constructor and UvData property."""
+    from molass_data import SAMPLE1
+    from molass.DataObjects import SecSaxsData as SSD
+
+    # Constructor alias
+    ssd_mon = SSD(SAMPLE1, uv_monitor=290)
+    assert ssd_mon.uv.pickat == 290
+    assert ssd_mon.uv.monitor == 290
+    assert ssd_mon.uv.uv_pickat == 290
+
+    # Property setter
+    ssd_mon.uv.monitor = 300
+    assert ssd_mon.uv.pickat == 300
+    assert ssd_mon.uv.uv_pickat == 300
