+++ molass/Peaks/EghPeeler.py
@@ -1,0 +1,162 @@
+"""
+Peaks.EghPeeler
+
+Sequential EGH (Exponentially-modified Gaussian Hybrid) peeling for
+automatic peak recognition and component count estimation.
+
+Algorithm:
+    1. Find tallest remaining peak (argmax of smoothed residual)
+    2. Fit EGH(H, mu, sigma, tau) via L-BFGS-B
+    3. Check significance: area(fitted) / total_area >= min_area_frac
+    4. Check shape: sigma >= min_sigma (reject noise spikes)
+    5. Subtract fitted EGH from residual
+    6. Repeat until no significant candidate remains
+
+This replaces the legacy ``recognize_peaks`` (greedy sequential subtraction
+with symmetric Gaussian initialization) from molass-legacy.
+"""
+import numpy as np
+from scipy.optimize import minimize as sp_minimize
+from scipy.signal import savgol_filter
+from molass.SEC.Models.Simple import egh, gaussian
+
+DEFAULT_MIN_AREA_FRAC = 0.02
+DEFAULT_MIN_SIGMA = 3.0
+TAU_BOUND_RATIO = 2.0
+MAX_STEPS = 10
+
+
+def _egh_area(x, H, mu, sigma, tau):
+    """Compute area under an EGH peak by numerical integration."""
+    y = egh(x, H, mu, sigma, tau)
+    return np.trapezoid(y, x)
+
+
+def _estimate_sigma(x, y, peak_idx, h_init, mu_init):
+    """Estimate initial sigma by expanding a Gaussian window until >50% of
+    points fall below the Gaussian envelope."""
+    sigma_init = 5.0
+    for s in range(3, len(x) // 2):
+        sl = slice(max(0, peak_idx - s), min(len(x), peak_idx + s))
+        gy = gaussian(x[sl], h_init, mu_init, s)
+        neg_frac = np.sum(y[sl] - gy < 0) / len(x[sl])
+        if neg_frac > 0.5:
+            sigma_init = float(s)
+            break
+    return sigma_init
+
+
+def egh_peel(x, y, num_components=None, min_area_frac=DEFAULT_MIN_AREA_FRAC,
+             min_sigma=DEFAULT_MIN_SIGMA, debug=False):
+    """Sequential EGH peeling to find peaks and estimate component count.
+
+    Parameters
+    ----------
+    x : array-like
+        Frame positions (1-D).
+    y : array-like
+        Intensity values (1-D, same length as *x*).
+    num_components : int or None
+        If given, peel exactly this many peaks (ignoring area significance).
+        If ``None``, peel until no significant candidate remains.
+    min_area_frac : float
+        Minimum area fraction (relative to total curve area) for a peak to
+        be considered significant.  Only used when ``num_components is None``.
+        Default 0.02 (2 %).
+    min_sigma : float
+        Minimum allowed sigma.  Peaks narrower than this are rejected as
+        noise spikes.  Default 3.0 frames.
+    debug : bool
+        If True, print diagnostic messages.
+
+    Returns
+    -------
+    peak_list : list of list
+        Each element is ``[H, mu, sigma, tau]`` — the fitted EGH parameters.
+        Sorted by ascending *mu* (retention time order).
+    """
+    x = np.asarray(x, dtype=float)
+    y = np.asarray(y, dtype=float)
+    residual = y.copy()
+    residual[residual < 0] = 0
+
+    total_area = np.trapezoid(y, x)
+    if total_area <= 0:
+        return []
+
+    max_steps = num_components if num_components is not None else MAX_STEPS
+    peak_list = []
+
+    for step in range(max_steps):
[…]
