+++ molass/LowRank/CoupledAdjuster.py
@@ -48,11 +48,12 @@ def make_component_curves(ssd, num_components, **kwargs):
 
     xr_peakpositions = kwargs.pop('xr_peakpositions', None)
     xr_icurve = ssd.xr.get_icurve()
-    if num_components is None:
-        num_components = len(xr_icurve.get_peaks())
+    # When num_components is None, let the EGH peeler auto-detect
     kwargs['data_matrix'] = ssd.xr.M
     kwargs['qv'] = ssd.xr.qv
     xr_ccurves = decompose_icurve_impl(xr_icurve, num_components, peakpositions=xr_peakpositions, **kwargs)
+    if num_components is None:
+        num_components = len(xr_ccurves)
 
     if ssd.uv is None:
         return xr_icurve, xr_ccurves, None, None
