### 2026-04-09 bc5cd50
feat: rigorous optimization APIs, floor-constrained optimization, EghPeeler improvements
- Decomposition: has_rigorous_results(), wait_for_rigorous_results(), list_rigorous_jobs()
- RigorousImplement: _compute_basic_floor(), pass basic_floor to legacy bridge
- CurrentStateUtils: JobInfo namedtuple, parse callback.txt for job listing
- LegacyBridgeUtils: basic_floor kwarg passthrough
- SecSaxsData/QuickImplement: optimize_rigorously() API wiring
- EghPeeler: tau estimation improvements
- tests/tutorial/11-rigorous_optimization.py: updated for new APIs
- tests/specific: EghPeeler test added
