### 2026-04-10 b52d248
feat: unified anomaly detection, visualization bands, rigorous optimizer improvements
