### 2026-04-12 70168eb
Add State of the field section to paper.md; update CURRENT-STATUS.md (review started); add response draft for Shimizu
