+++ paper.md
@@ -52,15 +52,19 @@ Analysis of SEC-SAXS data is inherently multi-step and complex. A typical workfl
 7. Folding state estimation – Kratky plot [@Kratky_1963]
 8. Electron density calculation 
 
-Molass Library currently implements steps 3–7. For steps 1 and 2, users may employ SAngler [@Shimizu:2016] or device-specific software, while for step 8, DENSS [@Grant:2018] is recommended. Although alternative tools exist, such as the proprietary program suite ATSAS [@Manalastas-Cantos:ge5081] and the open-source program BioXTAS RAW [@Hopkins:jl5075], Molass Library distinguishes itself by providing an open, scriptable, and modular platform. This design empowers researchers to flexibly tailor and extend their analysis pipelines within the Python ecosystem, thereby enhancing both reproducibility and adaptability.
+Molass Library currently implements steps 3–7. For steps 1 and 2, users may employ SAngler [@Shimizu:2016] or device-specific software, while for step 8, DENSS [@Grant:2018] is recommended.
+
+# State of the field
+
+Several software tools address SEC-SAXS data processing. The proprietary ATSAS suite [@Manalastas-Cantos:ge5081] and the open-source BioXTAS RAW [@Hopkins:jl5075] provide comprehensive GUI-based workflows widely used in the field. For the specific challenge of overlapping peak decomposition—required when chromatographic peaks overlap or significant interparticle interference occurs—CHROMIXS defers such analysis to "other methods" [@Panjkovich:2018], whereas EFAMIX defines quantitative conditions for reliable separation (signal-to-noise ratio ≥10³ for three-component separation; peak asymmetry parameter τ≤2; baseline width separation ≥2× for reliable decomposition) [@Konarev:2021]. REGALS employs a two-stage framework combining EFA with L1 regularization [@Meisburger:mf5050] and is therefore guided by assumptions underlying EFA in strongly overlapping cases. Molass Library takes a complementary approach: rather than extending these existing GUI-based tools, it provides a Python-first, Jupyter-centric platform in which overlapping peak decomposition is handled through explicit parametric elution curve modeling combined with global optimization, constrained by physicochemical consistency derived from SEC theory (size-dependent elution behavior) and SAXS analysis (radius of gyration consistency across elution fractions).
 
 # Software Design
 
-The MOLASS library is designed to support sustainable software development by providing an open framework that enables broad community participation and contribution. Rather than contributing to existing GUI-based tools like ATSAS or BioXTAS RAW, we chose to build a Python-first, Jupyter-centric solution designed to support AI-assisted maintenance and researcher extensibility. This design philosophy prioritizes explicit, readable code over graphical interfaces, making the analysis methodology transparent for both learning and AI-assisted maintenance. We are actively enhancing AI-readiness through systematic usability testing with AI agents and iterative improvements to API discoverability and inline documentation. The architecture emphasizes modularity and clear separation of concerns: elution curve models (EGH, SDM, EDM) are isolated as parametric functions; low-rank factorization uses standard linear algebra (Moore-Penrose pseudoinverse); and visualization is decoupled from computation. By integrating established packages (NumPy, SciPy, pybaselines, ruptures) rather than reimplementing core functionality, we reduce custom code volume and enhance long-term viability. This approach allows domain researchers to maintain and extend the code using AI-assisted development tools.
[…]
