### 2026-04-13 037ed28
feat: quiet mode, score analysis, convergence plot, rigorous optimization improvements
