+++ molass/Rigorous/CurrentStateUtils.py
@@ -10,19 +10,63 @@ from molass.SEC.Models.UvComponentCurve import UvComponentCurve
 from molass.Mapping.Mapping import Mapping
 
 JobInfo = namedtuple('JobInfo', ['id', 'iterations', 'best_fv', 'timestamp'])
-"""Metadata for a single rigorous optimization job.
+JobConvergence = namedtuple('JobConvergence', ['id', 'evals', 'fvs', 'best_fv', 'best_sv', 'timestamps'])
+ConvergenceInfo = namedtuple('ConvergenceInfo', [
+    'jobs', 'best_fv', 'best_sv', 'best_job_id', 'spread', 'trend', 'n_jobs',
+])
 
-Attributes
-----------
-id : str
-    Job folder name (e.g. ``'000'``, ``'001'``).
-iterations : int
-    Number of iterations recorded in ``callback.txt``.
-best_fv : float
-    Best (minimum) objective function value across all iterations.
-timestamp : datetime or None
-    Timestamp of the last recorded iteration, or ``None`` if unavailable.
+JobInfo.__doc__ = """Metadata for a single rigorous optimization job.
+
+Attributes: id, iterations, best_fv, timestamp.
 """
+
+JobConvergence.__doc__ = """Per-job convergence detail with full fv trajectory.
+
+Attributes: id, evals (list of eval counts), fvs (list of fv values),
+best_fv, best_sv, timestamps (list of datetimes).
+"""
+
+ConvergenceInfo.__doc__ = """Summary of convergence across all jobs.
+
+Attributes: jobs (list of JobConvergence), best_fv, best_sv, best_job_id,
+spread, trend ('improving'/'worsening'/'stable'), n_jobs.
+"""
+
+
+def fv_to_sv(fv):
+    """Convert objective function value (fv) to Score Value (SV, 0-100 scale).
+
+    The raw optimizer objective fv is negative and nonlinear, making it hard
+    to interpret. SV maps it to a human-readable 0-100 quality scale via:
+
+        SV = -200 / (1 + exp(-1.5 * fv)) + 100
+
+    Quality thresholds:
+        SV >= 80 : Good (green in plot_convergence)
+        SV 60-80 : Fair (orange)
+        SV <  60 : Poor (red)
+
+    Reference points:
+        fv = -3  -> SV ~= 98
+        fv = -1  -> SV ~= 64
+        fv =  0  -> SV =   0
+
+    Can also be applied to individual score values (not just the total fv)
+    to identify which objectives are bottlenecks.
+
+    Parameters
+    ----------
+    fv : float or array-like
+        Objective function value(s).
+
+    Returns
+    -------
+    float or ndarray
+        Score Value(s) on 0-100 scale.
+    """
+    import numpy as np
+    fv = np.asarray(fv, dtype=float)
+    return -200.0 / (1.0 + np.exp(-1.5 * fv)) + 100.0
 
 def construct_decomposition_from_results(run_info, **kwargs):      
     optimizer_folder = get_setting('optimizer_folder')
@@ -355,3 +399,228 @@ def wait_for_rigorous_results(analysis_folder, timeout=600, poll_interval=5):
         elapsed += poll_interval
 
     return True
+
+
+def read_convergence_data(analysis_folder):
+    """Read convergence data from all completed rigorous optimization jobs.
+
+    Parameters
+    ----------
+    analysis_folder : str
+        The same ``analysis_folder`` passed to ``optimize_rigorously()``.
+
+    Returns
+    -------
+    ConvergenceInfo
+        Summary with per-job trajectories and cross-job diagnostics.
+        Key fields for programmatic assessment:
+
[…]
