### 2026-04-14 4d57572
AI-friendliness: get_score_breakdown() on RunInfo (#87) Returns dict of individual objective function scores for diagnosis.
