### 2026-04-15 bf3c06b
feat: two-axis SDM variant system (pore_dist x rt_dist) - SdmColumn/SdmComponentCurve: pore_dist/rt_dist attributes, PDF dispatch - SdmOptimizer: reads variant from model_params, k bounds per rt_dist - FunctionCodeUtils.py: FUNCTION_CODE_MAP + detect_function_code() - RigorousSdmParams: conditional 6/7 element sdmcol_params - ComponentUtils: variant inference on load - Fix: map exponential->G1100 explicitly (not None)
