+++ molass/SEC/Models/SdmOptimizer.py
@@ -3,7 +3,11 @@ SEC.Models.SdmOptimizer.py
 """
 import numpy as np
 from scipy.optimize import minimize
-from molass_legacy.Models.Stochastic.DispersivePdf import dispersive_monopore_pdf, DEFUALT_TIMESCALE
+from molass.SEC.Models.SdmMonoPore import (
+    sdm_monopore_pdf,
+    sdm_monopore_gamma_pdf,
+    DEFAULT_TIMESCALE,
+)
 
 def optimize_sdm_xr_decomposition(decomposition, env_params, model_params=None, **kwargs):
     """ Optimize the SDM decomposition.
@@ -39,37 +43,54 @@ def optimize_sdm_xr_decomposition(decomposition, env_params, model_params=None, 
     rgv = np.asarray(decomposition.get_rgs())
 
     if model_params is None:
-        timescale = DEFUALT_TIMESCALE
+        timescale = DEFAULT_TIMESCALE
+        k_init = 2.0
+        pore_dist = 'mono'
+        rt_dist = 'gamma'
     else:
-        timescale = model_params.get('timescale', DEFUALT_TIMESCALE)
+        timescale = model_params.get('timescale', DEFAULT_TIMESCALE)
+        k_init = model_params.get('k', 2.0)
+        pore_dist = model_params.get('pore_dist', 'mono')
+        rt_dist = model_params.get('rt_dist', 'gamma')
+
+    if rt_dist == 'exponential':
+        k_init = 1.0   # not optimized for exponential
+        _pdf_func = sdm_monopore_pdf
+    else:
+        _pdf_func = sdm_monopore_gamma_pdf
 
     def estimate_initial_scales():
         scales = []
         for rg in rgv:
-            column = SdmColumn([N, T, me, mp, t0, t0, N0, poresize, timescale])
+            column = SdmColumn([N, T, me, mp, t0, t0, N0, poresize, timescale, k_init],
+                               pore_dist=pore_dist, rt_dist=rt_dist)
             ccurve = SdmComponentCurve(x, column, rg, scale=1.0)
             cy = ccurve.get_y()
-            k = np.argmax(cy)
-            scale = y[k]/cy[k]
+            idx = np.argmax(cy)
+            scale = y[idx]/cy[idx] if cy[idx] > 0 else 1.0
             scales.append(scale)
         return scales
 
     def objective_function(params, return_cy_list=False, plot=False):
-        N_, T_, x0_, tI_, N0_ = params[0:5]
-        rgv_ = params[5:5+num_components]
+        N_, T_, x0_, tI_, N0_, k_ = params[0:6]
+        rgv_ = params[6:6+num_components]
         rg_diff = np.diff(rgv_)
         non_ordered = np.where(rg_diff > 0)[0]
         order_penalty = np.sum(rg_diff[non_ordered]**2) * 1e3  # penalty for non-ordered rgv
         rhov = rgv_/poresize
         rhov[rhov > 1] = 1.0  # limit rhov to 1.0
-        scales_ = params[5+num_components:5+2*num_components]
+        scales_ = params[6+num_components:6+2*num_components]
         cy_list = []
         x_ = x - tI_
         t0 = x0_ - tI_
         for rho, scale in zip(rhov, scales_):
             ni = N_*(1 - rho)**me
             ti = T_*(1 - rho)**mp
-            cy = scale * dispersive_monopore_pdf(x_, ni, ti, N0_, t0, timescale=timescale)
+            if rt_dist == 'exponential':
+                cy = scale * _pdf_func(x_, ni, ti, N0_, t0, timescale=timescale)
+            else:
[…]
