### 2026-04-16 023a3bb
docs: add notebook workflow reference; broaden AI-friendliness definition
