### 2026-04-16 450b01f
feat: wire G1300 for SDM(lognormal) rigorous optimization (#94)
