+++ molass/Rigorous/ComponentUtils.py
@@ -17,24 +17,30 @@ def get_sdm_xr_ccurves(optimizer, xr_icurve, separated_params):
     xr_params = separated_params[0]
     rg_params = separated_params[2]
     sdmcol = separated_params[-1]
-    if len(sdmcol) >= 7:
-        N, K, x0, poresize, N0, tI, k_gamma = sdmcol[:7]
-    else:
-        N, K, x0, poresize, N0, tI = sdmcol
-        k_gamma = 1.0
-    T = K/N
-    me = mp = 1.5
     # Infer pore_dist/rt_dist from function code
-    model_name = optimizer.get_model_name()
     func_code = getattr(optimizer, 'function_code', None)
-    if func_code == 'G1200':
-        pore_dist, rt_dist = 'mono', 'gamma'
-    elif k_gamma != 1.0:
-        pore_dist, rt_dist = 'mono', 'gamma'
+    me = mp = 1.5
+    if func_code == 'G1300':
+        # Lognormal pore + gamma: sdmcol = [N, K, x0, mu, sigma, N0, tI, k]
+        N, K, x0, mu, sigma, N0, tI, k_gamma = sdmcol
+        T = K/N
+        column = SdmColumn([N, T, me, mp, x0, tI, N0, mu, sigma, k_gamma],
+                           pore_dist='lognormal', rt_dist='gamma')
     else:
-        pore_dist, rt_dist = 'mono', 'exponential'
-    column = SdmColumn([N, T, me, mp, x0, tI, N0, poresize, DEFUALT_TIMESCALE, k_gamma],
-                       pore_dist=pore_dist, rt_dist=rt_dist)
+        if len(sdmcol) >= 7:
+            N, K, x0, poresize, N0, tI, k_gamma = sdmcol[:7]
+        else:
+            N, K, x0, poresize, N0, tI = sdmcol
+            k_gamma = 1.0
+        T = K/N
+        if func_code == 'G1200':
+            pore_dist, rt_dist = 'mono', 'gamma'
+        elif k_gamma != 1.0:
+            pore_dist, rt_dist = 'mono', 'gamma'
+        else:
+            pore_dist, rt_dist = 'mono', 'exponential'
+        column = SdmColumn([N, T, me, mp, x0, tI, N0, poresize, DEFUALT_TIMESCALE, k_gamma],
+                           pore_dist=pore_dist, rt_dist=rt_dist)
     x = xr_icurve.x
     xr_ccurves = []
     for scale, rg in zip(xr_params, rg_params):
