### 2026-04-17 077cfea
fix: G1300 load path + score breakdown + Rg overlay (#103, #104)
