### 2026-04-17 5de854b
v0.9.3: SdmOptimizer improvements, dep bumps, study notebooks
