### 2026-04-20 710cfe9
fix: remove basic_floor from rigorous pipeline (#105, #106)
