### 2026-04-20 b9e3f72
fix: SDM lognormal optimizer initialization (#108)
