### 2026-04-21 f0f7b62
Add fast analytical moments for SDM lognormal-pore gamma (issue #113)
- Add sdm_lognormal_model_moments() in LognormalPore.py: 64-point Gauss-Legendre quadrature with hand-rolled lognormal PDF; ~50us/call (200x faster than full PDF)

- Add refine_lognormal_params_by_moments() in SdmEstimator.py: L-BFGS-B refinement of (t0, k, mu, sigma) against EGH per-component (M1, Var)

- Wire refinement into estimate_sdm_lognormal_from_monopore() (controlled by moment_refine kwarg, default True)

- SDM.py passes decomposition= so the refinement is enabled in the default lognormal pipeline

- Variance formu
[…]
