### 2026-04-21 39c0c1f
docs: update PROJECT_STATUS for issue #113 (fast SDM lognormal moments)
