### 2026-04-22 7844085
WIP: split-architecture design doc + in-process branch in RigorousImplement + ongoing tests
