+++ molass/LowRank/NumComponentsRecommender.py
@@ -1,0 +1,232 @@
+"""
+LowRank.NumComponentsRecommender — recommend ``num_components`` for SDM-style
+decompositions by detecting degeneracy when an extra component is added.
+
+Background
+----------
+Setting ``num_components`` one too high for the data is a common failure mode
+of SDM (and SDM-rigorous) optimization: the extra component's elution profile
+collapses onto an existing one, ``cond(C)`` blows up, and ``optimize_rigorously``
+later stalls at SV ≈ 51 ("Poor"). The diagnostic implemented here was
+verified across 7 datasets (SAMPLE1-4 + Apo + ATP + MY) in
+``molass-researcher/experiments/13_rigorous_optimization/13aa_num_components_diagnostic.ipynb``
+(see issue #116 for the full table).
+"""
+from collections import namedtuple
+import contextlib
+import io
+import warnings
+
+import numpy as np
+import pandas as pd
+
+
+Recommendation = namedtuple(
+    "Recommendation",
+    ["recommended_k", "reason", "metrics"],
+)
+"""Result of :meth:`Decomposition.recommend_num_components`.
+
+Attributes
+----------
+recommended_k : int or None
+    The recommended ``num_components``. ``None`` if no fit succeeded.
+reason : str
+    Human-readable justification for the choice.
+metrics : pandas.DataFrame
+    One row per ``k`` actually fitted, with columns
+    ``['k', 'residual', 'cond_C', 'max_cos', 'amp_ratio', 'flag_count', 'status']``.
+"""
+
+
+# Default thresholds for the three degeneracy flags. Verified in 13aa across
+# 7 datasets x k in {1,2,3}; no dataset trips a flag when k matches the true
+# number of components.
+DEFAULT_COND_THRESHOLD = 50.0
+DEFAULT_COS_THRESHOLD = 0.99
+DEFAULT_AMP_THRESHOLD = 0.20
+
+
+def _diagnose(decomp):
+    """Compute (residual, cond_C, max_cos, amp_ratio, n) for a Decomposition.
+
+    P is solved as ``M @ pinv(C)``; residual is ``||M - PC||_F / ||M||_F``.
+    """
+    C = np.array([c.y for c in decomp.xr_ccurves])  # (k, frames)
+    M = decomp.ssd.xr.M                              # (q, frames)
+    P = M @ np.linalg.pinv(C)                        # (q, k)
+
+    res = float(np.linalg.norm(M - P @ C) / np.linalg.norm(M))
+
+    try:
+        cond_C = float(np.linalg.cond(C))
+    except Exception:
+        cond_C = float("inf")
+
+    n = C.shape[0]
+    if n == 1:
+        max_cos = float("nan")
+    else:
+        norms = np.linalg.norm(C, axis=1)
+        gram = (C @ C.T) / np.outer(norms, norms)
+        np.fill_diagonal(gram, 0.0)
+        max_cos = float(np.max(np.abs(gram)))
+
+    # numpy>=2 renamed trapz -> trapezoid
+    _trapz = getattr(np, "trapezoid", getattr(np, "trapz"))
+    areas = _trapz(np.maximum(C, 0), axis=1)
+    if n == 1 or areas.max() == 0:
+        amp_ratio = float("nan")
+    else:
+        amp_ratio = float(areas.min() / areas.max())
+
+    return res, cond_C, max_cos, amp_ratio, n
+
+
+def _count_flags(row, cond_thr, cos_thr, amp_thr):
+    f = 0
+    if row["cond_C"] > cond_thr:
+        f += 1
+    if row["max_cos"] > cos_thr:
+        f += 1
+    if row["amp_ratio"] < amp_thr:
+        f += 1
[…]
