### 2026-04-22 d379be7
docs: update PROJECT_STATUS - split-architecture Phase 3 in progress
