### 2026-04-23 0bcb25a
fix #121: Rg overlay visibility - larger marker (s=12), viridis cmap, optional rg_score_threshold/rg_marker_size/rg_cmap kwargs
