### 2026-04-23 b81aea3
fix #120: plot_compact independent y-axes by default - add align_zero kwarg (default False); skip align_zero_y so each twinx trace autoscales and fills its own panel; pass align_zero=True for legacy shared-zero behavior
