### 2026-04-23 1f4984d
fix #122: filter Rg=0.0 as NaN in construct_rgcurve_from_list - SimpleGuinier returns Rg=0.0 for degenerate frames; only None was guarded before; 0.0 is physically impossible and corrupts the Guinier penalty in BasicOptimizer
