### 2026-04-23 d32bd5b
fix(RigorousImplement): correct mon_opt coordinate system for MplMonitor display
After create_optimizer_from_job + apply_x_shifts, xr_curve.x and uv_curve.x
are shifted to original frame numbers, but their splines are still built on
the 0-based local index domain ([0..759], [0..939]).  rg_curve.segments are
also in 0-based indices and are never shifted.

This mismatch caused three dashboard artefacts (biosaxs-dev/molass-legacy#21):
- Rg dots appeared ~550 frames too far left (0-based [277..457] vs correct
  [892..1004])
- XR panel xlim auto-extended left to accommodate the shifted dots, making
[…]
