### 2026-04-23 c67f2fc
feat(RunInfo): add monitor_snapshot_json_path and load_monitor_snapshot()
Adds two members to RunInfo for zero-guesswork access to the MplMonitor
JSON sidecar (written when MOLASS_MONITOR_SNAPSHOT=1):

  run_info.monitor_snapshot_json_path  -> str | None
      Returns the absolute path to <analysis_folder>/optimized/figs/
      mplmonitor_latest.json without requiring the caller to know the
      internal folder layout.  Returns None if analysis_folder is unset.

  run_info.load_monitor_snapshot()  -> dict
      Loads and returns the parsed JSON.  Raises FileNotFoundError with a
      clear me
[…]
