### 2026-04-23 fdf33fa
feat: add RunInfo.check_progress() delegating to CurrentStateUtils.check_progress()
