### 2026-04-23 f247018
feat: re-export CurrentStateUtils public API from molass.Rigorous; add See Also to RunInfo.check_progress()
