### 2026-04-23 68d693b
docs: record issue #123 (check_progress) in API_IMPROVEMENTS.md
