### 2026-04-23 b0f9f5a
test: add check_progress tests to test_plot_convergence.py (issue #123 follow-up)
