### 2026-04-23 22850bb
feat: add write_snapshot to check_progress; add RunInfo.load_progress_snapshot() and .progress_snapshot_json_path
