### 2026-04-23 095cd5e
docs: record issue #124 in API_IMPROVEMENTS.md
