### 2026-04-23 97d927f
refactor: add work_folder to RunInfo.__init__; add class docstring; remove monkey-patches
