### 2026-04-23 229ade3
docs: record issue #125 in API_IMPROVEMENTS.md
