### 2026-04-23 f833a1f
docs: update PROJECT_STATUS — April 23 RunInfo AI-friendliness sweep (#123-#125)
