### 2026-04-24 3aeee5e
v0.9.3 - close issues #126/#127/#128; RunInfo refactor and docs updates
