### 2026-04-26 9d77a5e
fix: revert Fix #21 — mon_opt curve patch broke #118 alignment (rg_curve used in objective_func, April 2026)
