### 2026-04-26 4097790
fix(#129): patch only mon_opt curve .x arrays for display, preserve #118 alignment
