### 2026-04-26 b2ea136
fix(#129): use parent optimizer for display rendering; no curve-state patching
