### 2026-04-26 7e90aea
docs(#117): fill Phase 0 baseline table with Apo 2c numbers
