### 2026-04-26 c482e28
docs: add Phase 2 step 3 (default progress UX patterns) to split-architecture design
