### 2026-04-26 2b2fa93
feat: add compare_optimization_paths() (Phase 2.5 of split-architecture)
