+++ molass/Rigorous/ComparePaths.py
@@ -1,0 +1,462 @@
+"""
+    Rigorous.ComparePaths.py
+
+    Side-by-side validation of `optimize_rigorously()` execution paths
+    (subprocess vs in-process), introduced as Phase 2.5 of the
+    split-architecture plan.  See
+    ``molass-library/Copilot/DESIGN_split_optimizer_architecture.md``.
+
+    Single-cell entry point for the `13h_split_architecture_validation`
+    workflow:
+
+        from molass.Rigorous import compare_optimization_paths
+
+        result = compare_optimization_paths(
+            decomp, rgcurve,
+            method='NS', niter=20,
+            paths=('subprocess', 'in_process'),
+            trimmed_ssd=trimmed,
+        )
+        result.summary_table()
+        result.plot_convergence()
+        result.plot_components()
+        result.assert_parity(fv_rtol=1e-2)
+
+    Copyright (c) 2026, SAXS Team, KEK-PF
+"""
+import os
+import time
+
+import numpy as np
+
+
+_PATH_ALIASES = {
+    'subprocess': False,
+    'subproc':    False,
+    'sub':        False,
+    'in_process': True,
+    'inprocess':  True,
+    'inproc':     True,
+    'inp':        True,
+}
+
+
+def _resolve_path(name):
+    """Map a path name to the ``in_process`` boolean."""
+    try:
+        return _PATH_ALIASES[name]
+    except KeyError as e:
+        raise ValueError(
+            "Unknown path %r; expected one of %s"
+            % (name, sorted(_PATH_ALIASES))
+        ) from e
+
+
+class PathResult:
+    """Result of a single ``optimize_rigorously`` call inside a comparison.
+
+    Attributes
+    ----------
+    label : str
+        Path label (``'subprocess'`` or ``'in_process'``).
+    in_process : bool
+        Resolved ``in_process`` flag passed to ``optimize_rigorously``.
+    analysis_folder : str
+        Absolute path to the analysis folder used by this run.
+    run_info : RunInfo
+        The ``RunInfo`` returned by ``optimize_rigorously``.
+    conv : object or None
+        Convergence info returned by ``Decomposition.plot_convergence``;
+        ``None`` if results never materialised within ``timeout``.
+    best : Decomposition or None
+        Best-result decomposition from ``load_best_rigorous_result``;
+        ``None`` if no result.
+    rgs : list[float] or None
+        Component Rg values from ``best.get_rgs()``; ``None`` if no result.
+    elapsed : float
+        Wall-clock seconds for the ``optimize_rigorously`` call.
+    completed : bool
+        ``True`` iff ``conv`` and ``best`` are both populated.
+    """
+
+    __slots__ = ('label', 'in_process', 'analysis_folder', 'run_info',
+                 'conv', 'best', 'rgs', 'elapsed', 'completed')
+
+    def __init__(self, label, in_process, analysis_folder, run_info,
+                 conv, best, rgs, elapsed, completed):
+        self.label = label
+        self.in_process = in_process
+        self.analysis_folder = analysis_folder
+        self.run_info = run_info
+        self.conv = conv
+        self.best = best
+        self.rgs = rgs
+        self.elapsed = elapsed
+        self.completed = completed
+
+    @property
+    def best_fv(self):
+        return None if self.conv is None else self.conv.best_fv
+
+    @property
+    def best_sv(self):
+        return None if self.conv is None else self.conv.best_sv
+
+
+class ComparisonResult:
+    """Result of running ``optimize_rigorously`` along multiple paths.
+
+    Returned by :func:`compare_optimization_paths`.
+
+    Attributes
+    ----------
+    results : list[PathResult]
+        One entry per path, in the order requested.
+    niter : int
[…]
