### 2026-04-26 b4377ab
feat(rigorous): add monitor=False to bypass MplMonitor for batch/comparison runs (#117)
