### 2026-04-26 14edf7b
feat(Rigorous): add RunRegistry for out-of-band run discovery
Drops RUN_MANIFEST.json into each analysis_folder and work_folder at run start, and appends to ~/.molass/runs.jsonl. Enables external observers (other VS Code sessions, AI assistants, shells) to find live rigorous optimization runs without parsing process trees.

New module: molass/Rigorous/RunRegistry.py

  - write_run_manifest / update_run_manifest / read_manifest

  - locate_recent_runs(since_minutes=60)

Wired into RigorousImplement.py at three points: after prepare_rigorous_folders (status=starting), after BackRunner.run (statu
[…]
