### 2026-04-26 6a83a72
docs(DESIGN): add Phase 3 split-architecture validation results
Apo 2c, NS, niter=20 (2026-04-26 via 13h notebook). assert_parity passes: |dfv|/|fv|=1.5%, |dSV|=0.60, max dRg=0.31 A. In-process is 13% slower (UltraNest/GIL, Phase 5+).
