### 2026-04-26 798761e
feat(Rigorous): add RunInfo.live_status() (closes #133)
One-call snapshot of where a rigorous run stands right now. Combines the scattered probe pattern (sv_history + check_progress + RunRegistry.read_manifest + subprocess_returncode + os.walk fallback for work_folder) into a single read-only dict.

Returns: phase ('pending'/'running'/'completed'/'failed'/'unknown'), n_evals, best_fv, best_sv, elapsed_s, analysis_folder, work_folder, subprocess_pid, subprocess_returncode, manifest.

Composes with ai-context-vscode#1 (aicKernelEval): external observers can now do aicKernelEval('run_sub.live_sta
[…]
