### 2026-04-26 8c4fede
feat(Rigorous): add ComparisonResult.live_status() convenience wrapper
compare_optimization_paths() consumes the underlying RunInfo objects, so callers had no obvious way to get the new live_status() dicts without spelling out result.by_label('subprocess').run_info.live_status(). Add a method that returns either a single dict (label= given) or {label: dict} for all paths.
