### 2026-04-26 18f0439
docs: surface live_status / RunRegistry / aicKernelEval in copilot-instructions
Adds 'Live run observability stack' subsection under section 9 listing the canonical one-call probe (RunInfo.live_status / ComparisonResult.live_status / aicKernelEval). Logs issues #131-133 and the RunRegistry/aicKernelEval infrastructure in API_IMPROVEMENTS.md.
