### 2026-04-26 fe7ee32
docs: update PROJECT_STATUS — Phase 3 validated, next task #132
