### 2026-04-27 563e080
feat: Phase 4 — flip in_process default to True, close #117 #119 (v0.9.4)
