+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "molass"
-version = "0.9.3"
+version = "0.9.4"
 
 dependencies = [
   "numpy",
