### 2026-04-27 05e418a
docs: mark Phase 4 complete in DESIGN doc
