### 2026-04-27 3b96874
fix: Decomposition.optimize_rigorously default in_process=True (matches make_rigorous_decomposition_impl)
