### 2026-04-27 f232f63
feat: optimize_rigorously(async_=True) + RunInfo.is_alive (issue #137)
