### 2026-04-27 61bc84c
Phase 4: add progress='dashboard' to optimize_rigorously() (ml#139)
