### 2026-04-27 e96da26
fix: pass niter to MplMonitor.for_run_info() for in-process dashboard (ml#139)
