### 2026-04-27 1766746
docs: update PROJECT_STATUS - Phase 5 dashboard fixes and kernel hang root cause
