### 2026-04-27 3490c91
fix: pass work_folder_callback to run_optimizer_in_process (#132)
Wire _on_folder_ready callback so run_info.work_folder is set as soon
as the job folder is allocated, not after the full run completes.
This enables MplMonitor watch thread to initialize job_state early and
render the upper convergence-plot panel during the run.
