### 2026-04-27 5f51671
fix: pass function_code to MplMonitor.for_run_info (#132 followup)
Without this, self.func_code was None on the monitor and draw_suptitle()
raised KeyError: None in get_model_name(). Now the dashboard title shows
the actual model name (e.g. EGH, SDM(gamma), etc.).
