### 2026-04-27 ed998b4
fix: clear jobs subfolders before async thread starts (clear_jobs=True default)
When async_=True, _allocate_work_folder() runs on the daemon thread and picks
the next non-empty slot (jobs/001, /002, ...) from prior runs.  The MplMonitor
for_run_info path (clear_jobs=False default) never cleared these because by the
time for_run_info is called, the thread has already allocated a folder.

Fix: clear all subdirectories of optimized/jobs/ BEFORE _thread.start() when
clear_jobs=True (the default).  This ensures every fresh run starts at jobs/000
and the watcher can locate callback.txt without knowi
[…]
