+++ molass/Rigorous/RigorousImplement.py
@@ -333,6 +333,20 @@ def make_rigorous_decomposition_impl(decomposition, rgcurve, analysis_folder=Non
 
         if async_:
             import threading
+
+            # Clear the jobs subfolder BEFORE starting the thread so that
+            # _allocate_work_folder() picks jobs/000 instead of accumulating
+            # jobs/001, /002, ... across repeated runs. (#132 followup)
+            if clear_jobs:
+                import shutil as _shutil
+                _jobs_dir = os.path.join(analysis_folder, "optimized", "jobs")
+                if os.path.isdir(_jobs_dir):
+                    for _entry in os.listdir(_jobs_dir):
+                        _ep = os.path.join(_jobs_dir, _entry)
+                        if os.path.isdir(_ep):
+                            _shutil.rmtree(_ep)
+                os.makedirs(_jobs_dir, exist_ok=True)
+
             _thread = threading.Thread(target=_run_in_process, daemon=True)
             run_info._async_thread = _thread
             _thread.start()
