### 2026-04-28 46fe7ec
AI-friendliness: RunInfo.diagnose() + copilot-instructions score-diagnosis (#145)
- molass/Rigorous/RunInfo.py: add diagnose(breakdown=None) method returning
  list of Diagnosis(score, status, reason, suggestion) namedtuples
- tests/specific/200_Rigorous/test_060_diagnose.py: 8 tests (all passing)
- Copilot/copilot-guidelines.md: Rule 12 AI-friendliness classes table
- .github/copilot-instructions.md: Score diagnosis paragraph in Convention 9
