+++ molass/Rigorous/RunInfo.py
@@ -35,7 +35,7 @@ class RunInfo:
     """
 
     def __init__(self, ssd, optimizer, dsets, init_params, monitor=None,
-                 analysis_folder=None, decomposition=None):
+                 analysis_folder=None, decomposition=None, rgcurve=None):
         self.ssd = ssd
         self.optimizer = optimizer
         self.dsets = dsets
@@ -43,6 +43,7 @@ class RunInfo:
         self.monitor = monitor
         self.analysis_folder = analysis_folder
         self.decomposition = decomposition
+        self.rgcurve = rgcurve          # cached to avoid recomputation in load_best/load_first
         # Set by RigorousImplement for in_process=True runs:
         self.work_folder = None
         self.in_process_result = None
@@ -62,6 +63,27 @@ class RunInfo:
         if t is None:
             return False
         return t.is_alive()
+
+    def __repr__(self):
+        if self._async_thread is not None:
+            state = "running" if self._async_thread.is_alive() else "done"
+        else:
+            state = "done"
+        parts = [f"state={state!r}"]
+        if self.analysis_folder is not None:
+            try:
+                from molass.Rigorous.CurrentStateUtils import list_rigorous_jobs
+                jobs = list_rigorous_jobs(self.analysis_folder)
+                if jobs:
+                    import math
+                    best_fv = min(j.best_fv for j in jobs)
+                    best_sv = -200 / (1 + math.exp(-1.5 * best_fv)) + 100
+                    n_evals = sum(j.iterations for j in jobs)
+                    parts.append(f"n_evals={n_evals}")
+                    parts.append(f"best_sv={best_sv:.1f}")
+            except Exception:
+                pass
+        return f"RunInfo({', '.join(parts)})"
 
     def get_current_decomposition(self, **kwargs):
         debug = kwargs.get('debug', False)
@@ -132,6 +154,11 @@ class RunInfo:
         poll_interval : float, optional
             Seconds between checks (default 5).
 
+            .. note::
+                For async in-process runs (``async_=True``), ``poll_interval``
+                is **silently ignored** — the implementation calls
+                ``thread.join(timeout)`` once with no looping.
+
         Returns
         -------
         bool
@@ -142,6 +169,27 @@ class RunInfo:
         ValueError
             If no ``analysis_folder`` was stored (e.g. RunInfo was created
             without one).
+
+        .. warning:: **Async in-process path** (``async_=True``)
+
+            This method calls ``_async_thread.join(timeout)`` once, which has
+            two failure modes for long BH/NS runs:
+
+            - **Default** ``timeout=600``: silently returns ``False`` for runs
+              longer than 10 minutes while the optimizer is still running.
+              Downstream calls to ``load_best()`` will then fail or load stale
+              results.
+            - ``timeout=0`` **(no limit)**: blocks the kernel's main thread
+              entirely.  No other cell — including ``live_status()``,
+              ``is_alive``, or any monitoring probe — can execute until the
+              optimizer finishes.
+
+            For interactive notebooks prefer :meth:`load_first`, which waits
+            only until the *first* result lands on disk and returns
+            immediately, or poll manually with::
+
+                if run_info.is_alive:
+                    print(run_info.live_status())
         """
         # For async in-process runs, join the background thread directly.
         if self._async_thread is not None:
@@ -220,8 +268,74 @@ class RunInfo:
                 "Cannot reconstruct result without the initial decomposition."
             )
         return load_rigorous_result(
-            decomp, self.analysis_folder, jobid=best.id, debug=debug
+            decomp, self.analysis_folder, jobid=best.id,
+            rgcurve=self.rgcurve, debug=debug
         )
+
+    def load_first(self, timeout=0, poll_interval=5, debug=False):
+        """Wait for the first rigorous result, then load the best job found so far.
+
[…]
