### 2026-04-29 1b5818d
fix: export uv_diff_spline for subprocess parity; RunInfo request_stop/compare_dsets; molass_legacy>=0.6.0; update PROJECT_STATUS
