+++ pyproject.toml
@@ -12,7 +12,7 @@ dependencies = [
   "matplotlib",
   "tqdm",
   "pybaselines>=1.2.0",
-  "molass_legacy>=0.5.6",
+  "molass_legacy>=0.6.0",
   "statsmodels",
   "scikit-learn",
   "toml",

[…]
