### 2026-04-30 bc17700
AI-friendliness: analysis_folder guard (#153), optimize_rigorously Notes docstring (#154), wait() UserWarning (#155); idempotency guard + max_trials in RigorousImplement; update copilot-instructions
