+++ molass/Rigorous/RunInfo.py
@@ -1,6 +1,14 @@
 """
 Rigorous.RunInfo.py
 """
+import weakref
+
+# Module-level weak reference to the currently-running in-process RunInfo.
+# Set when an async in-process run starts; cleared (to None) when it
+# finishes or when the kernel is restarted (module is reloaded → None).
+# Used by the idempotency guard in make_rigorous_decomposition_impl.
+_active_inprocess = None
+
 
 class RunInfo:
     """Handle returned by :func:`molass.Rigorous.make_rigorous_decomposition`
@@ -78,6 +86,65 @@ class RunInfo:
         if t is None:
             return False
         return t.is_alive()
+
+    @property
+    def current_work_folder(self):
+        """The currently-active (or most-recently-used) job folder.
+
+        Unlike :attr:`work_folder` — which is reset to ``None`` at the start
+        of each auto-resume trial — this property always returns the best
+        available answer:
+
+        1. ``self.work_folder`` if already set.
+        2. The most recently modified ``jobs/NNN`` sub-folder under
+           ``analysis_folder/optimized/jobs/`` (the highest ``NNN`` that
+           exists on disk).
+        3. ``None`` if neither is available.
+
+        This is the right folder to probe with ``aicKernelEval`` during a
+        multi-trial run (``max_trials > 0``).  Fixes issue #150.
+        """
+        import os
+        if self.work_folder is not None:
+            return self.work_folder
+        af = self.analysis_folder
+        if af is None:
+            return None
+        jobs_dir = os.path.join(af, 'optimized', 'jobs')
+        if not os.path.isdir(jobs_dir):
+            return None
+        try:
+            subdirs = sorted(
+                d for d in os.listdir(jobs_dir)
+                if os.path.isdir(os.path.join(jobs_dir, d))
+            )
+            if subdirs:
+                return os.path.join(jobs_dir, subdirs[-1])
+        except OSError:
+            pass
+        return None
+
+    def update_manifest_on_resume(self, work_folder):
+        """Update ``RUN_MANIFEST.json`` when a new auto-resume trial starts.
+
+        Called from ``MplMonitor._RunInfoSource._wf_callback`` as soon as
+        the new job folder is allocated, so that :meth:`live_status` reads
+        the correct ``work_folder`` and ``phase='running'`` for the new
+        trial.  Fixes issue #148.
+
+        Parameters
+        ----------
+        work_folder : str
+            Absolute path to the newly-allocated job folder.
+        """
+        af = self.analysis_folder
+        if af is None:
+            return
+        try:
+            from molass.Rigorous.RunRegistry import update_run_manifest
+            update_run_manifest(af, status='running', work_folder=work_folder)
+        except Exception:
+            pass
 
     def __repr__(self):
         if self._async_thread is not None:
@@ -208,6 +275,16 @@ class RunInfo:
         """
         # For async in-process runs, join the background thread directly.
         if self._async_thread is not None:
+            import warnings
+            warnings.warn(
+                "run_info.wait() on an async in-process run blocks the kernel "
+                "until the entire BH/NS run finishes (potentially hours).  "
+                "Use run_info.load_first() instead: it returns as soon as the "
+                "first result lands on disk, while the optimizer continues in "
+                "the background.",
+                UserWarning,
+                stacklevel=2,
+            )
             self._async_thread.join(timeout=timeout if timeout else None)
             if self._async_error is not None:
                 raise RuntimeError("Async optimizer failed") from self._async_error
@@ -1079,16 +1156,15 @@ class RunInfo:
             except Exception:
                 manifest = None
 
-        # If work_folder isn't set on the RunInfo, try the manifest, then
-        # fall back to a quick filesystem scan.  This is the same fallback
[…]
