### 2026-04-30 76c2985
test: RunInfo.reconnect() tests (issue #151); add DESIGN_terminate_inprocess.md
