+++ Copilot/DESIGN_terminate_inprocess.md
@@ -1,0 +1,268 @@
+# Design Discussion: Terminate Job for In-Process Runs
+
+**Status**: Draft — open for discussion  
+**Related**: molass-library#139 (Phase 2, where Terminate was hidden)  
+**Date**: April 28, 2026
+
+---
+
+## Current state
+
+The Terminate Job button is hidden when `MplMonitor` is driven by a `_RunInfoSource`
+(i.e. `optimize_rigorously(in_process=True, async_=True)`).  The reason given at the
+time of Phase 2 was:
+
+> "Resume and Terminate are not meaningful for thread-based runs:  
+> niter is fixed and threads cannot be killed cleanly."
+
+`_RunInfoSource.terminate()` is therefore a no-op:
+
+```python
+# molass_legacy/Optimizer/MplMonitor.py
+class _RunInfoSource:
+    def terminate(self):
+        """No-op: cannot cleanly kill a Python thread."""
+        pass
+```
+
+The `terminate_event` on `MplMonitor` is already checked in the watch loop and
+wired to `source.terminate()` — so the infrastructure is ready.  The missing piece
+is a live signal path from `terminate_event` → the optimizer thread.
+
+---
+
+## Why it matters
+
+A BH run with `niter=200` at ~2–3 min/step takes 6–10 hours.  If the user decides
+after 30 minutes that the result is already good enough (e.g. `best_sv > 80`), there
+is no way to stop the optimizer short of **restarting the kernel** — which destroys
+the live `RunInfo` object and any in-memory state.
+
+The dashboard shows a "Completed" status only when all `niter` steps finish.  No
+early-exit mechanism exists.
+
+---
+
+## What already exists
+
+| Component | Where | Role |
+|-----------|-------|------|
+| `terminate_event` | `MplMonitor` | `threading.Event`; set by Terminate button / `monitor.terminate()` |
+| Watch-loop check | `MplMonitor.watch_progress` L770 | calls `self.source.terminate()` when event is set |
+| GIL-release loop | `InProcessRunner.py` / molass-legacy#26 | main thread polls `thread.join(0.05)` → Ctrl+C lands cleanly |
+| BH callback param | `scipy.optimize.basinhopping` | accepts `callback(x, f, accepted)` → stops if returns `True` |
+
+---
+
+## The signal path that needs to be wired
+
+```
+User clicks "Terminate Job"
+    → MplMonitor.trigger_terminate()
+        → terminate_event.set()
+            → watch thread calls _RunInfoSource.terminate()
+                → ??? needs to reach optimizer.solve()
+                    → BH callback checks event → returns True → BH stops
+```
+
+Currently the `???` link is broken: `_RunInfoSource.terminate()` is a no-op.
+
+---
+
+## Design options
+
+### Option A — Cooperative flag via `terminate_event` passed to solver
+
+**Mechanism**:
+
+1. Store `terminate_event` on `RunInfo` (set by the monitor after construction, or
+   passed in at construction time).
+2. In `InProcessRunner.run_optimizer_in_process()`, pass a BH callback that checks it:
+
+```python
+# In InProcessRunner.py — BH callback
+def _make_terminate_callback(event):
+    def callback(x, f, accepted):
+        if event is not None and event.is_set():
+            return True   # signals BH to stop
+        return False
+    return callback
+```
+
+3. Pass this to `optimizer.solve()` → into `scipy.optimize.basinhopping(callback=...)`.
+4. Update `_RunInfoSource.terminate()` to call `run_info.terminate_event.set()`.
+
+**Pros**:
+- Clean and safe — BH exits after the current step completes, no mid-step abort.
+- No thread killing. No `ctypes`. No segfault risk.
[…]
