### 2026-04-30 b1ae2d6
fix: abspath analysis_folder before async thread starts (#156)
FileNotFoundError in MplMonitor when optimize_rigorously() is called
with a relative analysis_folder and async_=True / progress='dashboard'.

Root cause: the async optimizer thread calls os.chdir(work_folder),
changing the process-wide CWD to jobs/000/.  MplMonitor.__init__ then
reads the still-relative path from SerialSettings and opens
FileHandler(logpath), resolving against the new CWD -> doubled path.

Fix (two layers):
- RigorousImplement.py: os.path.abspath at entry of
  make_rigorous_decomposition_impl, before RunInfo / _job
[…]
