### 2026-04-30 fd6fd59
AI-friendliness: attach result.sv and result.fv in load_best() (#157)
After run_info.load_first() / load_best(), callers now get result.sv
and result.fv without having to re-parse callback.txt separately.

  result = run_info.load_first()
  print(f'SV: {result.sv:.1f}')   # now works
