### 2026-05-01 418a01c
refactor: remove load_first, fold wait logic into load_best (issue #158)
