+++ molass/Rigorous/RunInfo.py
@@ -51,7 +51,7 @@ class RunInfo:
         self.monitor = monitor
         self.analysis_folder = analysis_folder
         self.decomposition = decomposition
-        self.rgcurve = rgcurve          # cached to avoid recomputation in load_best/load_first
+        self.rgcurve = rgcurve          # cached to avoid recomputation in load_best
         # Set by RigorousImplement for in_process=True runs:
         self.work_folder = None
         self.in_process_result = None
@@ -266,8 +266,8 @@ class RunInfo:
               ``is_alive``, or any monitoring probe — can execute until the
               optimizer finishes.
 
-            For interactive notebooks prefer :meth:`load_first`, which waits
-            only until the *first* result lands on disk and returns
+            For interactive notebooks prefer :meth:`load_best`, which waits
+            only until the first result lands on disk and returns
             immediately, or poll manually with::
 
                 if run_info.is_alive:
@@ -279,7 +279,7 @@ class RunInfo:
             warnings.warn(
                 "run_info.wait() on an async in-process run blocks the kernel "
                 "until the entire BH/NS run finishes (potentially hours).  "
-                "Use run_info.load_first() instead: it returns as soon as the "
+                "Use run_info.load_best() instead: it returns as soon as the "
                 "first result lands on disk, while the optimizer continues in "
                 "the background.",
                 UserWarning,
@@ -300,29 +300,46 @@ class RunInfo:
             self.analysis_folder, timeout=timeout, poll_interval=poll_interval
         )
 
-    def load_best(self, debug=False):
-        """Load the best rigorous optimization result.
+    def load_best(self, timeout=0, poll_interval=5, debug=False):
+        """Load the best rigorous optimization result, waiting until one is available.
 
-        Combines ``list_rigorous_jobs()`` and ``load_rigorous_result()``
-        into one call: finds the job with the lowest objective function
-        value and reconstructs the ``Decomposition`` from it.
+        Safe to call immediately after :meth:`optimize_rigorously` — it blocks
+        until at least one job has completed, then returns the result with the
+        lowest objective function value found so far.  Re-running this cell while
+        the optimizer is still going always returns the best result available at
+        that moment.
 
         Parameters
         ----------
+        timeout : float, optional
+            Maximum seconds to wait for the first result (default ``0`` = no
+            limit).  When non-zero and no result appears within ``timeout``
+            seconds, raises :class:`TimeoutError`.
+        poll_interval : float, optional
+            Seconds between filesystem checks (default 5).
         debug : bool, optional
             If True, reload modules from disk.
 
         Returns
         -------
         Decomposition
-            A Decomposition built from the best optimized parameters.
+            A Decomposition built from the best optimized parameters available
+            at the time the call returns.  The ``result.sv`` and ``result.fv``
+            attributes are attached for convenience.
 
         Raises
         ------
         ValueError
             If no ``analysis_folder`` was stored.
-        FileNotFoundError
-            If no completed jobs are found.
+        TimeoutError
+            If ``timeout > 0`` and no result appears within that time.
+
+        Notes
+        -----
+        Interrupt with **Ctrl+C** to cancel the wait at any time.
+
+        If the optimizer is still running, the result reflects only the jobs
+        completed so far.  Re-call after the run finishes to get the final best.
 
         See Also
         --------
@@ -334,9 +351,8 @@ class RunInfo:
         ::
 
             run_info = decomp.optimize_rigorously(
-                analysis_folder="temp_analysis", niter=30)
-            run_info.wait(timeout=600)
-            result = run_info.load_best()
[…]
