### 2026-05-01 bdb8697
fix: auto-degrade progress='dashboard' when in_process=False (molass-library#159)
