+++ tests/specific/200_Rigorous/test_050_progress_dashboard.py
@@ -1,8 +1,11 @@
-"""
-Tests for the progress='dashboard' kwarg added in Phase 4 (molass-library#139).
+"""Tests for the progress kwarg in optimize_rigorously.
 
-These tests verify the validation guards only — the live MplMonitor dashboard
-cannot be meaningfully tested in a headless environment.
+molass-library#139: progress='dashboard' kwarg added in Phase 4.
+molass-library#159: progress is now deprecated/internal; invalid combinations
+auto-degrade silently instead of raising.  Tests updated accordingly.
+
+The live MplMonitor dashboard cannot be meaningfully tested in a headless
+environment.
 """
 import pytest
 
@@ -19,28 +22,38 @@ def test_unknown_progress_value_raises():
         )
 
 
-def test_dashboard_without_in_process_raises():
-    """progress='dashboard' requires in_process=True."""
+def test_dashboard_without_in_process_degrades_silently():
+    """progress='dashboard' with in_process=False auto-degrades; no ValueError.
+
+    molass-library#159: users should not need to pass progress=None explicitly.
+    The function must get past the progress guard (failing later on
+    decomposition=None), not at the progress check.
+    """
     from molass.Rigorous.RigorousImplement import make_rigorous_decomposition_impl
 
-    with pytest.raises(ValueError, match="requires in_process=True and async_=True"):
+    with pytest.raises(Exception) as exc_info:
         make_rigorous_decomposition_impl(
             decomposition=None, rgcurve=None,
             in_process=False, async_=True,
             progress='dashboard',
         )
+    assert "progress" not in str(exc_info.value).lower()
 
 
-def test_dashboard_without_async_raises():
-    """progress='dashboard' requires async_=True."""
+def test_dashboard_without_async_degrades_silently():
+    """progress='dashboard' with async_=False auto-degrades; no ValueError.
+
+    molass-library#159: users should not need to pass progress=None explicitly.
+    """
     from molass.Rigorous.RigorousImplement import make_rigorous_decomposition_impl
 
-    with pytest.raises(ValueError, match="requires in_process=True and async_=True"):
+    with pytest.raises(Exception) as exc_info:
         make_rigorous_decomposition_impl(
             decomposition=None, rgcurve=None,
             in_process=True, async_=False,
             progress='dashboard',
         )
+    assert "progress" not in str(exc_info.value).lower()
 
 
 def test_none_progress_passes_validation():
@@ -62,13 +75,14 @@ def test_none_progress_passes_validation():
 
 
 def test_decomposition_progress_kwarg_forwarded():
-    """Decomposition.optimize_rigorously accepts progress= and forwards it.
+    """Decomposition.optimize_rigorously forwards progress= to the impl.
 
-    Validation fires before any data access, so passing progress='bad'
-    raises ValueError even when decomposition/rgcurve are None.
+    The 'Unknown progress=' ValueError is raised inside make_rigorous_decomposition_impl.
+    We need analysis_folder!= None to get past the Decomposition-level guard first.
     """
     from molass.LowRank.Decomposition import Decomposition
 
     decomp = object.__new__(Decomposition)
     with pytest.raises(ValueError, match="Unknown progress="):
-        decomp.optimize_rigorously(progress='bad', in_process=True, async_=True)
+        decomp.optimize_rigorously(progress='bad', in_process=True, async_=True,
+                                   analysis_folder='dummy')
