### 2026-05-01 371ce69
docs: record issue #159 as done in API_IMPROVEMENTS.md
