### 2026-05-01 afa267b
fix: auto-degrade progress='dashboard' in Decomposition.py wrapper too (molass-library#159)
