### 2026-05-07 1b4bf56
fix: export in-process ElCurve y-values for subprocess (molass-legacy#38)
