### 2026-05-07 d1effbb
fix: export ip_xr_D.npy and ip_uv_U.npy for subprocess data matrix sync (molass-legacy#39)
