### 2026-05-07 d544823
fix: export ip_xr_E.npy for subprocess weight matrix sync (molass-legacy#40)
