### 2026-05-07 c9b6fc2
fix: export ip_xr_qvector.npy for subprocess override (molass-legacy#41)
