### 2026-05-07 62b7289
docs: update PROJECT_STATUS and copilot-instructions — molass-legacy#38–#42 fixed
