### 2026-05-08 e2536af
feat: sv_history_per_job + parse_sv_history_per_job (molass-library#161); fix: molass-legacy#52 in docs
