### 2026-05-11 582b5af
fix: anomaly bands in dashboard (#166); docstring #164/#165; set_anomaly_mask docs
