+++ molass/LowRank/Decomposition.py
@@ -920,7 +920,7 @@ class Decomposition:
                             trimmed_ssd=None,
                             clear_jobs=True, function_code=None,
                             in_process=True, monitor=True, async_=True, progress='dashboard',
-                            max_trials=0, debug=False,
+                            max_trials=0, debug=False, _dry_run=False,
                             **kwargs):
         """
         Perform a rigorous decomposition.
@@ -1136,7 +1136,7 @@ class Decomposition:
         if rgcurve is None:
             rgcurve = self.ssd.xr.compute_rgcurve()
 
-        return make_rigorous_decomposition_impl(self, rgcurve, analysis_folder=analysis_folder, method=method, niter=niter, frozen_components=frozen_components, trimmed_ssd=trimmed_ssd, clear_jobs=clear_jobs, function_code=function_code, in_process=in_process, monitor=monitor, async_=async_, progress=progress, max_trials=max_trials, debug=debug)
+        return make_rigorous_decomposition_impl(self, rgcurve, analysis_folder=analysis_folder, method=method, niter=niter, frozen_components=frozen_components, trimmed_ssd=trimmed_ssd, clear_jobs=clear_jobs, function_code=function_code, in_process=in_process, monitor=monitor, async_=async_, progress=progress, max_trials=max_trials, debug=debug, _dry_run=_dry_run)
 
     def load_best_rigorous_result(self, analysis_folder, rgcurve=None, debug=False):
         """Load the best rigorous optimization result from disk.
