### 2026-05-12 2195a26
chore: bump molass_legacy requirement to >=0.6.1
