### 2026-05-12 9992446
feat: uv_wavelength for trimmed_copy, rgcurve caching, v0.9.6 (#167, #168)
- trimmed_copy(uv_wavelength=(wl_min, wl_max)): clip UV to wavelength range
  (#167); bisect-based index computation, None for either end uses dataset
  default; test_030_uv_wavelength added
- quick_decomposition(rgcurve=...): accept pre-computed Rg curve to avoid
  redundant per-frame Guinier fits in multi-decomposition workflows (#168)
- Decomposition.get_rg_curve(): cache result as self._rgcurve so repeated
  calls return instantly (#168)
- bump version 0.9.5 → 0.9.6
