### 2026-05-12 0e9f0d7
fix: clear_jobs=False auto-loads best params as init_params (#169)
