### 2026-05-13 e171658
fix: store original in_folder on monitor before prepare_rigorous_folders overwrites it
