### 2026-05-13 20ae81c
fix: adaptive scale lower bound in SdmOptimizer (issue #170)
Hardcoded lower bound of 1e-3 for scale parameters caused the optimizer
to get stuck at the lower-bound wall for datasets with small XR amplitudes
(e.g. MY dataset: initial_scales=[5.67e-6, 1.13e-7]).

Replace with: max(min(initial_scales) * 1e-2, 1e-15), ensuring the
initial estimate is always inside the feasible region.

Before: optimized scales_: [0.001, 0.001]  (stuck at lower bound)
After:  optimized scales_: [0.434, 0.449]  (optimization succeeds)
