### 2026-05-13 d290953
fix: SDM mono-pore optimizer — poresize bounds, estimator clamping, adjust_rg formula
- Per-component scale lower bounds: max(physics_scale*0.20, initial_scale*0.10)
- Poresize as free parameter with tight bounds [poresize_physics*0.9, *1.1] (±10%)
- Physics-based start: SEC-consistent Rg ordering at poresize=100
- Position anchor penalty (amplitude-independent): prevents component collapse
- Estimator poresize clamped to physics value when out of bounds (avoids logit ±∞)
- Fix adjust_rg_and_poresize: sum of squared differences (was: square of sum)
- SdmEstimator: t0_upper = min_peak_frame - 1
[…]
