+++ molass/SEC/Models/SdmOptimizer.py
@@ -42,6 +42,21 @@ def optimize_sdm_xr_decomposition(decomposition, env_params, model_params=None, 
     N, T, me, mp, N0, t0, poresize = env_params
     rgv = np.asarray(decomposition.get_rgs())
 
+    # Quality-adaptive Rg anchoring: uses Guinier fit quality at each component's
+    # peak frame to weight soft constraints in the objective and set hard-bound width.
+    rgcurve = kwargs.get('rgcurve')
+    rg_anchor_scale = model_params.get('rg_anchor_scale', 1.0) if model_params else 1.0
+    rg_quality_threshold = model_params.get('rg_quality_threshold', 0.7) if model_params else 0.7
+    if rgcurve is not None:
+        jv_rg = np.array(rgcurve.frames)
+        sc_rg = np.array(rgcurve.scores)
+        rg_qualities = np.array([
+            sc_rg[np.argmin(np.abs(jv_rg - int(c.x[c.y.argmax()])))]
+            for c in decomposition.xr_ccurves
+        ])
+    else:
+        rg_qualities = np.ones(num_components)
+
     if model_params is None:
         timescale = DEFAULT_TIMESCALE
         k_init = 2.0
@@ -71,13 +86,21 @@ def optimize_sdm_xr_decomposition(decomposition, env_params, model_params=None, 
             scales.append(scale)
         return scales
 
+    # EGH peak frames are used as a soft position constraint in the objective.
+    # Scale: ~1e-5 is enough to overcome the ~0.037 advantage of the degenerate solution
+    # (where both components collapse to the same frame) over the separated solution.
+    egh_peak_frames = np.array([c.x[c.y.argmax()] for c in decomposition.xr_ccurves], dtype=float)
+    position_anchor_scale = model_params.get('position_anchor_scale', 1e-5) if model_params else 1e-5
+
     def objective_function(params, return_cy_list=False, plot=False):
         N_, T_, x0_, tI_, N0_, k_ = params[0:6]
         rgv_ = params[6:6+num_components]
         rg_diff = np.diff(rgv_)
         non_ordered = np.where(rg_diff > 0)[0]
         order_penalty = np.sum(rg_diff[non_ordered]**2) * 1e3  # penalty for non-ordered rgv
-        rhov = rgv_/poresize
+        # poresize is now an optimization variable (last in params vector)
+        poresize_ = params[6+2*num_components]
+        rhov = rgv_/poresize_
         rhov[rhov > 1] = 1.0  # limit rhov to 1.0
         scales_ = params[6+num_components:6+2*num_components]
         cy_list = []
@@ -104,14 +127,73 @@ def optimize_sdm_xr_decomposition(decomposition, env_params, model_params=None, 
                 plt.plot(x, cy, label='Component %d' % (i+1))
             plt.legend()
             plt.show()
-        error = np.sum((y - ty)**2) + order_penalty
+        # Quality-weighted soft Rg anchoring: pulls each component's Rg toward
+        # its EGH-estimated value in proportion to the Guinier fit quality.
+        # quality ≈ 1 → strong pull (stays near EGH Rg); quality ≈ 0 → unconstrained.
+        rg_anchor_penalty = np.sum(rg_qualities * ((rgv_ / rgv - 1) ** 2)) * rg_anchor_scale
+        # Position constraint: use theoretical centroid (mean of gamma, = x0 + N*T*(1-rho)^6)
+        # which is amplitude-independent — prevents collapse to degenerate solution even
+        # when one component's scale → 0. Peak position is the priority per user guidance.
[…]
