### 2026-05-13 fe2e23b
fix: physics-based UV initial scale + debug output (#171, #172)
