+++ molass/SEC/Models/UvOptimizer.py
@@ -48,10 +48,35 @@ def optimize_uv_decomposition(decomposition, xr_ccurves, **kwargs):
     mapping = decomposition.mapping
     a, b = mapping.slope, mapping.intercept
 
-    initial_guess = [a, b] + [1.0]*num_components
+    # Physics-based initial scale: ratio of UV to XR amplitude at XR component peak.
+    # Avoids the fixed-scale=1.0 local minimum that arises when model tails create
+    # Hessian cross-coupling (observed: SDM minor component, MY dataset, May 2026).
+    initial_scales = []
+    for xr_ccurve in xr_ccurves:
+        peak_xr_idx = int(np.argmax(xr_ccurve.y))
+        peak_xr_val = float(xr_ccurve.y[peak_xr_idx])
+        peak_xr_frame = float(xr_ccurve.x[peak_xr_idx])
+        # mapping converts XR frame → UV frame (a*xr + b = uv)
+        peak_uv_frame = a * peak_xr_frame + b
+        uv_idx = int(np.argmin(np.abs(x - peak_uv_frame)))
+        uv_val = float(y[uv_idx])
+        s0 = uv_val / peak_xr_val if peak_xr_val > 0 else 1.0
+        initial_scales.append(float(np.clip(s0, 1e-3, 10.0)))
+
+    if debug:
+        scale_str = '[' + ', '.join(f'{s:.3g}' for s in initial_scales) + ']'
+        print(f'[UV] initial_guess: a={a:.4g}  b={b:.4g}  scales={scale_str}')
+
+    initial_guess = [a, b] + initial_scales
     dx = (x[-1] - x[0])*0.1
     bounds = [(a*0.8, a*1.2), (b-dx, b+dx)] + [(1e-3, 10.0) for _ in range(num_components)]
     result = minimize(objective_function, initial_guess, bounds=bounds)
+
+    if debug:
+        converged_scales = list(result.x[2:])
+        residual_rms = float(np.sqrt(result.fun / len(y))) if len(y) > 0 else float('nan')
+        scale_str = '[' + ', '.join(f'{s:.3g}' for s in converged_scales) + ']'
+        print(f'[UV] converged:     a={result.x[0]:.4g}  b={result.x[1]:.4g}  scales={scale_str}  residual_rms={residual_rms:.4g}')
 
     new_mapping = Mapping(*result.x[0:2])
     new_uv_ccurves = []
