### 2026-05-13 79f465e
AI-friendliness: rename optimize_with_model to upgrade (#173)
