### 2026-05-14 4e693fe
feat: add shared_e mode to EdmOptimizer; add test_020_edm_optimizer
EdmOptimizer.shared_e=True reparameterizes the optimization so all
components share a single interstitial-volume parameter e, reducing
the risk of unphysical per-component e values (e.g. e=1.00 in 18c).

tests/specific/300_SEC_Models/test_020_edm_optimizer.py: two passing
tests for e_bounds and shared_e mode.
