+++ molass/SEC/Models/EdmOptimizer.py
@@ -17,14 +17,31 @@ def optimize_edm_xr_decomposition(decomposition, init_params, **kwargs):
         The initial parameters for the EDM components.
     kwargs : dict
         Additional parameters for the optimization process.
-        
+
+        position_anchor_scale : float, optional
+            Scale factor for the soft position constraint that keeps each
+            EDM component centroid near its EGH peak frame.  Default 1e-5.
+            Increase to enforce tighter position anchoring.
+
+        e_bounds : tuple or None, optional
+            Lower and upper bounds for the porosity parameter ``e`` (index 4
+            of each component's 7-parameter vector).  Default ``(0.0, 1.0)``.
+            Pass ``None`` to disable bounds on ``e`` (not recommended).
+
+        shared_e : bool, optional
+            If True, treat ``e`` as a single shared column parameter across
+            all components instead of fitting it independently per component.
+            This enforces the physical constraint that the interstitial
+            porosity e = Vm/(Vm+Vs) is a column property, not a
+            molecule-size-dependent property.  K_SEC (= ``a``) then carries
+            all molecule-size information.  Default False.
+
     Returns
     -------
     new_xr_ccurves : list of EdmComponentCurve
         The optimized EDM component curves.
     """
 
-    # N, T, N0, t0, poresize
     debug = kwargs.get('debug', False)
     if debug:
         from importlib import reload
@@ -46,19 +63,105 @@ def optimize_edm_xr_decomposition(decomposition, init_params, **kwargs):
             plt.show()
         debug_plot_params(x, y, init_params, "optimize: before minimize")
 
+    # EGH peak frames: used as soft position anchors so that the free
+    # EDM optimization cannot collapse one component into the other.
+    egh_peak_frames = np.array(
+        [c.x[c.y.argmax()] for c in decomposition.xr_ccurves], dtype=float
+    )
+    position_anchor_scale = kwargs.get('position_anchor_scale', 1e-5)
+
     shape = init_params.shape
 
-    def objective(p):
+    def objective(p, return_cy_list=False):
         cy_list = []
         for params in p.reshape(shape):
             cy = edm_impl(x, *params)
             cy_list.append(cy)
-        y_ = np.sum(cy_list, axis=0)
-        return np.sum((y_ - y)**2)
+        if return_cy_list:
+            return cy_list
+        ty = np.sum(cy_list, axis=0)
+        data_error = np.sum((ty - y) ** 2)
+        # Soft position constraint: penalise deviation of each component's
+        # centroid from the corresponding EGH peak frame.  Using the centroid
+        # (amplitude-independent distribution mean) prevents collapse even
+        # when one component's amplitude approaches zero.
+        position_penalty = 0.0
+        for cy, egh_peak in zip(cy_list, egh_peak_frames):
+            cy_abs_sum = np.sum(np.abs(cy))
+            if cy_abs_sum > 0:
+                centroid = np.sum(cy * x) / cy_abs_sum
+                position_penalty += (centroid - egh_peak) ** 2
+        return data_error + position_penalty * position_anchor_scale
 
-    result = minimize(objective, init_params.flatten())
+    # E_IDX: index of the porosity parameter 'e' in the 7-param vector
+    # [t0, u, a, b, e, Dz, cinj]
+    E_IDX = 4
+    e_bounds = kwargs.get('e_bounds', (0.0, 1.0))
+    e_bounds_val = e_bounds if e_bounds is not None else (None, None)
+
[…]
