### 2026-05-15 0d03821
feat: CEDM fast upgrade + rigorous optimization; rename optimize_with_model to upgrade (#173, #179)
