+++ molass/SEC/Models/EdmOptimizer.py
@@ -36,6 +36,34 @@ def optimize_edm_xr_decomposition(decomposition, init_params, **kwargs):
             molecule-size-dependent property.  K_SEC (= ``a``) then carries
             all molecule-size information.  Default False.
 
+        shared_column : bool, optional
+            If True, enforce the full constrained-EDM parameterisation where
+            ``t0``, ``u``, ``e``, ``Dz`` are shared column parameters and
+            only ``a`` (K_SEC), ``b``, and ``cinj`` are per-component.
+
+            This is the recommended mode for K_SEC identification.  With
+            over-parameterised (free) EDM, the optimiser can absorb peak
+            position into ``t0``/``u`` and leave ``a`` near zero for all
+            components, making Spearman(Rg, a) meaningless.  With constrained
+            EDM, peak positions are explained solely by ``a``-differences,
+            yielding Spearman(Rg, a) = -1.000 for well-separated SEC data.
+
+            Parameter vector layout::
+
+                [t0_sh, u_sh, e_sh, Dz_sh,
+                 a_0, b_0, cinj_0,
+                 a_1, b_1, cinj_1, ...]
+
+            Default bounds applied when ``shared_column=True``:
+            - ``e_bounds`` → ``(0.2, 0.85)`` (physical SEC total-porosity range)
+            - ``cinj_min`` → ``0.05`` (prevents component collapse to zero)
+
+            These defaults can be overridden via:
+            - ``e_bounds`` kwarg (same meaning as the top-level kwarg)
+            - ``cinj_min`` kwarg
+
+            Default False.
+
     Returns
     -------
     new_xr_ccurves : list of EdmComponentCurve
@@ -140,6 +168,117 @@ def optimize_edm_xr_decomposition(decomposition, init_params, **kwargs):
         for other_params in other_fitted:
             full_params = np.insert(other_params, E_IDX, e_fitted)
             new_xr_ccurves.append(EdmComponentCurve(x, full_params))
+        return new_xr_ccurves
+
+    shared_column = kwargs.get('shared_column', False)
+
+    if shared_column:
+        # --- Constrained-EDM mode ---
+        # Parameter vector: [t0_sh, u_sh, e_sh, Dz_sh,  a_0,b_0,cinj_0, a_1,b_1,cinj_1, ...]
+        # Shared indices in the 7-param vector: T0=0, U=1, E=4, DZ=5
+        # Per-component indices: A=2, B=3, CINJ=6
+        L_EDM = 30.0  # column length hardcoded in edm_func / edm_impl
+        N_SHARED = 4  # t0, u, e, Dz
+        N_PER_COMP = 3  # a, b, cinj
+        n_comp = shape[0]
+
+        # Shared-column bounds; prefer caller-supplied e_bounds over the tight default.
+        sc_e_bounds = kwargs.get('e_bounds', (0.2, 0.85))
+        cinj_min = kwargs.get('cinj_min', 0.05)
+
+        # --- analytical init for shared params ---
+        t0_sh = float(init_params[:, 0].min())
+        # u_sh: choose so that the void time is ~10% shorter than the fastest peak delay
+        peak_delay = float(egh_peak_frames.min()) - t0_sh
+        if peak_delay <= 0:
+            peak_delay = 1.0  # guard
+        void_time = peak_delay / 1.1  # leave 10% head-room so all a_init > 0
+        u_sh = L_EDM / void_time
+        Dz_sh = float(init_params[:, 5].mean())  # DZ_IDX = 5
+
+        # b=0 is the linear-limit start; per-component free-EDM b values are
+        # unreliable (can produce b≪0, causing gam3 overflow).
[…]
