### 2026-05-15 c63c14e
fix: CEDM model name in ComponentUtils; EDM default shared_column=True with DeprecationWarning
