### 2026-05-15 9051dca
fix: add get_cedm_xr_ccurves; reconstruct full 7-param vector for CEDM result loading
