### 2026-05-19 3c85d82
feat: SDM lognormal improvements — FftUtils autoresize, LognormalPore stability, SdmOptimizer sigma bound; bump v0.9.7
