+++ molass/MathUtils/FftUtils.py
@@ -25,13 +25,32 @@ def compute_standard_wCD(N):
 class FftInvPdf:
     def __init__(self, cf):
         self.cf = cf
-        self.N = N = 1024
-        self.w, self.C, self.D = compute_standard_wCD(N)
+        self.default_N = 1024
+        self.w, self.C, self.D = compute_standard_wCD(self.default_N)
+        # Cache for the resized grid (lazily populated on first large-t call).
+        # t is constant across all calls for a given dataset, so N_large is
+        # computed only once regardless of how many PDF evaluations are made.
+        self._large_N = None
+        self._large_w = self._large_C = self._large_D = None
 
     def __call__(self, t, *params):
-        N = self.N
-        cft = self.cf(self.w[N//2:], *params)
+        N = self.default_N
+        t_max = t[-1]  # t is always a sorted frame array (x - t0)
+        if t_max >= N:
+            # Auto-resize to the next power of 2 that covers the query range.
+            # Without this, UnivariateSpline would extrapolate outside [0, N-1],
+            # producing garbage PDF values (e.g. for unscaled lognormal models
+            # with raw frame values ~[455, 1393]).  See issue #181.
+            N = int(2 ** np.ceil(np.log2(t_max + 2)))
+            if N != self._large_N:
+                self._large_w, self._large_C, self._large_D = compute_standard_wCD(N)
+                self._large_N = N
+            w, C, D = self._large_w, self._large_C, self._large_D
+        else:
+            w, C, D = self.w, self.C, self.D
+
+        cft = self.cf(w[N//2:], *params)
         cft = np.concatenate([cft[::-1].conj(), cft])
-        pdfFFT = np.max([np.zeros(N), (self.C*np.fft.fft(self.D*cft)).real], axis=0)
+        pdfFFT = np.max([np.zeros(N), (C*np.fft.fft(D*cft)).real], axis=0)
         spline = UnivariateSpline(np.arange(N), pdfFFT, s=0)
         return spline(t)
