### 2026-05-21 eac6ae8
molass-legacy#65: expose ns_narrow_bounds through optimize_rigorously()
