+++ molass/LowRank/Decomposition.py
@@ -935,6 +935,12 @@ class Decomposition:
                 reload(molass.Rigorous.RigorousCedmParams)
             from molass.Rigorous.RigorousCedmParams import make_rigorous_initparams_impl
             return make_rigorous_initparams_impl(self, baseparams, debug=debug)
+        elif self.model == 'lkm':
+            if debug:
+                import molass.Rigorous.RigorousLkmParams
+                reload(molass.Rigorous.RigorousLkmParams)
+            from molass.Rigorous.RigorousLkmParams import make_rigorous_initparams_impl
+            return make_rigorous_initparams_impl(self, baseparams, debug=debug)
         else:
             raise ValueError(f"Decomposition.make_rigorous_initparams: Unsupported model '{self.model}'")
 
@@ -944,6 +950,7 @@ class Decomposition:
                             clear_jobs=True, function_code=None,
                             in_process=True, monitor=True, async_=True, progress='dashboard',
                             max_trials=0, debug=False, _dry_run=False,
+                            ns_narrow_bounds=True,
                             **kwargs):
         """
         Perform a rigorous decomposition.
@@ -1163,7 +1170,7 @@ class Decomposition:
         if rgcurve is None:
             rgcurve = self.ssd.xr.compute_rgcurve()
 
-        return make_rigorous_decomposition_impl(self, rgcurve, analysis_folder=analysis_folder, method=method, niter=niter, frozen_components=frozen_components, trimmed_ssd=trimmed_ssd, clear_jobs=clear_jobs, function_code=function_code, in_process=in_process, monitor=monitor, async_=async_, progress=progress, max_trials=max_trials, debug=debug, _dry_run=_dry_run)
+        return make_rigorous_decomposition_impl(self, rgcurve, analysis_folder=analysis_folder, method=method, niter=niter, frozen_components=frozen_components, trimmed_ssd=trimmed_ssd, clear_jobs=clear_jobs, function_code=function_code, in_process=in_process, monitor=monitor, async_=async_, progress=progress, max_trials=max_trials, debug=debug, _dry_run=_dry_run, ns_narrow_bounds=ns_narrow_bounds)
 
     def load_best_rigorous_result(self, analysis_folder, rgcurve=None, debug=False):
         """Load the best rigorous optimization result from disk.
