### 2026-05-21 0d48b3b
WIP: LKM model (LKM.py, LkmEstimator, LkmOptimizer, LkmLinear); copilot-instructions NS#65 note
