+++ molass/SEC/Models/LkmEstimator.py
@@ -1,0 +1,242 @@
+"""
+SEC.Models.LkmEstimator.py
+
+Estimate LKM (Lumped Kinetic Model) initial parameters from EGH component moments.
+
+The moment-matching approach follows the same design as SdmEstimator:
+  1. Compute numerical moments (M1, M2, M3) from each EGH component curve.
+  2. Match these to the LKM analytical cumulant expressions.
+
+LKM analytical cumulants (per component; Pe and t0 are shared across components):
+
+    kappa1_i = t0 * R_i                                                  (mean = tR_i)
+    kappa2_i = 2*tR_i**2/Pe + 2*t0*(R_i-1)/k_MT_i                      (variance)
+    kappa3_i = 12*tR_i**3/Pe**2
+               + 12*tR_i*t0*(R_i-1)/(Pe*k_MT_i)
+               + 6*t0*(R_i-1)/k_MT_i**2                                  (third cumulant)
+
+Factored optimization:
+    Given (t0, Pe), R_i = tR_i/t0 is determined, and k_MT_i can be solved
+    analytically from the kappa2_i equation.  The outer optimization over (t0, Pe)
+    then minimises the kappa3 residuals.
+
+    This construction guarantees that kappa1 and kappa2 are matched exactly by
+    the returned parameters; kappa3 is matched as closely as possible.
+
+In the large-Pe limit the factored solution reduces to the closed-form
+three-moment formulas:
+
+    k_MT_i  ≈  3 * M2_i / M3_i
+    t0      ≈  M1_i - 3 * M2_i**2 / (2 * M3_i)
+    R_i     =  M1_i / t0
+
+which are derived from the compound-Poisson characteristic function of the
+kinetics-only model (see molass-essence/chapters/45/characteristic-function.ipynb).
+"""
+
+import numpy as np
+from scipy.optimize import minimize
+
+# ── Constants ────────────────────────────────────────────────────────────────
+K_MT_MAX = 5000.0    # clamp when kinetics term is negligible (Pe-dominated peak)
+PE_MIN   = 10.0
+PE_MAX   = 20000.0   # cap at physically reasonable SEC range; BH can explore higher
+T0_FRAC  = 0.98      # t0 must be < T0_FRAC * min(tR)
+
+
+# ── Internal helpers ─────────────────────────────────────────────────────────
+
+def _compute_data_moments(xr_ccurves):
+    """
+    Compute (M1, M2, M3, scale) from each EGH component curve.
+
+    Uses discrete weighted sums (same convention as SdmEstimator).
+    """
+    moment_list = []
+    for ccurve in xr_ccurves:
+        cx, cy = ccurve.get_xy()
+        y_pos  = np.maximum(cy, 0)
+        W      = y_pos.sum()
+        m1     = (cx * y_pos).sum() / W
+        m2c    = (y_pos * (cx - m1)**2).sum() / W
+        m3c    = (y_pos * (cx - m1)**3).sum() / W
+        scale  = np.trapezoid(y_pos, cx)
+        moment_list.append((m1, m2c, m3c, scale))
+    return moment_list
+
+
+def _lkm_kappa3(t0, Pe, R, k_MT):
+    tR = t0 * R
+    return (12 * tR**3 / Pe**2
+            + 12 * tR * t0 * (R - 1) / (Pe * k_MT)
+            + 6  * t0 * (R - 1) / k_MT**2)
+
+
+def _k_MT_from_kappa2(t0, Pe, tR, kappa2):
+    """
+    Solve k_MT analytically from the kappa2 equation.
+
+    kappa2 = 2*tR**2/Pe  +  2*t0*(R-1)/k_MT
+    =>  k_MT = 2*t0*(R-1) / (kappa2 - 2*tR**2/Pe)
+
[…]
