+++ molass/MathUtils/FftUtils.py
@@ -23,6 +23,28 @@ def compute_standard_wCD(N):
     return w, C, D
 
 class FftInvPdf:
+    """
+    Numerically invert a characteristic function (CF) to obtain a PDF via FFT.
+
+    The FFT operates on a fixed integer grid ``[0, N]`` (default ``N=1024``).
+    The input time array passed to ``__call__`` must therefore be **pre-scaled**
+    so that the peak of the PDF falls within roughly ``[10, 100]`` on that grid.
+
+    Typical usage in a model wrapper::
+
+        _impl = FftInvPdf(my_cf)
+
+        def my_pdf(x, ..., timescale=None):
+            # Rule of thumb: map the peak position t_peak to ~80 on the grid.
+            if timescale is None:
+                timescale = 80.0 / t_peak
+            ts = timescale
+            return ts * _impl(ts * x, ...)   # inverse-scale the amplitude
+
+    The caller is responsible for both the forward scaling (``ts * x``) and the
+    inverse-amplitude scaling (``ts * result``) so that the returned values form
+    a proper PDF integrating to 1 over the original (unscaled) time axis.
+    """
     def __init__(self, cf):
         self.cf = cf
         self.default_N = 1024
@@ -34,8 +56,30 @@ class FftInvPdf:
         self._large_w = self._large_C = self._large_D = None
 
     def __call__(self, t, *params):
+        """
+        Evaluate the PDF at the pre-scaled time points ``t``.
+
+        Parameters
+        ----------
+        t : array-like
+            Sorted, non-negative, **pre-scaled** time values.  Must satisfy
+            ``t[-1] < N`` for best accuracy (auto-resize handles larger values
+            at the cost of coarser resolution).  See class docstring for the
+            recommended scaling convention.
+        *params
+            Additional parameters forwarded verbatim to the characteristic
+            function ``cf(w, *params)``.
+
+        Returns
+        -------
+        ndarray
+            PDF values at ``t``.  Integrates to 1 on the **pre-scaled** axis;
+            divide by ``timescale`` (or equivalently multiply by ``1/timescale``)
+            to obtain the PDF on the original time axis — or let the wrapper
+            handle this via ``ts * __call__(ts * x, ...)``.
+        """
         N = self.default_N
-        t_max = t[-1]  # t is always a sorted frame array (x - t0)
+        t_max = t[-1]  # t is a sorted, pre-scaled time array
         if t_max >= N:
             # Auto-resize to the next power of 2 that covers the query range.
             # Without this, UnivariateSpline would extrapolate outside [0, N-1],

[…]
