### 2026-05-22 7762459
fix: update ComponentUtils dispatch + LkmLinear docstring
- ComponentUtils.get_xr_ccurves(): EDM/CEDM → get_cedm_xr_ccurves (constrained),
  NEDM → get_edm_xr_ccurves (non-constrained); matches MODEL_NAME_DICT rename
- LkmLinear.lkm_pdf docstring: added STLC k convention note (k_MT = k_STLC/(1-e))
