### 2026-05-25 bbab53f
feat: add PoreEntryAnimation.py - single-molecule pore entry animation (Knox geometry -> K_SEC)
