+++ molass/SEC/PoreEntryAnimation.py
@@ -1,0 +1,338 @@
+"""
+    SEC.PoreEntryAnimation.py
+
+    Single-molecule pore-entry animation: Knox sector geometry -> K_SEC -> LKM k.
+
+    Complements ColumnSimulation.py (multi-particle elution) with a focused view of
+    what happens at the grain scale: one molecule diffusing through a 5-grain assembly,
+    entering and exiting pore sectors, accumulating dwell statistics.
+
+    The simulation closes the loop:
+        geometry -> K_SEC -> LKM mass-transfer rate k -> SDM n_pi
+
+    Copyright (c) 2024-2025, Molass Community
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.animation import FuncAnimation
+plt.rcParams['animation.embed_limit'] = 512
+
+from .ColumnElements import NewGrain, Particle, solvant_color
+from .StationaryMove import get_next_position_impl
+
+
+# ---------------------------------------------------------------------------
+# Default geometry (2+1+2 staggered, matches 3D random close-packing ~0.64)
+# ---------------------------------------------------------------------------
+
+_DEFAULT_GRAIN_CENTERS = [
+    (0.22, 0.20), (0.78, 0.20),   # bottom row
+    (0.50, 0.50),                  # middle
+    (0.22, 0.80), (0.78, 0.80),   # top row
+]
+
+
+def run_simulation(
+    r_mol=0.03,
+    R_grain=0.20,
+    num_pores=3,
+    grain_centers=None,
+    W=1.0, H=1.0,
+    D=0.003, dt=0.002,
+    n_steps=10_000,
+    v_drift=0.10,
+    seed=42,
+):
+    """
+    Run a single-molecule Brownian simulation through a grain assembly.
+
+    Parameters
+    ----------
+    r_mol : float
+        Molecule radius.
+    R_grain : float
+        Grain radius.
+    num_pores : int
+        Number of pore sectors per grain.  Determines sector half-angle
+        alpha = pi / (2 * num_pores) and Knox pore radius R_p = R_grain * alpha.
+    grain_centers : list of (float, float), optional
+        Centre coordinates of each grain.  Defaults to 2+1+2 staggered layout.
+    W, H : float
+        Simulation box width and height.
+    D : float
+        Diffusion coefficient.
+    dt : float
+        Time step.
+    n_steps : int
+        Number of simulation steps.
+    v_drift : float
+        Downward mobile-phase drift speed (mobile-phase steps only).
+    seed : int
+        Random seed.
+
+    Returns
+    -------
+    dict with keys:
+        positions       : ndarray (n_steps+1, 2)
+        states          : ndarray (n_steps+1,)  — grain index or -1 (mobile)
+        entry_times     : ndarray  — simulation time of each pore entry
+        dwell_times     : ndarray  — duration of each completed pore dwell
+        entry_grain_arr : ndarray  — which grain for each entry
+        grains          : list of NewGrain
+        R_p             : float    — Knox pore radius
+        rho             : float    — r_mol / R_p
+        K_SEC_theory    : float    — Knox (1-rho)^2 approximation
+        T_total         : float
+    """
+    if grain_centers is None:
+        grain_centers = _DEFAULT_GRAIN_CENTERS
+
+    grains = [NewGrain(gid, center, R_grain, num_pores)
+              for gid, center in enumerate(grain_centers)]
+
+    R_p          = np.pi * R_grain / (2 * num_pores)
+    rho          = r_mol / R_p
[…]
