### 2026-05-25 23f44b6
fix: PoreEntryAnimation suptitle clipped - swap tight_layout/subplots_adjust order
