### 2026-05-25 4909cc2
feat: PoreEntryAnimation Langmuir extension - k_ads/k_des wall-adsorption state
- run_simulation: add k_ads=0, k_des=1 parameters (backward-compatible)
- New 'wall-adsorbed' state: molecule frozen inside pore sector
  - probability k_ads*dt to adsorb from free-in-pore per step
  - probability k_des*dt to desorb back to free-in-pore per step
- New return keys: wall_bound (bool array), K_ads_theory, K_eff_theory
  - K_eff_theory = K_SEC_theory * (1 + K_ads_theory)
- get_pore_entry_animation: 3-colour molecule (royalblue/orchid/tomato)
  - tomato = wall-adsorbed state (new)
  - suptitle shows Lan
[…]
