+++ molass/SEC/PoreEntryAnimation.py
@@ -43,6 +43,8 @@ def run_simulation(
     n_steps=10_000,
     v_drift=0.10,
     seed=42,
+    k_ads=0.0,
+    k_des=1.0,
 ):
     """
     Run a single-molecule Brownian simulation through a grain assembly.
@@ -70,20 +72,27 @@ def run_simulation(
         Downward mobile-phase drift speed (mobile-phase steps only).
     seed : int
         Random seed.
+    k_ads : float
+        Langmuir adsorption rate (per unit time).  0 = pure SEC (no wall binding).
+    k_des : float
+        Langmuir desorption rate (per unit time).  K_ads = k_ads / k_des.
 
     Returns
     -------
     dict with keys:
         positions       : ndarray (n_steps+1, 2)
         states          : ndarray (n_steps+1,)  — grain index or -1 (mobile)
+        wall_bound      : ndarray (n_steps+1,) bool — True when wall-adsorbed
         entry_times     : ndarray  — simulation time of each pore entry
-        dwell_times     : ndarray  — duration of each completed pore dwell
+        dwell_times     : ndarray  — total pore dwell duration (free + adsorbed)
         entry_grain_arr : ndarray  — which grain for each entry
         grains          : list of NewGrain
         R_p             : float    — Knox pore radius
         rho             : float    — r_mol / R_p
         K_SEC_theory    : float    — Knox (1-rho)^2 approximation
         T_total         : float
+        K_ads_theory    : float    — k_ads / k_des equilibrium constant
+        K_eff_theory    : float    — K_SEC_theory * (1 + K_ads_theory)
     """
     if grain_centers is None:
         grain_centers = _DEFAULT_GRAIN_CENTERS
@@ -107,6 +116,8 @@ def run_simulation(
     positions[0] = pos
 
     entry_times, dwell_starts, dwell_times, entry_grain_list = [], [], [], []
+    wall_bound      = np.zeros(n_steps + 1, dtype=bool)
+    wall_bound_flag = False   # True when molecule is adsorbed to the pore wall
 
     for step in range(n_steps):
         dx      = rng.normal(0, sigma, 2)
@@ -148,26 +159,43 @@ def run_simulation(
                     else:
                         new_pos = pos.copy()
         else:
-            # In-pore phase
-            grain    = grains[in_grain]
-            particle = Particle(tuple(pos), r_mol)
-            nx, ny, inmobile = get_next_position_impl(
-                particle, grain,
-                pos[0], pos[1],
-                new_pos[0], new_pos[1])
-            new_pos = np.array([nx, ny])
+            # In-pore phase — two sub-states: free or wall-adsorbed
+            if wall_bound_flag:
+                # Wall-adsorbed: frozen in place, wait for desorption
+                new_pos = pos.copy()
+                if k_des > 0 and rng.random() < k_des * dt:
+                    wall_bound_flag = False       # desorb -> free in pore
+            else:
+                # Free in pore: try to adsorb, or normal Brownian step
+                if k_ads > 0 and rng.random() < k_ads * dt:
+                    wall_bound_flag = True         # adsorb -> frozen at pore wall
+                    new_pos = pos.copy()
+                else:
+                    grain    = grains[in_grain]
+                    particle = Particle(tuple(pos), r_mol)
+                    nx, ny, inmobile = get_next_position_impl(
+                        particle, grain,
+                        pos[0], pos[1],
+                        new_pos[0], new_pos[1])
+                    new_pos = np.array([nx, ny])
 
-            if inmobile:
[…]
