### 2026-05-25 cf2e2e1
feat: rename k_ads/k_des -> k_a/k_d, K_ads_theory -> K_ads in PoreEntryAnimation API
