+++ tests/specific/300_SEC_Models/test_070_langmuir_simulation.py
@@ -12,51 +12,51 @@ def _sim(**kwargs):
 
 
 def test_wall_bound_zero_when_no_adsorption():
-    """k_ads=0 → no wall-adsorbed steps, K_ads_theory=0."""
-    s = _sim(k_ads=0.0, k_des=1.0)
+    """k_a=0 → no wall-adsorbed steps, K_ads=0."""
+    s = _sim(k_a=0.0, k_d=1.0)
     assert s['wall_bound'].sum() == 0
-    assert s['K_ads_theory'] == 0.0
+    assert s['K_ads'] == 0.0
 
 
 def test_wall_bound_nonzero_with_adsorption():
-    """k_ads>0 → some wall-adsorbed steps (uses seed=42 that reliably enters pores)."""
-    s = _sim(k_ads=3.0, k_des=1.0, n_steps=10_000, seed=42)
+    """k_a>0 → some wall-adsorbed steps (uses seed=42 that reliably enters pores)."""
+    s = _sim(k_a=3.0, k_d=1.0, n_steps=10_000, seed=42)
     assert s['wall_bound'].sum() > 0
 
 
 def test_wall_bound_only_inside_pore():
     """wall_bound can only be True when the molecule is in a pore (states >= 0)."""
-    s = _sim(k_ads=2.0, k_des=1.0)
+    s = _sim(k_a=2.0, k_d=1.0)
     mobile_and_bound = s['wall_bound'] & (s['states'] < 0)
     assert mobile_and_bound.sum() == 0
 
 
 def test_K_ads_theory():
-    """K_ads_theory = k_ads / k_des."""
-    k_ads, k_des = 3.0, 2.0
-    s = _sim(k_ads=k_ads, k_des=k_des)
-    assert abs(s['K_ads_theory'] - k_ads / k_des) < 1e-10
+    """K_ads = k_a / k_d."""
+    k_a, k_d = 3.0, 2.0
+    s = _sim(k_a=k_a, k_d=k_d)
+    assert abs(s['K_ads'] - k_a / k_d) < 1e-10
 
 
 def test_K_eff_theory():
-    """K_eff_theory = K_SEC_theory * (1 + K_ads_theory)."""
-    k_ads, k_des = 2.0, 1.0
-    s = _sim(k_ads=k_ads, k_des=k_des)
-    expected = s['K_SEC_theory'] * (1 + s['K_ads_theory'])
+    """K_eff_theory = K_SEC_theory * (1 + K_ads)."""
+    k_a, k_d = 2.0, 1.0
+    s = _sim(k_a=k_a, k_d=k_d)
+    expected = s['K_SEC_theory'] * (1 + s['K_ads'])
     assert abs(s['K_eff_theory'] - expected) < 1e-10
 
 
 def test_K_eff_greater_than_K_SEC():
     """With adsorption, effective K > geometric K."""
-    s_sec = _sim(k_ads=0.0, k_des=1.0)
-    s_ads = _sim(k_ads=2.0, k_des=1.0)
+    s_sec = _sim(k_a=0.0, k_d=1.0)
+    s_ads = _sim(k_a=2.0, k_d=1.0)
     assert s_ads['K_eff_theory'] > s_sec['K_SEC_theory']
 
 
 def test_mean_dwell_longer_with_adsorption():
     """Adsorption delays exit → longer mean dwell time (statistical test)."""
-    s_sec = _sim(n_steps=10_000, k_ads=0.0, k_des=1.0, seed=42)
-    s_ads = _sim(n_steps=10_000, k_ads=3.0, k_des=1.0, seed=42)
+    s_sec = _sim(n_steps=10_000, k_a=0.0, k_d=1.0, seed=42)
[…]
