### 2026-05-26 08a75a6
feat: ns_narrow_bounds + ns_adaptive_nsteps in optimize_rigorously(); update copilot-instructions
