+++ molass/Rigorous/CurrentStateUtils.py
@@ -322,9 +322,24 @@ def list_rigorous_jobs(analysis_folder):
         if not fv_list:
             continue
 
+        # Guard against premature reads.  The very first callback.txt entry is
+        # always the init-params evaluation written before any search begins:
+        #   BH:  c=0, a=False  (init); subsequent entries have a=True when BH
+        #        accepts a jump.
+        #   NS:  a is ALWAYS False (hardcoded in SamplerCallback.__call__);
+        #        the counter rises with each NS iteration callback.
+        # Skip the job if only the single init entry exists (i.e. the search
+        # has not yet made its first move in either method). (issue #188)
+        if len(fv_list) <= 1:
+            continue  # only init entry — BH/NS has not started yet
+
+        # For BH: use accepted (a=True) entries to get the true BH best.
+        # For NS: all entries are a=False by design — use all of them.
+        accepted = [row for row in fv_list if row[2] is True]
+        best_entries = accepted if accepted else fv_list
         fv_arr = np.array(fv_list, dtype=object)
         iterations = len(fv_arr)
-        best_fv = float(min(row[1] for row in fv_list))
+        best_fv = float(min(row[1] for row in best_entries))
         # Last iteration's timestamp (column 3)
         timestamp = fv_list[-1][3]
 
