### 2026-05-27 5505bda
feat: ns_nsteps + ns_adaptive_nsteps params threaded to optimize_rigorously()
