### 2026-05-28 6987580
CMA support, issue #189 non-blocking BackRunner, rg-curve cleanup (#69)
- RigorousImplement.py, RunInfo.py: issue #189 — BackRunner path returns
  immediately; _subprocess_process stored; is_alive/wait() updated
- LegacyBridgeUtils.py: rg-curve cleanup (#69) — remove rg-curve/ export,
  use only rg_curve_parent/
- SecSaxsData.py: get_rg_curve() with caching (issue #168)
- Decomposition.py: minor fix
- .github/copilot-instructions.md: session notes
