+++ molass/Solvers/BH/__init__.py
@@ -1,0 +1,3 @@
+"""
+    molass.Solvers.BH
+"""
