+++ molass/Solvers/__init__.py
@@ -1,0 +1,7 @@
+"""
+    molass.Solvers
+
+    Active solver implementations (BH moved here from molass-legacy; CMA new).
+
+    Copyright (c) 2026, SAXS Team, KEK-PF
+"""

[…]
