### 2026-05-28 1bfc6aa
AI-friendliness: #190 get_rigorous_param_count(), #191 LKM docstrings
- Decomposition.get_rigorous_param_count(): wraps make_basecurves_from_decomposition
  + make_rigorous_initparams; works for all models (EGH/SDM/EDM/CEDM/LKM)
- detect_function_code docstring: lists all supported models with codes
  (None/G1100/G1200/G1300/G1400/G2020)
- optimize_rigorously() Notes: adds LKM two-step workflow example alongside
  existing EGH and EGH->SDM examples
