### 2026-05-28 86dda8c
upgrade() docstring: add LKM and CEDM; Notes example uses .upgrade('LKM')
