### 2026-05-28 4ae9d52
refactor: monitor=True is single dashboard control; deprecate progress param
