### 2026-05-28 eedd366
fix: cooperative stop in CMA every generation, not just on improvement (#170)
