### 2026-05-29 bd99fc5
fix: auto-fallback async_=False for CMA in_process to prevent kernel crash (#193)
