### 2026-05-29 fb722ac
fix: CMA async_ fallback → in_process=False (keeps dashboard, #193)
