### 2026-05-29 b637496
Fix load_best  so that it returns as expected for CMA
