### 2026-05-31 b544c35
fix: SDM lognormal upgrade and EDM/CEDM initial parameter unification
