+++ molass/SEC/Models/SdmOptimizer.py
@@ -365,7 +365,23 @@ def optimize_sdm_lognormal_xr_decomposition(decomposition, env_params, model_par
     env_params : tuple
         The environmental parameters ``(N, T, me, mp, N0, t0, mu, sigma)``.
     model_params : dict, optional
-        The parameters for the SDM model.
+        The parameters for the SDM model.  Recognized keys:
+
+        ``ln_pore_sigma`` : float or None (default 0.3)
+            Standard deviation of ``ln(r/r0)``, held constant during
+            optimization (removed from the free-parameter vector).
+            Invariant to the choice of length unit (a unit change only
+            shifts all ``ln(r/r0)`` values by a constant, which drops out
+            of the standard deviation).
+            The default 0.3 implies pores span a factor of
+            ``exp(0.3) ≈ 1.35`` around the median — a reasonable prior for
+            most SEC columns.  Pass ``None`` to make sigma a free parameter;
+            note that sigma is underdetermined when ``num_components=1``
+            and will drift to ``sigma_max`` if left free.
+        ``sigma_max`` : float (default 0.8)
+            Upper bound on sigma when it is free.
+        ``k`` : float (default 2.0)
+        ``rt_dist`` : str (default ``'gamma'``)
     kwargs : dict
         Additional parameters for the optimization process.
 
@@ -388,18 +404,23 @@ def optimize_sdm_lognormal_xr_decomposition(decomposition, env_params, model_par
     N, T, me, mp, N0, t0, mu_init, sigma_init = env_params
     rgv = np.asarray(decomposition.get_rgs())
 
+    # ln_pore_sigma default: 0.3 — std-dev of ln(r/r0), invariant to unit choice.
+    # Free sigma (None) is underdetermined for num_components=1 and drifts to sigma_max.
+    _LN_PORE_SIGMA_DEFAULT = 0.3
     if model_params is None:
         k_init = 2.0
         rt_dist = 'gamma'
         rg_penalty_weight = 1.0
         sigma_max = 0.8
         min_rg_gap = 0.5
+        ln_pore_sigma = _LN_PORE_SIGMA_DEFAULT
     else:
         k_init = model_params.get('k', 2.0)
         rt_dist = model_params.get('rt_dist', 'gamma')
         rg_penalty_weight = model_params.get('rg_penalty_weight', 1.0)
         sigma_max = model_params.get('sigma_max', 0.8)
         min_rg_gap = model_params.get('min_rg_gap', 0.5)  # Å — minimum required Rg gap between components
+        ln_pore_sigma = model_params.get('ln_pore_sigma', _LN_PORE_SIGMA_DEFAULT)
 
     if rt_dist == 'exponential':
         k_init = 1.0
@@ -432,9 +453,15 @@ def optimize_sdm_lognormal_xr_decomposition(decomposition, env_params, model_par
     rg_penalty_scale = rg_penalty_weight * initial_error
 
     def objective_function(params):
-        N_, T_, x0_, tI_, N0_, k_, mu_, sigma_ = params[0:8]
-        rgv_ = params[8:8 + num_components]
-        scales_ = params[8 + num_components:8 + 2 * num_components]
+        N_, T_, x0_, tI_, N0_, k_, mu_ = params[0:7]
+        if ln_pore_sigma is None:
+            sigma_ = params[7]
+            rgv_start = 8
+        else:
+            sigma_ = ln_pore_sigma
+            rgv_start = 7
+        rgv_ = params[rgv_start:rgv_start + num_components]
[…]

[…]
