### 2026-06-01 773588f
AI-friendliness: refine_lognormal_params_by_moments NM+Rg bound (#196)
- Replace L-BFGS-B with Nelder-Mead + penalty-based Rg bound
  mu_max = ln(2 * Rg_max): prevents degenerate Stage-3 decomposition
  maxiter=20000 for reliable convergence
- SdmComponentCurve: add SdmMonoColumnParams / SdmLognormalColumnParams namedtuples
- SdmOptimizer: Rg-aware RuntimeError hint when cond_C > 1e6
- LognormalPore: docstring note on me/mp explicit passing
Validated on SAMPLE1 (3 comps): residual 0.1801 -> 0.1442
