+++ molass/SEC/Models/SdmEstimator.py
@@ -2,7 +2,24 @@
 SEC.Models.SdmEstimator.py
 """
 import numpy as np
+from collections import namedtuple
 from scipy.optimize import minimize
+
+LognormalEnv = namedtuple(
+    'LognormalEnv',
+    ['N', 'T', 'me', 'mp', 'N0', 't0', 'mu', 'sigma']
+)
+"""Named-tuple form of the lognormal env-params 8-tuple ``(N, T, me, mp, N0, t0, mu, sigma)``.
+
+Returned by ``estimate_sdm_lognormal_from_monopore`` and accepted by
+``optimize_sdm_lognormal_xr_decomposition`` as its ``env_params`` argument.
+Backward-compatible with positional unpacking.
+
+Note: ``k`` (gamma shape) is *not* included here; it lives in
+``model_params={'k': ...}`` of the lognormal optimizer (default 2.0).
+Access ``me`` and ``mp`` via ``ln_env.me`` / ``ln_env.mp`` when calling
+``sdm_lognormal_model_moments(rg, N, T, N0, t0, k, mu, sigma, me=..., mp=...)``.
+"""
 
 def estimate_sdm_column_params(decomposition, **kwargs):
     """
@@ -285,7 +302,7 @@ def estimate_sdm_lognormal_from_monopore(mono_ccurves, xr_icurve, **kwargs):
             decomposition, N, T, N0, t0_adj, k, mu, sigma,
             me=me, mp=mp, debug=debug)
 
-    return N, T, me, mp, N0, t0_adj, mu, sigma
+    return LognormalEnv(N, T, me, mp, N0, t0_adj, mu, sigma)
 
 
 def refine_lognormal_params_by_moments(
@@ -295,13 +312,19 @@ def refine_lognormal_params_by_moments(
 
     Given an initial guess (typically from ``estimate_sdm_lognormal_from_monopore``
     which uses heuristic poresize geometric mean + sigma=0.3), this function
-    refines (mu, sigma, k, t0) by L-BFGS-B against the per-component (M1, Var)
+    refines (mu, sigma, k, t0) by Nelder-Mead against the per-component (M1, Var)
     extracted from the EGH decomposition.
+
+    The upper bound on ``mu`` is enforced as ``ln(2 × Rg_max)`` — the SDM
+    separation criterion.  Above this, K_SEC values compress across components
+    and elution curves converge, causing a degenerate Stage-3 decomposition.
+    The L-BFGS-B solver that was used previously could overshoot this bound,
+    accidentally yielding a bad starting point or causing a degenerate collapse;
+    Nelder-Mead with an explicit penalty respects the bound by construction.
 
     Uses the fast analytical moment evaluator
     :func:`molass.SEC.Models.LognormalPore.sdm_lognormal_model_moments`
-    (~50 us/call), so the full refinement (~100 evals × num_components) costs
-    only a few ms.
+    (~50 us/call), so the full refinement costs only a few tens of ms.
 
     Only (mu, sigma, k, t0) are refined; (N, T, N0) are held fixed because
     the mono-pore stage already constrains them well from M1+M2+M3 matching.
@@ -352,29 +375,36 @@ def refine_lognormal_params_by_moments(
 
     eps = 1e-30
 
+    # Rg-derived upper bound on mu: pore center must stay ≤ 2 × Rg_max
+    # so that K_SEC values remain distinct and Stage-3 elution curves stay
+    # separated.  Overshooting this bound causes a degenerate (rank-1) C
+    # matrix in Stage 3 and a RuntimeError.
+    rg_max = float(np.max(rgv))
+    mu_max = np.log(2.0 * rg_max)
+    mu_start = min(mu, mu_max - 0.05)
+
[…]
