### 2026-06-01 e3c4c27
feat: model_params maxiter support in optimize_sdm_lognormal_xr_decomposition
