### 2026-06-01 7d227c5
feat: post-NM NNLS rescaling in SDM mono/lognormal optimizers
After Nelder-Mead converges on column shape params, recompute component
scales analytically via NNLS on the converged PDF basis. This finds the
exact 1D-fit-optimal scales for the settled shape, correcting any drift
introduced during joint NM exploration.

Also: strip rgcurve at top of SDM.optimize_decomposition() (consolidated
from per-stage pop to a single pop before branching), so it cannot
propagate to any sub-function.
