### 2026-06-02 18c9054
fix: SDM pbar polish; EGH peak recognition unification (Level-A migration complete)
- SdmOptimizer.py: fill tqdm bar to 100% before close (prevents red on early NM convergence); close issue #198
- SDM.py: strip 'progress' from mono_kwargs in lognormal pipeline; close issue #197
- RecognizerSpecific.py: replace legacy recognize_peaks with EghPeeler.egh_peel (last library-side caller)
- RgCurveUtils.py, Reports/V1Report.py, ColumnSimulation.py: minor cleanups
- copilot-instructions.md: Jun 2026 unification trail entries
- Copilot/copilot-guidelines.md: updated
