### 2026-06-05 46e65d0
docs: plan ComplementaryView → plot_components refactor (2026-06-05)
