### 2026-06-08 cc684ba
feat: Complementary View → Plot Components refactor (Rg unification + cleanup)
- Bridge/SdAdapter.py: add make_ssd_from_corrected_sd, make_ssd_from_dsets, decomposition_from_optimizer_params
- Bridge/LegacyRgCurve.py: add missing parent attrs (ecurve, excl_spline, excl_info, X); docstring #203
- Guinier/RgCurveUtils.py: add progress_cb parameter to compute_rgcurve_info
- DataObjects/XrData.py: forward progress_cb to compute_rgcurve_info
- DataObjects/SecSaxsData.py: forward progress_cb to compute_rgcurve
- LowRank/Decomposition.py: copy_with_new_components accepts **kwargs; NOTE on _rgcurve lo
[…]
