+++ Copilot/copilot-guidelines.md
@@ -47,22 +47,6 @@ Use this magic phrase to initialize both technical context and behavioral rules:
 
     *Naming* and *workflow friction* tend to be general (applicable to any Python library). *Observability* and *interpretability* tend to be domain-specific (SEC-SAXS optimization semantics). When assessing what to improve next, look for gaps: e.g., a new module with opaque attribute names (Naming), or a new score with no physical explanation (Interpretability).
 
----
-
-## Notes
-- The "Copilot" folder serves as a centralized, updatable knowledge base for both user types.
-- Markdown format is recommended for clarity and ease of editing.
-- Content should be periodically reviewed and expanded as needed.
-
-## Environment Assumption
-
-- All users are assumed to work in Visual Studio Code (VS Code) with Agent mode enabled.
-- Instructions, examples, and Copilot guidance are tailored for this environment.
-
-## Attribution
-
-These guidelines were developed in collaboration with GitHub Copilot (GPT-4.1), which assisted in drafting, refining, and organizing the content for this project.
-
 13. **Repository Role Separation (Architecture Principle):** The two-repo structure has a clear intended division:
 
     - **`molass-legacy`** — maintained for the **tkinter GUI** and as a **historical record**. The GUI should be actively maintained. Legacy-only code that is not called by any GUI path or library path can be left as-is for reference.
@@ -88,6 +72,22 @@ These guidelines were developed in collaboration with GitHub Copilot (GPT-4.1), 
     The legacy `sd` (`SerialData`) and the library `ssd` (`SecSaxsData`) represent the same concept at different stages of development. The long-term goal is for `ssd` to fully replace `sd` as the authoritative data container, with the GUI eventually constructing and accepting `ssd` directly. This is a larger refactor than A–C above because `sd` is deeply embedded in the legacy GUI's internal data flow. Incremental steps: identify GUI paths that construct or pass `sd`, and replace them one by one with `ssd` equivalents.
 
     **Next planned step (2026-06-05)**: Replace `PeakEditor`'s "Complementary View" (legacy `ComplementaryView` dialog) with library `plot_components_impl`. Approach: in `show_peak_editor_impl` (OptimizerUtils.py), construct `ssd`+`Decomposition` alongside existing `sd`/`corrected_sd`, pass as optional `decomposition=` kwarg to `PeakEditor`; `show_complementary_view()` uses it when available, falls back to legacy otherwise. First step: check whether a `sd → ssd` conversion utility already exists. See `/memories/repo/complementary-view-refactor-plan.md` for full details.
+
+---
+
+## Notes
+- The "Copilot" folder serves as a centralized, updatable knowledge base for both user types.
+- Markdown format is recommended for clarity and ease of editing.
+- Content should be periodically reviewed and expanded as needed.
+
+## Environment Assumption
+
+- All users are assumed to work in Visual Studio Code (VS Code) with Agent mode enabled.
+- Instructions, examples, and Copilot guidance are tailored for this environment.
+
+## Attribution
+
+These guidelines were developed in collaboration with GitHub Copilot (GPT-4.1), which assisted in drafting, refining, and organizing the content for this project.
 
 ---
 
