### 2026-06-09 90a5720
feat: DE and NSGA2 solvers; SolverRegistry; **solver_kwargs API (#204 closed)
