+++ molass/Solvers/NSGA2/SolverNSGA2.py
@@ -1,0 +1,252 @@
+"""
+    molass.Solvers.NSGA2.SolverNSGA2
+
+    Multi-objective NSGA-II solver wrapping pymoo.
+
+    Instead of minimizing the synthesized scalar `fv`, this solver treats the 7
+    raw score components as separate objectives and lets NSGA-II maintain a
+    Pareto front across all of them simultaneously.
+
+    After the run, `synthesize()` is applied to each Pareto-optimal solution to
+    pick the best — but the search itself never saw the synthesized scalar.
+
+    This sidesteps the risk that `synthesize()`'s weighting creates a misleading
+    landscape that causes BH and CMA-ES to converge to wrong basins.
+
+    Interface matches SolverCMA / SolverDE so it can be wired into
+    BasicOptimizer.solve() with method='nsga2'.
+
+    Parameters are in the normalized [0, 10] space used throughout the
+    molass-legacy optimizer infrastructure.
+
+    Copyright (c) 2026, SAXS Team, KEK-PF
+"""
+import numpy as np
+from scipy.optimize import OptimizeResult
+
+# 7 score components (indices into the score_list returned by objective_func
+# with return_full=True):
+#   0  XR_2D_fitting
+#   1  XR_LRF_residual
+#   2  UV_2D_fitting
+#   3  UV_LRF_residual
+#   4  Guinier_deviation
+#   5  Kratky_smoothness
+#   6  SEC_conformance
+N_OBJ = 7
+
+# Index of negative_penalty in the full score_list_with_penalties (after the 7 major scores).
+# score_list_with_penalties order (from get_score_names):
+#   7 = mapping_penalty, 8 = negative_penalty, 9 = baseline_penalty, ...
+# We pass negative_penalty as an inequality constraint G ≤ 0 so NSGA-II penalises
+# solutions with negative scattering profiles, matching BH behaviour.
+# The constraint value is: G = negative_penalty (must be 0 to satisfy).
+_NEG_PENALTY_IDX = 8   # index 8 in score_list_with_penalties
+
+# Evaluation budget multiplier: max_fevals = niter * FEVALS_PER_NITER
+FEVALS_PER_NITER = 200
+
+
+class SolverNSGA2:
+    """NSGA-II solver using pymoo.
+
+    Treats the 7 raw score components as separate objectives.  After the run,
+    applies `synthesize()` to each Pareto member and returns the best as if it
+    were a single-objective result — so RunInfo / callback.txt / load_best()
+    all work unchanged.
+
+    Parameters
+    ----------
+    optimizer : BasicOptimizer
+        Fully constructed optimizer with `objective_func` and `minima_callback`.
+    pop_size : int
+        NSGA-II population size.  Default 100.
+    """
+
+    def __init__(self, optimizer, pop_size=100):
+        self.optimizer = optimizer
+        self.pop_size = pop_size
+
+    def minimize(self, objective, init_params, niter=100, seed=1234,
+                 bounds=None, narrow_bounds=False, show_history=False):
+        """Run NSGA-II minimization on the 7 raw score components.
+
+        Parameters
+        ----------
+        objective : callable
+            The synthesized scalar objective (used only for minima_callback
+            and as fallback).  NOT used for the NSGA-II search itself.
+        init_params : ndarray
+            Initial parameter vector in normalized [0, 10] space.
+        niter : int
+            Budget: max_fevals = niter * FEVALS_PER_NITER.
+        seed : int
+            RNG seed.
+        bounds, narrow_bounds, show_history
+            Same API as SolverBH / SolverCMA (bounds used; others ignored).
+
+        Returns
+        -------
+        scipy.optimize.OptimizeResult
[…]
