+++ molass/Solvers/NSGA2/__init__.py
@@ -1,0 +1,7 @@
+"""
+    molass.Solvers.NSGA2
+
+    Multi-objective NSGA-II solver using pymoo.
+
+    Copyright (c) 2026, SAXS Team, KEK-PF
+"""

[…]
