### 2026-06-09 77432a0
refactor: create Copilot/refactor/ folder; extract Rule 13 to ARCHITECTURE.md; add PLAN_subprocess_parity.md
