+++ Copilot/refactor/PLAN_subprocess_parity.md
@@ -1,0 +1,114 @@
+# Plan: Subprocess Parity — GUI Path Missing ip_*.npy Export
+
+**Issue**: molass-library#206  
+**Status**: Open  
+**Last updated**: Jun 9, 2026
+
+---
+
+## Background
+
+Issues #38–#42 (molass-legacy) established in-process/subprocess parity for BH:
+- `ip_xr_elcurve_y.npy`, `ip_uv_elcurve_y.npy` — EGH-fitted elution curves (#38)
+- `ip_xr_D.npy`, `ip_uv_U.npy` — 2D intensity matrices (#39)
+- `ip_xr_E.npy` — error matrix (#40)
+- `ip_xr_qvector.npy` — trimmed q-values (966 vs 972 elements) (#41)
+- `FixedBaselineOptimizer` removed from subprocess default (#42)
+
+Result: BH subprocess delta_fv < 0.001 for **notebook→subprocess** path.
+
+---
+
+## New observation (Jun 9, 2026)
+
+DE with `in_process=True` (notebook): **SV ≈ 84.37**  
+DE with `in_process=False` (GUI subprocess): **SV ≈ 78.56** — ~5.8 SV gap
+
+Confirmed on analysis-020: `optimized/` folder has **no ip_*.npy files**.
+
+---
+
+## Root cause
+
+There are two distinct code paths:
+
+**Path B (notebook → subprocess)** — *fixed by #38–42*:
+```
+decomp.optimize_rigorously(in_process=False)
+  → make_rigorous_decomposition_impl()
+    → prepare_rigorous_folders()   ← writes ip_*.npy ✅
+      → BackRunner.run()           ← subprocess reads ip_*.npy ✅
+```
+
+**Path C (GUI)** — *not fixed*:
+```
+GUI click → BackRunner.run()
+              → OptimizerMain.create_optimizer_from_job()
+                → get_sd_from_folder_impl()   ← legacy disk load, no ip_*.npy ❌
+```
+
+`prepare_rigorous_folders` is **never called** in Path C.  
+All 7 parity overrides silently skip → subprocess uses legacy-derived data.
+
+---
+
+## Divergence factors (all 7, now clearly mapped)
+
+| # | Factor | Override file | Loaded in | Fixed for Path B | Fixed for Path C |
+|---|---|---|---|---|---|
+| 1 | XR elution curve y | `ip_xr_elcurve_y.npy` | `OptDataSets.get_dsets_impl` | ✅ #38 | ❌ |
+| 2 | UV elution curve y + sy | `ip_uv_elcurve_y.npy` | `OptDataSets.get_dsets_impl` | ✅ #38+#51 | ❌ |
+| 3 | XR D matrix | `ip_xr_D.npy` | `OptDataSets.get_dsets_impl` | ✅ #39 | ❌ |
+| 4 | UV U matrix | `ip_uv_U.npy` | `OptDataSets.get_dsets_impl` | ✅ #39 | ❌ |
+| 5 | E matrix | `ip_xr_E.npy` | `OptDataSets.__init__` | ✅ #40 | ❌ |
+| 6 | qvector | `ip_xr_qvector.npy` | `OptimizerMain.create_optimizer_from_job` | ✅ #41 | ❌ |
+| 7 | FixedBaselineOptimizer | (removed) | `optimizer_main` | ✅ #42 | ✅ #42 |
+
+---
+
+## Fix options
+
+### Option A — Export ip_*.npy from the GUI side (shallow fix)
+
+In the GUI launch flow (before `BackRunner.run()`), reconstruct a library `SecSaxsData`
+from the loaded analysis folder and call a lightweight version of `prepare_rigorous_folders`
+that writes only the ip_*.npy files.
+
[…]
