### 2026-06-09 cc7f5c9
docs: add Factor 8 Rg investigation to PLAN_subprocess_parity; result: not culprit
