### 2026-06-10 9bbf769
fix: DE warm-start seeds entire population around init_params (Gaussian cloud)
Previously _WarmStartSampling only set X[0] = init_params and filled the
rest of the population randomly.  When a successive job loaded the previous
job's best params via _load_best_init_params, the 149 random members could
dominate the population best and drive DE into a worse basin — producing the
observed SV regression (83.35 → 82.55).

Fix: draw all population members from N(init_params, warm_sigma=1.0),
clipped to bounds.  Member 0 is set exactly to init_params so the known best
is always preserved in the initi
[…]
