### 2026-06-10 a0ac7dd
feat: GuiReplay.py — reconstruct LEG-GUI scenario in-process for notebook introspection
load_gui_scenario(analysis_folder) reads the files produced by a completed
GUI run (optimizer.log, init_params.txt, opt_settings.txt, ip_*.npy) and
reconstructs the optimizer in the calling process so it can be:
  - queried for fv and score_breakdown at any parameter vector
  - re-run with any solver for direct comparison against LIB-IN

Architecture alignment (ARCHITECTURE.md):
  - Lives in molass-library (library side)
  - Calls molass-legacy DsetsDebug layer for raw reconstruction
  - Serves the data-obj
[…]
