+++ molass/Rigorous/GuiReplay.py
@@ -1,0 +1,303 @@
+"""
+    molass.Rigorous.GuiReplay
+
+    Reconstruct a legacy GUI optimizer run in-process for notebook introspection.
+
+    This module bridges the gap between the opaque LEG-GUI subprocess path and
+    the notebook API.  It reads the files produced by a completed GUI run
+    (callback.txt, init_params.txt, optimizer.log, opt_settings.txt, ip_*.npy)
+    and reconstructs the optimizer in the calling process so it can be:
+
+    - queried for score_breakdown at any parameter vector
+    - re-run with any solver for direct comparison
+    - compared against LIB-IN results using compare_dsets()
+
+    Alignment with the refactor architecture:
+    - molass-library is the home for active computational code (ARCHITECTURE.md)
+    - This module calls molass-legacy's DsetsDebug layer for raw reconstruction
+    - Returns RunInfo so the result is consumable through the library's API
+    - Serves the data-object consolidation track: makes the GUI's sd-based
+      landscape directly inspectable, creating pressure to replace sd with ssd
+
+    Usage example
+    -------------
+    from molass.Rigorous.GuiReplay import load_gui_scenario
+
+    scenario = load_gui_scenario(r'C:\\PyTools\\reports\\analysis-005')
+    print('fv_init from GUI callback:', scenario.fv_init)
+    print('fv_init reproduced      :', scenario.eval_init())
+
+    # Re-run in-process with BH for comparison
+    run = scenario.run_inprocess(method='BH', niter=20)
+    run.wait()
+    bd = run.get_score_breakdown()
+
+    Copyright (c) 2026, SAXS Team, KEK-PF
+"""
+from __future__ import annotations
+
+import os
+import re
+import logging
+import numpy as np
+from collections import namedtuple
+
+logger = logging.getLogger(__name__)
+
+
+# ---------------------------------------------------------------------------
+# Public return type
+# ---------------------------------------------------------------------------
+
+class GuiScenario:
+    """Reconstructed GUI optimizer scenario, ready for in-process introspection.
+
+    Attributes
+    ----------
+    analysis_folder : str
+    work_folder : str
+        Path to the job folder (e.g. ``analysis_folder/optimized/jobs/000``).
+    optimizer : legacy optimizer object
+        Fully constructed and prepared (``prepare_for_optimization`` already
+        called with ``init_params``).
+    init_params : ndarray
+        Physical parameter vector from ``init_params.txt``.
+    fv_init : float
+        Objective value at ``init_params`` as recorded in ``callback.txt``.
+    n_components : int
+    class_code : str
+    """
+
+    def __init__(self, analysis_folder, work_folder, optimizer, init_params,
+                 fv_init, n_components, class_code):
+        self.analysis_folder = analysis_folder
+        self.work_folder = work_folder
+        self.optimizer = optimizer
+        self.init_params = init_params
+        self.fv_init = fv_init
+        self.n_components = n_components
+        self.class_code = class_code
+
+    def eval_init(self):
+        """Evaluate the objective at init_params in-process.
+
+        Returns the reproduced fv_init.  Compare with self.fv_init (from
+        callback.txt) to verify the reconstruction is faithful.
+        """
+        return float(self.optimizer.objective_func(self.init_params))
+
+    def get_score_breakdown(self):
+        """Return score breakdown at init_params."""
+        return self.optimizer.get_score_breakdown(self.init_params)
+
+    def run_inprocess(self, method='BH', niter=20, seed=1234,
+                      analysis_folder=None, **solver_kwargs):
+        """Re-run the optimizer in-process with any solver.
+
[…]
