### 2026-06-10 c5500f7
fix(GuiReplay): get_score_breakdown uses objective_func(return_full=True) directly
The legacy G0346 optimizer object returned by reconstruct_subprocess_optimizer
does not have get_score_breakdown() — that method lives on RunInfo, not on
the legacy objective function class.

Fix: call optimizer.objective_func(params, return_full=True) directly, then
build the {'fv': ..., 'scores': {name: val}} dict from the result.
Returns the same structure as RunInfo.get_score_breakdown().
