### 2026-06-10 e3f0cbb
feat(molass-library#209): compute_rg_curve_from_arrays — rg_curve independent of SSD/SD
Add compute_rg_curve_from_arrays(D, qv, E, jv=None) to RgCurveUtils.py.
Accepts raw numpy arrays (D matrix, q-vector, error matrix, frame indices)
instead of an XrData object — makes rg_curve computation independent of SSD
or SD.

Used by BackRunner.run() to recompute the library-quality rg_curve for LEG-GUI
subprocess runs, closing the Guinier_deviation gap (0.8 fv, ~2 SV) between
LIB-IN and LEG-GUI without requiring the GUI to construct an SSD object.

Also export ip_xr_jv.npy in prepare_rigorous_folders(
[…]
