+++ molass/Guinier/RgCurveUtils.py
@@ -23,6 +23,56 @@ inspect.getfullargspec = _cached_getfullargspec
 
 ADD_ALL_RESULTS = True
 
+def compute_rg_curve_from_arrays(D, qv, E, jv=None, progress_cb=None):
+    """Compute a library-quality RgCurve directly from numpy arrays.
+
+    This is the array-level entry point that makes rg_curve computation
+    independent of SSD or SD.  It uses the same SimpleGuinier pipeline as
+    ``compute_rgcurve_info`` / ``XrData.compute_rgcurve``, but accepts raw
+    arrays instead of an ``XrData`` object.
+
+    This function is used by ``BackRunner.run()`` to compute and export the
+    library rg_curve for LEG-GUI subprocess runs, closing the Guinier_deviation
+    gap between LIB-IN and LEG-GUI without requiring the GUI to construct an SSD.
+
+    Parameters
+    ----------
+    D : ndarray of shape (n_q, n_frames)
+        Corrected XR intensity matrix (e.g. from ``ip_xr_D.npy``).
+    qv : ndarray of shape (n_q,)
+        q-values in Å⁻¹ (e.g. from ``ip_xr_qvector.npy``).
+    E : ndarray of shape (n_q, n_frames)
+        Intensity error matrix (e.g. from ``ip_xr_E.npy``).
+    jv : ndarray of shape (n_frames,) or None
+        Original frame numbers.  If None, uses ``np.arange(n_frames)``.
+    progress_cb : callable or None
+        Same signature as in ``compute_rgcurve_info``.
+
+    Returns
+    -------
+    RgCurve
+        Library ``molass.Guinier.RgCurve.RgCurve`` object.
+    """
+    from molass_legacy.GuinierAnalyzer.SimpleGuinier import SimpleGuinier
+    from molass.Guinier.RgCurve import construct_rgcurve_from_list
+
+    n_frames = D.shape[1]
+    if jv is None:
+        jv = np.arange(n_frames)
+    rg_buffer = np.zeros(n_frames)
+    rginfo_list = []
+    for j in tqdm(range(n_frames)):
+        sg = SimpleGuinier(np.array([qv, D[:, j], E[:, j]]).T)
+        rg = sg.Rg
+        if rg is not None and rg > 0:
+            rg_buffer[j] = rg
+        if progress_cb is not None:
+            progress_cb(rg_buffer, j)
+        if sg.Rg is not None or ADD_ALL_RESULTS:
+            rginfo_list.append((int(jv[j]), sg))
+    return construct_rgcurve_from_list(rginfo_list)
+
+
 def compute_rgcurve_info(xrdata, progress_cb=None):
     """
     Computes Rg curve information from XR data.
