### 2026-06-10 a1e6bce
feat(molass-library#210): SecSaxsData.from_arrays() — array-level SSD constructor
Add @classmethod from_arrays(xr_M, xr_qv, xr_E, xr_jv=None, uv_M=None, ...)
that constructs an SSD directly from numpy arrays without reading from disk.

This is the entry point for the SD → SSD migration path (Phase 1 of the
three-layer refactor: unify SSD/SD → baseline → rg_curve).  Primary use:

  ssd = SecSaxsData.from_arrays(D, qv, E, jv, trimmed=True)
  rg  = ssd.xr.compute_rgcurve()   # library-quality rg_curve, no re-read

Used in molass-researcher/experiments/25_alternative_solvers/25d_path_comparison.ip
[…]
