### 2026-06-10 ca15bca
feat(molass-library#211): SecSaxsData.from_legacy_sd() — Phase 1 SD→SSD bridge
Add @classmethod from_legacy_sd(sd) that constructs an SSD from a legacy
SerialData object.  Reads sd.xray_array (n_q, n_frames, 3), converts to
library XrData layout, and delegates to from_arrays().

Usage (Phase 2 preview):
  ssd_raw      = SecSaxsData.from_legacy_sd(sd)
  ssd_trimmed  = ssd_raw.trimmed_copy()     # library q-trimming: strips beamstop rows
  ssd_corrected = ssd_trimmed.corrected_copy()
  rg = ssd_corrected.get_rg_curve()         # library-quality: 239 vs 121 valid frames
