### 2026-06-10 fd5a0ce
refactor(molass-library, #78): prepare_rigorous_folders writes to rg-curve/ directly
Eliminate rg_curve_parent/ — write rg_curve to rg-curve/ which is the single
authoritative folder read by the subprocess.

Also remove trust_rg_curve_folder setting (no longer needed after #78).
The subprocess trimming check now handles same-session runs correctly and
rejects cross-session mismatches with a clear warning.
