### 2026-06-11 b89cc74
Refactor: migrate GuinierDeviation and RgCurveUtils segment-analysis to molass/Guinier/
- New molass/Guinier/GuinierDeviation.py — canonical GuinierDeviation class
  with w(j) = q(j) * I_norm(j) combined weighting (fixes buffer-frame noise)
- New segment-analysis functions in molass/Guinier/RgCurveUtils.py:
  ValidBools, get_connected_curve_info, get_reconstructed_curve,
  compute_rg_curves, plot_rg_curves, convert_to_milder_qualities
- New molass/Testing/HeadlessPeakEditor.py — canonical headless PeakEditor
  exposed from molass.Testing.__init__
- molass/__init__.py: added __version__ attribu
[…]
