+++ molass/Guinier/RgCurveUtils.py
@@ -115,6 +115,203 @@ def compute_rgcurve_info(xrdata, progress_cb=None):
             rginfo_list.append((int(jv[j]), sg))
     return rginfo_list
 
+# ---------------------------------------------------------------------------
+# Segment-analysis utilities
+# (migrated from molass_legacy/GuinierTools/RgCurveUtils.py)
+# ---------------------------------------------------------------------------
+from collections import namedtuple
+
+# Two boolean masks over the rg_curve domain:
+#   all_frames: mask over the full frame range of rg_curve.x
+#   segment:    mask over concatenated active segments only (shorter)
+ValidBools = namedtuple('ValidBools', ['all_frames', 'segment'])
+
+VALID_QUIALTY_LIMIT = 0.01
+VALID_BASE_QUALITY = 0.3
+
+
+def convert_to_milder_qualities(qualities):
+    """Raise the quality floor so that low-quality frames still get some weight.
+
+    Maps raw qualities (0–1) to a compressed range (VALID_BASE_QUALITY–1):
+        out[i] = 0.3 + 0.7 * qualities[i]   for qualities[i] > 0.01
+
+    .. deprecated::
+        No longer used for Guinier_deviation weights (see molass-legacy issue #11).
+        Raw qualities are now used directly in ``GuinierDeviation``.
+        Kept for backward compatibility (e.g. ``max_mask`` threshold).
+    """
+    ret_qualities = qualities.copy()
+    valid = qualities > VALID_QUIALTY_LIMIT
+    ret_qualities[valid] = VALID_BASE_QUALITY + (1 - VALID_BASE_QUALITY) * qualities[valid]
+    return ret_qualities
+
+
+def get_connected_curve_info(rg_curve, debug=False):
+    """Extract concatenated x, y, Rg, and quality arrays from a segmented Rg curve.
+
+    Works with any object that exposes the segmented Rg interface:
+    ``get_curve_segments()``, ``.qualities``, ``.states``, ``.slices``, ``.x``.
+    This covers ``RgProcess.RgCurve``, ``RgProcess.RgCurveProxy``, and
+    ``Bridge.LegacyRgCurve``.
+
+    Returns
+    -------
+    x_ : ndarray
+        Concatenated frame positions of all active segments.
+    y_ : ndarray
+        Corresponding elution intensities.
+    rgv : ndarray
+        Corresponding smoothed Rg values.
+    qualities : ndarray
+        Concatenated raw Guinier quality scores (one per active frame).
+    valid_bools : ValidBools
+        Named-tuple with two boolean masks:
+        ``all_frames`` (len = len(rg_curve.x)) and
+        ``segment``    (len = len(qualities)).
+    """
+    segments = rg_curve.get_curve_segments()
+    qualities = np.concatenate(rg_curve.qualities)
+
+    x_list = []
+    y_list = []
+    rg_list = []
+    valid_bool_all = np.zeros(len(rg_curve.x), dtype=bool)
+    valid_bool_seg_list = []
+    k = 0
+    for state, slice_ in zip(rg_curve.states, rg_curve.slices):
+        if state == 0:
+            continue
+        x_, y_, rg_ = segments[k]
+        x_list.append(x_)
+        y_list.append(y_)
+        rg_list.append(rg_)
+        bvec = rg_curve.qualities[k] > VALID_QUIALTY_LIMIT
+        valid_bool_all[slice_] = bvec
+        valid_bool_seg_list.append(bvec)
+        k += 1
+
[…]
