### 2026-06-16 f1f393e
feat: GRM end-to-end integration (G1500)
- GrmComponentCurve: fix get_scale_param() — was returning Pe (params[0])
  instead of scale; UV amplitude ~750x too small in load_best().plot_components()
  (molass-library#209)
- GrmEstimator / GrmOptimizer: a_star_mean -> per-component a_star_list
- GrmComponentCurve: add get_scale_param() override returning self.scale
- ComponentCurve.get_scale_param(): add warning docstring for non-EGH subclasses
- Decomposition.optimize_rigorously(): pop model kwarg to avoid duplicate keyword
- ComponentUtils: add get_grm_xr_ccurves() for GRM load_rigorous_result 
[…]
