+++ molass/SEC/Models/GrmLinear.py
@@ -1,0 +1,183 @@
+"""
+SEC.Models.GrmLinear.py
+
+PDF of the General Rate Model (GRM) with a **linear isotherm**, computed via
+characteristic function (CF) inversion using FFT (PDE-based).
+
+The GRM is the most physically detailed of the three linear-isotherm SEC
+transport models available in this package:
+
+  EDM (EdmLinear) — axial dispersion only (k_MT → ∞ limit)
+  LKM (LkmLinear) — axial dispersion + lumped mass transfer
+  GRM (this file)  — axial dispersion + film mass transfer + pore diffusion
+
+Assumptions
+-----------
+- Linear isotherm: q = a_star * c  inside the pore  (no overloading)
+- Axial dispersion parametrised by the Peclet number Pe = u*L/D_ax
+- External (film) mass transfer with rate constant k_ext [length/time]
+- Intraparticle pore diffusion with effective diffusivity D_eff [length²/time]
+- Spherical particles of radius R_p
+- Intraparticle porosity eps_p absorbed into a_star:
+    a_star = eps_p + (1 - eps_p) * a_henry
+  where a_henry is the Henry coefficient for adsorption on the pore wall.
+  For non-porous particles set eps_p = 0; then a_star = a_henry.
+
+Model parameters
+----------------
+Pe      : Peclet number  Pe = u*L/D_ax
+t0      : dead time  t0 = L/u  (mobile-phase transit time)
+k_ext   : external film mass-transfer coefficient [length/time]
+R_p     : particle radius [length]
+D_eff   : effective intraparticle pore diffusivity [length²/time]
+          Large D_eff (film-only limit) → k_MT_eff = 3*F*k_ext / (R_p * R)
+a_star  : effective intraparticle retention parameter (see Assumptions above)
+F_ratio : phase ratio  F = (1 - ε) / ε  (ε = interstitial porosity)
+
+Retention factor
+----------------
+R_eff = 1 + F_ratio * a_star
+
+Transfer function (Laplace domain, Qamar 2014 eqs. 22–24, 37)
+---------------------------------------------------------------
+φ(s) = s + (3*F*k_ext/R_p) * (1 − f(s))
+
+f(s) = B_p / (B_p − 1 + ξ·coth(ξ))
+     where  ξ = sqrt(a_star * s / D_eff) * R_p
+            B_p = k_ext * R_p / D_eff
+
+H(s) = exp( Pe/2 * (1 − sqrt(1 + 4*t0*φ(s)/Pe)) )
+
+Moment formulas (Qamar 2014 Table 1, Dirichlet BCs, linear isotherm)
+----------------------------------------------------------------------
+μ₁     = t0 * R_eff
+σ²_GRM = 2*t0²*R_eff²/Pe  +  2*t0*(R_eff−1)²/(R_eff * k_MT_eff)
+
+where k_MT_eff = 3*F*k_ext / (R_p * R_eff)   [film-only limit, large D_eff]
+
+GRM ↔ LKM moment matching (Qamar 2014, Appendix C)
+----------------------------------------------------
+Given GRM k_MT_eff and R, the equivalent LKM k_MT is:
+  k_MT_LKM = R / (R−1) * k_MT_eff
+
+LKM initialisation from GRM result:
+  k_ext = k_MT_LKM * (R−1)/R * R_p / (3*F/R)
+        = k_MT_LKM * (R−1) * R_p / (3*F)
+
+References
+----------
+- Qamar et al. (2014), Comput. Chem. Eng. 71:383–399 (GRM CF formula)
+- Validated in molass-researcher/experiments/27_qamar_2014_paper/27a, 27b
[…]
