+++ molass/Rigorous/LegacyBridgeUtils.py
@@ -110,6 +110,8 @@ def construct_legacy_optimizer(dsets, baseline_objects, spectral_vectors, num_co
             pass  # CedmParams uses RigorousCedmParams for init; no legacy estimator needed
         elif model == "LKM":
             pass  # LkmParams derives bounds from init params directly; no legacy estimator needed
+        elif model == "GRM":
+            pass  # GrmParams derives bounds from init params directly; no legacy estimator needed
     
     return optimizer
 

[…]
