### 2026-06-17 29a5b81
feat: add refine_edm_per_component (opt-in lighter second XR pass for EDM)
refine_edm_per_component in EdmOptimizer.py:
- Fixes shared column (t0, u, e, Dz) from pass-1 result
- Refines per-component (a, b, cinj) via L-BFGS-B
- position_anchor_scale=1e-5 soft centroid penalty

EDM.py: wired as opt-in via refine_per_component=False (default)
- Enabled with upgrade(model='EDM', refine_per_component=True)
- Default off: SAMPLE4 (4-comp) regresses due to unconstrained b overflow

Benchmark (3-stage vs 4-stage, SAMPLE1-4):
- S1: neutral; S2/S3: fv -96k/-430k improvement; S4: regression (-83->-100 S
[…]
