### 2026-06-17 48acc18
feat: score_initial(), SDM use_guinier_rgs default, SDM optimizer improvements (v1.0.1)
New features:
- Decomposition.score_initial() — evaluate rigorous objective once at initial params,
  returns InitialScoreResult with .sv, .fv, .breakdown, .plot(), .diagnose()
- molass/Rigorous/InitialScore.py — InitialScoreResult class

SDM improvements:
- SDM.py: use_guinier_rgs=False as default for upgrade() (unstable Rg at init time)
- SdmEstimator.py, SdmOptimizer.py: estimator/optimizer refinements

pyproject.toml: v1.0.0 -> v1.0.1, molass_legacy dep >=1.6.7
