+++ molass/SEC/Models/SdmOptimizer.py
@@ -9,6 +9,20 @@ from molass.SEC.Models.SdmMonoPore import (
     DEFAULT_TIMESCALE,
 )
 
+
+def _proxy_rgs_from_peak_frames(decomposition, poresize_ref=100.0,
+                                rho_min=0.15, rho_max=0.35):
+    """Build Rg proxies from peak-frame order when Guinier Rg is disabled."""
+    xr_ccurves = decomposition.xr_ccurves
+    peak_frames = np.array([float(c.x[c.y.argmax()]) for c in xr_ccurves])
+    fmin, fmax = float(np.min(peak_frames)), float(np.max(peak_frames))
+    if abs(fmax - fmin) < 1e-12:
+        rho = np.full(len(peak_frames), (rho_min + rho_max) * 0.5)
+    else:
+        norm = (fmax - peak_frames) / (fmax - fmin)
+        rho = rho_min + norm * (rho_max - rho_min)
+    return np.asarray(rho * poresize_ref, dtype=float)
+
 def optimize_sdm_xr_decomposition(decomposition, env_params, model_params=None, **kwargs):
     """ Optimize the SDM decomposition.
 
@@ -41,12 +55,28 @@ def optimize_sdm_xr_decomposition(decomposition, env_params, model_params=None, 
     xr_icurve = decomposition.xr_icurve
     x, y = xr_icurve.get_xy()
     N, T, me, mp, N0, t0, poresize = env_params
-    rgv = np.asarray(decomposition.get_rgs())
+    # AI-friendliness switch shared with estimator:
+    # when False, avoid using potentially noisy Guinier Rgs as optimization
+    # references and use peak-frame proxy Rgs instead.
+    use_guinier_rgs = kwargs.get('use_guinier_rgs', None)
+    if use_guinier_rgs is None and model_params is not None:
+        use_guinier_rgs = model_params.get('use_guinier_rgs', False)
+    if use_guinier_rgs is None:
+        use_guinier_rgs = False
+    if use_guinier_rgs:
+        rgv = np.asarray(decomposition.get_rgs())
+    else:
+        rgv = _proxy_rgs_from_peak_frames(decomposition)
 
     # Quality-adaptive Rg anchoring: uses Guinier fit quality at each component's
     # peak frame to weight soft constraints in the objective and set hard-bound width.
     rgcurve = kwargs.get('rgcurve')
-    rg_anchor_scale = model_params.get('rg_anchor_scale', 1.0) if model_params else 1.0
+    if model_params and 'rg_anchor_scale' in model_params:
+        rg_anchor_scale = model_params['rg_anchor_scale']
+    else:
+        # In proxy-Rg mode, default anchor is off to avoid pulling results back
+        # toward Guinier-derived references.
+        rg_anchor_scale = 1.0 if use_guinier_rgs else 0.0
     rg_quality_threshold = model_params.get('rg_quality_threshold', 0.7) if model_params else 0.7
     if rgcurve is not None:
         jv_rg = np.array(rgcurve.frames)
@@ -99,6 +129,16 @@ def optimize_sdm_xr_decomposition(decomposition, env_params, model_params=None, 
         rg_diff = np.diff(rgv_)
         non_ordered = np.where(rg_diff > 0)[0]
         order_penalty = np.sum(rg_diff[non_ordered]**2) * 1e3  # penalty for non-ordered rgv
+        
+        # Minimum Rg separation penalty: prevent components from collapsing to identical Rgs
+        # when they're geometrically distinguished (e.g., separate SEC peaks with different Rgs).
+        # This prevents NNLS rank deficiency from driving minority components toward zero.
+        # Target: maintain at least 1 Å separation between adjacent components.
[…]
