### 2026-06-18 48920af
feat: replace detect_peaks with EGH peeling in recommend_decomposition_options
- _count_components_by_egh_peeling(): groups consecutive EGH peaks where
  spacing/sigma_sum < egh_overlap_threshold=1.3 into one component
- Fixes SAMPLE2 false-positive (detect_peaks returned 12 instead of 3)
- Validated safe window: SAMPLE3 max pair ratio=1.232, SAMPLE1 min=1.376
- SAMPLE1=3, SAMPLE2=3, SAMPLE3=1, SAMPLE4=2 with threshold=1.3
- SecSaxsData.recommend_decomposition_options: overlap_threshold/min_prominence
  replaced with egh_overlap_threshold=1.3
- Tests test_23--26 pass
