+++ molass/Decompose/Recommend.py
@@ -1,0 +1,251 @@
+"""
+Recommend decomposition options based on automated peak detection.
+
+This module is the backend for :meth:`SecSaxsData.recommend_decomposition_options`.
+Keeping the logic here allows it to grow (multi-channel heuristics, per-sample
+tuning, custom strategies) without cluttering the top-level data object.
+"""
+import numpy as np
+
+# ---------------------------------------------------------------------------
+# EGH-peeling-based peak counting
+# ---------------------------------------------------------------------------
+
+_EGH_OVERLAP_THRESHOLD = 1.3  # spacing / sigma_sum below this → EGH peaks merge into one cluster
+                               #
+                               # Physical meaning: SAMPLE3 contains a single heavily-overlapping
+                               # species whose elution peak is split by EGH peeling into 3 sub-peaks
+                               # (frames ≈156, 194, 206).  Those sub-peaks are NOT separate components —
+                               # they are artefacts of the sequential-subtraction algorithm fitting a
+                               # single asymmetric peak.  The cluster-merge step folds them back into
+                               # one component, restoring ncomp=1.
+                               #
+                               # Calibration (SAMPLE1–4):
+                               #   SAMPLE3 max pair ratio = 1.232  → merged  → ncomp = 1  ✓
+                               #   SAMPLE1 min pair ratio = 1.376  → kept    → ncomp = 3  ✓
+                               #   SAMPLE2 min pair ratio = 1.423  → kept    → ncomp = 3  ✓
+                               # Safe window: (1.232, 1.376).  1.3 sits near the centre.
+
+# ---------------------------------------------------------------------------
+# Interparticle detection helpers (Guinier low-q suppression test)
+# ---------------------------------------------------------------------------
+
+_LOWQ_SUPPRESSION_THRESHOLD = 0.995   # ratio < this → repulsive interparticle
+_PEAK_FRAMES_THRESHOLD      = 0.1     # fraction of max integrated intensity
+_GUINIER_Q_MAX_RATIO        = 1.3     # q·Rg upper bound for Guinier fit
+
+
+def _peak_frames_matrix(M, threshold=_PEAK_FRAMES_THRESHOLD):
+    """Return M restricted to frames where ∑_q I(q,j) > threshold × max.
+
+    Excludes buffer (non-peak) frames.  After baseline correction those frames
+    carry near-zero signal and only add noise to the SVD.
+    """
+    icurve = M.sum(axis=0)
+    return M[:, icurve > threshold * icurve.max()]
+
+
+def _guinier_func(q2, lnI0, Rg2):
+    """Linearised Guinier model: ln I = ln I₀ − Rg² q² / 3."""
+    return lnI0 - Rg2 * q2 / 3.0
+
+
+def _low_q_suppression_ratio(qv, P1, guinier_q_max_ratio=_GUINIER_Q_MAX_RATIO):
+    """Ratio P(q_min) / P_Guinier(q_min) for the leading SVD scattering profile.
+
+    Fits the Guinier model iteratively in the region q·Rg < *guinier_q_max_ratio*,
+    then compares the observed value at the lowest measured q to the extrapolation.
+
+    Returns
+    -------
+    ratio : float
+        < 1 → low-q suppression → repulsive interparticle effects present.
+        > 1 → no suppression → clean sample.
+        nan → fit failed (too few positive points or convergence error).
+    Rg : float
+        Fitted radius of gyration in the same units as ``qv`` (Å⁻¹ → Å).
+    fit_ok : bool
+    """
+    from scipy.optimize import curve_fit
+
[…]
