### 2026-06-18 c278c30
feat(EDM): two-pass upgrade on by default, b-bounds strategy for 4+ components
- refine_per_component now defaults to True (was False/opt-in)
- EDM.py: updated comment explaining the b-bounds / SAMPLE4 fix
- EdmOptimizer.refine_edm_per_component():
    * n_comp < 4: per-component b ceiling at Pass-1 value (b can decrease, not increase)
    * n_comp >= 4: hard b <= 0 cap (prevents EDM overflow in 4-component systems)
    * b_max kwarg for explicit global override
    * Docstring updated with full strategy explanation
- SdmOptimizer.py: use_guinier_rgs benchmark (SDM upgrade unaffected by rgcurv
[…]
