+++ molass/SEC/Models/EDM.py
@@ -49,11 +49,12 @@ class EDM:
         init_params = estimate_edm_init_params(decomposition, **kwargs)
         # Pass 1 (heavy): full shared-column optimisation — finds (t0, u, e, Dz, a, b, cinj).
         new_xr_ccurves = optimize_edm_xr_decomposition(decomposition, init_params, **kwargs)
-        # Pass 2 (lighter, optional): fix shared column (t0, u, e, Dz), refine per-component (a, b, cinj).
-        # Beneficial for SAMPLE2/SAMPLE3 (massive fv improvement) but regresses SAMPLE4 (4-comp).
-        # SAMPLE4 regression root cause: unconstrained b → EDM overflow for 4-component systems.
-        # Enable via refine_per_component=True once b-bounds / overflow handling is improved.
-        if kwargs.get('refine_per_component', False):
+        # Pass 2 (lighter): fix shared column (t0, u, e, Dz), refine per-component (a, b, cinj).
+        # b bounds: n_comp < 4 → b ≤ Pass-1 value per component; n_comp >= 4 → hard b ≤ 0.
+        # The n_comp >= 4 cap prevents EDM overflow without restricting 2/3-component systems.
+        # Validated cross-sample (SAMPLE1–4): SAMPLE4 SV -83→+79, SAMPLE2/3 fv dramatically better,
+        # SAMPLE1 neutral (−0.05 SV). Opt-out via refine_per_component=False.
+        if kwargs.get('refine_per_component', True):
             x, y = decomposition.xr_icurve.get_xy()
             new_xr_ccurves = refine_edm_per_component(new_xr_ccurves, x, y, **kwargs)
         if decomposition.uv is None:
