### 2026-06-18 19d8fe5
fix: UV scale bounds + cleanup debug prints in Rigorous*Params
UvOptimizer.py: remove hardcoded upper cap 10.0 on UV scales; use
data-driven upper_scale = max(initial_scales) * 3.0.  For datasets
where UV >> XR (e.g. SAMPLE4: UV≈14 OD, XR≈0.5 counts) the old cap
produced saturated scales ~5 instead of the physically correct ~14.

RigorousSdmParams.py: remove debug print statements + dead orig_rg_params.

RigorousEdmParams.py: fix wrong docstring (was copy of EGH header);
add explanation that the code works via polymorphism.
