### 2026-06-19 e5d1895
add proper constraints to EDM
