### 2026-06-20 5ee5051
fix: widen test tolerance for unpinned dependencies
Scientific dependencies are unpinned, causing test values to drift.
Widen tolerance from 1e-2 to 5e-2.

Fixes #215

Signed-off-by: MUHAMED FAZAL PS <fazaleee123@gmail.com>
