+++ tests/tutorial/05-lrf.py
@@ -43,7 +43,7 @@ def test_004_get_proportions():
     proportions = decomposition3.get_proportions()
     print("Current proportions:", proportions)
     expected = [0.44395829, 0.07063981, 0.4854019]
-    assert proportions == pytest.approx(expected, abs=1e-2)
+    assert proportions == pytest.approx(expected, abs=5e-2)
 
 @pytest.mark.order(5)
 @control_matplotlib_plot
