### 2026-06-20 d779119
fix: add CI, community files, and package metadata polish (#216 #217 #218)
