### 2026-06-20 0eb2739
fix: expose __version__ at module level (joss-paper branch)
