### 2026-06-20 4aea035
fix: add missing shapely dependency, remove duplicate numba
