### 2026-06-20 c356363
fix: require molass_legacy>=1.6.8 for Windows windll fix (#214)
