+++ _sources/source/molass.Rigorous.ComparePaths.rst
@@ -1,0 +1,7 @@
+Rigorous.ComparePaths
+===================================
+
+.. automodule:: molass.Rigorous.ComparePaths
+   :members:
+   :undoc-members:
+   :show-inheritance:
