+++ _sources/source/molass.Rigorous.RigorousLkmParams.rst
@@ -1,0 +1,7 @@
+Rigorous.RigorousLkmParams
+========================================
+
+.. automodule:: molass.Rigorous.RigorousLkmParams
+   :members:
+   :undoc-members:
+   :show-inheritance:
