+++ _sources/source/molass.Peaks.rst
@@ -12,6 +12,7 @@ Submodules
 .. toctree::
    :maxdepth: 4
 
+   molass.Peaks.EghPeeler
    molass.Peaks.PeakSimilarity
    molass.Peaks.Recognizer
    molass.Peaks.RecognizerSpecific
