+++ _sources/source/molass.Testing.rst
@@ -12,4 +12,5 @@ Submodules
 .. toctree::
    :maxdepth: 4
 
+   molass.Testing.HeadlessPeakEditor
    molass.Testing.PlotControl
