+++ _sources/source/molass.Solvers.CMA.rst
@@ -1,0 +1,15 @@
+Solvers.CMA
+==========================
+
+.. automodule:: molass.Solvers.CMA
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   molass.Solvers.CMA.SolverCMA
