+++ _sources/source/molass.Decompose.Recommend.rst
@@ -1,0 +1,7 @@
+Decompose.Recommend
+=================================
+
+.. automodule:: molass.Decompose.Recommend
+   :members:
+   :undoc-members:
+   :show-inheritance:
