### 2026-06-22 7af512d
CI: Add cross-platform testing (Windows+Ubuntu) and Python 3.14 support
- Test matrix: 2 OS × 3 Python versions (3.12, 3.13, 3.14)
- Addresses issue #214 (Windows compatibility)
- Forward compatibility with Python 3.14
- Add workflow_dispatch for manual triggering
