### 2026-06-22 9c6bd4a
CI: Add molass-legacy dependency to tests workflow
- Install molass-legacy from PyPI in CI
- Fixes test failures due to missing pythonpath dependency
- Required for tutorial tests that use rigorous optimization
