+++ .github/workflows/tests.yml
@@ -27,7 +27,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -e ".[testing]"
-          pip install molass_data
+          pip install molass_data molass-legacy
 
       - name: Run tutorial tests
         run: |
