### 2026-06-22 d2e9e0c
CI: Require molass-legacy>=1.6.9 for cross-platform fix
- molass-legacy 1.6.9 includes platform check for Windows-only windll
- Fixes AttributeError on Linux/macOS CI environments
- Completes fix for JOSS review issue #214 (cross-platform compatibility)
