### 2026-06-22 07fac9b
CI: Require molass-legacy>=1.6.10 for complete cross-platform fix
- 1.6.10 includes HookableMessageBox.py platform check
- Completes #214 fix (both CustomMessageBox and HookableMessageBox)
- All 6 CI jobs now passing (Ubuntu + Windows × Python 3.12/3.13/3.14)
