### 2026-06-22 94ec1f8
update to make it runnable in an isolated rum
