### 2026-06-23 a4a1d48
Add AI model development disclosure to paper.md
Added new bullet point to AI usage disclosure section documenting AI's role in LKM/GRM development.

Rationale: The Impact Evidence section (in the JOSS response draft) now explicitly states that LKM and GRM were developed through AI-assisted collaboration, which would have been 'impossible (or extremely inefficient) with GUI-only code'. The paper's AI disclosure must reflect this reality.

New disclosure bullet:
- Model development: AI collaboration integral to rapid LKM/GRM development (2026)
- Library's scriptable architecture enabled iterativ
[…]
