### 2026-06-23 b2ccb35
Add GRM reference (Qamar 2014) and note LKM reference TBD
Two model references updated in paper.md:
- LKM: Added '(reference TBD)' note - to be discussed later
- GRM: Added [@Qamar2014] citation

Added Qamar2014 entry to paper.bib:
- Placeholder entry for General Rate Model paper (2014)
- TODO note: Complete bibliographic details from reference_papers/2014, Shamsul Qamar.pdf
- Journal, volume, pages, DOI need to be extracted from PDF

Next step: Extract full bibliographic details from the PDF and complete the Qamar2014 entry. For LKM, need to identify appropriate reference paper (to be discuss
[…]
