+++ paper.bib
@@ -348,6 +348,15 @@ pages={2069-2080},
 doi={10.2298/TSCI201229179U},
 url= {https://doi.org/10.2298/TSCI201229179U},
 }
+
+@article{Qamar2014,
+author={Qamar, Shamsul},
+title={General Rate Model for column liquid chromatography: Analysis, parametrization and optimization},
+journal={TBD},
+year={2014},
+note={TODO: Complete bibliographic details from reference_papers/2014, Shamsul Qamar.pdf}
+}
+
 @article{Panjkovich:2018,
 author = "Panjkovich, Alejandro and Svergun, Dmitri I.",
 title = "{CHROMIXS: automatic and interactive analysis of chromatography-coupled small-angle X-ray scattering data}",
