### 2026-06-23 3130945
Add LKM reference (Felinger 2004)
Complete bibliographic entry for Lumped Kinetic Model:

@Article{Felinger2004}:
- Title: Equivalence of the microscopic and macroscopic models of chromatography: stochastic-dispersive versus lumped kinetic model
- Authors: Attila Felinger, Alberto Cavazzini, Francesco Dondi
- Journal: Journal of Chromatography A, 1043 (2004) 149-157
- DOI: 10.1016/j.chroma.2004.05.081

This paper explicitly discusses LKM and establishes its equivalence to SDM. Same first author (Felinger) as the SDM reference already cited (Felinger1999).

Both LKM and GRM references now compl
[…]
