+++ paper.bib
@@ -333,6 +333,19 @@ doi={10.1021/ac990412u},
 url={https://doi.org/10.1021/ac990412u}
 }
 
+@Article{Felinger2004,
+author={Felinger, Attila
+and Cavazzini, Alberto
+and Dondi, Francesco},
+title={Equivalence of the microscopic and macroscopic models of chromatography: stochastic–dispersive versus lumped kinetic model},
+journal={Journal of Chromatography A},
+year={2004},
+volume={1043},
+pages={149-157},
+doi={10.1016/j.chroma.2004.05.081},
+url={https://doi.org/10.1016/j.chroma.2004.05.081}
+}
+
 @Article{Ur2021,
 author={Ur Rehman, Jamil
 and Muneer, Adeel 
