### 2026-06-23 41d5a16
Complete GRM reference (Qamar 2014 preprint)
Updated Qamar2014 entry in paper.bib with full details from PDF:

@unpublished{Qamar2014}:
- Title: Analytical solutions and moment analysis of general rate model for linear liquid chromatography
- Authors: Shamsul Qamar, Javeria Nawaz Abbasi, Shumaila Javeed, Andreas Seidel-Morgenstern
- Year: 2014
- Type: Preprint submitted to Elsevier, August 2014
- Institution: Max Planck Institute for Dynamics of Complex Technical Systems, Magdeburg

Note: This is cited as a preprint. The document states 'Preprint submitted to Elsevier August 12, 2014' but no f
[…]
