+++ paper.bib
@@ -362,12 +362,11 @@ doi={10.2298/TSCI201229179U},
 url= {https://doi.org/10.2298/TSCI201229179U},
 }
 
-@article{Qamar2014,
-author={Qamar, Shamsul},
-title={General Rate Model for column liquid chromatography: Analysis, parametrization and optimization},
-journal={TBD},
+@unpublished{Qamar2014,
+author={Qamar, Shamsul and Abbasi, Javeria Nawaz and Javeed, Shumaila and Seidel-Morgenstern, Andreas},
+title={Analytical solutions and moment analysis of general rate model for linear liquid chromatography},
 year={2014},
-note={TODO: Complete bibliographic details from reference_papers/2014, Shamsul Qamar.pdf}
+note={Preprint submitted to Elsevier, August 2014. Max Planck Institute for Dynamics of Complex Technical Systems, Magdeburg, Germany}
 }
 
 @article{Panjkovich:2018,
