### 2026-06-23 af49a78
docs: fix README links and remove Discussions reference
- Remove GitHub Discussions link (not enabled)
- Consolidate support to GitHub Issues only
- Add link to AI Context Standard repo
- Remove outdated version number (v0.7)
- Fix Quick Start path (chapters/00/ not 01/)
- Fix copilot-guidelines branch (main not master)
