### 2026-06-23 389d0d0
docs: fix Quick Start link (underscore not hyphen)
