### 2026-06-24 0df201a
Complete paper.md revisions for KedoKudo response + Friday meeting prep
- paper.md: All 6 promised revisions complete (molass_legacy dependency, Research Impact Statement, DENSS bundling, etc.)
- paper-revision-status.md: Tracking document for promised changes
- FRIDAY-MEETING-AGENDA-20260626.md: Strategic analysis for main->joss-paper merge timing
