### 2026-06-25 b7e3b85
docs: clarify EDM/LKM/GRM hierarchy vs implementation scope in paper.md
