### 2026-06-25 dfdf427
docs: update Qamar2014 citation from preprint to published version
