+++ paper.bib
@@ -362,11 +362,16 @@ doi={10.2298/TSCI201229179U},
 url= {https://doi.org/10.2298/TSCI201229179U},
 }
 
-@unpublished{Qamar2014,
+@article{Qamar2014,
 author={Qamar, Shamsul and Abbasi, Javeria Nawaz and Javeed, Shumaila and Seidel-Morgenstern, Andreas},
 title={Analytical solutions and moment analysis of general rate model for linear liquid chromatography},
+journal={Chemical Engineering Science},
 year={2014},
-note={Preprint submitted to Elsevier, August 2014. Max Planck Institute for Dynamics of Complex Technical Systems, Magdeburg, Germany}
+volume={107},
+pages={192-205},
+issn={0009-2509},
+doi={10.1016/j.ces.2013.12.019},
+url={https://www.sciencedirect.com/science/article/pii/S0009250913008142}
 }
 
 @article{Panjkovich:2018,
