### 2026-06-25 94939cc
docs: fix Qamar 2014 citation in GrmLinear.py (Chem. Eng. Sci. not Comput. Chem. Eng.)
