### 2026-06-25 dccf87d
docs: add Felinger 2004 reference and SDM≡LKM equivalence note to LkmLinear.py
