### 2026-06-26 ae79bb0
docs: revise Research Impact Statement - library architecture benefits
- Clarify existing citations are for GUI version (predecessor)
- Add second paragraph: AI-assisted method development workflow
- Concrete validation: LKM/GRM development (impractical in legacy codebase)
- Removed private repo reference (molass-researcher)
- Removed redundant third paragraph (defensive tone)

Outcome: Clean two-paragraph statement showing what library architecture enables beyond GUI version.
