+++ _sources/source/molass.Rigorous.RigorousCedmParams.rst
@@ -3,5 +3,5 @@ Rigorous.RigorousCedmParams
 
 .. automodule:: molass.Rigorous.RigorousCedmParams
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
