+++ _sources/source/molass.Solvers.BH.BasinHopping.rst
@@ -3,5 +3,5 @@ Solvers.BH.BasinHopping
 
 .. automodule:: molass.Solvers.BH.BasinHopping
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
