+++ _sources/source/molass.Solvers.DE.SolverDE.rst
@@ -3,5 +3,5 @@ Solvers.DE.SolverDE
 
 .. automodule:: molass.Solvers.DE.SolverDE
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
