+++ _sources/source/molass.Rigorous.GuiReplay.rst
@@ -3,5 +3,5 @@ Rigorous.GuiReplay
 
 .. automodule:: molass.Rigorous.GuiReplay
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
