+++ _sources/source/molass.Rigorous.FunctionCodeUtils.rst
@@ -3,5 +3,5 @@ Rigorous.FunctionCodeUtils
 
 .. automodule:: molass.Rigorous.FunctionCodeUtils
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
