+++ _sources/source/molass.Solvers.BH.SolverBH.rst
@@ -3,5 +3,5 @@ Solvers.BH.SolverBH
 
 .. automodule:: molass.Solvers.BH.SolverBH
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
