+++ _sources/source/molass.SEC.Models.GrmLinear.rst
@@ -3,5 +3,5 @@ SEC.Models.GrmLinear
 
 .. automodule:: molass.SEC.Models.GrmLinear
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
