+++ _sources/source/molass.SEC.Models.LkmOptimizer.rst
@@ -3,5 +3,5 @@ SEC.Models.LkmOptimizer
 
 .. automodule:: molass.SEC.Models.LkmOptimizer
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
