+++ _sources/source/molass.SEC.Models.LKM.rst
@@ -3,5 +3,5 @@ SEC.Models.LKM
 
 .. automodule:: molass.SEC.Models.LKM
    :members:
-   :undoc-members:
    :show-inheritance:
+   :undoc-members:
