### 2026-06-29 c70e260
Document KedoKudo first pass responses and fix paper.md typo
- Created PAPER-UPDATE-POINTS.md tracking all 3 minor issues
- Minor Issue 1 (Author Contributions): Already addressed in review thread
- Minor Issue 2 (Research Impact): First para revised + LKM/GRM paragraph added
- Minor Issue 3 (Architecture/Dependency): Summary softened, molass_legacy explained, 5 models documented, DENSS bundling corrected
- Fixed typo in paper.md: 'algoirthm' -> 'algorithm'
- Added scikit-learn to dependencies list (bonus improvement)
