### 2026-06-29 91b1b97
docs: Response draft for KedoKudo review - paper updates complete
