### 2026-06-30 ffcb43a
docs: add research impact expectation for physical models
