### 2026-06-30 b7f6ae3
fix: use LaTeX subset symbols for proper PDF rendering
