### 2026-06-30 fcfd724
docs: fix Quick Start link in README (MyST URL structure)
