### 2026-06-30 e50a536
Revise MOLASS impact statement and library description
Updated the description of the MOLASS legacy code and its transition to the Molass Library, highlighting the benefits of a scriptable, notebook-based approach for SEC-SAXS analysis.
