### 2026-06-30 e4d66eb
Convert mathematical notation to LaTeX in State of the field section
