+++ paper.md
@@ -56,11 +56,11 @@ Molass Library currently implements steps 3–7. For steps 1 and 2, users may em
 
 # State of the field
 
-Several software tools address SEC-SAXS data processing. The proprietary ATSAS suite [@Manalastas-Cantos:ge5081] and the open-source BioXTAS RAW [@Hopkins:jl5075] provide comprehensive GUI-based workflows widely used in the field. For the specific challenge of overlapping peak decomposition—required when chromatographic peaks overlap or significant interparticle interference occurs—CHROMIXS defers such analysis to "other methods" [@Panjkovich:2018], whereas EFAMIX defines quantitative conditions for reliable separation (signal-to-noise ratio ≥10³ for three-component separation; peak asymmetry parameter τ≤2; baseline width separation ≥2× for reliable decomposition) [@Konarev:2021]. REGALS employs a two-stage framework combining EFA with L1 regularization [@Meisburger:mf5050] and is therefore guided by assumptions underlying EFA in strongly overlapping cases. Molass Library takes a complementary approach: rather than extending these existing GUI-based tools, it provides a Python-first, Jupyter-centric platform in which overlapping peak decomposition is handled through explicit parametric elution curve modeling combined with global optimization, constrained by physicochemical consistency derived from SEC theory (size-dependent elution behavior) and SAXS analysis (radius of gyration consistency across elution fractions).
+Several software tools address SEC-SAXS data processing. The proprietary ATSAS suite [@Manalastas-Cantos:ge5081] and the open-source BioXTAS RAW [@Hopkins:jl5075] provide comprehensive GUI-based workflows widely used in the field. For the specific challenge of overlapping peak decomposition—required when chromatographic peaks overlap or significant interparticle interference occurs—CHROMIXS defers such analysis to "other methods" [@Panjkovich:2018], whereas EFAMIX defines quantitative conditions for reliable separation (signal-to-noise ratio $\geq 10^3$ for three-component separation; peak asymmetry parameter $\tau \leq 2$; baseline width separation $\geq 2\times$ for reliable decomposition) [@Konarev:2021]. REGALS employs a two-stage framework combining EFA with L1 regularization [@Meisburger:mf5050] and is therefore guided by assumptions underlying EFA in strongly overlapping cases. Molass Library takes a complementary approach: rather than extending these existing GUI-based tools, it provides a Python-first, Jupyter-centric platform in which overlapping peak decomposition is handled through explicit parametric elution curve modeling combined with global optimization, constrained by physicochemical consistency derived from SEC theory (size-dependent elution behavior) and SAXS analysis (radius of gyration consistency across elution fractions).
 
 # Software Design
 
-The MOLASS library is designed to support sustainable software development through an open framework that enables broad community participation and contribution. This design philosophy prioritizes explicit, readable code over graphical interfaces, making the analysis methodology transparent for both learning and AI-assisted maintenance. We are actively enhancing AI-readiness through systematic usability testing with AI agents and iterative improvements to API discoverability and inline documentation. The architecture emphasizes modularity and clear separation of concerns: elution curve models (EGH, SDM, EDM, etc.) are isolated as parametric functions; low-rank factorization uses standard linear algebra (Moore-Penrose pseudoinverse); and visualization is decoupled from computation. By integrating established packages (NumPy, SciPy, pybaselines, ruptures) rather than reimplementing core functionality, we reduce custom code volume and enhance long-term viability. This approach allows domain researchers to maintain and extend the code using AI-assisted development tools.
[…]
