### 2026-06-30 472c468
Fix Konarev 2021 citation: correct SNR criterion from 'three-component' to 'multiple components'
- paper.md: Changed 'for three-component separation' to 'for reliable separation of multiple components' (line 59)
- Konarev 2021 shows three-component systems work at 10^2 photon counts; 10^3 threshold applies to four-component systems
- Updated response draft to document this correction
