### 2026-06-30 eff614b
Clarify Konarev 2021 peak separation criterion: 'baseline width separation' → 'peak center-to-center distance ≥ (2× peak width)'
Verified against Konarev 2021 lines 737-738: 'the distance between peak maxima is smaller than twice the individual peak width'.

Changes:
- paper.md line 59: Replaced ambiguous 'baseline width separation' with explicit 'peak center-to-center distance ≥ (2× peak width)'
- response draft: Documented this clarification in Section 3
