### 2026-07-01 0eb7371
Refine text for clarity in paper.md
Reworded sections for clarity and consistency in the Molass Library description and acknowledgements.
