### 2026-07-01 7d41c9a
fix: resolve merge conflict markers (anomaly_mask naming)
