+++ molass/Bridge/SdProxy.py
@@ -16,15 +16,9 @@ def _resolve_neg_peak_exclude(xr):
 
     When auto-detecting, frames where the recognition curve is negative are flagged.
     """
-<<<<<<< HEAD
-    if not getattr(xr, 'allow_negative_peaks', False):
-        return None
-    np_mask = getattr(xr, 'negative_peak_mask', None)
-=======
     if not getattr(xr, 'has_anomaly_mask', False):
         return None
     np_mask = getattr(xr, 'anomaly_mask', None)
->>>>>>> 53675d65bb5c75a3c302a9677d2fa94ef773bd9a
     jv = xr.jv
     if np_mask is None:
         return xr.get_recognition_curve().y < 0
