+++ molass/DataObjects/SsMatrixData.py
@@ -27,14 +27,6 @@ class SsMatrixData:
     baseline_method : str
         The method used for baseline correction. Default is 'linear'.
     allow_negative_peaks : bool
-<<<<<<< HEAD
-        If True, frames in the negative-peak region are excluded from LPM's
-        anchor pool before baseline fitting, preventing contamination from
-        physically real negative-peak frames.  Default False.
-    negative_peak_mask : array-like of bool, slice, or None
-        Explicit mask of frames to exclude from LPM fitting when
-        ``allow_negative_peaks=True``.  If None (default), the mask is
-=======
         Deprecated alias for ``has_anomaly_mask``.
     has_anomaly_mask : bool
         If True, frames in the anomaly region are excluded from LPM's
@@ -46,7 +38,6 @@ class SsMatrixData:
     anomaly_mask : array-like of bool, slice, or None
         Explicit mask of frames to exclude from LPM fitting when
         ``has_anomaly_mask=True``.  If None (default), the mask is
->>>>>>> 53675d65bb5c75a3c302a9677d2fa94ef773bd9a
         derived automatically from the recognition curve (frames where y < 0).
     """
     def __init__(self, M, iv, jv, E=None,
@@ -76,13 +67,8 @@ class SsMatrixData:
         self.E = E      # may be None
         self.moment = moment
         self.baseline_method = baseline_method
-<<<<<<< HEAD
-        self.allow_negative_peaks = allow_negative_peaks
-        self.negative_peak_mask = negative_peak_mask  # None = auto-detect from recognition curve
-=======
         self.has_anomaly_mask = allow_negative_peaks
         self.anomaly_mask = negative_peak_mask  # None = auto-detect from recognition curve
->>>>>>> 53675d65bb5c75a3c302a9677d2fa94ef773bd9a
 
     @property
     def data(self):
@@ -133,13 +119,8 @@ class SsMatrixData:
                             Ecopy,
                             moment=None,  # note that moment is not copied
                             baseline_method=self.baseline_method,
-<<<<<<< HEAD
-                            allow_negative_peaks=self.allow_negative_peaks,
-                            negative_peak_mask=self.negative_peak_mask,
-=======
                             allow_negative_peaks=self.has_anomaly_mask,
                             negative_peak_mask=self.anomaly_mask,
->>>>>>> 53675d65bb5c75a3c302a9677d2fa94ef773bd9a
                             )
 
     def get_icurve(self, pickat):
@@ -214,55 +195,26 @@ class SsMatrixData:
         """Get the baseline method for this data object."""
         return self.baseline_method
 
-<<<<<<< HEAD
-    def set_allow_negative_peaks(self, value=True, mask=None):
-        """Declare that this dataset contains physically real negative peaks.
-
-        When True, ``get_baseline2d()`` excludes negative-peak frames from
-=======
     def set_anomaly_mask(self, mask=None):
         """Declare that this dataset contains anomalous frames to exclude from baseline fitting.
 
         When set, ``get_baseline2d()`` excludes the specified frames from
->>>>>>> 53675d65bb5c75a3c302a9677d2fa94ef773bd9a
         LPM's anchor pool before baseline fitting.  LPM itself is unchanged;
         only the frame set it operates on is filtered.
 
         Parameters
         ----------
-<<<<<<< HEAD
-        value : bool, optional
-            Default True.
-=======
->>>>>>> 53675d65bb5c75a3c302a9677d2fa94ef773bd9a
         mask : array-like of bool, slice, or None, optional
             Explicit mask of frames to exclude (True = exclude).  If None
             (default), the mask is derived automatically at fitting time from
[…]
