### 2026-07-01 9325819
fix: validate ranks length in update_xr_ranks()
Prevents IndexError when ranks list length doesn't match num_components.
Added clear error message explaining expected format.

Fixes issue where update_xr_ranks([2]) on 3-component decomposition
caused IndexError at line 618 in get_xr_components().
