### 2026-07-03 9a7611d
fix: update num_components when peak recognition finds fewer peaks
When peak recognition returns fewer peaks than requested, num_components
must be updated to match the actual peak count. Otherwise, bounds array
(sized by num_components) mismatches init_params (sized by actual peaks),
causing ValueError in scipy.optimize.minimize.

Fixes scipy bounds error in:
- correction_survey.ipynb
- folder_walker.ipynb

Issue: molass-technical notebook execution migration
