### 2026-07-03 ab6b04e
chore: remove obsolete diagnostic/utility scripts
- _check_tau.py: temporary EghPeeler diagnostic with hardcoded Dropbox paths
- read_nb_outputs.py: one-off notebook parser for deleted scipy_inspect_issue.ipynb
