### 2026-07-03 6eee7d5
chore: remove obsolete root-level files
- _issue_cma_async_crash.md: issue #193 investigation notes (now documented in copilot-instructions.md)
- example_test.py: old test example (9 months, superseded by test suite docs)
- run_rg_test.py: quick smoke test (likely issue #168, functionality covered by test suite)
