### 2026-07-03 3f87da8
docs: improve Sphinx conf.py (version, type hints, intersphinx, theme)
