+++ docs/conf.py
@@ -19,6 +19,11 @@ project = 'Molass Library'
 copyright = '2025, Molass Community'
 author = 'Molass Community'
 
+# Get version from package
+from molass import get_version
+release = str(get_version(toml_only=True))
+version = '.'.join(release.split('.')[:2])  # X.Y from X.Y.Z
+
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
@@ -30,6 +35,10 @@ extensions = [
     'myst_parser',
     "sphinx.ext.intersphinx",  # <-- Add this line
 ]
+
+# Type hints configuration
+autodoc_typehints = 'description'
+autodoc_typehints_description_target = 'documented'
 
 autoclass_content = 'both'
 # napoleon_google_docstring = True
@@ -58,6 +67,9 @@ add_module_names = False  # Removes the "molass." prefix from module names in ti
 html_theme_options = {
     "repository_url": "https://github.com/biosaxs-dev/molass-library",
     "use_repository_button": True,
+    "show_toc_level": 2,
+    "navigation_depth": 4,
+    "pygment_dark_style": "monokai",
 }
 
 html_theme = 'sphinx_book_theme'
@@ -67,6 +79,9 @@ html_favicon = "_static/molamola.png"
 
 # Intersphinx mapping for external documentation cross-references
 intersphinx_mapping = {
+    "python": ("https://docs.python.org/3", None),
+    "numpy": ("https://numpy.org/doc/stable/", None),
     "matplotlib": ("https://matplotlib.org/stable/", None),
+    "scipy": ("https://docs.scipy.org/doc/scipy/", None),
 }
 
