### 2026-07-06 e42f445
fix: preserve _rgcurve cache in upgrade() (issue #220)
- Modified copy_with_new_components() to automatically transfer _rgcurve
- Added test_030_upgrade_preserves_rgcurve to verify behavior
- Eliminates redundant Rg curve computation during model comparisons
- Aligns with expensive-object caching pattern from issue #168
