+++ tests/generic/200_LRF/test_010_LRF.py
@@ -30,6 +30,37 @@ def test_020_num_components():
     decomposition = ssd.quick_decomposition(num_components=3)
     decomposition.plot_components(debug=is_interactive())
 
+def test_030_upgrade_preserves_rgcurve():
+    """Test that upgrade() preserves the _rgcurve cache (issue #220).
+    
+    The Rg curve is expensive to compute. When upgrading from EGH to a physical
+    model (SDM/EDM/LKM/GRM), the cache should be preserved automatically so
+    subsequent score_initial() calls don't recompute it.
+    """
+    ssd = corrected_ssd_instance
+    
+    # Get Rg curve and create initial decomposition
+    rgcurve = ssd.get_rg_curve()
+    decomp_egh = ssd.quick_decomposition(num_components=3, rgcurve=rgcurve)
+    
+    # Verify EGH has cached rgcurve
+    assert hasattr(decomp_egh, '_rgcurve')
+    assert decomp_egh._rgcurve is not None
+    egh_rgcurve_id = id(decomp_egh._rgcurve)
+    
+    # Upgrade to SDM
+    decomp_sdm = decomp_egh.upgrade(model='SDM')
+    
+    # Verify SDM preserved the cache
+    assert hasattr(decomp_sdm, '_rgcurve'), "upgrade() should preserve _rgcurve cache"
+    assert decomp_sdm._rgcurve is not None, "_rgcurve should not be None after upgrade"
+    assert id(decomp_sdm._rgcurve) == egh_rgcurve_id, "_rgcurve should be the same object"
+    
+    # Verify it works for other models too
+    decomp_edm = decomp_egh.upgrade(model='EDM')
+    assert hasattr(decomp_edm, '_rgcurve')
+    assert id(decomp_edm._rgcurve) == egh_rgcurve_id
+
 if __name__ == "__main__":
     test_010_default()
     # plt.show()
