### 2026-07-07 cdba3e2
feat: Add RunInfo.score_optimized() for symmetric score visualization API (#221)
