### 2026-07-07 98f67de
refactor: Symmetric score API - rename InitialScoreResult→Score, score_initial/optimized→score (#221)
- Rename class InitialScoreResult → Score (backward compat alias kept)
- Rename Decomposition.score_initial() → score() (deprecated alias kept)
- Rename RunInfo.score_optimized() → score() (deprecated alias kept)
- Update all docstrings and examples
- Update test_110_score_optimized.py to use new names
- Context is encoded in object type, not method name
