### 2026-07-07 40af98d
fix: Remove stray docstring marker in test_110_score_optimized.py
