### 2026-07-07 698ce08
refactor: Rename InitialScore.py → Score.py, make_initial_score_impl → make_score_impl
- File rename preserves git history
- Function rename completes the symmetry refactoring
- All imports updated (Decomposition, RunInfo, tests)
- Backward compatibility aliases for both names
