### 2026-07-07 63aae6b
fix: test_score_symmetric_api was checking class instead of instances (#221)
