### 2026-07-07 b398ca2
fix: test_score_symmetric_api - remove nonexistent trimmed_ssd attribute
decomp.score() has trimmed_ssd=None default and auto-creates it,
so no argument needed. RunInfo doesn't expose trimmed_ssd.
