### 2026-07-07 2ca4566
docs: add memory maintenance workflow to workflow_notes.md
Principle: When introducing code deprecations or API refactoring, always
check and update user memory (/memories/) to prevent AI assistants from
using outdated patterns.

Workflow:
1. Search memory: grep -r 'deprecated_api' /memories/
2. Update affected files
3. Document change with date/reason
4. Route pattern to memory (preferences) vs git docs (API info)

Recent example: score_initial/score_optimized deprecation (2026-07-07)
- Memory check: no references found
- Git docs: workflow_notes.md documents aliases
- Result: AI assistants d
[…]
