### 2026-07-08 14aa7cc
Phase 2E: GRM c_inj embedded scale implementation
Add c_inj parameter to GRM model (mirrors LKM implementation):
- GrmLinear.py: Add c_inj, t_inj parameters to grm_pdf function
- GrmComponentCurve.py: Replace scale with c_inj, update get_scale_param()
- GrmEstimator.py: Return c_inj instead of scale_list
- GrmOptimizer.py: Use c_inj for NNLS basis and component curves
- CurrentStateUtils.py: Add 'grm' to embedded scale branch

Completes unified scale pattern refactoring for all 5 models.
