### 2026-07-08 204fc62
Fix LKM/GRM rigorous params: use get_scale_param() instead of .scale
After Phase 2 refactoring, LkmComponentCurve and GrmComponentCurve no
longer have a .scale attribute (replaced with c_inj). RigorousLkmParams
and RigorousGrmParams were still accessing ccurve.scale, causing
AttributeError.

Fix: Use ccurve.get_scale_param() which returns c_inj for LKM/GRM.

Fixes error in 29a_sample1_five_models.ipynb cell 22 (LKM score()).
