### 2026-07-08 f3da859
Fix LKM/GRM rigorous params: add c_inj to column params
After Phase 2 refactoring, legacy LkmParams/GrmParams expect c_inj in
the column parameter layout, but RigorousLkmParams/RigorousGrmParams
were not updated to include it.

Legacy expects:
- LKM: [Pe, t0, c_inj, R_0, k_MT_0, ...] (3 + 2*nc params)
- GRM: [Pe, t0, R_p, D_eff, c_inj, R_0, k_ext_0, ...] (5 + 2*nc params)

Library was sending:
- LKM: [Pe, t0, R_0, k_MT_0, ...] (2 + 2*nc) ❌
- GRM: [Pe, t0, R_p, D_eff, R_0, k_ext_0, ...] (4 + 2*nc) ❌

Fix: Extract c_inj from first component and add to column params.

Fixes AssertionError in 29a 
[…]
