### 2026-07-09 2661101
UV/XR scale ratio unification (Fixes 1-4)
Architecture unification across all 5 elution models:
- UV parameters now represent ε_i/k ratios (species properties)
- XR parameters represent c_inj (injection concentrations)
- Physics: UV_i = (ε_i/k) × XR_i where ratios are preserved

Phase 5: UV optimizer ratio preservation (Fix 1)
- UvOptimizer.py: compute UV/XR ratios from old decomp, apply to new XR scales
- SDM.py, LKM.py, GRM.py, EDM.py: added preserve_ratios=True parameter

Phase 6: Model upgrades use ratio-preserving UV optimizer
- All optimize_uv_decomposition calls pass preserve_ratios=Tru
[…]
