### 2026-07-10 674ee82
Fix LKM/GRM UV scale bugs; fix DE budget, model= error, n_callbacks rename
- RigorousLkmParams.py, RigorousGrmParams.py: remove /xr_params division (UV ratio=1.0)
- UvOptimizer.py: use get_scale_param() instead of .scale in step 2
- SolverDE.py: FEVALS_PER_NITER 200→1000 (issue #229, fair BH comparison)
- Decomposition.py: optimize_rigorously(model=...) now raises TypeError (issue #230)
- RunInfo.py, CurrentStateUtils.py: n_evals→n_callbacks with deprecated alias (issue #231)
- test_plot_convergence.py: updated assertions for n_callbacks
- Copilot docs: DESIGN/MIGRATION/ASSESSMENT updated for 
[…]
