### 2026-07-10 50955a9
Add pymoo to dependencies for DE solver support
