+++ pyproject.toml
@@ -25,6 +25,7 @@ dependencies = [
   "ipywidgets",
   "ipykernel",
   "cma",
+  "pymoo",
   "shapely",
 ]
 
