### 2026-07-10 9835d58
Switch DE solver from pymoo to scipy.optimize.differential_evolution
