+++ molass/Solvers/DE/SolverDE.py
@@ -1,7 +1,7 @@
 """
     molass.Solvers.DE.SolverDE
 
-    Differential Evolution solver wrapping pymoo.
+    Differential Evolution solver using scipy.optimize.differential_evolution.
 
     Drop-in replacement for SolverBH / SolverCMA in the rigorous optimization
     pipeline. Interface matches SolverCMA.minimize() so it can be wired into
@@ -17,10 +17,10 @@
         max_fevals = niter * FEVALS_PER_NITER
         max_gen    = max_fevals // pop_size   (≥ 1)
 
-    With the default pop_size=None the population is sized automatically as
-    max(20, 5 * n_var) — i.e. 150 for a 30-param problem, 250 for 50 params.
+    With the default pop_size=None the population is sized automatically by
+    scipy (typically 15).
 
-    Unlike CMA-ES (pycma), pymoo's DE is pure Python so it does NOT trigger
+    Unlike CMA-ES (pycma), scipy's DE is NumPy-based and does NOT trigger
     the ProactorEventLoop / BLAS C-extension race that caused the molass-library
     #193 async crash.  It is therefore safe to run with in_process=True and
     async_=True.
@@ -28,7 +28,7 @@
     Copyright (c) 2026, SAXS Team, KEK-PF
 """
 import numpy as np
-from scipy.optimize import OptimizeResult
+from scipy.optimize import differential_evolution, OptimizeResult
 
 # Evaluation budget multiplier: max_fevals = niter * FEVALS_PER_NITER.
 # BH's niter counts hops; each hop runs a full Nelder-Mead local optimization
@@ -41,31 +41,32 @@ FEVALS_PER_NITER = 6600
 
 
 class SolverDE:
-    """Differential Evolution solver wrapping pymoo DE.
+    """Differential Evolution solver using scipy.
 
     Parameters
     ----------
     optimizer : BasicOptimizer
         Fully constructed optimizer that provides ``minima_callback``.
     pop_size : int or None
-        Population size.  None → max(20, 5 * n_var) (auto-sized).
-    variant : str
-        DE variant string understood by pymoo, e.g. ``"DE/rand/1/bin"``.
-    CR : float
+        Population size multiplier (scipy uses popsize * n_var).
+        None → scipy default (15).
+    strategy : str
+        DE strategy, e.g. ``"best1bin"``, ``"rand1bin"``.
+        See scipy.optimize.differential_evolution for full list.
+    recombination : float
         Crossover probability (0–1).
-    F : float or tuple
-        Differential weight (mutation factor).  A tuple ``(F_min, F_max)``
+    mutation : float or tuple
+        Differential weight (mutation factor).  A tuple ``(min, max)``
         enables dithering.
     """
 
-    def __init__(self, optimizer, pop_size=None, variant="DE/rand/1/bin",
-                 CR=0.5, F=0.5, warm_sigma=1.0):
+    def __init__(self, optimizer, pop_size=None, strategy="best1bin",
+                 recombination=0.7, mutation=0.5):
         self.optimizer = optimizer
-        self._pop_size = pop_size
-        self.variant = variant
-        self.CR = CR
-        self.F = F
-        self.warm_sigma = warm_sigma
+        self._pop_size = pop_size if pop_size is not None else 15
+        self.strategy = strategy
+        self.recombination = recombination
+        self.mutation = mutation
 
     def minimize(self, objective, init_params, niter=100, seed=1234,
                  bounds=None, narrow_bounds=False, show_history=False):
@@ -98,126 +99,61 @@ class SolverDE:
             ``.nit`` — number of DE generations completed
             ``.nfev``— total function evaluations
         """
-        from pymoo.algorithms.soo.nonconvex.de import DE
[…]
