+++ pyproject.toml
@@ -25,7 +25,6 @@ dependencies = [
   "ipywidgets",
   "ipykernel",
   "cma",
-  "pymoo",
   "shapely",
 ]
 
