### 2026-07-13 e88856d
fix: Phase 1c SDM UV/XR scale unification (UvOptimizer preserve_ratios)
