### 2026-07-13 d9ce5e4
feat: score() warns when Rg curve not pre-computed; get_rg_curve() inherits lazily from parent
