### 2026-07-13 4db7933
fix: Score.plot() returns PlotResult instead of Figure (prevents double display in Jupyter)
