### 2026-07-13 ec27fb3
fix: RigorousSdmParams uv_params no longer divides by xr_params (Phase 1c)
