### 2026-07-13 0fe2b6d
docs: update MIGRATION_unified_uv_params - Phase 1c complete (SDM G1100/G1200/G1300)
