### 2026-07-13 a166765
fix: clamp x0 to bounds in SolverDE.minimize
scipy differential_evolution raises ValueError if any x0 entry falls
outside the specified bounds.  SolverBH (and the old G1xxx-based
legacy path) never checked this, so parameters estimated to be outside
the hardcoded KT_BOUND / other column-param bounds caused DE to fail
silently (swallowed by BasicOptimizer's except-ValueError block, which
re-raised as 'Unknown method: de').

Root cause for SAMPLE1/SDM: K = N*T = 228.4 < KT_BOUND[0] = 500
→ norm_K = (228.4 - 500) / 150 = -1.81 (outside [0, 10])

Fix: clip init_params (norm_params) to [bounds_lo, 
[…]
