### 2026-07-13 e6e061a
docs: update MIGRATION_unified_uv_params.md with 2026-07-13 status
- Mark SolverDE x0 clamp and BasicOptimizer debug cleanup as done
- Add reload note (RigorousImplement does not reload G1xxx objectives)
- Add reference to molass-legacy#84 (KT_BOUND too narrow)
