### 2026-07-13 586e43d
fix: UvComponentCurve.scale must be ratio in load_rigorous_result / construct_decomp
UvComponentCurve.get_y() = self.scale * xr_ccurve.get_y()
xr_ccurve.get_y() already returns the ABSOLUTE XR curve (amplitude embedded).
Therefore self.scale must be the UV/XR ratio (epsilon_i/k) directly.

load_rigorous_result() was doing:
    xr_scale = xr_ccurve.get_scale_param()
    uv_scale = scale * xr_scale  # 'ratio -> absolute'
    UvComponentCurve(x, mapping, xr_ccurve, uv_scale)
  -> get_y() = (ratio * xr_scale) * xr_scale * PDF = ratio * xr_scale^2 * PDF  WRONG

The factor xr_scale is:
  EGH: H ~ 0.
[…]
