### 2026-07-13 7ffae10
fix: EGH uv_params are absolute UV heights, not ratios
G0346/G0367 (EGH objective) stores uv_params[k] as the absolute UV
peak height H_uv = ratio * H_xr, while G1100-G1500 (SDM/LKM/GRM/EDM,
Phase 1c convention) stores uv_params[k] as the UV/XR ratio directly.

CurrentStateUtils.py was treating EGH uv_params as ratios, causing
UvComponentCurve.get_y() = H_uv * H_xr * pdf(t) -- 10x too small.

Fix: detect function code (G0346/G0367) and divide by H_xr to recover
the ratio before constructing UvComponentCurve.

After fix: optimized_egh.uv_ccurves[i].scale ~ initial ratio (~2.2),
and UV_peak = ra
[…]
