### 2026-07-13 44eb4e3
docs: record EGH UV scale fix in MIGRATION_unified_uv_params.md
