### 2026-07-13 99bc550
feat: seed_params kwarg + RunInfo.best_params property
Add two small API additions enabling the BH→DE two-stage strategy:

1. RunInfo.best_params (new property)
   Reads analysis_folder/optimized/jobs/*/callback.txt, finds the
   global-minimum fv across all jobs, returns the corresponding param
   vector as a NumPy array.  Useful for seeding a follow-up optimizer
   from a prior run's result.

2. optimize_rigorously(seed_params=None) kwarg
   When provided, overrides the estimator-derived init_params before
   optimizer.prepare_for_optimization().  Takes priority over the
   clear_jobs=False 
[…]
