### 2026-07-14 52368ea
feat: make DE convergence tolerance configurable via tol= kwarg (de_tol)
SolverDE.__init__ now accepts tol=None (default: 0.01, matching scipy).
Wired to the de_tol SerialSettings key via Registry.settings_keys.
