### 2026-07-14 ae8cfcd
feat: add Decomposition.get_rigorous_initparams(); SolverDE mutation/recombination/tol kwargs
Decomposition.get_rigorous_initparams():
  Returns the estimator-derived init-param vector without running the
  optimizer.  Reuses make_basecurves_from_decomposition (same as
  get_rigorous_param_count).  Enables restartable notebooks that seed
  DE from the decomposition's own estimates, without a prior BH run.

SolverDE:
  - Uses mutation/recombination/strategy as canonical param names
    (matching what Registry now sends after the molass-legacy rename).
  - Removed obsolete F/CR/variant aliases.

[…]
