### 2026-07-15 488153b
feat: LumpingConstraint API for optimize_rigorously()
Formalise the lumping constraint discovered in experiment series 31
(molass-researcher/experiments/31_lumping_constraints/) as a proper API.

New: molass/Rigorous/LumpingConstraint.py
  - LumpingConstraint class: escape-zone soft-ramp penalty
  - Computes K-means group boundaries + escape zones from initial decomp
  - __call__(lrf_info) -> float penalty (weight × total_penetration)
  - Exported from molass.Rigorous

API change: optimize_rigorously(constraints=None)
  - Decomposition.optimize_rigorously accepts constraints=[...]
  - Passed t
[…]
