+++ molass/Rigorous/RigorousImplement.py
@@ -162,7 +162,7 @@ def _load_best_init_params(analysis_folder, init_params):
     return None
 
 
-def make_rigorous_decomposition_impl(decomposition, rgcurve, analysis_folder=None, niter=20, method="BH", frozen_components=None, frozen_param_groups=None, trimmed_ssd=None, clear_jobs=True, function_code=None, in_process=True, monitor=True, async_=True, progress='dashboard', max_trials=0, debug=False, _dry_run=False, ns_narrow_bounds=True, ns_adaptive_nsteps=False, ns_nsteps=None, solver_kwargs=None, seed_params=None):
+def make_rigorous_decomposition_impl(decomposition, rgcurve, analysis_folder=None, niter=20, method="BH", frozen_components=None, frozen_param_groups=None, trimmed_ssd=None, clear_jobs=True, function_code=None, in_process=True, monitor=True, async_=True, progress='dashboard', max_trials=0, debug=False, _dry_run=False, ns_narrow_bounds=True, ns_adaptive_nsteps=False, ns_nsteps=None, solver_kwargs=None, seed_params=None, constraints=None):
     """
     Make a rigorous decomposition using a given RG curve.
 
@@ -452,6 +452,13 @@ def make_rigorous_decomposition_impl(decomposition, rgcurve, analysis_folder=Non
             if _resume_init is not None:
                 init_params = _resume_init
         optimizer.prepare_for_optimization(init_params)
+
+        # Inject pluggable constraint hooks (e.g. LumpingConstraint).
+        # Constraints survive module reloads because _constraints is an
+        # instance attribute; BasicOptimizer.compute_fv reads it via
+        # getattr(self, '_constraints', []).
+        if constraints:
+            optimizer._constraints = list(constraints)
 
     # run optimization (outside _quiet — subprocess launch message is useful)
     from molass_legacy.Optimizer.Scripting import run_optimizer
