### 2026-07-16 2d03c15
fix: repair 8 pre-existing test failures in tests/specific/200_Rigorous
Four independent issues, all pre-dating the current session:

1. test_020 (_FakeCurve missing .x/.y):
   SdmEstimator._proxy_rgs_from_peak_frames was updated to access c.x/c.y
   directly, but the test mock was never updated. Add self.x = x; self.y = y
   to _FakeCurve.__init__.

2. test_050 (progress deprecation not fully applied):
   Issue #159 deprecated progress= and removed validation, but two tests
   still expected ValueError('Unknown progress='). Updated both to verify
   that unknown progress values are silently i
[…]
