### 2026-07-16 fa12e7f
feat: auto-apply LumpingConstraint for DE with 3+ components (closes #234)
When optimize_rigorously(method='DE') is called with a decomposition that
has 3 or more components and no explicit constraints, automatically build
LumpingConstraint(decomposition) from the current peak positions.

Motivation: DE's randomised population can drift two components into the
same elution zone, producing a degenerate collapsed solution.  31g showed
without this constraint DE converged to SV=81.44 with wrong lumping [0,0,1];
with it, SV=84.46 with correct [1,1,2].

Opt-out: pass constraints=[] to disable. BH/N
[…]
