### 2026-07-16 f88bd61
feat: disable DE tol-convergence when constraints are active (closes #234)
When constraints are present (auto-applied or explicit), DE's population
range metric can drop quickly once all candidates are inside the valid zone,
causing tol=0.01 to fire after just 10-20 callbacks instead of the full
niter budget.

Fix: when constraints is non-empty and method='DE', merge de_tol=0 into
solver_kwargs before calling set_optimizer_settings.  tol=0 in scipy's
differential_evolution means the convergence criterion std(energies) <= 0
never fires on floating-point data, so DE always runs the full maxiter.
[…]
