### 2026-07-16 01854b4
fix: always reload RigorousImplement in optimize_rigorously (not only in debug mode)
Previously, the if debug: guard meant that a live kernel that imported
RigorousImplement before a code change would keep the stale cached module
for every subsequent call.  This caused the auto-LumpingConstraint (added
in fa12e7f) to be silently skipped in notebooks that were already running.

Removing the guard matches the pattern used inside make_rigorous_decomposition_impl
itself (which always reloads LegacyBridgeUtils, FunctionCodeUtils, etc.)
and ensures that edits to RigorousImplement take effect without
[…]
