### 2026-07-16 c2b06f2
fix: auto-LumpingConstraint above _stack; use _parent (EGH) as boundary source (#234)
Two bugs found by the 31i experiment (SV=78.5 vs expected ~84.5):

1. Warning swallowed: the auto-apply block was inside 'with _stack:' which
   enters warnings.simplefilter('ignore').  Moved above _stack (same placement
   rule as the Pattern-A trimmed_ssd warning).  The constraint injection
   (optimizer._constraints) stays inside _stack where optimizer is available.

2. Wrong boundary source: LumpingConstraint(decomp_sdm) uses SDM-model curves
   whose peak positions differ from the EGH fit, producing zone
[…]
