### 2026-07-16 c8c96e5
fix: use _source_decomp for auto-LumpingConstraint boundaries (#234)
_parent fix (c2b06f2) didn't work: quick_decomposition(rgcurve=rgcurve) sets
_rgcurve on the EGH decomp, so copy_with_new_components takes the if-branch
(copies _rgcurve) instead of the else-branch (sets _parent). Result: upgraded
decomps have no _parent, so the fallback was decomp_sdm itself -- giving
wrong SDM-model peak positions for zone boundaries.

Fix: always set _source_decomp in copy_with_new_components regardless of
_rgcurve status. Use _source_decomp (preferred) > _parent (fallback) >
decomposition itself in Rigoro
[…]
