### 2026-07-16 ee0576a
fix #236, add #237 #238: clear_jobs outside async_ guard; loaded_job property; source_decomp property
#236: Move clear_jobs rmtree from inside 'if async_:' to before the
if/else split so sync (async_=False) in-process runs also clear stale
job folders. Without this, repeated async_=False calls accumulated
jobs/001, /002, ... and load_best() could return a stale bad-SV result.

#237: Store best.id in self._loaded_job inside load_best() and expose
as the loaded_job property on RunInfo. Callers no longer need to grep
stdout to know which job was loaded.

#238: Add public source_decomp property on
[…]
