### 2026-07-16 552d1db
fix #232: implement get_peak_top_x() for SDM, EDM, GRM component curves
SDM and EDM had NotImplementedError (with SDM's message copy-pasted into
EDM). GRM had get_peak_position() but no get_peak_top_x(), so the base
class returned params[1]=t0 (wrong). All three now use x[argmax(get_y())]
via a one-liner, matching the LKM pattern.
