### 2026-07-16 7a88e05
fix #233: gate noisy prints in construct_decomposition_from_results behind debug=False
The two progress lines ('Loading current decomposition...' and 'Using job id...')
now only print when debug=True is passed to get_current_decomposition().
The existing debug kwarg is already threaded through from RunInfo.get_current_decomposition().
