### 2026-07-16 3eef7a1
docs: mark Phase 2 and Phase 3 complete in MIGRATION_unified_uv_params.md
