### 2026-07-16 68dc0a3
fix: skip c_inj when reading R/k_MT/k_ext in get_lkm/grm_xr_ccurves (fix #239)
