### 2026-07-16 880a784
docs: add Linux prerequisites and CONTRIBUTING link in Quick Start (#240, #241)
