### 2026-07-16 3f52654
chore: bump version to 1.0.2; require molass_legacy>=1.6.11
