+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "molass"
-version = "1.0.1"
+version = "1.0.2"
 
 dependencies = [
   "numpy",
@@ -12,7 +12,7 @@ dependencies = [
   "matplotlib",
   "tqdm",
   "pybaselines>=1.2.0",
-  "molass_legacy>=1.6.10",
+  "molass_legacy>=1.6.11",
   "statsmodels",
   "scikit-learn",
   "toml",
