### 2026-07-17 4b2ccd7
revert: #235 remove Lampinen code, soft penalty retained
Experiments 31k/31l showed soft penalty SV=85.11 vs Lampinen SV=72.39.
Root cause: LHS init ~87% infeasible, wrong basin. Full revert:
- LumpingConstraint.py: remove violation() and bind_optimizer()
- RigorousImplement.py: simplify de_tol block, remove tight-init auto-set, remove debug print
- SolverDE.py: remove NonlinearConstraint block
- test_200_lumping_constraint.py: remove TestViolationMethod
- copilot-instructions.md: document findings

12 tests pass. Closes #235.
