### 2026-07-21 faa0e79
chore: bump version to 1.0.3 (dev/ongoing-work, post-JOSS v1.0.2 archive)
