### 2026-07-21 f7fd566
docs: fix low-rank hyphenation and k-means phrasing in paper.md
