### 2026-07-21 dc3e5b6
docs: restore 'by the k-means algorithm' per Shimizu
