### 2026-07-22 0f77bde
docs: add Verification section to ARCHITECTURE.md
Documents the GuiSimUtils framework (MockEditor, SimpleLrfSource,
evaluate_init) and the standard workflow for verifying GUI consistency
at each migration step without Tkinter.

Links to the Experiment 33 notebook series in molass-researcher.
