### 2026-07-22 318b664
fix(construct_legacy_optimizer): match any SDM variant with model.startswith('SDM')
Previously only matched model == 'SDM' exactly, so passing model='SDM(mono)'
or model='SDM(lognormal)' left optimizer.params_type.estimator as None,
causing AttributeError in prepare_for_optimization -> get_param_bounds -> get_colparam_bounds.
