### 2026-07-22 ee32857
fix(SolverDE): normalize GUI-label strategy to scipy format via _STRATEGY_MAP
OptStrategyDialog stores 'DE/best/1/bin' / 'DE/rand/1/bin' in SerialSettings.
scipy.optimize.differential_evolution only accepts short names like 'best1bin'.
Add _STRATEGY_MAP and apply it in __init__ so any GUI label is transparently
converted before being passed to scipy.

Fixes subprocess crash: ValueError: Please select a valid mutation strategy
Diagnosed via molass-researcher/experiments/33_gui_consistency/33b_sdm_model.ipynb
