### 2026-07-22 db2843a
fix(SdAdapter): add make_ssd_from_sd neutral alias (molass-library#242)
