### 2026-07-23 f706162
molass-library#243: add mu_max to optimize_sdm_lognormal_xr_decomposition model_params
Add mu_max key to the model_params dict handled by
optimize_sdm_lognormal_xr_decomposition.

Changes:
- model_params is None branch: mu_max = 8.0 (no constraint, preserves old behaviour)
- model_params dict branch: mu_max = model_params.get('mu_max', 8.0)
- objective guard: mu_ > 8.0 → mu_ > mu_max
- bounds: (2.0, 8.0) → (2.0, min(mu_max, 8.0))

Rationale: when the initial T has wrong k-scaling, the optimizer drifts
mu to compensate (e.g. poresize ~165 Å for SAMPLE1 Rg_max=40 Å), compressing
all K_SEC values
[…]
