### 2026-07-23 ddd399f
SdmOptimizer: add position anchor to lognormal objective
Prevents the 'lumping' problem where a component drifts from its
initial chromatographic group into a different SEC peak during Stage 4
optimization of optimize_sdm_lognormal_xr_decomposition.

Root cause: the lognormal objective had order_penalty, gap_penalty,
and rg_penalty but no peak-position anchor — unlike the mono optimizer
which already has position_penalty anchored to EGH peak frames.

Fix: add intensity-weighted centroid penalty (position_anchor_scale=1e-5)
anchored to the EGH-derived peak frames.  Centroid is a smooth function
[…]
