### 2026-07-23 eb96de3
SdmEstimator: fix T/k parameterization mismatch in estimate_sdm_lognormal_from_monopore
SDM(mono) stores T as the Gamma MEAN (= k_mono x theta).
SDM(lognormal) optimizer treats T as the Gamma SCALE (mean = k x T_ln).

For matching peak positions: T_ln = T_mono / k_optimizer (k_optimizer=2.0).

The function was returning T_mono as the LognormalEnv T parameter, causing
Stage 4 to start with a column whose peak timing is 2x too long. The optimizer
then explored around a wrong basin (component 1 drifts to lump 2 for SAMPLE1).

Fix:
- Compute T_ln = T / k_for_shift (k_for_shift=2.0, lognormal optim
[…]
