### 2026-07-23 e7c8488
fix: SDM(lognormal) position anchor — adaptive scale + M_1 position formula
- SdmOptimizer.py: add _adaptive_anchor_scale() helper shared by both SDM
  optimizers. A full inter-lump drift costs exactly initial_error, making
  the anchor self-calibrating regardless of data amplitude or frame rate.
  Override via model_params['position_anchor_scale'].
- SdmOptimizer.py (lognormal): replace centroid-based position with
  M_1 = tI + sdm_lognormal_model_moments(...)[0]. Amplitude-independent
  (robust when scale->0); unifies formula with mono optimizer.
- SDM.py: skip moment_refine when ln_pore_sig
[…]
