### 2026-07-24 554d02e
fix: port joss-paper code fixes to dev/ongoing-work
Port two bug fixes that were applied to joss-paper during JOSS review
but never ported to main/dev:

1. CurveDecomposer: update num_components when peak recognition finds
   fewer peaks than requested (prevents bounds/init_params size mismatch
   in scipy.optimize.minimize). Fix 9a7611d from joss-paper (2026-07-03).

2. Decomposition.update_xr_ranks: validate ranks length matches
   num_components, raise ValueError with clear message. Fix 9325819
   from joss-paper (2026-07-01).

3. Update test_020_update_xr_ranks.py to assert ValueError is r
[…]
