+++ tests/generic/210_Ranks/test_020_update_xr_ranks.py
@@ -1,31 +1,29 @@
-"""Tests for Decomposition.update_xr_ranks() method
-
-NOTE: Current behavior (as of 2026-07-03):
-- update_xr_ranks() accepts any list without validation
-- get_xr_matrices() uses fallback: `ranks or [1] * num_components`
-- No error is raised for length mismatch
-
-This test documents current behavior. Future enhancement could add validation.
-"""
+"""Tests for Decomposition.update_xr_ranks() method"""
 import pytest
 from molass_data import SAMPLE3
 from molass.DataObjects import SecSaxsData as SSD
 
 
 def test_update_xr_ranks_sets_attribute():
-    """update_xr_ranks sets xr_ranks attribute without validation"""
+    """update_xr_ranks sets ranks when length matches num_components"""
     ssd3 = SSD(SAMPLE3)
     d = ssd3.quick_decomposition()
-    
-    # SAMPLE3 has 3 components
+
     assert d.num_components == 3
-    
-    # Can set any ranks list (no validation currently)
-    d.update_xr_ranks([2])
-    assert d.xr_ranks == [2]
-    
+
     d.update_xr_ranks([2, 1, 1])
     assert d.xr_ranks == [2, 1, 1]
+
+
+def test_update_xr_ranks_raises_on_length_mismatch():
+    """update_xr_ranks raises ValueError when ranks length != num_components"""
+    ssd3 = SSD(SAMPLE3)
+    d = ssd3.quick_decomposition()
+
+    assert d.num_components == 3
+
+    with pytest.raises(ValueError, match="Length of ranks"):
+        d.update_xr_ranks([2])  # length 1 != num_components 3
 
 
 def test_get_xr_matrices_with_ranks():
