### 2026-07-24 d20df00
fix: resolve unresolved merge conflict in init.prompt.md
Accept main's version: agent=copilot, full ai-context-vscode install
instructions. The conflict was left unresolved from a prior merge.
