+++ .github/prompts/init.prompt.md
@@ -1,9 +1,5 @@
 ﻿---
-<<<<<<< HEAD
-agent: ask
-=======
 agent: copilot
->>>>>>> 53675d65bb5c75a3c302a9677d2fa94ef773bd9a
 description: セッション初期化（PROJECT_STATUS.md の読み込みと確認）
 alwaysApply: true
 ---
@@ -14,13 +10,6 @@ Respond in English regardless of this file's language.
 
 `.github/vscode-version.txt` を読んでください。
 
-<<<<<<< HEAD
-- **バージョン番号が記載されている場合**: そのバージョンを使用してステップ2へ
-- **バージョン番号が未記載またはファイルが存在しない場合**: 以下をユーザーに表示してください：
-
-  > ⚠️ **VS Code バージョンの記録をお願いします**  
-  > メニュー → **ヘルプ** → **バージョン情報** → 最初の行の番号（例: `1.114.0-insider`）を  
-=======
 - **`Auto-updated by vscode-version-recorder extension` という行がある場合**: 拡張機能が正常動作中です。そのバージョンを使用してステップ2へ。
 - **バージョン番号はあるが上記の行がない場合**: そのバージョンを使用してステップ2へ。その後、以下をユーザーに表示：
 
@@ -40,7 +29,6 @@ Respond in English regardless of this file's language.
 
   > ⚠️ **VS Code バージョンの記録をお願いします**  
   > メニュー → **ヘルプ** → **バージョン情報** → 最初の行の番号（例: `1.115.0-insider`）を  
->>>>>>> 53675d65bb5c75a3c302a9677d2fa94ef773bd9a
   > `.github/vscode-version.txt` の末尾に追記してください。
 
 ## ステップ2: alwaysApply 対応チェック
