### 2026-07-24 15c0c65
chore: remove personal preferences file; sync .gitignore to main
