+++ preferences/RESPONSE_PREFERENCES.md
@@ -1,6 +1,0 @@
-# AI Assistant Response Preferences
-
-- Always provide code or markdown in a **single code block** using four backticks and the appropriate language tag.
-- Do not split suggestions into multiple blocks unless explicitly requested.
-- Place explanations after the code block.
-- Include file paths as comments when relevant.
