### 2026-07-24 a30885c
feat: add mu_min constraint to lognormal SDM optimizer (molass-library#244)
Add mu_min parameter to optimize_sdm_lognormal_xr_decomposition.
Constrains pore-size lower bound; defaults to 2.0 (no-op, backward-compatible).

Callers pass mu_min=ln(poresize_bounds[0]) so the optimizer stays in the
physical pore-size range and avoids the degenerate poresize≈Rg_max basin that
previously produced SV≈59 instead of SV≈67.
