### 2026-07-24 e78a8a3
restrict the dependency on molass_legacy
