+++ pyproject.toml
@@ -12,7 +12,7 @@ dependencies = [
   "matplotlib",
   "tqdm",
   "pybaselines>=1.2.0",
-  "molass_legacy>=1.6.10",
+  "molass_legacy>=1.6.12",
   "statsmodels",
   "scikit-learn",
   "toml",
