### 2026-07-29 b7d4448
chore: bump version to 1.0.4
