### 2026-07-31 ea4d7da
fix: make_basecurves_from_decomposition reads unified_baseline_type (closes #244)
