### 2026-07-31 e690efb
fix: pre-existing test failures — is_interactive import + EdmOptimizer e_bounds=None crash
