+++ tests/specific/910_Anomaly/test_010_Bubbles.py
@@ -2,6 +2,7 @@
     test Bubbles
 """
 import os
+from molass.Testing import is_interactive
 from molass import get_version
 get_version(toml_only=True)     # to ensure that the current repository is used
 from molass.Local import get_local_settings
