### 2026-07-31 b86a80c
docs: record absolute frame coordinates design principle in ARCHITECTURE.md
