### 2026-07-31 10d789a
export ip_uv_elcurve_x.npy for SSD-native UV frame alignment in subprocess
