### 2026-07-31 1325d7b
export ip_xr_elcurve_x.npy for SSD-native XR frame alignment in subprocess
