+++ paper.md
@@ -18,7 +18,7 @@ affiliations:
     index: 1
   - name: Photon Factory, Institute of Materials Structure Science, High Energy Accelerator Research Organization (KEK)
     index: 2
-  - name: Life Science Research Infrastructure Group, R&D of Technology and Systems for Synchrotron Radiation Applications Division, RIKEN SPring-8 Center
+  - name: Life Science Research Infrastructure Group, R&D of Technology and Systems for Synchrotron Radiation Applications Division, RIKEN SPring-8 Center, RIKEN
     index: 3
     ror: 01d1kv753
 date: 2 October 2025
