### 2026-08-03 751b25e
chore: bump version to 1.0.4 for JOSS release
