+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "molass"
-version = "1.0.2"
+version = "1.0.4"
 
 dependencies = [
   "numpy",
