### 2026-08-04 1c46980
refactor: move Tkinter show() out of Score and InitialEstimator (Tkinter lives in molass_gui)
