### 2026-08-04 63173fb
perf: skip .npy subprocess overrides when in_process=True
