+++ molass/Rigorous/LegacyBridgeUtils.py
@@ -116,7 +116,7 @@ def construct_legacy_optimizer(dsets, baseline_objects, spectral_vectors, num_co
     
     return optimizer
 
-def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_ssd=None, debug=False):
+def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_ssd=None, debug=False, export_npy=True):
     from molass_legacy._MOLASS.SerialSettings import get_setting, set_setting
     if analysis_folder is None:
         analysis_folder = get_setting('analysis_folder')
@@ -162,36 +162,24 @@ def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_
     dsets = make_dsets_from_decomposition(decomposition, rgcurve, data_ssd=data_ssd, debug=debug)
     basecurves, baseparams = make_basecurves_from_decomposition(decomposition, data_ssd=data_ssd, debug=False)
 
-    # Export in-process ElCurve y-values so subprocess uses same curves (molass-legacy#38).
-    # The subprocess re-derives ElCurves via legacy smoothing (get_xr/uv_data_separate_ly),
-    # producing ~4% difference from the EGH-fitted curves used by the in-process path.
-    # Exporting y-values here lets OptDataSets.get_dsets_impl override after loading from disk.
-    np.save(os.path.join(optimizer_folder, 'ip_xr_elcurve_y.npy'), dsets[0][0].y)
-    np.save(os.path.join(optimizer_folder, 'ip_xr_elcurve_x.npy'), dsets[0][0].x)
-    np.save(os.path.join(optimizer_folder, 'ip_uv_elcurve_y.npy'), dsets[2][0].y)
-    # Export UV x-axis (frame numbers) so subprocess can align the UV curve with the correct
-    # frame range.  With the SSD-native prepare_rg_curve path the in-process UV curve spans
-    # ~2249 UV-native absolute frames, while the legacy subprocess loads only ~645 frames
-    # (trimmed to match the XR window).  Without this export the ip_uv_elcurve_y override
-    # causes a length mismatch → ValueError in InterpolatedUnivariateSpline.
-    np.save(os.path.join(optimizer_folder, 'ip_uv_elcurve_x.npy'), dsets[2][0].x)
-    # Export in-process data matrices and error matrix so subprocess uses the same corrected
-    # 2D data (molass-legacy#39).  The subprocess loads D/U via sd.get_xr/uv_data_separate_ly()
-    # (legacy correction) which differs from molass-library ssd.xr.M / ssd.uv.M.
-    # E = ssd.xr.E is also different: it feeds compute_weight_info(1/(E+D/100)) in
-    # OptDataSets.__init__ → W_ → xrDw, so even with identical D the weight matrix differs.
-    np.save(os.path.join(optimizer_folder, 'ip_xr_D.npy'), dsets[0][1])
-    np.save(os.path.join(optimizer_folder, 'ip_uv_U.npy'), dsets[2][1])
-    np.save(os.path.join(optimizer_folder, 'ip_xr_E.npy'), dsets.E)
-    # Export in-process qvector so subprocess uses the same trimmed q-values
-    # (molass-legacy#41: subprocess sd.qvector has 972 elements from the full
-    # raw data; in-process corrected SSD has 966 elements after trimming.
-    # The difference shifts GuinierDeviation's xr_index bisection → different fv.)
-    _export_ssd = data_ssd if data_ssd is not None else decomposition.ssd
-    np.save(os.path.join(optimizer_folder, 'ip_xr_qvector.npy'), _export_ssd.xr.q_values)
[…]
