### 2026-08-05 b4a5db1
feat: Option E recipe-based subprocess (RecipeRunner + pipeline_recipe param)
- RecipeRunner.py: new library-side recipe optimizer builder (testable in notebook)
- Decomposition.optimize_rigorously(): added pipeline_recipe param (fixes solver_kwargs leak)
- RigorousImplement: pass pipeline_recipe through to prepare_rigorous_folders
- LegacyBridgeUtils: skip ip_*.npy when pipeline_recipe is not None
- ARCHITECTURE.md: updated for molass-gui perspective change (new UI layer)
