+++ molass/Rigorous/LegacyBridgeUtils.py
@@ -116,7 +116,7 @@ def construct_legacy_optimizer(dsets, baseline_objects, spectral_vectors, num_co
     
     return optimizer
 
-def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_ssd=None, debug=False, export_npy=True):
+def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_ssd=None, debug=False, export_npy=True, pipeline_recipe=None):
     from molass_legacy._MOLASS.SerialSettings import get_setting, set_setting
     if analysis_folder is None:
         analysis_folder = get_setting('analysis_folder')
@@ -162,8 +162,8 @@ def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_
     dsets = make_dsets_from_decomposition(decomposition, rgcurve, data_ssd=data_ssd, debug=debug)
     basecurves, baseparams = make_basecurves_from_decomposition(decomposition, data_ssd=data_ssd, debug=False)
 
-    # subprocess-only overrides: skip entirely for in_process=True runs
-    if export_npy:
+    # subprocess-only overrides: skip entirely for in_process=True or recipe mode
+    if export_npy and pipeline_recipe is None:
         # Export in-process ElCurve y-values so subprocess uses same curves (molass-legacy#38).
         np.save(os.path.join(optimizer_folder, 'ip_xr_elcurve_y.npy'), dsets[0][0].y)
         np.save(os.path.join(optimizer_folder, 'ip_xr_elcurve_x.npy'), dsets[0][0].x)
@@ -195,7 +195,7 @@ def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_
     # Export parent's UV diff_spline so subprocess uses same baseline (molass-legacy#34).
     uv_base_curve = basecurves[0]
     _ds_exported = False
-    if export_npy and hasattr(uv_base_curve, 'diff_spline') and uv_base_curve.diff_spline is not None:
+    if export_npy and pipeline_recipe is None and hasattr(uv_base_curve, 'diff_spline') and uv_base_curve.diff_spline is not None:
         if hasattr(uv_base_curve, 'curve1') and uv_base_curve.curve1 is not None:
             _ds_x = uv_base_curve.curve1.x
         else:
@@ -205,6 +205,16 @@ def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_
         np.save(os.path.join(optimizer_folder, 'uv_diff_spline_x.npy'), _ds_x)
         np.save(os.path.join(optimizer_folder, 'uv_diff_spline_y.npy'), _ds_y)
         _ds_exported = True
+
+    # Export pipeline recipe for subprocess reconstruction (Option E).
+    # Only when export_npy=True (subprocess mode); skip for in_process mode.
+    _recipe_exported = False
+    if export_npy and pipeline_recipe is not None:
+        import json
+        recipe_path = os.path.join(optimizer_folder, 'recipe.json')
+        with open(recipe_path, 'w') as _f:
+            json.dump(pipeline_recipe, _f, indent=2)
+        _recipe_exported = True
 
     # Diagnostic log (molass-legacy#34): confirm what the parent actually wrote.
     # Written to optimizer_folder (not rg_folder) so it survives if rg_folder is cleared.
@@ -217,6 +227,9 @@ def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_
         _f.write(f"  uv_diff_spline exported: {_ds_exported}\n")
         if _ds_exported:
             _f.write(f"    x range [{_ds_x[0]:.1f}, {_ds_x[-1]:.1f}] len={len(_ds_x)}\n")
[…]
