### 2026-08-05 a736307
ai-friendliness: RunInfo.stop(), monitor=True warning for subprocess
- RunInfo.stop(): combines request_stop() + subprocess SIGTERM (was missing, caused silent failure)
- RigorousImplement: UserWarning when monitor=True with in_process=False (silently ignored before)
