### 2026-08-06 e0eabb7
feat: recipe subprocess Steps 3-5 + AI-friendliness fixes (#247-249)
Steps 3-5 of recipe subprocess migration:
- Step 3: auto-recipe + DeprecationWarning when in_process=False, pipeline_recipe=None
- Step 4: remove export_npy param and ip_*.npy block from prepare_rigorous_folders()
- Step 5: monitor=True for subprocess via MplMonitor.for_run_info() external watcher

RecipeRunner:
- model field: decomp.upgrade(model) for non-EGH in subprocess
- method field: LumpingConstraint auto-applied for DE with 3+ comps
- LumpingConstraint uses egh_decomp (pre-upgrade) for reliable boundaries

RigorousImp
[…]
