+++ molass/Rigorous/RigorousImplement.py
@@ -19,6 +19,7 @@ import os
 import numpy as np
 from importlib import reload
 
+
 def _set_identity_restrict_lists(ssd):
     """Set identity (full-range) restrict_lists for already-exported data.
 
@@ -162,6 +163,19 @@ def _load_best_init_params(analysis_folder, init_params):
     return None
 
 
+def _build_auto_recipe(decomposition, method='BH'):
+    """Build a minimal recipe dict from a decomposition for backward-compat auto-construction."""
+    model = decomposition.xr_ccurves[0].model  # 'egh', 'sdm', 'lkm', etc.
+    return {
+        "num_components": decomposition.num_components,
+        "model": model,
+        "method": method.lower(),
+        "decomp_params": {},
+        "trim_params": {},
+        "baseline_params": {},
+    }
+
+
 def make_rigorous_decomposition_impl(decomposition, rgcurve, analysis_folder=None, niter=20, method="BH", frozen_components=None, frozen_param_groups=None, trimmed_ssd=None, clear_jobs=True, function_code=None, in_process=True, monitor=True, async_=True, progress='dashboard', max_trials=0, debug=False, _dry_run=False, ns_narrow_bounds=True, ns_adaptive_nsteps=False, ns_nsteps=None, solver_kwargs=None, seed_params=None, constraints=None, pipeline_recipe=None):
     """
     Make a rigorous decomposition using a given RG curve.
@@ -200,9 +214,23 @@ def make_rigorous_decomposition_impl(decomposition, rgcurve, analysis_folder=Non
         of spawning a subprocess.  The library-prepared optimizer (with the
         live dsets, base curves, and spectral vectors built above) is the one
         that runs — no re-derivation from disk, no parent/subprocess divergence.
-        Set ``False`` to use the legacy subprocess path (required by the tkinter
-        GUI; available as an escape hatch for notebook users who need process
-        isolation).
+        Set ``False`` to use the recipe-based subprocess path, which rebuilds
+        the SSD pipeline inside the subprocess from ``recipe.json``.
+
+        **Subprocess lifecycle** (``in_process=False``):
+
+        * The subprocess is an independent OS process.  It is **not** killed
+          by a Jupyter kernel interrupt (SIGINT).
+        * On normal kernel restart (SIGTERM) an ``atexit`` handler terminates
+          the subprocess automatically.
+        * On force-kill (SIGKILL) the subprocess becomes an orphan and
+          continues until it finishes.  The next ``optimize_rigorously()``
+          call creates a new ``MplMonitor`` which reads
+          ``active_processes.json`` and terminates the orphan.
+        * To stop a running subprocess explicitly::
+
+              run_info.stop()   # cooperative signal + p.terminate()
+
         See ``molass-library/Copilot/DESIGN_split_optimizer_architecture.md``.
     monitor : bool, optional
         Controls the ``MplMonitor`` ipywidgets dashboard.  When True
@@ -344,6 +372,9 @@ def make_rigorous_decomposition_impl(decomposition, rgcurve, analysis_folder=Non
     if _dry_run:
         return None
 
+    # Capture user-supplied constraints before auto-apply may add LumpingConstraint.
+    _original_user_constraints = constraints
+
     # Auto-apply LumpingConstraint for DE with 3+ components (molass-library#234).
     # PLACEMENT: must be ABOVE `with _stack:` — the stack enters
     # warnings.simplefilter("ignore") and would swallow this warning silently.
@@ -378,6 +409,38 @@ def make_rigorous_decomposition_impl(decomposition, rgcurve, analysis_folder=Non
         solver_kwargs = dict(solver_kwargs or {})
         solver_kwargs.setdefault('de_tol', 0)
 
+    # #249: warn when USER-supplied constraints cannot be transferred to recipe subprocess.
[…]
