+++ molass/Rigorous/LegacyBridgeUtils.py
@@ -116,7 +116,7 @@ def construct_legacy_optimizer(dsets, baseline_objects, spectral_vectors, num_co
     
     return optimizer
 
-def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_ssd=None, debug=False, export_npy=True, pipeline_recipe=None):
+def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_ssd=None, debug=False, pipeline_recipe=None):
     from molass_legacy._MOLASS.SerialSettings import get_setting, set_setting
     if analysis_folder is None:
         analysis_folder = get_setting('analysis_folder')
@@ -162,24 +162,6 @@ def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_
     dsets = make_dsets_from_decomposition(decomposition, rgcurve, data_ssd=data_ssd, debug=debug)
     basecurves, baseparams = make_basecurves_from_decomposition(decomposition, data_ssd=data_ssd, debug=False)
 
-    # subprocess-only overrides: skip entirely for in_process=True or recipe mode
-    if export_npy and pipeline_recipe is None:
-        # Export in-process ElCurve y-values so subprocess uses same curves (molass-legacy#38).
-        np.save(os.path.join(optimizer_folder, 'ip_xr_elcurve_y.npy'), dsets[0][0].y)
-        np.save(os.path.join(optimizer_folder, 'ip_xr_elcurve_x.npy'), dsets[0][0].x)
-        np.save(os.path.join(optimizer_folder, 'ip_uv_elcurve_y.npy'), dsets[2][0].y)
-        np.save(os.path.join(optimizer_folder, 'ip_uv_elcurve_x.npy'), dsets[2][0].x)
-        # Export in-process data matrices and error matrix (molass-legacy#39, #40).
-        np.save(os.path.join(optimizer_folder, 'ip_xr_D.npy'), dsets[0][1])
-        np.save(os.path.join(optimizer_folder, 'ip_uv_U.npy'), dsets[2][1])
-        np.save(os.path.join(optimizer_folder, 'ip_xr_E.npy'), dsets.E)
-        # Export in-process qvector (molass-legacy#41).
-        _export_ssd = data_ssd if data_ssd is not None else decomposition.ssd
-        np.save(os.path.join(optimizer_folder, 'ip_xr_qvector.npy'), _export_ssd.xr.q_values)
-        # Export frame numbers (jv) for BackRunner / compute_rg_curve_from_arrays.
-        np.save(os.path.join(optimizer_folder, 'ip_xr_jv.npy'), _export_ssd.xr.jv)
-    else:
-        _export_ssd = data_ssd if data_ssd is not None else decomposition.ssd
     # Export the Rg curve to rg-curve/ — the single authoritative folder for rg_curve.
     # rg_curve_parent/ has been eliminated (molass-legacy#78): both prepare_rigorous_folders()
     # and BackRunner.run() write directly to rg-curve/.  The subprocess reads rg-curve/ and
@@ -192,24 +174,9 @@ def prepare_rigorous_folders(decomposition, rgcurve, analysis_folder=None, data_
     os.makedirs(rg_curve_folder)
     rgcurve_.export(rg_curve_folder)
 
-    # Export parent's UV diff_spline so subprocess uses same baseline (molass-legacy#34).
-    uv_base_curve = basecurves[0]
-    _ds_exported = False
-    if export_npy and pipeline_recipe is None and hasattr(uv_base_curve, 'diff_spline') and uv_base_curve.diff_spline is not None:
[…]
