### 2026-08-17 def8c22
feat: RunInfo.rg_history / sv_history_raw for GUI/notebook progress-chart parity
- CurrentStateUtils.parse_rg_history(): per-component Rg trajectories, read
  directly from callback.txt x= vectors via optimizer.params_type.get_rg_start_index()
  (Rg is a free optimizer parameter, not re-derived via Guinier)
- CurrentStateUtils.parse_sv_history_raw(): per-evaluation SV without the
  running-minimum accumulation (includes rejected trials)
- RunInfo.rg_history / RunInfo.sv_history_raw properties
- RunInfo.plot_rg_history() new method; plot_sv_history() updated to overlay
  the raw scatter behind 
[…]
