### 2026-08-19 6b5ccda
feat: ParamsTable module for model-agnostic Show Parameters extraction; rg_alpha_by_score/rg_alpha_power in plot_components
