+++ molass/Rigorous/ParamsTable.py
@@ -1,0 +1,205 @@
+"""
+Rigorous.ParamsTable.py
+
+Model-agnostic "Show Parameters" data extractor.
+
+Promoted from molass-researcher/experiments/34_ssd_rigorous_gui/34k_show_parameters_design.ipynb,
+where it was validated against EGH, SDM, CEDM, LKM and GRM (SAMPLE1). Replaces the legacy
+per-model ``*ParamsSheet`` classes' hand-built row/col grids with one shared extractor plus
+one small per-model namer for the trailing ``model_colparams`` tuple.
+
+This module only builds the data (a tidy ``pandas.DataFrame``); rendering (notebook display,
+``tksheet.Sheet``, etc.) is left to the caller.
+"""
+import numpy as np
+import pandas as pd
+
+_EGH_XR_LABELS = ["h", "mu (tR)", "sigma", "tau"]
+
+
+def common_param_rows(xr_params, xr_baseparams, rgs, mapping, uv_params, uv_baseparams,
+                       mappable_range, xr_labels=None):
+    """Build the model-agnostic rows: xr/rg, xr baseline, uv, uv baseline, mapping, range.
+
+    Everything except ``model_colparams`` (see the per-model ``*_colparam_rows`` functions)
+    is identical across EGH/SDM/EDM/CEDM/LKM/GRM.
+
+    Parameters
+    ----------
+    xr_params, xr_baseparams, rgs, mapping, uv_params, uv_baseparams, mappable_range :
+        The first seven elements of ``optimizer.split_params_simple(params)``.
+    xr_labels : list of str, optional
+        Column names for ``xr_params``. Defaults to EGH's ``[h, mu (tR), sigma, tau]``
+        truncated to the actual column count -- pass explicitly for models whose
+        per-component params mean something else (e.g. CEDM's ``[a, b, c_inj]``).
+
+    Returns
+    -------
+    list of dict
+        Rows with keys ``section``, ``label``, ``component``, ``value``.
+    """
+    rows = []
+    xr_params = np.asarray(xr_params)
+    if xr_params.ndim == 1:
+        # SDM/EDM/CEDM/LKM/GRM: one scale per component (flat), not an EGH-style
+        # (component, param) matrix -- np.atleast_2d would misread this as a
+        # single row instead of n rows, silently dropping components 2..n.
+        xr_params = xr_params.reshape(-1, 1)
+    n = xr_params.shape[0]
+    if xr_labels is None:
+        xr_labels = _EGH_XR_LABELS[:xr_params.shape[1]]
+
+    for i in range(n):
+        for j, label in enumerate(xr_labels):
+            rows.append({"section": "xr", "label": label, "component": i + 1, "value": xr_params[i, j]})
+        rows.append({"section": "xr", "label": "rg", "component": i + 1, "value": rgs[i]})
+
+    for j, label in enumerate(["slope", "intercept", "fouling"][:len(xr_baseparams)]):
+        rows.append({"section": "xr_baseline", "label": label, "component": None, "value": xr_baseparams[j]})
+
+    uv_params = np.atleast_1d(uv_params)
+    for i, v in enumerate(uv_params):
+        rows.append({"section": "uv", "label": "h", "component": i + 1, "value": v})
+
+    uv_bp_labels = ["L", "x0", "k", "b", "s1", "s2", "diff_ratio", "fouling"][:len(uv_baseparams)]
+    for j, label in enumerate(uv_bp_labels):
+        rows.append({"section": "uv_baseline", "label": label, "component": None, "value": uv_baseparams[j]})
+
+    for j, label in enumerate(["slope", "intercept"]):
[…]
