### 2026-08-20 2469c4b
AI-friendliness: has_rg_curve property, score() progress_cb phases, thread-safety warning (#250, #251, #252)
