### 2026-08-20 4faab89
test: mark slow rigorous-optimization tests with @pytest.mark.slow, document -m 'not slow'
