### 2026-08-20 47607b5
fix: forward rg_alpha_by_score/rg_alpha_power in plot_components_impl; change rg_cmap/rg_alpha_by_score/rg_alpha_power defaults to YlGn/True/2
