### 2026-08-20 d58f0a9
docs: record decision to reverse optimize_rigorously() in_process default to False
