+++ Copilot/refactor/DESIGN_default_in_process_reversal.md
@@ -1,0 +1,117 @@
+# DESIGN: Reverse `optimize_rigorously()`'s `in_process` default to `False`
+
+**Status**: Decision made (2026-08-20) — implementation pending
+**Context**: Follow-up to `DESIGN_split_optimizer_architecture.md` (April 22, 2026) and
+`DESIGN_inprocess_monitor.md` (April 27, 2026). Triggered by a molass-gui "Continue in
+Notebook…" feature discussion that surfaced a real performance/architecture question about
+which mode should be the *default* for `Decomposition.optimize_rigorously()`.
+
+---
+
+## How this came up
+
+While simplifying molass-gui's generated "Continue in Notebook…" export (dropping
+GUI-specific parameters in favor of library defaults, per a general "options for power
+users, defaults for general users" principle adopted this session), the notebook stopped
+passing `in_process=False` and started relying on the library default (`True`). This raised
+the question: should the exported notebook instead match the GUI's own default
+(`in_process=False`, subprocess via `pipeline_recipe`), and if so, should that be achieved
+by changing the *library's* default instead of the GUI's or the notebook's?
+
+## Evidence gathered
+
+1. **Original rationale for `in_process=True` as the library/notebook default**
+   (`DESIGN_split_optimizer_architecture.md`, April 22, 2026): subprocess was designed for
+   the *tkinter* constraint (`mainloop()` must not block). It was inherited uniformly by the
+   library/notebook path "because it was already there," even though notebooks have no such
+   constraint (kernel I/O runs on a separate iopub thread). This caused a real bug class
+   (#117/#118/#119 — subprocess and parent built divergent optimizers, producing different
+   `fv` for the same params). The split architecture fixed this by making the notebook path
+   in-process by construction (no re-derivation, no divergence possible).
+
+2. **molass-gui's own history** (recovered from a past chat session via `session_store_sql`,
+   session `8500c8a8-6f7f-4e40-9eeb-516b93dab422`, turn 6): the GUI's "Use subprocess"
+   checkbox defaults to checked *not* for crash isolation, but for **measured performance**:
+   the GUI's live dashboard re-evaluates the full `objective_func(params, plot=True)` every
+   3 seconds to redraw, estimated at **~13-28% total overhead** on a 30-minute run (watch-loop
+   evaluations ~10-20%, `_objective_lock` contention ~2-5%, Tkinter event loop ~1-3%),
+   translating to roughly **13-15% faster wall-clock time** for subprocess mode on long runs.
+
+3. **This overhead is a library-level property, not GUI-specific.** Re-reading
+   `MplMonitor.update_plot()` (`molass-legacy/molass_legacy/Optimizer/MplMonitor.py`), the
+   in-process "external watcher" dashboard (`DESIGN_inprocess_monitor.md`'s Phase 5) reads
+   `callback.txt` cheaply for the best params, but then **still calls the full
+   `objective_func()`** on the live, actively-running optimizer to render the UV/XR/Score
+   panels (`display_optimizer = self.monitor_optimizer or self.optimizer`). The code's own
+   comment confirms this: *"Issue #50: prevent concurrent objective_func access when using
+   the live in-process optimizer... The lock is held through both `plot_job_state()` and
+   `_build_monitor_snapshot_json()` since both call `objective_func()`."* So any caller using
+   `in_process=True, monitor=True` — GUI or notebook — pays this same lock-contention /
+   redraw cost during long monitored runs.
+
+4. **The state-divergence risk that motivated in-process (#117/#118/#119) has a separate,
+   already-validated fix for subprocess mode**: `pipeline_recipe` (Option E, "recipe-based
+   subprocess," implemented and validated 2026-08-05/06 — see the superseded top section of
+   `/memories/repo/molass-gui-prototype-status.md`). Instead of the subprocess re-deriving
[…]
