### 2026-08-20 fc07e50
feat: reverse optimize_rigorously() in_process default to False (subprocess+auto-recipe) -- see DESIGN_default_in_process_reversal.md
