### 2026-08-20 f51c9df
docs: mark in_process default reversal core change as implemented
