### 2026-08-20 825d9e7
docs: reconcile design doc - correct which tests are actually affected, mark completed steps
