+++ Copilot/refactor/DESIGN_default_in_process_reversal.md
@@ -1,8 +1,7 @@
 # DESIGN: Reverse `optimize_rigorously()`'s `in_process` default to `False`
 
-**Status**: Core change implemented (2026-08-20, commit `fc07e50`) -- ripple-effect audit (step 4
-below) not yet done. User explicitly accepted backward-compatibility influence rather than
-gating the change on a full audit first.
+**Status**: Core change implemented and verified (2026-08-20, commit `fc07e50`). See "Status"
+section at the end for what's deliberately left undone.
 **Context**: Follow-up to `DESIGN_split_optimizer_architecture.md` (April 22, 2026) and
 `DESIGN_inprocess_monitor.md` (April 27, 2026). Triggered by a molass-gui "Continue in
 Notebook…" feature discussion that surfaced a real performance/architecture question about
@@ -83,55 +82,46 @@ default), while keeping that document's underlying architectural fix (determinis
 recipe-based subprocess reconstruction, avoiding #117-#119-style divergence) as the
 mechanism that makes subprocess-by-default safe to reinstate.
 
-## Required implementation changes (not yet done)
+## Implementation changes
 
-1. `Decomposition.optimize_rigorously()` (`molass/LowRank/Decomposition.py`): change
-   `in_process=True` → `in_process=False` in the signature.
-2. `make_rigorous_decomposition_impl()` (`molass/Rigorous/RigorousImplement.py`): same
-   default flip; the function's `**kwargs`/default handling needs to auto-construct
-   `pipeline_recipe` (from current decomposition/trim/baseline/model state) when the caller
-   doesn't supply one, matching what `upgraded_view.py` already builds explicitly today —
-   this makes the auto-construct path the *normal* default flow, not a deprecated fallback.
-3. The existing `DeprecationWarning` text ("Prefer `in_process=True` (the default)...") is
-   now backwards and must be rewritten or removed — it was written when `in_process=False`
-   was the legacy/deprecated path; after this change it becomes the default, sanctioned path.
-4. Audit ripple effects before landing:
-   - Tests that assert `in_process=True` as default behavior (e.g.
-     `tests/specific/200_Rigorous/test_090_pattern_a_warning.py`,
-     `test_100_cma_async_fallback.py`) will need updating.
-   - Tutorials/docs (`tests/tutorial/11-rigorous_optimization.py`) and any
-     molass-researcher notebooks relying on the current default's in-process behavior
-     (variables staying alive across cells, no subprocess startup cost) need review.
-   - The CMA async-crash auto-fallback logic (`in_process=True, async_=True` → falls back
-     to `in_process=False`) becomes a no-op once `in_process=False` is already the default —
-     should be simplified/removed rather than left as dead code.
-5. molass-gui side: once the library default matches the GUI's own behavior, the GUI's "Use
-   subprocess" checkbox and the notebook export's explicit `in_process=`/`pipeline_recipe=`
-   parameters can both be simplified/dropped, since both now get the right behavior "for
-   free" from the library default — achieving the original goal (GUI and notebook
-   equivalent, notebook stays simple) without either side needing to fight the library's
-   own default.
+1. ✅ `Decomposition.optimize_rigorously()` (`molass/LowRank/Decomposition.py`): changed
+   `in_process=True` → `in_process=False` in the signature; docstring rewritten.
+2. ✅ `make_rigorous_decomposition_impl()` (`molass/Rigorous/RigorousImplement.py`): same
+   default flip; docstring rewritten. `pipeline_recipe` auto-construction (`_build_auto_recipe`)
+   already existed and is now the normal default flow, not a fallback.
+3. ✅ The old `DeprecationWarning` ("Prefer `in_process=True` (the default)...") was **removed
[…]
