### 2026-08-20 46258ec
docs: mark in_process default reversal fully complete (library + molass-gui side)
