### 2026-08-20 5cc6a68
fix: async_=False did not block for subprocess path -- optimize_rigorously() returned before job completed
