### 2026-08-20 72a59e5
docs: record the async_=False subprocess regression found by actually running the ripple-effect audit
