### 2026-08-21 97dbf87
fix: restore() factory (issue #222) + live_status() stuck-at-running bug
- RunInfo.restore(decomposition, analysis_folder, rgcurve=None): module-level
  factory wrapping RunInfo.reconnect(), attaching decomposition/ssd/rgcurve so
  cross-session code gets the same load_best()/live_status() interface as a
  live run. Decomposition.load_best_rigorous_result() now emits a
  DeprecationWarning pointing to restore(). Both were referenced by
  tests/specific/200_Rigorous/test_120_restore.py but never implemented,
  causing a pre-existing collection ImportError.

- live_status(): fixed a bug where th
[…]
