### 2026-08-21 5a50799
chore: bump version to 1.0.6
