+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "molass"
-version = "1.0.5"
+version = "1.0.6"
 
 dependencies = [
   "numpy",
