### 2026-08-21 62714e1
fix: mirror de_tol=0 override in subprocess DE+LumpingConstraint path (#253)
