### 2026-08-24 e40abeb
Add num_jobs successive-jobs loop, DE tight-init reseed fix, load_analysis_session, monitor=True notebook guard
- optimize_rigorously(num_jobs=N, on_round_start=, stop_check=): run N full
  rounds, each reseeding from the global best (clear_jobs=True/False), the
  mechanism validated in molass-researcher experiment 36.
- SolverDE tight-population-init (_de_use_tight_init) now actually gets wired
  up on reseeded DE rounds, in both the in-process path (RigorousImplement.py)
  and the subprocess path (RecipeRunner.py) -- fixes #259: DE reseeding
  previously only touched 1 of ~500 population mem
[…]
