+++ molass/LowRank/Decomposition.py
@@ -1142,6 +1142,9 @@ class Decomposition:
                             seed_params=None,
                             constraints=None,
                             pipeline_recipe=None,
+                            num_jobs=1,
+                            on_round_start=None,
+                            stop_check=None,
                             **kwargs):
         """
         Perform a rigorous decomposition.
@@ -1219,6 +1222,30 @@ class Decomposition:
             automatically used as ``init_params`` for the new trial (resuming
             from the best known point rather than the original decomp params,
             see issue #169).
+        num_jobs : int, optional
+            Run ``num_jobs`` full, un-truncated rounds instead of one, each
+            round reseeding ``init_params`` from the best params found across
+            *all* previous rounds (the same mechanism as calling this method
+            repeatedly with ``clear_jobs=(round == 0)``). Default 1 (today's
+            single-round behavior, unchanged).
+
+            Requires ``async_=False`` — the call blocks until every round
+            completes. Validated monotonic/no-regression on real BH and DE
+            runs; see ``molass-researcher/experiments/36_bh_cutomize_trial``.
+            Distinct from ``max_trials`` (which is ``monitor=True``-only and
+            reseeds from just the previous trial's own best, not the global
+            best — see issue #257).
+        on_round_start : callable, optional
+            Only used when ``num_jobs > 1``. Called as
+            ``on_round_start(round_idx, num_jobs, run_info)`` immediately
+            after each round's ``run_info`` becomes available (before that
+            round blocks) — use it to poll/display progress for the round
+            currently in flight, e.g. from a GUI's own polling loop.
+        stop_check : callable, optional
+            Only used when ``num_jobs > 1``. Called with no arguments before
+            each round launches; return a truthy value to abort the remaining
+            rounds (the round already in flight is not interrupted by this --
+            call ``run_info.stop()`` for that).
         in_process : bool, optional
             If False (default), use the recipe-based subprocess path: the optimizer
             runs in a separate OS process, isolating it from crashes and from the
@@ -1395,6 +1422,38 @@ class Decomposition:
 
         if rgcurve is None:
             rgcurve = self.get_rg_curve()  # cached; avoids redundant Guinier fit on repeated calls (#248)
+
+        if num_jobs > 1:
+            if async_:
+                raise ValueError(
+                    "num_jobs > 1 requires async_=False -- each round blocks until it "
+                    "completes (and reseeds the next round from its best params) before "
+                    "returning. For a non-blocking multi-job sequence, loop over "
+                    "optimize_rigorously(clear_jobs=(i == 0), ...) yourself in a background "
+                    "thread instead."
+                )
+            if max_trials:
+                import warnings
+                warnings.warn(
+                    "num_jobs and max_trials both requested -- max_trials's own "
+                    "(per-round, monitor=True only) auto-resume would additionally fire "
+                    "inside each of the num_jobs rounds. Prefer one or the other.",
+                    UserWarning, stacklevel=2,
+                )
+            run_info = None
+            for round_idx in range(num_jobs):
+                if stop_check is not None and stop_check():
+                    break
+                # async_=True internally so run_info exists (for on_round_start) before
+                # this round's wait() blocks -- the overall call still blocks the caller
+                # per the async_=False contract enforced above.
[…]
