+++ molass/Rigorous/RecipeRunner.py
@@ -6,6 +6,76 @@ this function directly, bypassing the subprocess machinery entirely.
 """
 
 
+def rebuild_decomposition_from_recipe(analysis_folder):
+    """Rebuild (ssd, trimmed, decomp, recipe) from an analysis_folder's recipe.json.
+
+    Replays the same SSD -> trim -> correct -> quick_decomposition -> upgrade
+    pipeline used when the analysis_folder was first created, using recipe.json
+    plus any existing job's in_folder.txt as the source of truth. Shared by
+    create_optimizer_from_recipe (subprocess path) and by session-reconstruction
+    helpers (e.g. CurrentStateUtils.load_analysis_session) that need the initial
+    decomposition without building a full legacy optimizer.
+
+    Parameters
+    ----------
+    analysis_folder : str
+        Same value passed to optimize_rigorously(analysis_folder=...).
+
+    Returns
+    -------
+    ssd, trimmed, decomp, recipe
+        ``decomp`` is the initial (estimator) decomposition -- not yet attached
+        to any rigorous-optimization result. Use ``decomp.load_rigorous_result(
+        analysis_folder)`` to attach the best on-disk result.
+    """
+    import os, json
+
+    analysis_folder = os.path.abspath(analysis_folder)
+    optimizer_folder = os.path.join(analysis_folder, "optimized")
+    recipe_file = os.path.join(optimizer_folder, 'recipe.json')
+    if not os.path.exists(recipe_file):
+        raise FileNotFoundError(f"recipe.json not found at {recipe_file}")
+    with open(recipe_file) as f:
+        recipe = json.load(f)
+
+    jobs_folder = os.path.join(optimizer_folder, "jobs")
+    in_folder = None
+    if os.path.isdir(jobs_folder):
+        for jobid in sorted(os.listdir(jobs_folder)):
+            candidate = os.path.join(jobs_folder, jobid, 'in_folder.txt')
+            if os.path.exists(candidate):
+                with open(candidate) as f:
+                    in_folder = f.read().strip()
+                break
+    if in_folder is None:
+        raise FileNotFoundError(f"No in_folder.txt found under any job in {jobs_folder}")
+
+    from molass.DataObjects import SecSaxsData as SSD
+
+    n_components = recipe.get('num_components', 3)
+    model = recipe.get('model', 'egh').lower()
+    pore_dist = recipe.get('pore_dist', None)
+    ln_pore_sigma = recipe.get('ln_pore_sigma', None)
+    trim_params = recipe.get('trim_params', {})
+    baseline_params = recipe.get('baseline_params', {})
+    decomp_params = dict(recipe.get('decomp_params', {}))
+    decomp_params['num_components'] = n_components
+
+    ssd = SSD(in_folder)
+    trimmed = ssd.trimmed_copy(**trim_params)
+    corrected = trimmed.corrected_copy(**baseline_params)
+    decomp = corrected.quick_decomposition(**decomp_params)
+    if model != 'egh':
+        upgrade_kwargs = {}
+        if pore_dist is not None:
+            upgrade_kwargs['pore_dist'] = pore_dist
+            if ln_pore_sigma is not None:
+                upgrade_kwargs['model_params'] = {'ln_pore_sigma': ln_pore_sigma}
+        decomp = decomp.upgrade(model=model, **upgrade_kwargs)
+
+    return ssd, trimmed, decomp, recipe
+
+
 def create_optimizer_from_recipe(work_folder, class_code):
     """Rebuild the SSD pipeline from recipe.json and return a ready optimizer.
 
@@ -22,50 +92,18 @@ def create_optimizer_from_recipe(work_folder, class_code):
[…]

[…]
