### 2026-08-25 9bfa58f
fix(#260): capture frozen_components/frozen_param_groups/niter/function_code/seed_params in recipe.json
- _build_auto_recipe() now serializes these into recipe.json (optional keys
  omitted when not supplied, backward compatible)
- function_code auto-detection moved earlier so it can be captured
- RecipeRunner.create_optimizer_from_recipe() applies frozen_components/
  frozen_param_groups read from recipe.json to the reconstructed optimizer,
  and passes function_code to decomp.score()
- add tests/specific/200_Rigorous/test_260_recipe_schema.py
