### 2026-08-25 ce1ffcb
refactor(#255): unify DE/LumpingConstraint auto-apply into ConstraintDefaults
- new molass/Rigorous/ConstraintDefaults.py::get_constraint_and_overrides()
  is the single source of truth for method=='DE' and n_components>=3 ->
  auto-apply LumpingConstraint + de_tol=0 override
- RigorousImplement.py and RecipeRunner.py both call it instead of each
  re-deriving the condition and its consequences independently -- #253's
  bug was exactly a safety override added to one copy and not the other
- add tests/specific/200_Rigorous/test_255_constraint_defaults.py
