### 2026-08-25 #10403 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] Please add countries to author affiliations
- [ ] Please add citations for [numpy](https://numpy.org/citing-numpy/), [scipy](https://scipy.org/citing-scipy/), [matplotlib](https://matplotlib.org/stable/project/citing.html), and [scikit-learn](https://scikit-learn.org/stable/about.html#citing-scikit-learn)
