### 2026-08-25 1092623
refactor(#257): extract find_global_best_params() shared by clear_jobs=False resume
- new molass.Rigorous.RigorousImplement.find_global_best_params(jobs_dir, init_params=None)
  scans every completed job's callback.txt for the global-best params/fv
- _load_best_init_params() is now a thin wrapper around it (behavior unchanged,
  verified by existing test_085_clear_jobs_best_init.py, 5/5 pass)
- exposed publicly so molass-legacy's MplMonitor can share the same scan for its
  max_trials auto-resume / Resume Job button, instead of re-deriving its own
  weaker (just-completed-trial-only) reseed lo
[…]
