### 2026-08-25 0cfbff7
Address JOSS AEiC (rkurchin) final checks: add countries to affiliations, cite NumPy/SciPy/Matplotlib/scikit-learn
- All 3 affiliations now end with ', Japan'
- paper.bib already had Harris2020/Virtanen2020/Hunter:2007 entries for
  NumPy/SciPy/Matplotlib but they were never cited with [@key] in the text --
  added the citations
- Added a new Pedregosa2011 bib entry for scikit-learn and cited it
