+++ paper.bib
@@ -239,6 +239,15 @@ doi={10.1038/s41592-019-0686-2},
 url={https://doi.org/10.1038/s41592-019-0686-2}
 }
 
+@article{Pedregosa2011,
+author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and Vanderplas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, {\'E}douard},
+title={Scikit-learn: Machine Learning in {P}ython},
+journal={Journal of Machine Learning Research},
+year={2011},
+volume={12},
+pages={2825-2830}
+}
+
 @Article{Hunter:2007,
   Author    = {Hunter, J. D.},
   Title     = {Matplotlib: A 2D graphics environment},
