### 2026-08-25 eaa0f9a
chore: bump version to 1.0.7
