+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "molass"
-version = "1.0.6"
+version = "1.0.7"
 
 dependencies = [
   "numpy",
