### 2026-08-26 0552c65
Update pybaselines citation to match Zenodo record (v1.2.1, DOI 10.5281/zenodo.16790579)
