+++ paper.bib
@@ -265,8 +265,11 @@ pages={2825-2830}
 
 @software{pybaselines,
   author = {Erb, Donald},
-  doi = {10.5281/zenodo.5608581},
+  doi = {10.5281/zenodo.16790579},
   title = {{pybaselines}: A {Python} library of algorithms for the baseline correction of experimental data},
+  version = {v1.2.1},
+  publisher = {Zenodo},
+  year = 2025,
   url = {https://github.com/derb12/pybaselines}
 }
 
