### 2026-08-26 50beb19
fix: degrade monitor=True when active matplotlib backend is a GUI toolkit (Tcl_AsyncDelete)
