### 2026-08-26 8ddedf7
docs: add molass-gui reference to README
