### 2026-08-30 #10403 @rkurchin
Congrats to @nshimizu0721 and thanks to handling editor @zhubonan and reviewers @marshallmcdonnell and @KedoKudo. JOSS wouldn't exist without your volunteer efforts!
