### 2026-08-31 #10403 @nshimizu0721
Dear @zhubonan, @KedoKudo, @marshallmcdonnell, @rkurchin, and everyone involved,

Thank you very much for the time and effort you devoted to reviewing, editing, and guiding our submission through to publication. We sincerely appreciate your support throughout the review process.

Best regards,
Masatsuyo Takahashi and Nobutaka Shimizu
